% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Activation}
\alias{Activation}
\title{Activation}
\format{A flowSet containing two flowFrame objects:}
\usage{
Activation
}
\description{
An experiment performed to assess the in vitro activation of OT-I and OT-II transgenic T cells co-cultured with
OVA-presenting dendritic cells (DCs). This data set is comprised of two samples saved as a \code{flowSet} object.
}
\details{
Pooled DCs & T cells were labelled as follows:
- Hoechst 33258, a DNA interchelating dye which fluoresces in Alexa Fluor 405-A and Alexa Fluor 430-A channels
- CD11c-APC-Cy7 to identify DCs within the samples.
- Va2-PE to identify cells which express the OVA-specific TCR (i.e. OT-I and OT-II T cells).
- CD8-FITC to identify the CD8+ OT-I transgenic T cells (Alexa Fluor 488-A).
- CD4-Alexa Fluor 700 to identify the CD4+ OT-II transgenic T cells (Alexa Fluor 700-A)
- CD44-APC to assess T cell activation (Alexa Fluor 647-A)
- CD69-PerCP Cy5.5 to assess T cell activation (7-AAD-A)
}
\keyword{datasets}
