% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\docType{methods}
\name{axisLabels,flowFrame-method}
\alias{axisLabels,flowFrame-method}
\title{Get Appropriate Axes Labels for Transformed Channels - flowFrame Method}
\usage{
\S4method{axisLabels}{flowFrame}(x, channel, transList)
}
\arguments{
\item{x}{an object of class \code{flowFrame}.}

\item{channel}{name of the channel.}

\item{transList}{object of class \code{"transformList"} or
\code{"transformerList"} generated by estimateLogicle containing the
transformations applied to the flowFrame.}
}
\value{
list containing axis labels and breaks.
}
\description{
Get Appropriate Axes Labels for Transformed Channels - flowFrame Method
}
