% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\docType{methods}
\name{axisLabels,GatingHierarchy-method}
\alias{axisLabels,GatingHierarchy-method}
\title{Get Appropriate Axes Labels for Transformed Channels - GatingHierarchy Method}
\usage{
\S4method{axisLabels}{GatingHierarchy}(x, channel)
}
\arguments{
\item{x}{\code{GatingHiearchy}.}

\item{channel}{\code{character} channel name.}
}
\value{
when there is transformation function associated with the given
  channel, it returns a list of that contains positions and labels to draw on
  the axis otherwise returns NULL.
}
\description{
Get Appropriate Axes Labels for Transformed Channels - GatingHierarchy Method
}
