% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{axisLabels}
\alias{axisLabels}
\title{Get Appropriate Axes Labels for Transformed Channels - flowWorkspace}
\usage{
axisLabels(x, ...)
}
\arguments{
\item{x}{object of class \code{flowFrame} or \code{GatingHierarchy}.}

\item{...}{additional arguments.}
}
\value{
list containing axis labels and breaks.
}
\description{
Get Appropriate Axes Labels for Transformed Channels - flowWorkspace
}
