% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{checkAlias}
\alias{checkAlias}
\title{Check Supplied Alias}
\usage{
checkAlias(alias, gate_type)
}
\arguments{
\item{alias}{vector indicating the names of the populations to be gated.}

\item{gate_type}{vector indicating the type(s) of gate(s) to be constructed.}
}
\value{
Stops the gating process if alias is missing or \code{length(alias) != length(gate_type)}.
}
\description{
Check Supplied Alias
}
