% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\docType{methods}
\name{checkChannels,GatingSet-method}
\alias{checkChannels,GatingSet-method}
\title{Check Supplied Channels - GatingSet Method}
\usage{
\S4method{checkChannels}{GatingSet}(x, channels, plot = TRUE)
}
\arguments{
\item{x}{an object of class
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{channels}{vector of channel names (e.g. c("PE-A","APC-A")) or marker
names.}

\item{plot}{logical indicating whether the channels will be used to construct
a plot, set to TRUE by default to accept 1 or 2 channels.}
}
\value{
vector of valid channel names.
}
\description{
\code{checkChannels} will check whether the supplied channels are valid for
the supplied \code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
