% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{checkChannels}
\alias{checkChannels}
\title{Check Channels Supplied to drawGate.}
\usage{
checkChannels(x, channels)
}
\arguments{
\item{x}{an object of class \code{flowFrame}, \code{flowSet} or \code{GatingSet}.}

\item{channels}{vector of channel names (e.g. c("PE-A","APC-A")).}
}
\value{
Stop gating process if the supplied channels do ot meet the requirements for drawGate.
}
\description{
\code{checkChannels} will check whether the supplied channels are valid for the \code{flowSet} or \code{GatingSet} and that only 1or 2 channels are supplied.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
