% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{checkCSV}
\alias{checkCSV}
\title{Check .csv file exists in working directory}
\usage{
checkCSV(name)
}
\arguments{
\item{name}{name of the csv file to be checked.}
}
\value{
TRUE/FALSE if file exists
}
\description{
Check .csv file exists in working directory
}
