% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{checkFile}
\alias{checkFile}
\title{Check File Exists in Working Directory}
\usage{
checkFile(name)
}
\arguments{
\item{name}{filename including file extension to be checked.}
}
\value{
TRUE/FALSE if file exists in the current working directory.
}
\description{
Check File Exists in Working Directory
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
