% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{checkOSGD}
\alias{checkOSGD}
\title{Check Operating System & Open New Graphics Device}
\usage{
checkOSGD()
}
\description{
Check Operating System & Open New Graphics Device
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
