% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{checkTemplate}
\alias{checkTemplate}
\title{Check gatingTemplate for Existing Entry}
\usage{
checkTemplate(parent, alias, gtfile)
}
\arguments{
\item{parent}{name of the parent population.}

\item{alias}{name of the population of interest.}

\item{gtfile}{csv file name of the gatingTemplate.}
}
\value{
stops the gating process if an entry already exists in the gtfile for this population.
}
\description{
Check gatingTemplate for Existing Entry
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
