% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-statistics.R
\name{computePopStats}
\alias{computePopStats}
\title{Compute Population Level Statistics for a GatingSet.}
\usage{
computePopStats(gs, parent = NULL, alias = NULL, stat = "median",
  channels = NULL, inv.trans = TRUE, trans = NULL)
}
\arguments{
\item{gs}{a \code{GatingSet} object defining populations for which statistics should be calulated.}

\item{parent}{name of the parent population to use for frequency calculations.}

\item{alias}{a vector containing the \code{node} names of the populations of interest.}

\item{stat}{specify the statistic of interest by name, can be either "mean", "median" or "count". By default we use the more robust
median fluorescent intensity (MedFI).}

\item{channels}{a vector of channel names for which population statistics should be calculated, all fluorescent channels by default.}

\item{trans}{object of class transformerList to be used for inverse logicle transformation.}
}
\value{
a \code{list} object containing the calculated statistics for specified populations in all samples.
}
\description{
\code{computePopStats} function allows easy extraction of population level statistics from a \code{GatingSet} object,
including mean, geometirc mean and median fluorescent intensities as well as population counts and frequencies. 
\code{PopStats} searches \code{GatingSet} object for logicle transformation parameters and if present performs the inverse
logicle transformation prior to calculation of relevant statistic. Users can also supply their own transformerList to the
"trans" argument which will be used for the inverse transformation. By default if no channels are supplied \code{computePopStats} 
returns statistics for channels containing \code{markernames}, if no \code{markernames} are found statistics are performed on all 
channels.
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
\keyword{MFI,}
\keyword{count,}
\keyword{frequency}
\keyword{geometric}
\keyword{mean,}
\keyword{median,}
\keyword{population,}
\keyword{statistics,}
