% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSpillover-methods.R
\docType{methods}
\name{computeSpillover,flowSet-method}
\alias{computeSpillover,flowSet-method}
\title{Compute Spillover Matrix - flowSet Method}
\usage{
\S4method{computeSpillover}{flowSet}(x, transList = NULL,
  cmfile = NULL, spfile = "Spillover Matrix.csv", ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowSet-class]{flowSet}}
containing pre-gated single stain compensation controls and a universal
unstained control. Currently, computeSpillover does not pre-gate samples to
obtain a homogeneous cell population for downstream calculations. We
therefore recommmend pre-gating samples based on FSC and SSC parameters
prior to passing them to computeSpillover (i.e. \code{x} should contain
events for single cells only). Passing raw files to computeSpillover will
result in inaccurate calculations of fluorescent spillover matrix.}

\item{transList}{object of class
\code{\link[flowCore:transformList-class]{transformList}} generated by
\code{estimateLogicle} to transform fluorescent channels for gating.
\code{transList} is required if logicle transformation has already been
applied to \code{x} using estimateLogicle. \code{computeSpillover} will
automatically call \code{\link[flowCore:logicleTransform]{estimateLogicle}}
internally to transform channels prior to gating, if \code{transList} is
supplied it will be used for the transformation instead.}

\item{cmfile}{name of .csv file containing the names of the samples in a
column called "name" and their matching channel in a column called
"channel". \code{computeSpillover} will the guide you through the channel
selection process and generate a channel match file called "Compensation
Channels.csv" automatically. If you already have a complete cmfile and
would like to bypass the channel selection process, simply pass the name of
the cmfile to this argument (e.g. "Compensation Channels.csv").}

\item{spfile}{name of the output spillover csv file, set to \code{"Spillover
Matrix.csv"} by default.}

\item{...}{additional arguments passed to \code{\link{plotCyto1d,flowFrame-method}}
and \code{\link{plotCyto1d,flowSet-method}}.}
}
\value{
spillover matrix object and \code{"Spillover Matrix.csv"} file.
}
\description{
\code{computeSpillover} uses the method described by Bagwell & Adams 1993 to
calculate the fluorescent spillover matrix using a reference universal
unstained control and single stain compensation controls.
}
\details{
Calculate spillover matrix using
\code{\link[flowCore:flowSet-class]{flowSet}} containing gated single stain
compensation controls and an unstained control. \code{computeSpillover}
begins by the user selecting which fluorescent channel is associated with
each control from a dropdown menu. Following channel selection,
\code{computeSpillover} runs through each control and plots the density
distribution of the unstained control in red and the compensation control in
blue. Users can then gate the positive signal for spillover calculation using
an interval gate. The percentage spillover is calculated based on the median
fluorescent intensities of the stained populations and the universal
unstained sample. The computed spillover matrix is returned as an r object
and written to a named .csv file for future use.
}
\examples{
\dontrun{
fs <- Activation
gate1 <- drawGate(fs, alias = "Cells", channels = c("FSC-A","SSC-A"))
Cells <- Subset(fs, g1[[1]])
gate2 <- drawGate(Cells, alias = "Singlets", channels = c("FSC-A","FSC-H"))
Singlets <- Subset(Cells, g2[[1]])
spill <- computeSpillover(Singlets, spfile = "Example Spillover Matrix.csv")
}

}
\references{
C. B. Bagwell \& E. G. Adams (1993). Fluorescence spectral
  overlap compensation for any number of flow cytometry parameters. in:
  Annals of the New York Academy of Sciences, 677:167-184.
}
\seealso{
\code{\link{plotCyto1d,flowFrame-method}}

\code{\link{plotCyto1d,flowSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
