% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSpillover-methods.R
\docType{methods}
\name{computeSpillover,GatingSet-method}
\alias{computeSpillover,GatingSet-method}
\title{Compute Spillover Matrix - GatingSet Method}
\usage{
\S4method{computeSpillover}{GatingSet}(x, parent = NULL,
  transList = NULL, cmfile = NULL, spfile = "Spillover Matrix.csv",
  ...)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} containing pre-gated
single stain compensation controls and a universal unstained control.
Currently, computeSpillover does not pre-gate samples to obtain a
homogeneous cell population for downstream calculations. We therefore
recommmend pre-gating samples based on FSC and SSC parameters prior to
passing them to computeSpillover and indicate the population of interest
using the \code{parent} argument.}

\item{parent}{name of the pre-gated population to use for downstream
calculations, set to the last node of the GatingSet by default (e.g.
"Single Cells").}

\item{transList}{object of class
\code{\link[flowCore:transformList-class]{transformerList}} generated by
\code{estimateLogicle} to transform fluorescent channels for gating.
\code{transList} is required if logicle transformation has already been
applied to \code{x} using estimateLogicle. \code{computeSpillover} will
automatically call \code{\link[flowCore:logicleTransform]{estimateLogicle}}
internally to transform channels prior to gating, if \code{transList} is
supplied it will be used for the transformation instead.}

\item{cmfile}{name of .csv file containing the names of the samples in a
column called "name" and their matching channel in a column called
"channel". \code{computeSpillover} will the guide you through the channel
selection process and generate a channel match file called "Compensation
Channels.csv" automatically. If you already have a complete cmfile and
would like to bypass the channel selection process, simply pass the name of
the cmfile to this argument (e.g. "Compensation Channels.csv").}

\item{spfile}{name of the output spillover csv file, set to \code{"Spillover
Matrix.csv"} by default.}

\item{...}{additional arguments passed to \code{\link{plotCyto1d,flowFrame-method}}
and \code{\link{plotCyto1d,flowSet-method}}.}
}
\value{
spillover matrix object and \code{"Spillover Matrix.csv"} file.
}
\description{
\code{computeSpillover} uses the method described by Bagwell & Adams 1993 to
calculate the fluorescent spillover matrix using a reference universal
unstained control and single stain compensation controls.
}
\details{
Calculate spillover matrix using
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} containing gated
single stain compensation controls and an unstained control.
\code{computeSpillover} uses the method described by Bagwell & Adams 1993 to
calculate fluorescent spillover values using single stain compensation
controls and a universal unstained control. \code{computeSpillover} begins by
the user selecting which fluorescent channel is associated with each control
from a dropdown menu. Following channel selection, \code{computeSpillover}
runs through each control and plots the density distribution of the unstained
control in red and the compensation control in blue. Users can then gate the
positive signal for spillover calculation using an interval gate. The
percentage spillover is calculated based on the median fluorescent
intensities of the stained populations and the universal unstained sample.
The computed spillover matrix is returned as an r object and written to a
named .csv file for future use.
}
\examples{
\dontrun{
fs <- Activation
gs <- GatingSet(fs)
drawGate(gs, 
         parent = "root", 
         parent = "Cells", 
         channels = c("FSC-A","SSC-A"), 
         gtfile = "Example gatingTemplate.csv")
drawGate(gs, 
         parent = "Cells", 
         parent = "Singlets", 
         channels = c("FSC-A","FSC-H"), 
         gtfile = "Example gatingTemplate.csv")
spill <- computeSpillover(gs, 
                          parent = "Single Cells", 
                          spfile = "Example Spillover Matrix.csv")
}

}
\references{
C. B. Bagwell \& E. G. Adams (1993). Fluorescence spectral
  overlap compensation for any number of flow cytometry parameters. in:
  Annals of the New York Academy of Sciences, 677:167-184.
}
\seealso{
\code{\link{plotCyto1d,flowFrame-method}}

\code{\link{plotCyto1d,flowSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
