% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStats-methods.R
\docType{methods}
\name{computeStats,flowFrame-method}
\alias{computeStats,flowFrame-method}
\title{Compute Statistics - flowFrame Method}
\usage{
\S4method{computeStats}{flowFrame}(x, channels = NULL,
  transList = NULL, stat = "median", adjust = 1.5)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}}.}

\item{channels}{names of of channels for which statistic should be
calculated, set to all channels by default.}

\item{transList}{object of class
\code{\link[flowCore:transformList-class]{transformList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} used to transform
the fluorescent channels of x. TransList is required to apply the inverse
transformation such that the required statistics are returned on the
original linear scale.}

\item{stat}{name of the statistic to calculate, options include
\code{"count"}, \code{"freq"}, \code{"median"}, \code{"mode"},
\code{"mean"}, \code{"geo mean"}, \code{"CV"} or \code{CVI}.}

\item{adjust}{smoothing parameter passed to
\code{\link[stats:density]{density}} when calculating mode, set to 1.5 by
default.}
}
\description{
Calculate and export flow cytometry statistics for a flowFrame.
}
\seealso{
\code{\link{computeStats,flowSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
