% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeStats-methods.R
\name{computeStats}
\alias{computeStats}
\title{Compute Statistics}
\usage{
computeStats(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{...}{additional method-specific arguments.}
}
\description{
Calculate and export flow cytometry statistics.
}
\seealso{
\code{\link{computeStats,flowFrame-method}}

\code{\link{computeStats,flowSet-method}}

\code{\link{computeStats,GatingSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
