% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{cyto_annotate}
\alias{cyto_annotate}
\title{Interactively edit pData Information for a flowSet or GatingSet}
\usage{
cyto_annotate(x, file = NULL)
}
\arguments{
\item{x}{object of class \code{flowSet} or \code{GatingSet}.}

\item{file}{name of csv file containing experimental information.}
}
\value{
NULL and update pData for the \code{flowSet} or \code{GatingSet}.
}
\description{
Interactively edit pData Information for a flowSet or GatingSet
}
\examples{
\dontrun{
library(CytoRSuiteData)

# Load in samples
fs <- Activation

# Edit pData in table editor
cyto_annotate(fs)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
