% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_apply}
\alias{cyto_apply}
\alias{cyto_apply.flowSet}
\title{Apply a function over cytometry objects}
\usage{
cyto_apply(x, FUN, ...)

\method{cyto_apply}{flowSet}(x, FUN, ...)
}
\arguments{
\item{x}{object of class \code{flowSet} or \code{cytoset}.}

\item{FUN}{function to apply to each \code{flowFrame} or \code{cytoframe} in
the supplied \code{flowSet} or \code{cytoset}.}

\item{...}{additional arguments passed to
\code{\link[flowCore:fsApply]{fsApply}}.}
}
\description{
Apply a function over cytometry objects
}
\examples{
library(CytoExploreRData)

# Activation dataset
fs <- Activation

# Sample each flowFrame
fs <- cyto_apply(fs, cyto_sample)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
