% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_array_analyse.R
\name{cyto_array_analyse}
\alias{cyto_array_analyse}
\title{Analyse cytokine bead array data}
\arguments{
\item{x}{GatingSet.}

\item{array}{array details.}

\item{model}{model to use.}

\item{channel}{detection channel.}

\item{format}{either long or wide return.}

\item{save_as}{csv file for results}

\item{popup}{whether standard curves should plotted in pop-up window, set to
FALSE by default.}

\item{...}{additional arguments passed to model.}
}
\description{
Analyse cytokine bead array data
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
