% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_beads_sample}
\alias{cyto_beads_sample}
\title{Sample GatingSet to bead count}
\usage{
cyto_beads_sample(x, beads = NULL, bead_count = NULL)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{beads}{name of the gated bead population to use for the calculation. If
not supplied internal checks will be made for populations named "Single
Beads" or "Beads".}

\item{bead_count}{minimal bead count to down sample to, set to the minimum
bead count in samples by default. The bead count must be less than or equal
to the minimum bead count among the samples.}
}
\description{
\code{cyto_beads_sample} can be used on samples containing beads, to
downsamples each sample in a GatingSet based on a specific \code{bead_count}.
For example, if Sample A contains 100 beads and 75000 events, and Sample B
contains 50 beads and 5000 events. \code{cyto_beads_sample} will downsample
each sample to contain 50 beads and therefore 5000 events for Sample A and
3750 events for Sample B.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
