% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\docType{methods}
\name{cyto_channel_check,flowFrame-method}
\alias{cyto_channel_check,flowFrame-method}
\title{Check Supplied Channels - flowFrame Method}
\usage{
\S4method{cyto_channel_check}{flowFrame}(x, channels, plot = TRUE)
}
\arguments{
\item{x}{an object of class
\code{\link[flowCore:flowFrame-class]{flowFrame}}.}

\item{channels}{vector of channel names (e.g. c("PE-A","APC-A")) or marker
names.}

\item{plot}{logical indicating whether the channels will be used to construct
a plot, set to TRUE by default to accept 1 or 2 channels.}
}
\value{
vector of valid channel names.
}
\description{
\code{cyto_channel_check} will check whether the supplied channels are valid for
the supplied \code{\link[flowCore:flowFrame-class]{flowFrame}}.
}
\examples{
library(CytoRSuiteData)

# Add samples to flowSet
fs <- Activation

# Assign marker names
chnls <- c("PE-A", "Alexa Fluor 488-A", "Alexa Fluor 700-A", "Alexa Fluor 647-A", "7-AAD-A")
markers <- c("Va2", "CD8", "CD4", "CD44", "CD69")
names(markers) <- chnls
markernames(fs) <- markers

cyto_channel_check(fs[[1]], channels = c("CD4", "CD8"), plot = TRUE)
cyto_channel_check(fs[[1]], channels = c("CD4", "CD8", "CD44", "CD69"), plot = FALSE)
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
