% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channels-helpers.R
\name{cyto_channel_match}
\alias{cyto_channel_match}
\title{Match each single colour control to a fluorescent channel}
\usage{
cyto_channel_match(
  x,
  channels = NULL,
  peaks = FALSE,
  file = NULL,
  save_as = NULL,
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{channels}{names of the possible channels or markers to be matched to
each control, set to all area fluorescent parameters by default.}

\item{peaks}{logical to indicate whether peak detectors should be
automatically assigned using \code{cyto_peaks_assign()}, set to FALSE by
default.}

\item{file}{name of a CSV file from which the channel matching information
should be inherited. If not supplied, \code{cyto_channel_match()} will
automatically search for a file named \code{"Control-Details.csv"} or
create interactively create such a file.}

\item{save_as}{name of a CSV file to which the channel matching should be
written for downstream use, set to \code{"Control-Details.csv"}
prefixed with the date by default. Users can set custom file names here,
but the file name should contain \code{"Control-Details.csv"} in order
to be automatically detected by CytoExploreR within
\code{cyto_spillover_compute()}, \code{cyto_spillover_edit()},
\code{cyto_spillover_spread()} and \code{cyto_plot_compensation()}.}
}
\value{
a data.frame written to a CSV file containing information matching
  each file name to a channel. This channel matching information is also
  automatically added to the \code{cyto_details()} of the supplied samples
  where it can be easily accessed by CytoExploreR downstream.
}
\description{
Match each single colour control to a fluorescent channel
}
\examples{
library(CytoExploreRData)

# Compensation GatingSet
gs <- GatingSet(Compensation)

# Channel matching
cyto_channel_match(gs)

}
\seealso{
\code{\link{cyto_peaks_assign}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
