% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channel_match.R
\name{cyto_channel_match2}
\alias{cyto_channel_match2}
\title{Matches each compensation control to a fluorescent channel}
\usage{
cyto_channel_match2(
  x,
  channels = NULL,
  save_as = NULL,
  strip = TRUE,
  ignore.case = TRUE,
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{channels}{names of the possible channels or markers to be matched to
each control, set to all area fluorescent parameters by default.}

\item{save_as}{name of a CSV file to which the channel matching should be
written for downstream use, set to \code{"Compensation-Details.csv"}
prefixed with the date by default. Users can set custom file names here,
but the file name should contain \code{"Compensation-Details.csv"} in order
to be automatically detected by CytoExploreR within
\code{cyto_spillover_compute()}, \code{cyto_spillover_edit()},
\code{cyto_spillover_spread()} and \code{cyto_plot_compensation()}.}

\item{strip}{logical indicating whether overlapping characters in file names
should be stripped prior to matching, set to TRUE by default for more
accurate matching.}

\item{ignore.case}{logical indicating whether all case insensitive matches
should be found, set to TRUE by default.}

\item{...}{additional arguments passed to \code{link{grep}} when performing
character matching.}
}
\value{
a data.frame written to a CSV file containing information matching
  each file name to a channel. This channel matching information is also
  automatically added to the \code{cyto_details()} of the supplied samples
  where it can be easily accessed by CytoExploreR downstream.
}
\description{
Matches each compensation control to a fluorescent channel
}
\examples{
library(CytoExploreRData)

# Compensation GatingSet
gs <- GatingSet(Compensation)

# Channel matching
cyto_channel_match(gs)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
