% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channels-helpers.R
\name{cyto_channel_match_check}
\alias{cyto_channel_match_check}
\title{Check and prepare data stored in channel match files}
\usage{
cyto_channel_match_check(
  x,
  parent = "root",
  channel_match = NULL,
  save_as = NA
)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{vector of parental populations to extract from each sample in
\code{x}, set to the \code{"root"} node by default.}

\item{channel_match}{a vector of channels one per sample or the name of a CSV
file which indicates the name of the channel associated with each sample If
\code{channel_match} is not supplied or the channels are not annotated in
\code{cyto_details(x)} the user will be asked to interactively enter this
information usng \code{cyto_channel_match()}.}

\item{save_as}{passed to \code{cyto_channel_match()} to control whether the
channel annotation should be written to a CSV file.}
}
\value{
data.frame containing the updated experimental details.
}
\description{
Check and prepare data stored in channel match files
}
\examples{
\dontrun{
library(CytoExploreRData)

# Compensation GatingSet
gs <- GatingSet(Compensation)

# Annotate channel matches in experiment details
cyto_channel_match_check(gs)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
