% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\docType{methods}
\name{cyto_channel_select,flowFrame-method}
\alias{cyto_channel_select,flowFrame-method}
\title{Select Fluorescent Channel for Compensation Controls - flowFrame Method}
\usage{
\S4method{cyto_channel_select}{flowFrame}(x)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}}.}
}
\value{
selected channel associated with the supplied flowFrame.
}
\description{
Select Fluorescent Channel for Compensation Controls - flowFrame Method
}
\examples{
\dontrun{
library(CytoRSuiteData)

# Load in samples
fs <- Activation

# Select a channel from dropdown menu
cyto_channel_select(fs[[1]])
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
