% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\docType{methods}
\name{cyto_channel_select,flowSet-method}
\alias{cyto_channel_select,flowSet-method}
\title{Select Fluorescent Channel for Compensation Controls - flowSet Method}
\usage{
\S4method{cyto_channel_select}{flowSet}(x)
}
\arguments{
\item{x}{object of class
\code{\link[flowCore:flowSet-class]{flowSet}} containing compensation
controls.}
}
\value{
vector of channels in order of flowSet.
}
\description{
Select Fluorescent Channel for Compensation Controls - flowSet Method
}
\examples{
\dontrun{
library(CytoRSuiteData)

# Load in samples
fs <- Activation

# Select channel for each sample from dropdown menu
cyto_channel_select(fs)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
