% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\docType{methods}
\name{cyto_channel_select,GatingSet-method}
\alias{cyto_channel_select,GatingSet-method}
\title{Select Fluorescent Channel for Compensation Controls - GatingSet Method}
\usage{
\S4method{cyto_channel_select}{GatingSet}(x)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} containing
compensation controls.}
}
\value{
vector of channels in order of GatingSet.
}
\description{
Select Fluorescent Channel for Compensation Controls - GatingSet Method
}
\examples{
\dontrun{
library(CytoRSuiteData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Select channel for each sample from dropdown menu
cyto_channel_select(gs)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
