% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channels-helpers.R
\name{cyto_channel_select}
\alias{cyto_channel_select}
\title{Select Fluorescent Channel for Compensation Controls}
\usage{
cyto_channel_select(x)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[flowWorkspace:cytoset]{cytoset}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} containing
compensation controls.}
}
\value{
vector of channels in order of compensation Control samples.
}
\description{
Select Fluorescent Channel for Compensation Controls
}
\examples{
\dontrun{
library(CytoExploreRData)

# Compensation Gatingset
gs <- load_gs(system.file("extdata/Compensation-GatingSet",
                         package = "CytoExploreRData"))

# Select a channel for each control from dropdown menu
cyto_channel_select(gs)
}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
