% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channels-helpers.R
\name{cyto_channels<-}
\alias{cyto_channels<-}
\title{Replace channel names}
\usage{
cyto_channels(x) <- value
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class] {flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}},
\code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSey}}.}

\item{value}{vector of new column names to replace the old ones.}
}
\description{
Replace channel names
}
\examples{
library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Channels
cyto_channels(fs)

# Update first FSC-A to FSC
cyto_channels(fs)[1] <- "FSC"

}
\seealso{
\code{\link{cyto_channels}}

\code{\link{cyto_fluor_channels}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
