% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channels-helpers.R
\name{cyto_channels}
\alias{cyto_channels}
\title{Extract channel names}
\usage{
cyto_channels(x, select = NULL, exclude = NULL, append = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{select}{vector of channel names to select.}

\item{exclude}{vector of channel names to exclude.}

\item{append}{logical indicating whether the name of the channel should be
appended to the marker names in the form \code{<marker> channel}, set to
FALSE by default.}

\item{...}{additional arguments passed to \code{\link[base:grep]{grepl}} for
character matching. For exact character string matching to override the
default which ignores character case, set \code{fixed} to TRUE.}
}
\value{
vector of channel names.
}
\description{
Simply a wrapper around \code{colnames} to extract the channels associated
with a \code{cytoframe}, \code{cytoset} \code{GatingHierarchy} or
\code{GatingSet}.
}
\examples{
library(CytoExploreRData)

# Activation GatingSet
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                          package = "CytoExploreRData"))

# GatingSet
cyto_channels(gs)

# GatingHierarchy
cyto_channels(gs[[1]], select = "Alexa")

# cytoset
cs <- cyto_data_extract(gs, "root")[["root"]]
cyto_channels(cs)

# cytoframe
cyto_channels(cs[[1]], exclude = c("FSC","SSC"))

}
\seealso{
\code{\link{cyto_fluor_channels}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
