% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channels-helpers.R
\name{cyto_channels_extract}
\alias{cyto_channels_extract}
\alias{cyto_channels_extract.default}
\alias{cyto_channels_extract.flowFrame}
\alias{cyto_channels_extract.flowSet}
\alias{cyto_channels_extract.GatingHierarchy}
\alias{cyto_channels_extract.GatingSet}
\title{Extract channels associated with certain markers}
\usage{
cyto_channels_extract(x, channels, plot = FALSE)

\method{cyto_channels_extract}{default}(x, channels, plot = FALSE)

\method{cyto_channels_extract}{flowFrame}(x, channels, plot = FALSE)

\method{cyto_channels_extract}{flowSet}(x, channels, plot = FALSE)

\method{cyto_channels_extract}{GatingHierarchy}(x, channels,
  plot = FALSE)

\method{cyto_channels_extract}{GatingSet}(x, channels, plot = FALSE)
}
\arguments{
\item{x}{an object of class
\code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{channels}{vector of channel and/or marker names (e.g. c("Alexa Fluor
700-A","CD8")).}

\item{plot}{logical indicating whether the channels will be used to construct
a plot, set to FALSE by default. If set to TRUE an additional check will be
performed to ensure that only 1 or 2 \code{channels} are supplied.}
}
\value{
A vector of valid channel names.
}
\description{
\code{cyto_channels_extract} will check whether the supplied channels or
marker names are valid for the
\code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} and return a vector of
valid channel names. \code{cyto_channels_extract} is particularly useful for
determining which channel(s) are associated with particular marker(s).
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
