% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channels-helpers.R
\name{cyto_channels_restrict}
\alias{cyto_channels_restrict}
\title{Remove channels of a flowFrame or flowSet}
\usage{
cyto_channels_restrict(x)
}
\arguments{
\item{x}{object of class flowFrame or flowSet.}
}
\value{
flowFrame or flowSet object with unused channels removed.
}
\description{
\code{cyto_channels_restrict} removes channels from the flowFrame or flowSet
which do not have markers assigned. The FSC, SSC and Time parameters are
always retained irrespective of marker assignment.
}
\examples{

library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Channels
cyto_channels(fs)

# Remove unused channels
fs <- cyto_channels_restrict(fs)

# Channels removed
cyto_channels(fs)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
