% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_check}
\alias{cyto_check}
\title{Check a flowFrame, flowSet, GatingHierarchy or GatingSet has been supplied}
\usage{
cyto_check(x)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} to be checked.}
}
\value{
TRUE or FALSE if object meets this class criteria.
}
\description{
Check a flowFrame, flowSet, GatingHierarchy or GatingSet has been supplied
}
\examples{
# Load in CytoExploreRData to acess data
library(CytoExploreRData)

# Valid object
cyto_check(Activation)

# Invalid list object
cyto_check(list(Activation))

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
