% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_class}
\alias{cyto_class}
\title{Get or check class of a cytometry object}
\usage{
cyto_class(x, expect = NULL, class = FALSE)
}
\arguments{
\item{x}{either a \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}},
\code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} for which the class
should be checked or returned.}

\item{expect}{a class of object supplied as a character string, if supplied a
logical value will be returned indicating whether \code{x} is of the
expected class.}

\item{class}{indicates whether the class of \code{x} should be matched based
on the objects class (exact class match) or based on its reference class.
See examples for more details.}
}
\value{
class of \code{x} or a logical indicating whether \code{x} matches
  the expected class.
}
\description{
Get or check class of a cytometry object
}
\examples{
library(CytoExploreRData)

# flowSet
fs <- Activation
cyto_class(fs)
cyto_class(fs, expect = "flowSet")

# GatingSet
gs <- GatingSet(fs)
cyto_class(gs)
cyto_class(gs, expect = "GatingSet")

# cytoset
cs <- cyto_extract(gs, "root")
cyto_class(cs)
cyto_class(cs, expect = "flowSet", class = FALSE) # reference class

# cytoframe
cf <- cs[[1]]
cyto_class(cf)
cyto_class(cf, expect = "flowFrame", class = FALSE) # reference class

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
