% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_coerce}
\alias{cyto_coerce}
\title{Coerce cytoframes in a cytoset}
\usage{
cyto_coerce(
  x,
  display = 1,
  seed = NULL,
  barcode = FALSE,
  format = "cytoset",
  name = "merge",
  ...
)
}
\arguments{
\item{x}{\code{\link[flowWorkspace:cytoset]{cytoset}}.}

\item{display}{passed to \code{\link{cyto_sample}} to control the number of
events to retain in the coerced \code{cytoframe}, set to 1 by default to
retain all events.}

\item{seed}{numeric passed to \code{\link{set.seed}} to return the same
sample with each run, set to NULL by default for random sampling.}

\item{barcode}{logical indicating whether the samples should be barcoded
prior to coercion, set to FALSE by default.}

\item{format}{can be either 1 - matrix, 2 - cytoframe or 3 - cytoset to
control the format of the returned data, set to \code{"cytoset"} by
default.}

\item{name}{character string indicating the name to use for the coerced
object, set to \code{"merge"} by default.}

\item{...}{additional arguments passed to \code{\link{cyto_sample}}.}
}
\value{
a sampled and coerced \code{matrix}, \code{cytoframe} or
  \code{cytoset}.
}
\description{
\code{cyto_coerce} is an efficient method for simultaneously coercing and
downsampling elements of a \code{cytoset}. Basically, the individual
\code{cytoframe} elements are sampled prior to coercion using a combination
of \code{cyto_sample_n} and \code{cyto_sample}. \code{format} provides
control over the format in which the coerced data should be returned.
}
\examples{
library(CytoExploreRData)

# Activation GatingSet
gs <- cyto_load(
 system.file(
   "extdata/Activation-GatingSet",
   package = "CytoExploreRData"
   )
 )

# Activation cytoset
cs <- cyto_data_extract(gs, "root")[["root"]]

# Coerce & Sample
cyto_coerce(cs, display = 50000, seed = 56)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
