% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_compensate}
\alias{cyto_compensate}
\alias{cyto_compensate.flowFrame}
\alias{cyto_compensate.flowSet}
\alias{cyto_compensate.GatingSet}
\title{Apply fluorescence compensation to samples}
\usage{
cyto_compensate(x, ...)

\method{cyto_compensate}{flowFrame}(x, spillover = NULL, select = 1)

\method{cyto_compensate}{flowSet}(x, spillover = NULL, select = 1)

\method{cyto_compensate}{GatingSet}(x, spillover = NULL, select = 1)
}
\arguments{
\item{x}{object of class \code{flowFrame}, \code{flowSet} or
\code{GatingSet}.}

\item{spillover}{name of the spillover matrix csv file (e.g.
"Spillover-Matrix.csv") saved in the current working directory.}

\item{select}{index or name of the sample from which the spillover matrix
should be extracted when no spillover matrix file is supplied to
\code{spillover}. To compensate each sample individually using their stored
spillover matrix file, set \code{select} to NULL.}
}
\value{
a compensated \code{flowFrame}, \code{flowSet} or \code{GatingSet}
  object.
}
\description{
\code{cyto_compensate} will apply a saved spillover matrix csv file to all
samples. The csv file must contain the fluorescent channels as both the
colnames (i.e. first row) and rownames (i.e. first column). If no
\code{spillover} csv file is supplied, the spillover matrix will be extracted
directly from the first element of \code{x}. To select a different sample for
spillover matrix extraction supply the index or name of the sample to the
\code{select} argument.
}
\examples{
library(CytoRSuiteData)

# Apply stored spillover matrix to flowSet
cyto_compensate(Activation)

# Save spillover matrix in correct format
spill <- Activation[[1]]@description$SPILL
rownames(spill) <- colnames(spill)
write.csv(
  spill,
  "Spillover-Matrix.csv"
)

# Apply saved spillover matrix csv file to flowSet
cyto_compensate(Activation, "Spillover-Matrix.csv")

# Apply stored spillover matrix to GatingSet
gs <- GatingSet(Activation)
cyto_compensate(gs)

# Apply saved spillover matrix csv file to GatingSet
cyto_compensate(gs, "Spillover-Matrix.csv")
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
