% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_convert}
\alias{cyto_convert}
\title{Convert objects to cytoframe/cytoset objects}
\usage{
cyto_convert(x, ...)
}
\arguments{
\item{x}{object of class \code{flowFrame}, \code{flowSet}, \code{cytoframe}
\code{cytoset}, \code{matrix}, \code{data.frame} or \code{data.table}.}

\item{...}{additional arguments passed to
\code{flowWorkspace::load_cytoframe_from_fcs} to customise how data is read
from the written FCS files.}

\item{y}{either convert to a \code{"cytoframe"} or \code{"cytoset"}. For
backwards compatibility the default for flowFrames/cytoframes is
\code{"cytoframe"} and the default for flowSets/cytosets is
\code{"cytoset"}.}

\item{save_as}{location to FCS file in the case of flowFrames/cytoframes or a
directory name for flowSets/cytosets where FCS files should be written
during conversion, defaults to a tempfile if not specified manually.}
}
\value{
a cytoframe or cytoset.
}
\description{
Convert objects to cytoframe/cytoset objects
}
\examples{
library(CytoExploreRData)

# flowSet to cytoset
cs <- cyto_convert(Activation)

}
\seealso{
\code{\link[flowWorkspace:convert]{flowFrame_to_cytoframe}}

\code{\link[flowWorkspace:convert]{flowSet_to_cytoset}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
