% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_copy}
\alias{cyto_copy}
\title{Copy a cytoset, cytoframe or GatingSet}
\usage{
cyto_copy(x, type = "HDF5")
}
\arguments{
\item{x}{cytoframe, cytoset or GatingSet to be copied.}

\item{type}{\code{TRUE}/\code{"HDF5"}/\code{"disk"} for a disk-backed copy
(default) or \code{"memory"} for an in-memory MemCytoFrame-backed copy.}
}
\value{
copied cytoframe, cytoset or GatingSet.
}
\description{
Copy a cytoset, cytoframe or GatingSet
}
\examples{
library(CytoExploreRData)

# Activation Gatingset
gs <- load_gs(
  system.file(
    "extdata/Activation-GatingSet",
    package = "CytoExploreRData"
  )
)

gs_copy <- cyto_copy(gs)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
