% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_data_extract}
\alias{cyto_data_extract}
\title{Extract data from cytosets, GatingHierarchies or GatingSets}
\usage{
cyto_data_extract(
  x,
  parent = "root",
  select = NULL,
  copy = FALSE,
  format = "cytoset",
  channels = NULL,
  markers = FALSE,
  split = FALSE,
  trans = NA,
  inverse = FALSE,
  events = 1,
  coerce = FALSE,
  barcode = FALSE,
  overwrite = NULL,
  seed = NULL,
  path = "auto"
)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the parent population to extract from
\code{GatingHierarchy} or \code{GatingSet} objects.}

\item{select}{named list containing experimental variables to be used to
select samples using \code{\link{cyto_select}}.}

\item{copy}{logical indicating whether a deep copy of the extracted data
should be returned, set to FALSE by default. Users should set this argument
to TRUE if they wish to keep the original data unchanged.}

\item{format}{can be either \code{"cytoframe"}, \code{"cytoset"} or
\code{"matrix"} to indicate in what format the data should be returned in
the lists, set to \code{"cytoset"} by default.}

\item{channels}{names of the markers or channels for which data should be
extracted, set to all channels by default.}

\item{markers}{logical indicating whether the column names of the raw data
matrices should be replaced with the marker names where possible, set to
FALSE by default.}

\item{split}{logical indicating whether cytosets should be split into
cytosets containing a single sample each, set to FALSE by default.}

\item{trans}{object of class transformerList required to inverse
transformations applied to the data when \code{inverse = TRUE}.}

\item{inverse}{logical indicating whether transformations applied to the data
should be inversed, set to FALSE by default.}

\item{events}{passed to \code{\link{cyto_sample}} to extract a subset of
events from each sample. \code{sample} indicates the number or proportion
of events to extract from each sample, set to \code{1} by default to
extract all events.}

\item{coerce}{logical to indicate whether the data should be merged into a
single object prior to returning the data, set to FALSE by default. If
\code{coerce} is TRUE, \code{sample} controls the proportion or number of
events to keep in the merged object.}

\item{barcode}{logical to control whether \code{cyto_barcode} should be
called on the data prior to export. This argument also accepts options that
can be passed to the \code{type} argument of \code{cyto_barcode} to control
the type of barcoding to perform.}

\item{overwrite}{logical passed to \code{cyto_barcode} to provide a
non-interactive way of controlling how existing barcodes should be handled.}

\item{seed}{values used to \code{set.seed()} internally to return the same
extract the same subset of events on each run when \code{sample} is
supplied.}

\item{path}{can be either \code{"auto"} or \code{"full"} to control whether
the extracted list should be labelled with the shortest or complete path to
each population in parent.}
}
\value{
either a list of cytosets, list of cytoframe lists or a list of raw
  data matrix lists.
}
\description{
\code{cyto_data_extract} will extract populations specified by \code{parent}
and return a list of cytosets. \code{cyto_data_extract} can also optionally
perform additional tasks on the extracted cytosets, such subsetting channels,
inversing transformations and returning raw data matrices.
}
\examples{
# Load in CytoExploreRData to access data
library(CytoExploreRData)

# Activation GatingSet
gs <- cyto_load(system.file("extdata/Activation-GatingSet",
                package = "CytoExploreRData"))

# Extract cytosets & inverse transform
cyto_data_extract(gs[[1]],
parent = c("CD4 T Cells", "CD8 T Cells"),
inverse = TRUE)

# Extract raw data matrices
mt <- cyto_data_extract(gs,
parent = "root",
format = "matrix",
channels = c("CD4", "CD8"))
mt[["Activation_1.fcs"]]

}
\seealso{
\code{\link{cyto_select}}

\code{\link{cyto_barcode}}

\code{\link{cyto_coerce}}

\code{\link{cyto_sample}}

\code{\link{cyto_transform}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
