% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_details}
\alias{cyto_details}
\title{Extract experiment details from flowSet or GatingSet}
\usage{
cyto_details(x)
}
\arguments{
\item{object}{of class \code{flowSet} or \code{GatingSet}.}
}
\value{
experiment details as data.frame.
}
\description{
Simply a autocomplete-friendly wrapper around
\code{\link[flowWorkspace:pData]{pData}}.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
