% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_details_edit}
\alias{cyto_details_edit}
\title{Interactively edit cyto_details for a cytoset or GatingSet}
\usage{
cyto_details_edit(x, file = NULL, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{file}{name of csv file containing the experimental information
associated with each sample in the supplied data. This file must contain a
column called \code{"name"} which contains the names of the samples. This
argument can be set to NA if you don't wish to write the updated experiment
details to file.}

\item{...}{not in use.}
}
\value{
NULL and return \code{flowSet} or \code{GatingSet} with updated
  experimental details.
}
\description{
Interactively edit cyto_details for a cytoset or GatingSet
}
\examples{
\dontrun{
# Load in CytoExploreRData to access data
library(CytoExploreRData)

# Load in samples
fs <- Activation

# Edit cyto_details in table editor
cyto_details_edit(fs)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
