% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_dist.R
\name{cyto_dist}
\alias{cyto_dist}
\title{Compute distance matrices between samples or distributions}
\usage{
cyto_dist(
  x,
  parent = "root",
  select = NULL,
  alias = NULL,
  channels = NULL,
  type = "aitchison",
  merge_by = "name",
  scale = "range",
  trans = NA,
  inverse = FALSE,
  events = NA,
  k = 30,
  grid = c(14, 14),
  heatmap = TRUE,
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the population upon which comparisons should be made,
set to the \code{"root"} node by default.}

\item{select}{vector or list of sample selection criteria passed to
\code{cyto_select()} to extract the samples for which the specified
\code{type} statistic should be computed, set to NULL by default to use all
samples in the supplied \code{cytoset} or \code{GatingSet}.}

\item{alias}{names of the populations to use when comparing the compositions
of samples using the \code{Aitchison} distance.}

\item{channels}{names of the channel(s) or marker(s) for which comparisons on
the \code{parent} population should be made.}

\item{type}{indicates the type of metric to use when comparing samples. For
compositional comparisons when \code{alias} is supplied, \code{type} is set
to \code{"Aitchison"} to compute the Aitchison distances. If
\code{channels} are specified comparison of distributions across samples
can be made using either the default \code{"Wasserstein"} distance,
\code{"Jensen-Shannon"} divergence or \code{"Kolmogorov-Smirnov"} test. If
\code{alias} and \code{channels} are not specified, comparisons of the
\code{parent} populations can be made either using the default \code{"kNN"}
or alternatively \code{"SOM"} entropies.}

\item{merge_by}{a vector of experiment variables by which samples should be
merged prior to computing the desired statistic, set to \code{"name"} by
default to compare individual samples to one another.}

\item{trans}{an object of class \code{transformerList} containing the
transformer definitions for transformations already applied to the supplied
cytoset.}

\item{inverse}{logical indicating whether inverse data transformations should
be applied to the data from the \code{parent} population prior to computing
the desired statistic, set to FALSE by default.}

\item{events}{indicates the number or proportion of events to extract from
the \code{parent} population in each sample prior to computing the
statistic specified by \code{type}. Setting \code{events = NA} will extract
the minimum number of events across groups from each group.}

\item{heatmap}{logical indicating whether to plot a heatmap for each computed
distance matrix, set to TRUE by default.}

\item{...}{additional arguments passed to \code{cyto_plot_heatmap()}.}
}
\value{
a distance matrix for compositional or sample comparisons, or a list
  of distance matrices for each channel for distribution comparisons.
}
\description{
Compute distance matrices between samples or distributions
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
