% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_empty}
\alias{cyto_empty}
\title{Construct an empty cytoframe}
\usage{
cyto_empty(name = NULL, channels = NULL, ...)
}
\arguments{
\item{name}{name to add to the constructed cytoframe}

\item{channels}{channels to include in the constructed cytoframe}

\item{...}{additional arguments passed to
\code{\link[flowCore:flowFrame-class]{flowFrame}}.}
}
\description{
Construct an empty cytoframe
}
\examples{
# Construct empty cytoframe
cyto_empty(name = "Test.csv", 
channels = c("FSC-A", "SSC-A", "PE-A"))

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
