% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_export}
\alias{cyto_export}
\title{Export cytometry data for use in flowJo or Cytobank}
\usage{
cyto_export(x, save_as = NULL, ...)
}
\arguments{
\item{x}{object of class GatingSet to export.}

\item{save_as}{the name of a flowJo workspace file with .wsp extension or a
Cytobank xml file with .xml extension.}

\item{...}{additional arguments passed to \code{gatingset_to_cytobank} or
\code{gatingset_to_flowjo}.}
}
\description{
Simply a wrapper around CytoML \code{gatingset_to_cytobank} and
\code{gatingset_to _flowjo} to export your CytoExploreR analyses in a format
accepted by these platforms. Users will need to have \code{docker} running on
their computers to export flowJo workspace files.
}
\examples{
\dontrun{
library(CytoExploreRData)
library(CytoML)

# Activation GatingSet
gs <- GatingSet(Activation)

# Compensation
gs <- cyto_compensate(gs)

# Transformations
gs <- cyto_transform(gs)

# Gating
gs <- cyto_gatingTemplate_apply(gs, Activation_gatingTemplate)

# Export to cytobank xml
cyto_export(gs, "cytobank.xml")

# Export to flowjo workspace
cyto_export(gs, "flowjo.wsp")
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
