% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_exprs}
\alias{cyto_exprs}
\alias{cyto_exprs.flowFrame}
\alias{cyto_exprs.flowSet}
\title{Extract raw data matrix from cytoframe or cytoset}
\usage{
cyto_exprs(x, ...)

\method{cyto_exprs}{flowFrame}(x, channels = NULL, markers = FALSE, drop = TRUE, ...)

\method{cyto_exprs}{flowSet}(x, channels = NULL, markers = FALSE, drop = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoframe]{cytoframe}} or
\code{\link[flowWorkspace:cytoset]{cytoset}}.}

\item{...}{additional arguments passed to data extraction from matrix.}

\item{channels}{vector of channel or marker names for which raw data should
be extracted.}

\item{markers}{logical indicating whether the column names of the extracted
matrix should be converted to markers where possible, set to FALSE by
default.}

\item{drop}{logical to control whether single channel extractions should be
converted to vectors by default, set to TRUE by default.}
}
\value{
matrix or list of matrices containing raw data.
}
\description{
A convenient wrapper around \code{\link[flowCore:exprs]{exprs}} which
simplifies extraction of raw data from certain parameters. \code{cyto_exprs}
is primarily for use within CytoExploreR and users should instead use the
\code{cyto_data_extract} API.
}
\examples{
#' library(CytoExploreRData)

# Activation Gatingset
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                          package = "CytoExploreRData"))

# cytoset
cs <- cyto_data_extract(gs, "T Cells")[["T Cells"]]

# matrices
cyto_exprs(cs, c("CD44", "CD69"))

}
\seealso{
\code{\link{cyto_data_extract}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
