% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_extract}
\alias{cyto_extract}
\title{Extract a valid flowFrame or flowSet}
\usage{
cyto_extract(
  x,
  parent = NULL,
  select = NULL,
  copy = FALSE,
  raw = FALSE,
  channels = NULL,
  ...
)
}
\arguments{
\item{x}{object of class \code{flowFrame}, \code{flowSet},
\code{GatingHierarchy} or \code{GatingSet}.}

\item{parent}{name of the parent population to extract from
\code{GatingHierarchy} or \code{GatingSet} objects.}

\item{select}{named list containing experimental variables to be used to
select samples using \code{\link{cyto_select}}.}

\item{copy}{logical indicating whether a deep copy of the extracted data
should be returned.}

\item{raw}{logical indicating whether a list of raw data matrices should be
returned instead of a flowFrame or flowSet.}

\item{channels}{names of the markers or channels for which data should be
extracted, set to all channels by default.}

\item{...}{additional arguments passed to
\code{\link[flowWorkspace:gh_pop_get_data]{gh_pop_get_data}} or
\code{\link[flowWorkspace:gh_pop_get_data]{gs_pop_get_data}}.}
}
\value{
either a \code{flowFrame} or a \code{cytoset}  by default. A list of
  raw data matrices when raw is set to TRUE.
}
\description{
\code{cyto_extract} is essentially a wrapper for
\code{\link[flowWorkspace:gh_pop_get_data]{gs_pop_get_data}} which also
accepts \code{\link[flowCore:flowFrame-class]{flowFrame}} or
\code{\link[flowCore:flowSet-class]{flowSet}} objects. The \code{parent}
population is extracted from
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} objects whilst
\code{flowFrame} or \code{flowSet} objects are returned as is.
}
\examples{

# Load in CytoExploreRData to access data
library(CytoExploreRData)

# GatingSet
gs <- GatingSet(Activation)

# Extract cytoframe
cf <- cyto_extract(gs[[1]], parent = "root")

# Extract cytoset
cs <- cyto_extract(gs, parent = "root")

# Extract raw data matrices
mt <- cyto_extract(gs, parent = "root", raw = TRUE)
mt[["Activation_1.fcs"]]
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
