% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_fcs_pregating.R
\name{cyto_fcs_pregating}
\alias{cyto_fcs_pregating}
\title{Import, pregate and extract population of interest from your samples to make the files smaller.
By default it imports all of the ".fcs" and ".FCS" files from specified directory.
The function skips the compensation and transformation - pregating can be only done using FSC and SSC channels.}
\usage{
cyto_fcs_pregating(
  path = ".",
  gatingTemplate = "gatingTemplate_pregating.csv",
  output = NULL,
  gates = c(cyto_gate_draw(gs, parent = "root", alias = "Cells", channels = c("FSC-A",
    "SSC-A")), cyto_gate_draw(gs, parent = "Cells", alias = "Single Cells", channels =
    c("FSC-A", "FSC-H"))),
  stats = c(cyto_stats_compute(gs, parent = "root", alias = "Cells", stat = "freq"),
    cyto_stats_compute(gs, parent = "Cells", alias = "Single Cells", stat = "freq")),
  control = NULL,
  output_gate = "Single Cells"
)
}
\arguments{
\item{path}{path points to the location of the .fcs files to read in. Preferably
the name of folder in current working directory}

\item{gatingTemplate}{name of a gatingTemplate csv file to be used for gate saving.
Set to "gatingTemplate_pregating.csv" by default}

\item{output}{name of the folder to which the written FCS files should be
saved, set to NULL by default to save the files to the current working
directory}

\item{gates}{sets gates to use for exporting the population of interest. Set to c(
cyto_gate_draw(gs, parent = "root", alias = "Cells", channels = c("FSC-A", "SSC-A")),
cyto_gate_draw(gs, parent = "Cells", alias = "Single Cells", channels = c("FSC-A", "FSC-H"))
) by default}

\item{stats}{defines the statistics to be exported to control csv sheet. Set to stats = c(
cyto_stats_compute(gs, parent = "root", alias = "Cells", stat = "freq"),
cyto_stats_compute(gs, parent = "Cells", alias = "Single Cells", stat = "freq")
) by default}

\item{control}{name of the file compiling the frequency of each gate for all of the samples. Set to NULL by default - no file created}
}
\value{
new fcs files extracted from subpopulation of interest
}
\description{
Import, pregate and extract population of interest from your samples to make the files smaller.
By default it imports all of the ".fcs" and ".FCS" files from specified directory.
The function skips the compensation and transformation - pregating can be only done using FSC and SSC channels.
}
\examples{

#Load CytoExploreR
library(CytoExploreR)

# Get path to Activation .fcs files in CytoExploreRData
datadir <- system.file("extdata", package = "CytoExploreRData")
path <- paste0(datadir, "/Activation")

# Load and preprocess the fcs files
cyto_fcs_pregating(
path,
output = "pregated",
control = "pregated"
)

}
\author{
Mieszko Lachota
}