% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_file_search.R
\name{cyto_file_name}
\alias{cyto_file_name}
\title{Create a unique name for a file}
\usage{
cyto_file_name(x, ext = NULL, ...)
}
\arguments{
\item{x}{desired name for a file optionally including the extension, if the
extension is not manually supplied to \code{ext}.}

\item{ext}{the file extension prefixed by \code{"."}.}

\item{...}{not in use.}
}
\value{
a unique file name.
}
\description{
Create a unique name for a file
}
\examples{
cyto_file_name("Experiment-Details.csv")

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
