% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_file_search.R
\name{cyto_file_search}
\alias{cyto_file_search}
\title{Search for CSV file with required information in current directory}
\usage{
cyto_file_search(
  x,
  dir = ".",
  colnames = NULL,
  rownames = NULL,
  ignore.case = TRUE,
  data.table = FALSE,
  type = NULL,
  files = NULL,
  ...
)
}
\arguments{
\item{x}{regular expression to use when searching for files by name in
specified directory.}

\item{dir}{location where the search should be performed.}

\item{colnames}{minimal requirement for column names in the file.}

\item{rownames}{minimal requirement for row names in the file.}

\item{data.table}{logical indicating whether the imported data should be
returned as a \code{data.table} instead of a \code{data.frame}, set to
FALSE by default.}

\item{type}{for internal use only to specify the type of data for which the
search has been performed, simply used to provide informative messages when
multiple files are located.}

\item{files}{optional vector of file names in the current working directory
from which data should be imported. Supplying the file names manually will
bypass the internal file search but still perform the same checks on the
imported data.}

\item{...}{additional named arguments indicating values to find in particular
columns of the file (e.g. name = c("Activation_1.fcs")).}

\item{ignore.cse}{logical passed to \code{grepl()} to indicate whether to
ignore the case when matching the file name, set to TRUE by default.}
}
\value{
NULL or a list named with the file name containing a
  \code{data.frame} or \code{data.table} of the data imported from the
  selected file.
}
\description{
\code{cyto_file_search()} performs a search in the current working directory
for a CSV file matching the supplied file name as well as the specified
column and row name requirements. If multiple files are located and
CytoExploreR is run in interactive mode, the user will be prompted to select
the file manually. Otherwise, \code{cyto_file_search()} will resort to
importing data from the first file name which matches all the requirements.
\code{cyto_file_search()} will automatically import the data from the
requested file and return it in the form of a \code{data.frame} or
\code{data.table} for downstream use.
}
\examples{
\dontrun{
  # search for experiment details of samples
  cyto_file_search(
    "Experiment-Details.csv$",
    rownames = c("Activation_1.fcs", "Activation_2.fcs")
  )
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
