% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_filter}
\alias{cyto_filter}
\title{Select samples based on experiment variables}
\usage{
cyto_filter(x, ...)
}
\arguments{
\item{x}{object of class \code{flowSet} or \code{GatingSet}.}

\item{...}{}
}
\value{
\code{flowSet} or \code{GatingSet} restricted to samples which meet
  the selection criteria.
}
\description{
Select samples based on experiment variables
}
\examples{
library(CytoRSuite)

# Look at experiment details
pData(Activation)

# Select Stim-C samples with 0 and 0.5 OVA concentrations
fs <- cyto_filter(Activation, Treatment == "Stim-C", OVAConc \%in\% c(0,0.5))
  
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
