% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\docType{methods}
\name{cyto_fluor_channels,flowFrame-method}
\alias{cyto_fluor_channels,flowFrame-method}
\title{Extract Fluorescent Channels - flowFrame Method}
\usage{
\S4method{cyto_fluor_channels}{flowFrame}(x)
}
\arguments{
\item{x}{object \code{\link[flowCore:flowFrame-class]{flowFrame}}.}
}
\value{
vector of fluorescent channels.
}
\description{
Extract Fluorescent Channels - flowFrame Method
}
\examples{
library(CytoRSuiteData)

# Load in samples
fs <- Activation

# Get fluorescent channels
cyto_fluor_channels(fs[[1]])
}
\seealso{
\code{\link{cyto_fluor_channels,flowSet-method}}

\code{\link{cyto_fluor_channels,GatingSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
