% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\docType{methods}
\name{cyto_fluor_channels,flowSet-method}
\alias{cyto_fluor_channels,flowSet-method}
\title{Extract Fluorescent Channels - flowSet Method}
\usage{
\S4method{cyto_fluor_channels}{flowSet}(x)
}
\arguments{
\item{x}{object \code{\link[flowCore:flowSet-class]{flowSet}}.}
}
\value{
vector of fluorescent channels.
}
\description{
Extract Fluorescent Channels - flowSet Method
}
\examples{
library(CytoRSuiteData)

# Load in samples
fs <- Activation

# get fluorescent channels
cyto_fluor_channels(fs)
}
\seealso{
\code{\link{cyto_fluor_channels,flowFrame-method}}

\code{\link{cyto_fluor_channels,GatingSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
