% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\docType{methods}
\name{cyto_fluor_channels,GatingHierarchy-method}
\alias{cyto_fluor_channels,GatingHierarchy-method}
\title{Extract Fluorescent Channels - GatingHierarchy Method}
\usage{
\S4method{cyto_fluor_channels}{GatingHierarchy}(x)
}
\arguments{
\item{x}{object
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}}.}
}
\value{
vector of fluorescent channels.
}
\description{
Extract Fluorescent Channels - GatingHierarchy Method
}
\examples{
library(CytoRSuiteData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Get fluorescent channels
cyto_fluor_channels(gs[[1]])
}
\seealso{
\code{\link{cyto_fluor_channels,flowFrame-method}}

\code{\link{cyto_fluor_channels,flowSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
