% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\docType{methods}
\name{cyto_fluor_channels,GatingSet-method}
\alias{cyto_fluor_channels,GatingSet-method}
\title{Extract Fluorescent Channels - GatingSet Method}
\usage{
\S4method{cyto_fluor_channels}{GatingSet}(x)
}
\arguments{
\item{x}{object \code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}
}
\value{
vector of fluorescent channels.
}
\description{
Extract Fluorescent Channels - GatingSet Method
}
\examples{
library(CytoRSuiteData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Get fluorescent channels
cyto_fluor_channels(gs)
}
\seealso{
\code{\link{cyto_fluor_channels,flowFrame-method}}

\code{\link{cyto_fluor_channels,flowSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
