% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{cyto_fluor_channels}
\alias{cyto_fluor_channels}
\title{Extract Fluorescent Channels}
\usage{
cyto_fluor_channels(x)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}
}
\description{
Extract Fluorescent Channels
}
\seealso{
\code{\link{cyto_fluor_channels,flowFrame-method}}

\code{\link{cyto_fluor_channels,flowSet-method}}

\code{\link{cyto_fluor_channels,GatingSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
