% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_function-helpers.R
\name{cyto_func_call}
\alias{cyto_func_call}
\title{Call a function on a set of arguments}
\usage{
cyto_func_call(FUN, args = NULL, ...)
}
\arguments{
\item{FUN}{a function or name of a function to be called. Namespaced
functions are also supported (i.e. "CytoExploreR::cyto_plot").}

\item{args}{a list of named arguments on which \code{FUN} should be called.
Alternatively, these arguments can be supplied separately by names to
\code{...} below.}

\item{...}{named arguments that will be passed to the specified function when
a named list of arguments is not supplied to \code{args}.}
}
\value{
output of the specified function when once called on the supplied
  arguments.
}
\description{
This function is used within CytoExploreR to call functions from external
packages that are not included as imports.
}
\examples{
library(CytoExploreRData)

# Activation Gatingset
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                          package = "CytoExploreRData"))

cyto_fun_call("cyto_channels", gs)

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
