% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_function-helpers.R
\name{cyto_func_execute}
\alias{cyto_func_execute}
\title{Call a function on its formal arguments}
\usage{
cyto_func_execute(FUN, args = NULL, drop = NULL, ...)
}
\arguments{
\item{FUN}{a function or name of a function to be called. Namespaced
functions are also supported (i.e. "CytoExploreR::cyto_plot").}

\item{args}{a list of named arguments on which \code{FUN} should be called.
Alternatively, these arguments can be supplied separately by names to
\code{...} below.}

\item{drop}{a vector of arguments to be removed from the returned vector of
argument names.}

\item{...}{named arguments that will be passed to the specified function when
a named list of arguments is not supplied to \code{args}.}
}
\value{
output of the specified function when once called on the supplied
  arguments.
}
\description{
\code{cyto_func_execute()} is similar to \code{cyto_fun_call()} with the
exception that \code{FUN} is only called on arguments returned by
\code{cyto_fun_args()}. This mean the function will execute successfully if
erroneous arguments are supplied but additional arguments won't be passed
through \code{...}.
}
\examples{
#' library(CytoExploreRData)

# Activation Gatingset
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                          package = "CytoExploreRData"))

cyto_fun_execute("cyto_channels", gs)

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
