% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_function-helpers.R
\name{cyto_func_match}
\alias{cyto_func_match}
\title{Convert function name to valid function}
\usage{
cyto_func_match(FUN, descend = TRUE)
}
\arguments{
\item{FUN}{a function or name of a function to be called. Namespaced
functions are also supported (i.e. "CytoExploreR::cyto_plot").}

\item{descend}{logical to control whether to search past non-function
objects.}
}
\value{
a valid function.
}
\description{
Convert function name to valid function
}
\examples{
cyto_func_match("uwot::umap")

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
