% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_function-helpers.R
\name{cyto_func_name}
\alias{cyto_func_name}
\title{Get the name of a function as a charcter string}
\usage{
cyto_func_name(FUN)
}
\arguments{
\item{FUN}{character string or function for which the function name should be
extracted.}
}
\value{
a character string with the name of a function.
}
\description{
Get the name of a function as a charcter string
}
\examples{
cyto_func_name("stats::density")

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
