% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gate-helpers.R
\name{cyto_gate_apply}
\alias{cyto_gate_apply}
\title{Apply a list of gates to a cytoframe or cytoset}
\usage{
cyto_gate_apply(x, gate = NA, negate = FALSE, skip = FALSE)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoframe]{cytoframe}} or
\code{\link[flowWorkspace:cytoset]{cytoset}}.}

\item{gate}{a filters, rectangleGate, polygonGate, ellipsoidGate, quadGate or
list of these objects.}

\item{negate}{logical flag indicating whether events outside of the supplied
gates should be included as a separate population, set to FALSE by default.}

\item{skip}{logical indicating whether gating should be skipped for gates
when\code{negate = TRUE}, set to FALSE by default. Setting this option to
TRUE will return a list containing only the negated population.}
}
\value{
a list of cytoframes or cytosets per gate.
}
\description{
\code{cyto_gate_apply()} is a helper function used within CytoExploreR to
handle all gating operations. In particular it ensures appropriate handling
of \code{quadGate} and \code{rectangleGates} associated with
\code{quadGates}.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
