% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gate_auto.R
\name{cyto_gate_auto}
\alias{cyto_gate_auto}
\title{CytoExploreR wrapper around openCyto automated gating methods}
\usage{
cyto_gate_auto(
  x,
  parent = NULL,
  alias = NULL,
  channels = NULL,
  type = "flowClust",
  merge_by = "all",
  gatingTemplate = NULL,
  group_by = NULL,
  slot = NULL,
  events = 1,
  input = "flowFrame",
  inverse = FALSE,
  scale = FALSE,
  seed = 2022,
  ...
)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the parent population to gate using the specified
automated gating algorithm supplied to \code{type}.}

\item{alias}{names(s) to use for the newly gated populations.}

\item{channels}{names of the channels or markers to be be passed to the
automated gating algorithm.}

\item{type}{the name of the automated gating algorithm to use, options
include \code{"mindensity"}, \code{"quantileGate"}, \code{"tailgate"},
\code{"boundary"}, \code{"singletGate"}, \code{"flowClust"},
\code{"quadGate.tmix"} or \code{"flowDensity"}, set to \code{"flowClust"}
by default.}

\item{merge_by}{a vector of experiment variables to merge the data into
groups prior to applying the automated gating algorithm supplied to
\code{type}, set to \code{"all"} by default to merge all samples prior to
gating. If \code{merge_by} is used, the specified automated gating
algorithm will be applied separately to each group of merged samples.}

\item{gatingTemplate}{name of \code{gatingTemplate} csv file to which the
\code{gatingTemplate} entries for the \code{GatingSet} method should be
saved, set to \code{cyto_gatingTemplate_active()} by default.}

\item{group_by}{included for compatibility with \code{openCyto} but is
treated in the same way as \code{merge_by}.}

\item{slot}{a character string or function to apply to the output of a custom
automated gating algorithm to extract the gate object after gating. The
purpose of \code{slot} is to ensure that a valid gate type (i.e. a
rectangleGate) is passed to \code{openCyto} during the gating step.}

\item{events}{indicates the number of events to pass to the automated gating
algorithm.}

\item{input}{indicates the format in which the data should be passed to the
automated gating algorithm passed to \code{type}, options include
\code{"flowFrame"}, \code{"cytoframe"}, \code{"flowSet"}, \code{"cytoset"}
or \code{"matrix"}.}

\item{inverse}{logical to indicate whether inverse data transformations
should be applied to the data prior to passing it to the automated gating
algorithm specified by \code{type}, set to FALSE by default.}

\item{scale}{optional argument to scale each channel prior to passing data to
the specified automated gating algorithm, options include \code{"range"},
\code{"mean"}, \code{"median"} or \code{"zscore"}. Set to \code{FALSE} by
default to bypass the data scaling step.}

\item{seed}{numeric to set seed when \code{events != 1}.}

\item{...}{additional arguments passed to the automated gating function or
\code{openCyto::gs_add_gating_method()}.}
}
\value{
a GatingHierarchy or GatingSet with automated gates added and updated
  in the gatingTemplate.
}
\description{
CytoExploreR wrapper around openCyto automated gating methods
}
\seealso{
\code{\link{cyto_gate_draw}}

\code{\link{cyto_gate_clust}}

\code{\link{cyto_gatingTemplate_edit}}

\code{\link{cyto_gate_remove}}

\code{\link{cyto_gate_rename}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
