% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gate_clean.R
\name{cyto_gate_clean}
\alias{cyto_gate_clean}
\title{Apply anomaly detection algorithms to gate high quality cytometry data}
\usage{
cyto_gate_clean(
  x,
  parent = "root",
  alias = "clean",
  type = "PeacoQC",
  channels = NULL,
  gatingTemplate = NULL,
  input = "flowFrame",
  slot = NULL,
  ...
)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the parental population to sample, set to the
\code{"root"} node by default.}

\item{alias}{name of the population containing high quality events to be
created in the GatingSet, set to \code{"clean"} by default.}

\item{type}{the method to use when cleaning the data, options include
\code{"flowAI"}, \code{"flowClean"}, \code{"flowCut"} or \code{"PeacoQC"},
set to \code{"flowAI"} by default. Custom functions are also supported
through \code{type} if they return a logical vector indicating the events
to gate. The \code{slot} argument provides flexibility over how the
resulting logical vector is obtained from the object returned by the custom
gating function.}

\item{channels}{vector of channels/markers over which anomaly detection
algorithms should be applied to gate high quality events, set to all
channels except \code{"Time"}, \code{"Event-ID"} and \code{"Sample-ID'} by
default.}

\item{gatingTemplate}{name of \code{gatingTemplate} csv file to which the
\code{gatingTemplate} entries for the \code{GatingSet} method should be
saved, set to \code{cyto_gatingTemplate_active()} by default.}

\item{input}{allows flexibility over how the data should be formatted prior
to passing it a custom gating function, set to \code{flowFrame} by default
to match the current behaviour in \code{openCyto}.}

\item{slot}{provides flexibility over how the resulting logical vector is
obtained from the object returned by the custom gating function.}

\item{...}{additional arguments to passed to the gating function.}
}
\value{
a \code{GatingHierarchy} or \code{GatingSet} with new node containing
  high quality events and an updated \code{gatingTemplate}.
}
\description{
Apply anomaly detection algorithms to gate high quality cytometry data
}
\examples{
\dontrun{
library(CytoExploreRData)

# Prepare Activation GatingSet
gs <- GatingSet(Activation)
gs <- cyto_compensate(gs)
gs <- cyto_transform(gs)

# Write gatingTemplate to file
cyto_gatingTemplate_create("gatingTemplate.csv", active = TRUE)

# Gate high quality events
cyto_gate_clean(
  gs,
  parent = "root",
  alias = "clean",
  type = "PeacoQC",
  gatingTemplate = "gatingTemplate.csv"
)
}

}
\seealso{
\code{\link{cyto_slot}}

\code{\link{cyto_clean}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
