% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gate_clust.R
\name{cyto_gate_clust}
\alias{cyto_gate_clust}
\title{Use clustering algorithm to gate populations}
\usage{
cyto_gate_clust(
  x,
  parent = NULL,
  alias = NULL,
  channels = NULL,
  type = "FlowSOM",
  merge_by = "all",
  gatingTemplate = NULL,
  group_by = NULL,
  slot = NULL,
  input = "flowFrame",
  inverse = FALSE,
  ...
)
}
\arguments{
\item{x}{object of class \code{GatingSet}.}
}
\description{
Use clustering algorithm to gate populations
}
