% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gate-helpers.R
\name{cyto_gate_convert}
\alias{cyto_gate_convert}
\alias{cyto_gate_convert.rectangleGate}
\alias{cyto_gate_convert.polygonGate}
\alias{cyto_gate_convert.ellipsoidGate}
\alias{cyto_gate_convert.quadGate}
\alias{cyto_gate_convert.filters}
\alias{cyto_gate_convert.list}
\title{Convert Between 1D and 2D Gate Objects}
\usage{
\method{cyto_gate_convert}{rectangleGate}(x, channels = NULL)

\method{cyto_gate_convert}{polygonGate}(x, channels = NULL)

\method{cyto_gate_convert}{ellipsoidGate}(x, channels = NULL)

\method{cyto_gate_convert}{quadGate}(x, channels = NULL)

\method{cyto_gate_convert}{filters}(x, channels = NULL)

\method{cyto_gate_convert}{list}(x, channels = NULL)
}
\arguments{
\item{x}{gate object(s) to be converted.}

\item{channels}{indicates the required dimensions of the gate for plotting.}
}
\value{
modified gate object with appropriate dimensions.
}
\description{
Useful function to convert between 1D and 2D gates once a new plot has been
called.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
