% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gate_draw-methods.R
\name{cyto_gate_draw.flowFrame}
\alias{cyto_gate_draw.flowFrame}
\title{cyto_gate_draw flowFrame Method.}
\usage{
\method{cyto_gate_draw}{flowFrame}(x, alias = NULL, channels = NULL,
  type = NULL, display = 1, overlay = NA, axis = "x",
  label = TRUE, plot = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}}.}

\item{alias}{the name(s) of the populations to be gated. If multiple
population names are supplied (e.g. \code{c("CD3,"CD4)}) multiple gates
will be returned. \code{alias} is \code{NULL} by default which will halt
the gating routine.}

\item{channels}{vector of channel names to use for plotting, can be of length
1 for 1-D density histogram or length 2 for 2-D scatter plot.}

\item{type}{vector of gate type names used to construct the gates. Multiple
gate types are supported but should be accompanied with an \code{alias}
argument of the same length (i.e. one \code{type} per \code{alias}).
Supported gate types include \code{polygon, rectangle, ellipse, threshold,
boundary, interval, quadrant and web} which can be abbreviated as upper or
lower case first letters as well. Default \code{type} is \code{"interval"}
for 1D gates and \code{"polygon"} for 2D gates.}

\item{display}{numeric [0,1] to control the percentage of events to be
plotted. Specifying a value for \code{display} can substantial improve
plotting speed for less powerful machines.}

\item{overlay}{a \code{flowFrame}, \code{flowSet} or list of flowFrames to
overlay. To improve plotting speed, overlays are subjected to the same
\code{display} requirements as the underlying \code{flowFrame}.}

\item{axis}{indicates whether the \code{"x"} or \code{"y"} axis should be
gated for 2-D interval gates.}

\item{label}{logical indicating whether to include
\code{\link{cyto_plot_label}} for the gated population(s), \code{TRUE} by
default.}

\item{plot}{logical indicating whether a plot should be drawn, set to
\code{TRUE} by default.}

\item{...}{additional arguments for \code{\link{cyto_plot,flowFrame-method}}.}
}
\value{
a \code{\link[flowCore:filters-class]{filters}} list containing the
  drawn gate objects.
}
\description{
Manually draw gates around populations for analysis of flow cytometry data.
}
\examples{
\dontrun{
library(CytoRSuiteData)

# Load in samples
fs <- Activation

# draw gates using cyto_gate_draw - add contour lines & overlay control
gt <- cyto_gate_draw(fs[[4]],
  channels = c("FSC-A", "SSC-A"),
  alias = "Cells",
  type = "ellipse",
  contour_lines = 15,
  overlay = fs[[1]]
)

# gt is a filters object containing the contructed ellipsoidGate
gt[[1]]
}

}
\seealso{
\code{\link{cyto_plot,flowFrame-method}}

\code{\link{cyto_gate_draw,flowSet-method}}

\code{\link{cyto_gate_draw,GatingSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
