% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gate_draw-methods.R
\name{cyto_gate_draw.flowSet}
\alias{cyto_gate_draw.flowSet}
\title{cyto_gate_draw flowSet Method}
\usage{
\method{cyto_gate_draw}{flowSet}(x, alias = NULL, channels = NULL,
  type = NULL, display = NULL, group_by = "all", select = NULL,
  axis = "x", label = TRUE, plot = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowSet-class]{flowSet}}.}

\item{alias}{the name(s) of the populations to be gated. If multiple
population names are supplied (e.g. \code{c("CD3,"CD4)}) multiple gates
will be returned. \code{alias} is \code{NULL} by default which will halt
the gating routine.}

\item{channels}{vector of channel names to use for plotting, can be of length
1 for 1-D density histogram or length 2 for 2-D scatter plot.}

\item{type}{vector of gate type names used to construct the gates. Multiple
gate types are supported but should be accompanied with an \code{alias}
argument of the same length (i.e. one \code{type} per \code{alias}).
Supported gate types are \code{polygon, rectangle, ellipse, threshold,
boundary, interval, quadrant and web} which can be abbreviated as upper or
lower case first letters as well. Default \code{type} is \code{"interval"}
for 1D gates and \code{"polygon"} for 2D gates.}

\item{display}{numeric [0,1] to control the percentage of events to be
plotted. Specifying a value for \code{display} can substantial improve
plotting speed for less powerful machines.}

\item{group_by}{vector of pData column names (e.g.
c("Treatment","Concentration") indicating how the samples should be grouped
prior to gating, set to "all" by default to construct a single gate for all
samples. If group_by variables are supplied a different gate will be
constructed for each group.}

\item{select}{designates which samples will be plotted and used for
determining the best location to set the drawn gate(s). Filtering steps
should be comma separated and wrapped in a list (e.g. list(Treatment ==
"Stim-A", OVAConc %in% c(0,0.5))).}

\item{axis}{indicates whether the \code{"x"} or \code{"y"} axis should be
gated for 2-D interval gates.}

\item{label}{logical indicating whether to include
\code{\link{cyto_plot_label}} for the gated population(s), \code{TRUE} by
default.}

\item{plot}{logical indicating whether a plot should be drawn, set to
\code{TRUE} by default.}

\item{...}{additional arguments for \code{\link{cyto_plot,flowSet-method}}.}
}
\value{
a \code{\link[flowCore:filters-class]{filters}} object or list of
  \code{\link[flowCore:filters-class]{filters}} objects when grouping is
  applied.
}
\description{
Manually draw gates around populations for analysis of flow cytometry data.
}
\examples{
\dontrun{
library(CytoRSuiteData)

# Load in samples
fs <- Activation

# draw gates using cyto_gate_draw - add contour lines & overlay control
cyto_gate_draw(fs,
  channels = c("FSC-A", "SSC-A"),
  alias = "Cells",
  type = "polygon",
  contour_lines = 15,
  overlay = fs[[1]]
)

# gt is a filters object containing the contructed polygonGate
gt[[1]]
}

}
\seealso{
\code{\link{cyto_plot,flowSet-method}}

\code{\link{cyto_gate_draw,flowFrame-method}}

\code{\link{cyto_gate_draw,GatingSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
