% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gate_draw.R
\name{cyto_gate_draw}
\alias{cyto_gate_draw}
\alias{cyto_gate_draw.GatingSet}
\alias{cyto_gate_draw.flowSet}
\alias{cyto_gate_draw.flowFrame}
\title{Interactively draw and construct gate objects}
\usage{
\method{cyto_gate_draw}{GatingSet}(
  x,
  parent = "root",
  alias = NULL,
  channels = NULL,
  type = NULL,
  gatingTemplate = NULL,
  overlay = NA,
  group_by = "all",
  select = NULL,
  negate = FALSE,
  display = 25000,
  axis = "x",
  label = TRUE,
  plot = TRUE,
  popup = TRUE,
  axes_limits = "machine",
  gate_point_shape = 16,
  gate_point_size = 1,
  gate_point_col = "red",
  gate_point_col_alpha = 1,
  gate_line_type = 1,
  gate_line_width = 2.5,
  gate_line_col = "red",
  gate_line_col_alpha = 1,
  ...
)

\method{cyto_gate_draw}{flowSet}(
  x,
  alias = NULL,
  channels = NULL,
  type = NULL,
  overlay = NA,
  group_by = "all",
  select = NULL,
  negate = FALSE,
  display = 25000,
  axis = "x",
  label = TRUE,
  plot = TRUE,
  popup = TRUE,
  axes_limits = "machine",
  gate_point_shape = 16,
  gate_point_size = 1,
  gate_point_col = "red",
  gate_point_col_alpha = 1,
  gate_line_type = 1,
  gate_line_width = 2.5,
  gate_line_col = "red",
  gate_line_col_alpha = 1,
  ...
)

\method{cyto_gate_draw}{flowFrame}(
  x,
  alias = NULL,
  channels = NULL,
  type = NULL,
  overlay = NA,
  negate = FALSE,
  display = 25000,
  axis = "x",
  label = TRUE,
  plot = TRUE,
  popup = TRUE,
  axes_limits = "machine",
  gate_point_shape = 16,
  gate_point_size = 1,
  gate_point_col = "red",
  gate_point_col_alpha = 1,
  gate_line_type = 1,
  gate_line_width = 2.5,
  gate_line_col = "red",
  gate_line_col_alpha = 1,
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the \code{parent} population to extract for gating when
a \code{GatingSet} object is supplied.}

\item{alias}{the name(s) of the populations to be gated. If multiple
population names are supplied (e.g. \code{c("CD3,"CD4)}) multiple gates
will be returned. \code{alias} is \code{NULL} by default which will halt
the gating routine.}

\item{channels}{vector of channel names to use for plotting, can be of length
1 for 1-D density histogram or length 2 for 2-D scatter plot.}

\item{type}{vector of gate type names used to construct the gates. Multiple
gate types are supported but should be accompanied with an \code{alias}
argument of the same length (i.e. one \code{type} per \code{alias}).
Supported gate types are \code{polygon, rectangle, ellipse, threshold,
boundary, interval, quadrant and web} which can be abbreviated as upper or
lower case first letters as well. Default \code{type} is \code{"interval"}
for 1D gates and \code{"polygon"} for 2D gates.}

\item{gatingTemplate}{name of \code{gatingTemplate} csv file to which the
\code{gatingTemplate} entries for the \code{GatingSet} method should be
saved.}

\item{overlay}{name(s) of the populations to overlay or a \code{flowFrame},
\code{flowSet}, \code{list of flowFrames} or \code{list of flowSets}
containing populations to be overlaid onto the plot(s).}

\item{group_by}{vector of \code{\link{cyto_details}} column names (e.g.
c("Treatment","Concentration") indicating how the samples should be grouped
prior to gating, set to the length of x by default to construct a single
gate for all samples. If group_by is supplied a different gate will be
constructed for each group. This argument only applies to \code{flowSet} or
\code{GatingSet} objects.}

\item{select}{designates which samples will be plotted and used for
determining the best location to set the drawn gate(s). Filtering steps
should be comma separated and wrapped in a list. Refer to
\code{\link{cyto_select}}.}

\item{negate}{logical indicating whether a gatingTemplate entry should be
made for the negated population (i.e. all events outside the constructed
gates), set to FALSE by default. If negate is set to TRUE, a name for the
negated population MUST be supplied at the end of the alias argument.}

\item{display}{fraction or number of events to display in the plot during the
gating process, set to 25 000 events by default.}

\item{axis}{indicates whether the \code{"x"} or \code{"y"} axis should be
gated for 2-D interval gates.}

\item{label}{logical indicating whether to include
\code{\link{cyto_plot_label}} for the gated population(s), \code{TRUE} by
default.}

\item{plot}{logical indicating whether a plot should be drawn, set to
\code{TRUE} by default.}

\item{popup}{logical indicating whether the plot should be constructed in a
pop-up window, set to TRUE by default.}

\item{axes_limits}{options include \code{"auto"}, \code{"data"} or
\code{"machine"} to use optimised, data or machine limits respectively. Set
to \code{"machine"} by default to use entire axes ranges. Fine control over
axes limits can be obtained by altering the \code{xlim} and \code{ylim}
arguments.}

\item{gate_point_shape}{shape to use for selected gate points, set to
\code{16} by default to use filled circles. See
\code{\link[graphics:par]{pch}} for alternatives.}

\item{gate_point_size}{numeric to control the size of the selected gate
points, set to 1 by default.}

\item{gate_point_col}{colour to use for the selected gate points, set to
"red" by default.}

\item{gate_point_col_alpha}{numeric [0,1] to control the transparency of the
selected gate points, set to 1 by default to use solid colours.}

\item{gate_line_type}{integer [0,6] to control the line type of gates, set to
\code{1} to draw solid lines by default. See
\code{\link[graphics:par]{lty}} for alternatives.}

\item{gate_line_width}{numeric to control the line width(s) of gates, set to
\code{2.5} by default.}

\item{gate_line_col}{colour to use for gates, set to \code{"red"} by default.}

\item{gate_line_col_alpha}{numeric [0,1] to control the transparency of the
selected gate lines, set to 1 by default to use solid colours.}

\item{...}{additional arguments for \code{\link{cyto_plot}}.}
}
\value{
\code{flowFrame} and \code{flowSet} methods return a list of flowCore
  gate objects per group in group_by, whilst the \code{GatingSet} applies the
  constructed gates directly to the \code{GatingSet} and adds appropriate
  entries into the specified \code{gatingTemplate}. The \code{GatingSet}
  method does not return the constructed gates but instead invisibly returns
  the \code{gatingTemplate} entries.
}
\description{
\code{cyto_gate_draw} allows to users to interactive draw gates around
populations which are returned as \code{flowCore} gate objects. The flowFrame
and flowSet methods simply return the constructed gates as a list of
flowCore-compatible gate objects, whilst the GatingSet method automatically
applies the constructed gates to the GatingSet and saves the constructed
gates in an \code{openCyto}
\code{\link[openCyto:gatingTemplate-class]{gatingTemplate}} for future use.
See \code{\link{cyto_gate_edit}}, \code{\link{cyto_gate_remove}} and
\code{\link{cyto_gate_rename}} to manipulate constructed gates and modify
their entries in the gatingTemplate.
}
\examples{
\dontrun{
# Gate drawing requires an interactive R session
library(CytoExploreRData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(fs))
gs <- transform(gs, trans)

# Gate using cyto_gate_draw
gt_gating(Activation_gatingTemplate, gs)

# draw gates using cyto_gate_draw
cyto_gate_draw(gs,
  parent = "Dendritic Cells",
  channels = c("Alexa Fluor 488-A", "Alexa Fluor 700-A"),
  alias = c("CD8+ DC", "CD4+ DC"),
  gatingTemplate = "Example-gatingTemplate.csv",
  type = "rectangle",
  contour_lines = 15
)

# Constructed gate applied directly to GatingSet
cyto_nodes(gs)
}

}
\seealso{
\code{\link{cyto_plot}}

\code{\link{cyto_gate_edit}}

\code{\link{cyto_gate_remove}}

\code{\link{cyto_gate_rename}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
