% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gate-helpers.R
\name{cyto_gate_edit}
\alias{cyto_gate_edit}
\title{Edit gates that exist in a GatingSet and gatingTemplate}
\usage{
cyto_gate_edit(
  x,
  parent = NULL,
  alias = NULL,
  channels = NULL,
  type = NULL,
  gatingTemplate = NULL,
  overlay = NA,
  merge_by = "all",
  group_by = NULL,
  select = NULL,
  negate = FALSE,
  display = 50000,
  axis = "x",
  label = TRUE,
  plot = TRUE,
  title = "",
  axes_trans = NA,
  axes_limits = "machine",
  gate_point_shape = 16,
  gate_point_size = 1,
  gate_point_col = "red",
  gate_point_col_alpha = 1,
  gate_line_type = 1,
  gate_line_width = 2.5,
  gate_line_col = "red",
  gate_line_col_alpha = 1,
  label_text_size = 1,
  label_text_font = 2,
  label_text_col = "black",
  label_text_col_alpha = 1,
  label_fill = "white",
  label_fill_alpha = 0.75,
  seed = 42,
  ...
)
}
\arguments{
\item{x}{an object of class \code{GatingSet}.}

\item{parent}{name of the parental population.}

\item{alias}{name(s) of the gate to edit (e.g. "Single Cells").}

\item{channels}{name(s) of the channel(s) used to construct the gate(s). This
argument is not necessary but is included to allow conversion of
\code{cyto_gate_draw} code to \code{cyto_gate_remove} code by simply
changing \code{"draw"} to \code{"remove"}.}

\item{type}{vector of gate type names used to construct the gates. Multiple
\code{types} are supported but should be accompanied with an \code{alias}
argument of the same length (i.e. one \code{type} per \code{alias}).
Supported \code{gate_types} are \code{polygon, rectangle, ellipse,
threshold, boundary, interval, quadrant and web} which can be abbreviated
as upper or lower case first letters as well. Default \code{type} is
\code{"polygon"}.}

\item{gatingTemplate}{name of the \code{gatingTemplate} csv file (e.g.
"gatingTemplate.csv") where the gate is saved.}

\item{overlay}{name(s) of the populations to overlay or a \code{flowFrame},
\code{flowSet}, \code{list of flowFrames} or \code{list of flowSets}
containing populations to be overlaid onto the plot(s). Only overlaid
flowSet objects are subjected to sampling by \code{display}.}

\item{merge_by}{vector of pData column names (e.g.
c("Treatment","Concentration") indicating how the samples should be grouped
prior to gating, set to the length of x by default to construct a single
gate for all samples. If merge_by is supplied a different gate will be
constructed for each group.}

\item{group_by}{same as \code{merge_by} included for backwards compatibility
with older versions of CytoExploreR, \code{group_by} was renamed to
\code{merge_by} in CytoExploreR v2.0.0 to maintain consistency with
\code{cyto_plot}. Users should therefore use \code{merge_by} as support for
\code{group_by} will be ended in the future.}

\item{select}{vector containing the indices of samples within gs to use for
plotting.}

\item{negate}{logical indicating whether a gatingTemplate entry should be
made for the negated population (i.e. all events outside the constructed
gates), set to FALSE by default. If negate is set to TRUE, a name for the
negated population MUST be supplied at the end of the alias argument.}

\item{display}{fraction or number of events to display in the plot during the
gating process, set to 25 000 events by default.}

\item{axis}{indicates whether the \code{"x"} or \code{"y"} axis should be
gated for 2-D interval gates.}

\item{label}{logical indicating whether to include
\code{\link{cyto_plot_label}} for the gated population(s), \code{TRUE} by
default.}

\item{plot}{logical indicating whether a plot should be drawn, set to
\code{TRUE} by default.}

\item{axes_limits}{options include \code{"auto"}, \code{"data"} or
\code{"machine"} to use optimised, data or machine limits respectively. Set
to \code{"machine"} by default to use entire axes ranges. Fine control over
axes limits can be obtained by altering the \code{xlim} and \code{ylim}
arguments.}

\item{gate_point_shape}{shape to use for selected gate points, set to
\code{16} by default to use filled circles. See
\code{\link[graphics:par]{pch}} for alternatives.}

\item{gate_point_size}{numeric to control the size of the selected gate
points, set to 1 by default.}

\item{gate_point_col}{colour to use for the selected gate points, set to
"red" by default.}

\item{gate_point_col_alpha}{numeric [0,1] to control the transparency of the
selected gate points, set to 1 by default to use solid colours.}

\item{gate_line_type}{integer [0,6] to control the line type of gates, set to
\code{1} to draw solid lines by default. See
\code{\link[graphics:par]{lty}} for alternatives.}

\item{gate_line_width}{numeric to control the line width(s) of gates, set to
\code{2.5} by default.}

\item{gate_line_col}{colour to use for gates, set to \code{"red"} by default.}

\item{gate_line_col_alpha}{numeric [0,1] to control the transparency of the
selected gate lines, set to 1 by default to use solid colours.}

\item{label_text_size}{numeric to control the size of text in the plot
labels, set to 1 by default.}

\item{label_text_font}{numeric to control the font of text in plot labels,
set to 2 for bold font by default. See \code{\link[graphics:par]{font}} for
alternatives.}

\item{label_text_col}{colour(s) to use for text in plot labels, set to
\code{"black"} by default.}

\item{label_text_col_alpha}{numeric [0, 1] to control the transparency of the
text colour, set to 1 by default to remove transparency.}

\item{label_fill}{fill colour(s) to use for labels, set to "white" by
default.}

\item{label_fill_alpha}{numeric to control background fill transparency of
label, set to 0.6 by default to introduce some transparency.}

\item{seed}{numeric passed to \code{\link{set.seed}} to ensure that the same
sampling is applied with each \code{\link{cyto_plot}} call, set to an
arbitrary numeric by default. This behaviour can be turned off by setting
this argument to NULL.}

\item{...}{additional arguments for \code{\link{cyto_plot}}.}

\item{popup}{logical indicating whether the plot should be constructed in a
pop-up window, set to TRUE by default.}
}
\value{
an object of class \code{GatingSet} with edited gate applied, as well
  as gatingTemplate file with edited gate saved.
}
\description{
Edit gates that exist in a GatingSet and gatingTemplate
}
\examples{
\dontrun{
library(CytoExploreRData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- cyto_compensate(gs)

# Transform fluorescent channels
gs <- cyto_transform(gs)

# Gate using cyto_gate_draw
gt <- Activation_gatingTemplate
gt_gating(gt, gs)

# Edit CD4 T Cells Gate - replace gatingTemplate name
cyto_gate_edit(gs,
  parent = "T Cells",
  alias = "CD4 T Cells",
  gatingTemplate = "gatingTemplate.csv"
)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
