% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gate-helpers.R
\name{cyto_gate_extract}
\alias{cyto_gate_extract}
\title{Extract Saved Gate(s) from GatingHierarchy, GatingSet or gatingTemplate.}
\usage{
cyto_gate_extract(
  x,
  parent = NULL,
  alias,
  select = NULL,
  merge_by = "name",
  path = "auto",
  bool = TRUE,
  ...
)
}
\arguments{
\item{x}{object of class \code{GatingHierarchy}, \code{GatingSet},
\code{gatingTemplate} or the name of a gatingTemplate CSV file.}

\item{parent}{name of the parental population.}

\item{alias}{name of the population for which the gate must be extracted.}

\item{select}{passed to \code{cyto_select}.}

\item{merge_by}{passed to \code{cyto_groups} to split \code{GatingSet} into
groups, gates will be extracted from the first GatingHierarchy within each
group.}

\item{path}{passed to \code{cyto_nodes()}.}

\item{bool}{logical indicating whether booleanFilters should be converted
into complementFilters, set to TRUE by default. Only negated booleanFilters
are supported.}

\item{...}{not in use.}
}
\description{
Extract Saved Gate(s) from GatingHierarchy, GatingSet or gatingTemplate.
}
\examples{
library(CytoExploreRData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- cyto_compensate(gs)

# Transform fluorescent channels
gs <- cyto_transform(gs)

# Gate using cyto_gate_draw
gt <- Activation_gatingTemplate
gt_gating(gt, gs)

# gatingTemplate
gtfile <- system.file("extdata",
  "Activation-gatingTemplate.csv",
  package = "CytoExploreRData"
)

# Extract T Cells gate
cyto_gate_extract(gs, "Live Cells", "T Cells", gatingTemplate = gtfile)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
