% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gate-helpers.R
\name{cyto_gate_indices}
\alias{cyto_gate_indices}
\title{Extract event level gate indices for samples in a GatingSet}
\usage{
cyto_gate_indices(x, parent = "root", select = NULL, nodes = NULL)
}
\arguments{
\item{x}{object of class GatingHierarchy or GatingSet.}

\item{parent}{name of the parent population from which the gate indices
should be returned, set to the \code{"root"} node by default.}

\item{select}{list of selection criteria passed to \code{cyto_select()} to
restrict the samples prior to extracting the population indices for each
node.}

\item{nodes}{names of the populations for which the indices should be
returned, set to all nodes except the \code{"root"} node by default.}
}
\value{
a named list containing a matrix of indices for each of the supplied
  nodes.
}
\description{
Extract event level gate indices for samples in a GatingSet
}
\examples{
\dontrun{
library(CytoExploreRData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- cyto_compensate(gs)

# Transform fluorescent channels
gs <- cyto_transform(gs)

# Gate using cyto_gate_draw
gs <- cyto_gatingTemplate_apply(gs, Activation_gatingTemplate)

# Remove T Cells population - replace gatingTemplate name
cyto_gate_indices(
  gs,
  select = 32,
  nodes = c("T Cells", "CD4 T Cells", "CD8 T Cells")
)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
