% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gate-helpers.R
\name{cyto_gate_remove}
\alias{cyto_gate_remove}
\title{Remove Gate(s) and Edit gatingTemplate csv File}
\usage{
cyto_gate_remove(
  x,
  parent = NULL,
  alias = NULL,
  channels = NULL,
  type = NULL,
  gatingTemplate = NULL,
  ...
)
}
\arguments{
\item{x}{object of class \code{GatingHierarchy} or \code{GatingSet}.}

\item{parent}{name of the parent population from which to remove the gate.
This argument is not necessary but is included to allow conversion of
\code{cyto_gate_draw} code to \code{cyto_gate_remove} code by simply
changing \code{"draw"} to \code{"remove"}.}

\item{alias}{name(s) of the population(s) to remove (e.g. "Single Cells"). By
default all descendant populations will be removed as well.}

\item{channels}{names of the channel(s) used to gate the population. This
argument is not necessary but is included to allow conversion of
\code{cyto_gate_draw} code to \code{cyto_gate_remove} code by simply
changing \code{"draw"} to \code{"remove"}.}

\item{type}{gate type(s) used to for the gates to be removed. This argument
is not necessary but is included to allow conversion of
\code{cyto_gate_draw} code to \code{cyto_gate_remove} code by simply
changing \code{"draw"} to \code{"remove"}.}

\item{gatingTemplate}{name of the \code{gatingTemplate} csv file (e.g.
"gatingTemplate.csv").}

\item{...}{additional \code{cyto_gate_draw} arguments that will be ignored.}
}
\value{
an object of class \code{GatingSet} with gate and children removed
  and updated gatingTemplate to reflect these changes.
}
\description{
Remove Gate(s) and Edit gatingTemplate csv File
}
\examples{
\dontrun{
library(CytoExploreRData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- cyto_compensate(gs)

# Transform fluorescent channels
gs <- cyto_transform(gs)

# Gate using cyto_gate_draw
gt <- Activation_gatingTemplate
gt_gating(gt, gs)

# Remove T Cells population - replace gatingTemplate name
cyto_gate_remove(gs, "T Cells", gatingTemplate = "gatingTemplate.csv")
}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
