% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gate-helpers.R
\name{cyto_gate_rename}
\alias{cyto_gate_rename}
\title{Rename Gates}
\usage{
cyto_gate_rename(x, alias = NULL, names = NULL,
  gatingTemplate = NULL)
}
\arguments{
\item{x}{object of class \code{GatingHierarchy} or \code{GatingSet}.}

\item{alias}{current names of the gates to be changed.}

\item{names}{new names to use for \code{alias}.}

\item{gatingTemplate}{name of the gatingTemplate csv file to be edited.}
}
\value{
update gate names in \code{GatingSet} or \code{GatingHierarchy} and
  update the gatingTemplate accordingly.
}
\description{
Rename Gates
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
