% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gate_sample.R
\name{cyto_gate_sample}
\alias{cyto_gate_sample}
\title{Gate a sample of a node in a GatingHierarchy or GatingSet}
\usage{
cyto_gate_sample(
  x,
  parent = "root",
  alias = NULL,
  merge_by = "name",
  events = 1,
  gatingTemplate = NULL,
  seed = 56,
  group_by = NULL,
  ...
)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the parental population to sample, set to the
\code{"root"} node by default.}

\item{alias}{name of the sampled population to be created in the GatingSet.}

\item{merge_by}{vector of \code{\link{cyto_details}} column names (e.g.
c("Treatment","Concentration")) indicating how the samples should be grouped
prior to sampling, set to "name" by default to downsample each sample
separately.}

\item{events}{numeric to control the number of events to downsample each
group in \code{merge_by}, set to 1 by default to include all events.
\code{events} can be supplied per group otherwise the same degree of
sampling will be applied to all groups. \code{cyto_gate_sample()} calls
\code{cyto_sample_n()} on each group to determine the proportionate number
of events to gate in each sample.}

\item{gatingTemplate}{name of \code{gatingTemplate} csv file to which the
\code{gatingTemplate} entries for the \code{GatingSet} method should be
saved, set to \code{cyto_gatingTemplate_active()} by default.}

\item{seed}{numeric passed to \code{\link{set.seed}} to ensure that the same
sampling is applied with each call to \code{cyto_gate_sample()}, set to an
arbitrary numeric by default. This behaviour can be turned off by setting
this argument to NULL.}

\item{group_by}{same as \code{merge_by} included for backwards compatibility
with older versions of CytoExploreR, \code{group_by} was renamed to
\code{merge_by} in CytoExploreR v2.0.0 to maintain consistency with
\code{cyto_plot}. Users should therefore use \code{merge_by} as support for
\code{group_by} will be ended in the future.}

\item{...}{not in use.}
}
\value{
a \code{GatingHierarchy} or \code{GatingSet} with new sampled node
  added and updated \code{gatingTemplate}.
}
\description{
\code{cyto_gate_sample()} provides a way to downsample events within nodes in
a GatingHierarchy or GatingSet and store these events as a new node.
}
\examples{
\dontrun{
library(CytoExploreRData)

# Prepare Activation GatingSet
gs <- GatingSet(Activation)
gs <- cyto_compensate(gs)
gs <- cyto_transform(gs)
gs <- cyto_gatingTemplate_apply(gs, Activation_gatingTemplate)

# Write gatingTemplate to file
cyto_gatingTemplate_write(Activation_gatingTemplate, "gatingTemplate.csv")

# Add sampled T Cells node - 500 events per Treatment group
cyto_gate_sample(
  gs,
  parent = "T Cells",
  alias = "T Cells Sample",
  merge_by = "Treatment",
  events = 500,
  gatingTemplate = "gatingTemplate.csv"
)
}

}
\seealso{
\code{\link{cyto_sample}}

\code{\link{cyto_gatingTemplate_active}}

\code{\link{cyto_gatingTemplate_edit}}

\code{\link{cyto_gate_remove}}

\code{\link{cyto_gate_rename}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
