% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_som.R
\name{cyto_gate_som}
\alias{cyto_gate_som}
\title{Map data to a gated self-organising map (SOM)}
\usage{
cyto_gate_som(
  x,
  parent = "root",
  som = NULL,
  alias = NULL,
  channels = NULL,
  hidden = TRUE,
  ...
)
}
\arguments{
\item{x}{an object of class
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the parent population which should be gated using the
SOM, ideally the same population used to train the SOM in
\code{cyto_som()}.}

\item{som}{object of class \code{GatingSet} containing the codebook vectors
for the trained SOM.}

\item{alias}{name(s) of the populations gated on the SOM that should be
transferred back onto the original data, set to the \code{descendants} of
the \code{"root"} node by default.}

\item{channels}{name(s) of the channel(s) or marker(s) that should be used to
map the data to the SOM, ideally the same set of channels used to train the
SOM in \code{cyto_som()}.}

\item{hidden}{logical to indicate whether hidden nodes be included as
descendants in \code{alias} when \code{alias} is not manually supplied, set
to TRUE by default.}

\item{...}{additional arguments passed to \code{cyto_som()} to train a new
SOM when a \code{som} is not supplied manually.}
}
\value{
a GatingSet containing the original data with gates inherited from
  the gated SOM.
}
\description{
\code{cyto_som()} provides a way to compress the original data into a more
manageable format for downstream processing. \code{cyto_gate_som()} transfers
the gates from a gated SOM \code{GatingSet} produced by \code{cyto_som()}
back onto the original data stored in a separate \code{GatingSet}.
}
\seealso{
\code{\link{cyto_som}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
