% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gate_transfer.R
\name{cyto_gate_transfer}
\alias{cyto_gate_transfer}
\alias{cyto_gate_transfer.GatingHierarchy}
\alias{cyto_gate_transfer.GatingSet}
\title{Transfer all gates from one GatingHierarchy/GatingSet to another}
\usage{
\method{cyto_gate_transfer}{GatingHierarchy}(from, to)

\method{cyto_gate_transfer}{GatingSet}(from, to)
}
\arguments{
\item{from}{object of class
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} from which the gates
should be extracted.}

\item{to}{object of class
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} which should inherit
the transferred gates.}
}
\value{
a GatingHierarchy or GatingSet containing the transferred gates.
}
\description{
Transfer all gates from one GatingHierarchy/GatingSet to another
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
