% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gate-helpers.R
\name{cyto_gate_transform}
\alias{cyto_gate_transform}
\alias{cyto_gate_transform.rectangleGate}
\alias{cyto_gate_transform.polygonGate}
\alias{cyto_gate_transform.ellipsoidGate}
\alias{cyto_gate_transform.quadGate}
\alias{cyto_gate_transform.multiRangeGate}
\alias{cyto_gate_transform.filters}
\alias{cyto_gate_transform.list}
\title{Transform gates objects}
\usage{
\method{cyto_gate_transform}{rectangleGate}(x, trans = NULL, inverse = FALSE, ...)

\method{cyto_gate_transform}{polygonGate}(x, trans = NULL, inverse = FALSE, ...)

\method{cyto_gate_transform}{ellipsoidGate}(x, trans = NULL, inverse = FALSE, ...)

\method{cyto_gate_transform}{quadGate}(x, trans = NULL, inverse = FALSE, ...)

\method{cyto_gate_transform}{multiRangeGate}(x, trans = NULL, inverse = FALSE, ...)

\method{cyto_gate_transform}{filters}(x, trans = NULL, inverse = FALSE, ...)

\method{cyto_gate_transform}{list}(x, trans = NULL, inverse = FALSE, ...)
}
\arguments{
\item{x}{object of class rectangleGate, polygonGate, ellipsoidGate, quadGate,
filters or a list of these gate objects.}

\item{trans}{list of transformers to apply to the gate co-ordinates.}

\item{inverse}{logical indicating whether the inverse transformations should
be applied, set to FALSE by default.}

\item{...}{not in use.}
}
\value{
transformed gate object.
}
\description{
Transform gates objects
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
