% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gate-helpers.R
\name{cyto_gate_type}
\alias{cyto_gate_type}
\title{Get Gate Type from Saved Gate.}
\usage{
cyto_gate_type(gates)
}
\arguments{
\item{gates}{an object of class \code{filters} containing the gates from
which the \code{type(s)} must be obtained.}
}
\value{
vector of gate type names for supplied gates.
}
\description{
Get Gate Type from Saved Gate.
}
\examples{
library(CytoRSuiteData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(gs))
gs <- transform(gs, trans)

# Gate using cyto_gate_draw
gt <- Activation_gatingTemplate
gating(gt, gs)

# Get gate type used for T Cells gate
gate_type(getGate(gs, "Cells")[[1]])
gate_type(getGate(gs, "T Cells")[[1]])
gate_type(getGate(gs, "CD69+ CD4 T Cells")[[1]])
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
