% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gateTemplate.R
\name{cyto_gateTemplate}
\alias{cyto_gateTemplate}
\alias{cyto_gateTemplate.GatingHierarchy}
\alias{cyto_gateTemplate.GatingSet}
\title{Create a gateTemplate for a GatingHierarchy or GatingSet}
\usage{
\method{cyto_gateTemplate}{GatingHierarchy}(x)

\method{cyto_gateTemplate}{GatingSet}(x)
}
\arguments{
\item{x}{an object of class
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}
}
\value{
a list with class \code{gateTemplate} containing all the gates
  applied to the GatingHierarchy or GatingSet.
}
\description{
\code{cyto_gateTemplate()} creates on object of class \code{gateTemplate}
containing a template for all the gates applied to the supplied
GatingHierarchy or GatingSet. \code{gateTemplate(s)} can then be applied to
other GatingHierarchies or GatingSets using \code{cyto_gateTemplate_apply()}.
These functions are used internally to transfer gates when we need to create
a new GatingHierarchy or GatingSet containing a subset of the original data.
Users should not use these functions directly and instead transfer gates
using gatingTemplates and \code{cyto_gatingTemplate_apply()}.
}
\seealso{
\code{\link{cyto_gateTemplate_apply}}

\code{\link{cyto_gatingTemplate_apply}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
