% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gateTemplate.R
\name{cyto_gateTemplate_apply}
\alias{cyto_gateTemplate_apply}
\title{Apply a set of gates from another GatingHierarchy or GatingSet}
\usage{
cyto_gateTemplate_apply(x, gateTemplate = NULL)
}
\arguments{
\item{x}{an object of class
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} which will inherit
the gates stored in the \code{gateTemplate}.}

\item{gateTemplate}{object created by calling \code{cyto_gateTemplate()} on
the reference GatingHierarchy or GatingSet. Alternatively, the reference
GatingHierarchy or GatingSet can be supplied and a call will be made to
\code{cyto_gateTemplate()} to create the gateTemplate(s).}
}
\value{
a GatingHierarchy or GatingSet with gates from \code{gateTemplate}
  applied.
}
\description{
\code{cyto_gateTemplate_apply()} is designed to be used internally within
CytoExploreR to allow the transfer of gates from one GatingHierarchy or
GatingSet to another. \code{gateTemplate} must be generated by calling
\code{cyto_gateTemplate()} on the reference GatingHierarchy or GatingSet.
Users should instead use gatingTemplates and
\code{cyto_gatingTemplate_apply()} transfer gates between GatingHierarchies
or GatingSets.
}
\seealso{
\code{\link{cyto_gateTemplate}}

\code{\link{cyto_gatingTemplate_apply}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
