% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gatingTemplate-helpers.R
\name{cyto_gatingTemplate}
\alias{cyto_gatingTemplate}
\title{Extract a gatingTemplate from a GtaingHierrachy or GatingSet}
\usage{
cyto_gatingTemplate(x, select = 1, data.table = FALSE)
}
\arguments{
\item{x}{an object of class
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{select}{passed to \code{cyto_select()} to select the GatingHierarchy
from which the gatingTemplate should be extracted, set to 1 by default to
select the first GatingHierarchy.}

\item{data.table}{logical indicating whether the extracted gatingTemplate
should be returned as a \code{data.table}, set to FALSE by default.}
}
\value{
gatingTemplate as either a \code{data.frame} or \code{data.table}.
}
\description{
Simply a wrapper around
\code{\link[openCyto:gh_generate_template]{gh_generate_template()}} to
extract a gatingTemplate from either a GatingHierarchy or GatingSet.
}
\examples{
library(CytoExploreRData)

# Activation GatingSet
gs <- GatingSet(Activation)
gs <- cyto_compensate(gs)
gs <- cyto_transform(gs)
gs <- cyto_gatingTemplate_apply(gs)

# Extract gatingTemplate
gt <- cyto_gatingTemplate_apply(gs)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
