% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingTemplate-helpers.R
\name{cyto_gatingTemplate_active}
\alias{cyto_gatingTemplate_active}
\title{Active gatingTemplate}
\usage{
cyto_gatingTemplate_active()
}
\value{
name of the last and therefore active gatingTemplate assigned by
  \code{cyto_gatingTemplate_select}.
}
\description{
Active gatingTemplate
}
\seealso{
\code{\link{cyto_gatingTemplate_select}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
