% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gatingTemplate-helpers.R
\name{cyto_gatingTemplate_apply}
\alias{cyto_gatingTemplate_apply}
\title{Apply gates saved in gatingTemplate to a GatingHierarchy or GatingSet}
\usage{
cyto_gatingTemplate_apply(
  x,
  gatingTemplate = NULL,
  active = TRUE,
  overwrite = NULL,
  nodes = NULL,
  ...
)
}
\arguments{
\item{x}{object of class \code{GatingHierarchy} or \code{GatingSet}.}

\item{gatingTemplate}{name of the gatingTemplate csv file which contains the
gates to be applied to \code{x}. This can alos be an object of class
\code{gatingTemplate} if a \code{GatingSet} object is supplied.}

\item{active}{logical indicating whether the applied gatingTemplate should be
set as the current active gatingTemplate set to TRUE by default.}

\item{overwrite}{logical indicating whether existing gates should be removed
prior to applying the gatingTemplate, only required when running
CytoExploreR non-interactively.}

\item{nodes}{names of the populations to gate, set to NULL by default to
include all nodes in the \code{gtaingTemplate}.}

\item{...}{additional arguments passed to
\code{\link[openCyto:gt_gating]{gt_gating()}} including \code{start} and
\code{stop.at} to only apply particular gates from the gatingTemplate.}
}
\value{
NULL and update the GatingHierarchy/ GatingSet in the global
  environment with the gates in the gatingTemplate.
}
\description{
A convenient wrapper for
\code{\link[openCyto:gatingTemplate-class]{gatingTemplate}} and
\code{\link[openCyto:gt_gating]{gt_gating}} to apply a gatingTemplate csv
file to a GatingSet.
}
\examples{
\dontrun{
library(CytoExploreRData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(fs))
gs <- transform(gs, trans)

# Apply gatingTemplate in working directory to GatingSet
cyto_gatingTemplate_apply(gs, "gatingTemplate.csv")
}

}
