% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gatingTemplate-helpers.R
\name{cyto_gatingTemplate_convert}
\alias{cyto_gatingTemplate_convert}
\title{Convert Old gatingTemplates to New Format}
\usage{
cyto_gatingTemplate_convert(gs, gatingTemplate = NULL)
}
\arguments{
\item{gs}{object of class GatingSet.}

\item{gatingTemplate}{name of the gatingTemplate csv file to convert.}
}
\value{
converted gatingTemplate saved to .csv file.
}
\description{
Convert Old gatingTemplates to New Format
}
\examples{
\dontrun{
library(CytoRSuiteData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(fs))
gs <- transform(gs, trans)

# Convert gatingTemplate CytoRSuite <= 0.9.5 to new format
cyto_gatingTemplate_convert(gs, "gatingTemplate.csv")

# Updated gatingTemplate will work as expected
gt <- gatingTemplate("gatingTemplate.csv")
gt_gating(gt, gs)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
