% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gatingTemplate-helpers.R
\name{cyto_gatingTemplate_create}
\alias{cyto_gatingTemplate_create}
\title{Create an empty gatingTemplate csv file}
\usage{
cyto_gatingTemplate_create(gatingTemplate = NULL)
}
\arguments{
\item{gatingTemplate}{name of the gatingTemplate csv file to create.}
}
\description{
Create an empty gatingTemplate csv file
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
