% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gatingTemplate-helpers.R
\name{cyto_gatingTemplate_edit}
\alias{cyto_gatingTemplate_edit}
\title{Interactively Edit a gatingTemplate}
\usage{
cyto_gatingTemplate_edit(x, gatingTemplate = NULL, ...)
}
\arguments{
\item{x}{object of class \code{GatingSet} which has been gated with the
supplied gatingTemplate.}

\item{gatingTemplate}{name of the gatingTemplate csv file to be edited.}

\item{...}{not in use.}
}
\value{
update gatingTemplate csv file and update the GatingSet accordingly.
}
\description{
\code{cyto_gatingTemplate_edit} provides an interactive interface to editing
the openCyto gatingTemplate. This function is intended to aid in adding
boolean and reference gates to the gatingTemplate. Users should NOT modify
existing entries in the gatingTemplate.
}
\examples{
\dontrun{
library(CytoExploreR)

# gs is a GatingSet object
cyto_gatingTemplate_edit(gs, "gatingTemplate.csv")
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
