% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gatingTemplate-helpers.R
\name{cyto_gatingTemplate_parse}
\alias{cyto_gatingTemplate_parse}
\title{Convert parsed gatingTemplate into a format compatible with CytoExploreR}
\usage{
cyto_gatingTemplate_parse(gatingTemplate, data.table = TRUE, ...)
}
\arguments{
\item{gatingTemplate}{gatingTemplate object or the name of a gatingTemplate
csv file to read in and parse.}

\item{data.table}{logical indicating whether the parsed gatingTemplate should
be returned as a \code{data.table}, set to TRUE by default.}

\item{...}{not in use}
}
\value{
parsed gatingTemplate as either a \code{data.table} or
  \code{data.frame}.
}
\description{
Convert parsed gatingTemplate into a format compatible with CytoExploreR
}
\examples{
\dontrun{
library(CytoExploreRData)
# Parse Activation gatingTemplate
cyto_gatingTemplate_parse(Activation_gatingTemplate)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
