% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gatingTemplate-helpers.R
\name{cyto_gatingTemplate_read}
\alias{cyto_gatingTemplate_read}
\title{Read in a gatingTemplate object or gatingTemplate csv file}
\usage{
cyto_gatingTemplate_read(
  gatingTemplate = NULL,
  data.table = FALSE,
  parse = FALSE,
  ...
)
}
\arguments{
\item{gatingTemplate}{gatingTemplate object or the name of a gatingTemplate
csv file to read in.}

\item{data.table}{logical indicating whether the gatingTemplate should be
read into a \code{data.table}, set to FALSE by default to use
\code{data.frame}.}

\item{parse}{logical indicating whether attempts should be made to parse
\code{alias = "*"} into a valid set of comma separated aliases, set to
FALSE by default. Required to appropriately parse gating entries created by
\code{cyto_gate_clust()}.}

\item{...}{additional arguments passed to
\code{\link[data.table:fread]{fread}}.}
}
\description{
Read in a gatingTemplate object or gatingTemplate csv file
}
\examples{
\dontrun{
library(CytoExploreRData)

# gatingTemplate
gt <- cyto_gatingTemplate_read(Activation_gatingTemplate)

# write gatingTemplate
cyto_gatingTemplate_write(gt, "gatingTemplate.csv")

# gatingTemplate csv file
gt <- cyto_gatingTemplate_read("gatingTemplate.csv")
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
