% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gatingTemplate-helpers.R
\name{cyto_gatingTemplate_update}
\alias{cyto_gatingTemplate_update}
\title{Convert CytoRSuite gatingTemplate to be compatible with  CytoExoloreR}
\usage{
cyto_gatingTemplate_update(gatingTemplate = NULL,
  save_as = "Updated-gatingTemplate.csv")
}
\arguments{
\item{gatingTemplate}{name of the gatingTemplate csv file to convert.}

\item{save_as}{name for the updated gatingTemplate csv file, set to "Updated
gatingTemplate.csv" by default.}
}
\value{
converted gatingTemplate saved to .csv file.
}
\description{
Convert CytoRSuite gatingTemplate to be compatible with  CytoExoloreR
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
