% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gatingTemplate-helpers.R
\name{cyto_gatingTemplate_write}
\alias{cyto_gatingTemplate_write}
\title{Write gatingTemplate to csv file}
\usage{
cyto_gatingTemplate_write(gatingTemplate = NULL, save_as = NULL, ...)
}
\arguments{
\item{gatingTemplate}{data.frame, data.table, gatingTemplate object or the
name of a gatingTemplate csv file.}

\item{save_as}{name of the csv file to write the gatingtemplate to.}

\item{...}{additional arguments passed to
\code{\link[data.table:fwrite]{fwrite}}.}
}
\description{
Write gatingTemplate to csv file
}
\examples{
\dontrun{
library(CytoExploreRData)

# gatingTemplate
gt <- cyto_gatingTemplate_write(Activation_gatingTemplate)

# write gatingTemplate
cyto_gatingTemplate_write(gt, "gatingTemplate.csv")
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
