% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_extensions.R
\name{cyto_ggplot}
\alias{cyto_ggplot}
\title{Convert plots created by cyto_plot() into ggplot2 objects}
\usage{
cyto_ggplot(..., grob = FALSE)
}
\arguments{
\item{...}{arguments passed to \code{\link{cyto_plot}}.}

\item{grob}{logical indicating whether a grob should be returned instead of a
ggplot2 object, set to FALSE by default.}
}
\value{
a plot produced by \code{cyto_plot()} converted to either a grob or
  ggplot2 object.
}
\description{
Much of the analysis and data visualisation downstream of CytoExploreR will
be performed by the ggplot2 package. \code{cyto_ggplot} makes it easy to
convert plots generated by \code{cyto_plot()} into grob or ggplot objects
that can easily be arranged with other ggplot objects using the
\code{cowplot} or \code{patchwork} packages.
}
\examples{
\dontrun{
library(CytoExploreRData)

gs <- GatingSet(Activation)

g <- cyto_ggplot(
gs[1],
parent = "root",
channels = c("FSC-A", "SSC-A")
)
g

}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
