% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_groups}
\alias{cyto_groups}
\title{Retrieve details about experimental groups}
\usage{
cyto_groups(x, group_by = "all", details = FALSE)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{group_by}{names of cyto_details variables to use for grouping For more
control over the group order, specify the factor levels for each variable
in a list (e.g. list(Treatment = c("Stim-A","Stim-C","Stim-B", "Stim-D"))).}

\item{details}{logical indicating whether the split experimental details
should be returned instead of the group names, set to FALSE by default.}
}
\value{
names of experimental groups or a list of experiment details per
  experimental group.
}
\description{
Retrieve details about experimental groups
}
\examples{
library(CytoExploreRData)

# Activation Gatingset
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                          package = "CytoExploreRData"))

# Treatment & OVAConc group names
cyto_groups(gs,
group_by = list(Treatment = c("Stim-A", "Stim-B", "Stim-C", "Stim-D"),
OVAConc = c(0, 5, 50, 500)))

# Treatment & OVAConc group details
cyto_groups(gs,
group_by = list(Treatment = c("Stim-A", "Stim-B", "Stim-C", "Stim-D"),
OVAConc = c(0, 5, 50, 500)),
details = TRUE)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
