% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_import}
\alias{cyto_import}
\title{Import data from other platforms using CytoML}
\usage{
cyto_import(
  path = ".",
  type = "flowJo",
  markers = TRUE,
  details = TRUE,
  barcode = TRUE,
  gatingTemplate = NULL,
  ...
)
}
\arguments{
\item{path}{path to xml or wsp file and associated fcs files.}

\item{type}{indicates the type of data to import, options include
\code{"flowJo"}, \code{"cytobank"} or \code{"diva"}. Set to \code{"flowJo"}
by default.}

\item{markers}{logical indicating whether a call should be made to
\code{cyto_markers_edit} to update the markers associated with channels in
the loaded samples, set to TRUE by default. The name of the csv to which
these details will be supplied can also be passed to this argument.}

\item{details}{logical indicating whether a call should be made to
\code{cyto_details_edit} to update the experimental details associated with
the loaded samples, set to TRUE by default. The name of the csv to which
these details will be supplied can also be passed to this argument.}

\item{barcode}{logical indicating whether events in each sample should be
barcoded for compatibility with \code{cyto_plot()}, set to TRUE by default.}

\item{gatingTemplate}{passed to \code{cyto_gatingTemplate_generate()} to
create a CytoExploreR friendly gatingTemplate based on gates applied to the
loaded samples.}

\item{...}{additional arguments passed to \code{cytobank_to_gatingset()},
\code{flowjo_to_gatingset} or \code{diva_to_gatingset}. Refer to
documentation in CytoML package for more details.}
}
\value{
a GatingSet and write new gatingTemplate to CSV file.
}
\description{
Import data from other platforms using CytoML
}
\examples{
\dontrun{
# flowJo
gs <- cyto_import(
  "flowjo-samples",
  gatingTemplate = "gatingTemplate.csv"
)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
