% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_impute.R
\name{cyto_impute}
\alias{cyto_impute}
\title{Impute values or labels using k nearest neighbours}
\usage{
cyto_impute(
  train = NULL,
  test = NULL,
  labels = NULL,
  parent = "root",
  channels = NULL,
  k = 5,
  method = "dist",
  scale = "range",
  ...
)
}
\arguments{
\item{train}{object of class \code{matrix}, \code{data.frame},
\code{cytoframe}, \code{cytoset}, \code{GatingHierarchy} or
\code{GatingSet} containing the training data.}

\item{test}{object of class \code{matrix}, \code{data.frame},
\code{cytoframe}, \code{cytoset}, \code{GatingHierarchy} or
\code{GatingSet} containing the testing data.}

\item{labels}{a factor of labels or a vector of values for every cell within
the training data. Alternatively, multiple labels can be supplied in the
form of a matrix or a named list.}

\item{parent}{name of the population to extract from \code{train} and
\code{test} if \code{GatingHierarchy} or \code{GatingSet} objects are
supplied, set to the \code{"root"} node by default.}

\item{channels}{names of the markers or channels to use to constructed the
kNN graph, must be supplied manually as no default has been set.}

\item{k}{number of nearest neighbours to compute when constructing the kNN
graph, set to 5 by default.}

\item{method}{determines whether to use \code{"dist"} or \code{"vote"} when
performing kNN classification, set to \code{"dist"} by default.}

\item{scale}{method to use to scale the channels of training and testing data
prior to constructing the kNN graph, set to \code{"range"} by default.}

\item{...}{not in use.}
}
\value{
a mtrix containing the imputed labels or values.
}
\description{
Impute values or labels using k nearest neighbours
}
\examples{
\dontrun{

# Traning and testing data
ind <- 1:nrow(iris)
train_ind <- sample(ind, ceiling(0.8 * nrow(iris)))
test_ind <- ind[!ind \%in\% train_ind]

# labels
cyto_impute(
  iris[train_ind, ],
  iris[test_ind, ],
  factor(iris[train_ind, 5]),
  channels = colnames(iris)[1:4]
)

# values
cyto_impute(
  iris[train_ind, ],
  iris[test_ind, ],
  iris[train_ind, 4],
  channels = colnames(iris)[1:3]
)

}

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
