% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_infinity_screen.R
\name{cyto_infinity_screen}
\alias{cyto_infinity_screen}
\title{Use infinityFlow to integrate multiplexed parallel cytometry datasets}
\usage{
cyto_infinity_screen(
  path = NULL,
  backbone_selection_file = NULL,
  annotation = NA,
  isotype = NA,
  save_as = NULL,
  ...
)
}
\arguments{
\item{path}{points to the location where the FCS files (one per well) are
stored, replaces the \code{path_to_fcs} argument in \code{infinity_flow()}.}

\item{backbone_selection_file}{name of a CSV file containing columns
\code{"name"}, \code{"desc"} and \code{"type"}, where type indicates
whether the named parameter is a part of the \code{"backbone"}, or instead
an \code{"exploratory"} parameter. Parameters can be optionally excluded by
setting \code{type} to \code{"discard"}. Users need not create this file by
hand as CytoExploreR will allow the user to interactively edit this
information when required.}

\item{annotation}{Named character vector. Elements should be the targets of
the exploratory antibodies, names should be the name of the FCS file where
that exploratory antibody was measured. Users need not supply this
information manually as CytoExploreR will allow users to interactively
enter this information.}

\item{isotype}{Named character vector. Elements should be the isotype used in
each of the wells (e.g. IgG2) and the corresponding isotype should be
present in \emph{annotation} (e.g. Isotype_IgG2, with this capitalization
exactly). Autofluorescence measurements should be listed here as "Blank".
Users need not supply this information manually as CytoExploreR will allow
users to interactively enter this information.}

\item{save_as}{name of the directory to which the output FCS files should be
written, replaces the \code{path_to_output} argument in
\code{infinity_flow()}.}

\item{...}{additional arguments passed to \code{infinity_flow()}, refer to
the \code{infinityFlow} package documentation for more details.}
}
\value{
raw and background-corrected imputed expression data for every
  Infinity antibody, new FCS files written to \code{save_as} along with plots
  of UMAP projections of the backbone antibody panel.
}
\description{
\code{cyto_infinity_screen()} is a CytoExploreR wrapper around the
\code{infinity_flow()} function in the \code{infinityFlow} package that uses
machine learning to integrate multiplexed parallel cytometry datasets. Users
should note that \code{cyto_infinity_screen()} takes on the path to the FCS
files rather than a \code{cytoset} or \code{GatingSet}. Furthermore,
\code{cyto_infinity_screen()} does not return a \code{cytoset} or
\code{GatingSet}, instead the new FCS files written to \code{save_as} should
be read in afterwards using \code{cyto_setup} as per usual.
}
\references{
Becht E, Headley M, Newell E, Gottardo R (2020). Infinity Flow:
  comprehensive single-cell protein profiling via massively parallel flow
  cytometry and machine learning. J Immunol 204:159.2
}
\seealso{
\code{\link{cyto_setup}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
