% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_interactive}
\alias{cyto_interactive}
\title{Check for interactive CytoExploreR session}
\usage{
cyto_interactive(interactive = NULL)
}
\arguments{
\item{interactive}{logical requesting for the current CytoExploreR session to 
be run interactively or not.}
}
\value{
logical indicating whether the current CytoExploreR session is 
interactive or not.
}
\description{
Check for interactive CytoExploreR session
}
\examples{
\dontrun{
# Run CytoExploreR non-interactively
cyto_interactive(FALSE)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
