% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_keyword}
\alias{cyto_keyword}
\title{Extract a keyword from a cytoframe, cytoset or GatingSet}
\usage{
cyto_keyword(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[FlowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[FlowWorkspace:GatingSet-class]{GatingSet}}.}

\item{...}{additional arguments passed to \code{flowCore::keyword()} or
\code{flowWorkspace::keyword()} including \code{keyword} which indicates
the name of a specific keyword to extract.}
}
\value{
a list of extract keywords.
}
\description{
Extract a keyword from a cytoframe, cytoset or GatingSet
}
\examples{
library(CytoExploreRData)

cyto_keyword(
  Activation[[1]],
  "SPILL"
)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
