% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_list}
\alias{cyto_list}
\title{Convert cytoset to a list of cytoframes or cytosets}
\usage{
cyto_list(x, format = "cytoframe")
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}}.}

\item{format}{either \code{"cytoframe"} or \code{"cytoset"} to indicate the
class of each list element, set to \code{"cytoframe"} by default.}
}
\value{
a list of individual cytoframe or cytoset objects.
}
\description{
Convert cytoset to a list of cytoframes or cytosets
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
