% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_load}
\alias{cyto_load}
\title{Load .fcs files into ncdfFlowSet}
\usage{
cyto_load(path = ".", ...)
}
\arguments{
\item{path}{points to the location of the .fcs files to read in.}

\item{...}{additional arguments passed to read.ncdfFlowSet.}
}
\value{
\code{ncdfFlowSet}
}
\description{
\code{cyto_load} is a convenient wrapper around
\code{\link[base:list.files]{list.files}} and
\code{\link[ncdfFlow:read.ncdfFlowSet]{read.ncdfFlowSet}} which makes it easy
to load .fcs files into a ncdfFlowSet.
}
\examples{
library(CytoRSuiteData)

# Get path to Activation .fcs files in CytoRSuiteData
datadir <- system.file("extdata", package = "CytoRSuiteData")
path <- paste0(datadir,"/Activation")

# Load in .fcs files into ncdfFlowSet
fs <- cyto_load(path)

# fs is a ncdfFlowSet
class(fs)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
