% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_map.R
\name{cyto_map}
\alias{cyto_map}
\alias{cyto_map.flowFrame}
\title{Create dimension-reduced maps of cytometry data}
\usage{
cyto_map(x, ...)

\method{cyto_map}{flowFrame}(
  x,
  channels,
  sample = 50000,
  method = "tSNE",
  save = FALSE,
  save_as = "cyto_map",
  plot = TRUE,
  seed,
  ...
)
}
\arguments{
\item{x}{object of class \code{flowFrame} or \code{flowSet}.}

\item{...}{additional arguments passed to the called dimension reduction
function. Links to the documentation for these functions can be found
below.}

\item{channels}{vector of channels names indicating the channels that should
be used by the dimension reduction algorithm to compute the 2-dimensional
map, set to all channels by default. Restricting the number of channels can
greatly improved processing speed but may result in poorer resolution.}

\item{sample}{number of events to map, set to 50 000 events by default.}

\item{method}{direction reduction method to use to generate the map,
supported options include "PCA", "tSNE" and "UMAP".}

\item{save}{logical indicating whether the mapped \code{flowFrame} or
\code{flowSet} should be saved as .fcs file(s) in a folder in the  current
working directory.}

\item{save_as}{name of the folder to save the .fcs files to when save is
TRUE, set to "cyto_map" by default.}

\item{plot}{logical indicating whether the constructed map should be plotted
using \code{cyto_plot}.}

\item{seed}{integer to set seed prior to mapping to ensure consistent results
between runs.}
}
\description{
Create dimension-reduced maps of cytometry data
}
\seealso{
\code{\link[stats:prcomp]{PCA}}

\code{\link[Rtsne:Rtsne]{tSNE}}

\code{\link[umap:umap]{UMAP}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
