% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_map_compare.R
\name{cyto_map_compare}
\alias{cyto_map_compare}
\title{Compare cross-entropies computed for tSNE and UMAP embeddings}
\usage{
cyto_map_compare(
  x,
  select = NULL,
  parent = "root",
  merge_by = "name",
  channels = NULL,
  base_test = "ks",
  plot = TRUE,
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{names(s) for populations for which the comparisons should be
performed when a GatingHierarchy or GatingSet is supplied, set to the
\code{"root"} node by default.}

\item{merge_by}{merging criteria passed to \code{cyto_merge_by()} to split
the data into groups for comparison, set to \code{"name"} by default to
compare every sample to one another.}

\item{channels}{name of the channel containing the cross entropy values
computed by \code{cyto_map()}, defaults to a channel containing
\code{"entropy"} in its name.}

\item{base_test}{test type}

\item{plot}{logical indicating whether the dendrogram should be plotted
showing the relationships between the groups specified by \code{merge_by},
set to TRUE by default.}

\item{...}{not in use.}

\item{sample}{selection criteria passed to \code{cyto_select()} to restrict
the samples for comparison, set to NULL by default to include all samples
in the comparisons.}
}
\value{
list with elements \code{p} a matrix containing pairwise adjusted p
  values, \code{dist} a matrix containing the pairwise distance metrics and
  \code{hclust} an object of class \code{hclust} containing the results from
  heirarchical clustering.
}
\description{
\code{cyto_map_compare()} implements a refined version of the cross-entropy
test described by Roca et al. 2021. Before users attempt to use
\code{cyto_map_compare()}, cross-entropies should be computed per cell within
\code{cyto_map()} by setting \code{cross_entropy = TRUE}. Subsequently,
comparisons of cross-entropies can be made within \code{cyto_map_compare()}
across multiple populations and/or groups as specified by the \code{parent}
and \code{merge_by} arguments.
}
\references{
Roca C. et al. (2021) A Cross Entropy test allows quantitative
  statistical comparison of t-SNE and UMAP representations. arXiv:2112.04172
}
\seealso{
\code{\link{cyto_select}}

\code{\link{cyto_merge_by}}

\code{\link{cyto_groups}}

\code{\link{cyto_map}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
