% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channels-helpers.R
\name{cyto_markers<-}
\alias{cyto_markers<-}
\title{Replace marker names}
\usage{
cyto_markers(x) <- value
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}},
\code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSey}}.}

\item{value}{named vector of channels with their associated markers.}
}
\description{
Replace marker names
}
\examples{
library(CytoExploreRData)

# Activation GatingSet
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                         package = "CytoExploreRData"))

# GatingSet
cyto_channels(gs)

# Set FSC-A marker to FSC
cyto_markers(gs) <- c("FSC" = "FSC-A")

}
\seealso{
\code{\link{cyto_markers}}

\code{\link{cyto_markers_edit}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
