% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channels-helpers.R
\name{cyto_markers}
\alias{cyto_markers}
\alias{cyto_markers.GatingSet}
\alias{cyto_markers.GatingHierarchy}
\alias{cyto_markers.flowSet}
\alias{cyto_markers.flowFrame}
\alias{cyto_markers.list}
\title{Extract marker names}
\usage{
\method{cyto_markers}{GatingSet}(x, select = NULL, exclude = NULL, ...)

\method{cyto_markers}{GatingHierarchy}(x, select = NULL, exclude = NULL, ...)

\method{cyto_markers}{flowSet}(x, select = NULL, exclude = NULL, ...)

\method{cyto_markers}{flowFrame}(x, select = NULL, exclude = NULL, ...)

\method{cyto_markers}{list}(x, select = NULL, exclude = NULL, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} or a list of these
objects.}

\item{select}{vector of channels or markers for which the channel/marker
combinations should be returned.}

\item{exclude}{vector of channels or markers for which the channel/marker
combinations should not be returned.}

\item{...}{additional arguments passed to \code{\link[base:grep]{grepl}} for
character matching. For exact character string matching to override the
default which ignores character case, set \code{fixed} to TRUE.}
}
\value{
vector of marker names or NULL if no markers have been assigned.
}
\description{
Extract marker names
}
\examples{

# Load in CytoExploreRData to access data
library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Activation GatingSet
gs <- GatingSet(fs)

# flowFrame
cyto_markers(fs[[1]])

# flowSet
cyto_markers(fs)

# GatingHierarchy
cyto_markers(gs[[1]])

# GatingSet
cyto_markers(gs)

}
\seealso{
\code{\link{cyto_channels}}

\code{\link{cyto_fluor_channels}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
