% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channels-helpers.R
\name{cyto_markers_extract}
\alias{cyto_markers_extract}
\alias{cyto_markers_extract.flowFrame}
\alias{cyto_markers_extract.flowSet}
\alias{cyto_markers_extract.GatingHierarchy}
\alias{cyto_markers_extract.GatingSet}
\title{Extract marker names for certain channels}
\usage{
cyto_markers_extract(x, markers, plot = FALSE)

\method{cyto_markers_extract}{flowFrame}(x, channels, plot = FALSE)

\method{cyto_markers_extract}{flowSet}(x, channels, plot = FALSE)

\method{cyto_markers_extract}{GatingHierarchy}(x, channels, plot = FALSE)

\method{cyto_markers_extract}{GatingSet}(x, channels, plot = FALSE)
}
\arguments{
\item{x}{an object of class
\code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{plot}{logical indicating whether the channels will be used to construct
a plot, set to FALSE by default. If set to TRUE an additional check will be
performed to ensure that only 1 or 2 \code{channels} are supplied.}

\item{channels}{vector of channel and/or marker names (e.g. c("Alexa Fluor
700-A","CD8")).}
}
\value{
A vector of marker names.
}
\description{
\code{cyto_markers_extract} will check whether the supplied channels or
marker names are valid for the
\code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} and return a vector of
marker names. The name of the channel will be returned if there is no
associated marker found.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
