% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_match}
\alias{cyto_match}
\title{Get index of samples matching experimental criteria}
\usage{
cyto_match(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{...}{named list containing experimental variables to be used to select
samples or named arguments containing the levels of the variables to
select. See below examples for use cases. Neagtive exclusion indices can be
returned by setting \code{exclude = TRUE}.}
}
\value{
vector of indices that representat the location of the selected
  sample(s) within the supplied \code{cytoset} or \code{GatingSet}.
}
\description{
Get index of samples matching experimental criteria
}
\examples{
library(CytoExploreRData)

# Activation Gatingset
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                          package = "CytoExploreRData"))

# Look at experiment details
cyto_details(gs)

# Indices of Stim-C samples with 0 and 500 nM OVA concentrations
cyto_match(gs,
  Treatment = "Stim-C",
  OVAConc = c(0, 500)
)

# Indices of Stim-A and Stim-C treatment groups
cyto_match(
  gs,
  list("Treatment" = c("Stim-A", "Stim-C"))
)

# Exclusion indices of Stim-D treatment group
cyto_match(gs,
  Treatment = "Stim-D",
  exclude = TRUE
)

}
\seealso{
\code{\link{cyto_select}}

\code{\link{cyto_filter}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
