% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_names<-}
\alias{cyto_names<-}
\title{Replacement method for cyto_names}
\usage{
cyto_names(x) <- value
}
\arguments{
\item{x}{object of class flowFrame, flowSet, GatingHierarchy or GatingSet.}

\item{value}{vector of replacement names.}
}
\description{
Replacement method for cyto_names
}
\examples{

library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Sample names
cyto_names(fs)

# Change first sample name 
cyto_names(fs)[1] <- "first_sample"

# Activation GatingSet
gs <- GatingSet(fs)

# Change last sample name
cyto_names(gs)[length(gs)] <- "last_sample"

# Updated sample names
cyto_names(gs)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
