% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_names}
\alias{cyto_names}
\title{Extract sample names}
\usage{
cyto_names(x, select = NULL, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{select}{sample selection criteria passed to \code{cyto_select()} to
restrict the set of samples prior to extracting the sample names.}

\item{...}{not in use.}
}
\value{
sample names associated with the supplied object.
}
\description{
Simply a convenient and autocomplete-friendly wrapper around
\code{\link[flowCore:identifier-methods]{identifier}}
\code{\link[flowWorkspace:sampleNames]{sampleNames}} to extract the sample
names from flowFrame, flowSet GatingHierarchy or GatingSet. Anonymous
\code{\link[flowCore:flowFrame-class]{flowFrame}} identifiers will be
converted to \code{"Combined Events"}.
}
\examples{
library(CytoExploreRData)

# Activation GatingSet
gs <- cyto_load(
 system.file(
   "extdata/Activation-GatingSet",
   package = "CytoExploreRData"
   )
 )

# Activation cytoset
cs <- cyto_data_extract(gs, "root")[["root"]]

# GatingSet
cyto_names(gs)

# cytoset
cyto_names(cs)

}
\seealso{
\code{\link{cyto_select}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
