% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_names_parse}
\alias{cyto_names_parse}
\title{Parse file names into experiment details}
\usage{
cyto_names_parse(x, vars = NULL, split = "_", exclude = NULL, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{vars}{vector containing the names of the variables to be added to
\code{cyto_details}, set to var_1, var_2 ... by default.}

\item{split}{delimiter to split the file name into fragments, set to "_" by
default.}

\item{exclude}{vector of indices indicating which text fragments should not
be included in the experiment details, set to NULL by default.}

\item{...}{additional arguments passed to
\code{\link[base:strsplit]{strsplit}}.}
}
\value{
cytoset or GatingSet with \code{cyto_details} updated with new
  variables extracted from the file names.
}
\description{
Reward users that use consistent names for files by parsing file names into
new variables and adding them to \code{cyto_details}. Variable construction
may not be perfect if certain files are named differently, but it will give a
starting point for editing these values using \code{cyto_details_edit}
without having to manually enter a lot of these details.
}
\examples{
library(CytoExploreRData)

# Parse file names to variables
fs <- cyto_names_parse(Activation,
vars = c("sample_type", "sample_id"),
split = "_")

# Updated experiment details
cyto_details(fs)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
