% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_nodes}
\alias{cyto_nodes}
\title{Extract Names of Gated Populations in GatingHierarchy or GatingSet}
\usage{
cyto_nodes(x, ...)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{...}{additional arguments passed to
\code{\link[flowWorkspace:gs_get_pop_paths]{gs_get_pop_paths}}.}
}
\value{
character vector of gated node/population names.
}
\description{
\code{cyto_nodes} is simply an autocomplete-friendly wrapper for
\code{\link[flowWorkspace:gs_get_pop_paths]{gs_get_pop_paths}}.
}
