% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_nodes_ancestor}
\alias{cyto_nodes_ancestor}
\title{Get name of most recent common ancestor shared by nodes}
\usage{
cyto_nodes_ancestor(x, nodes = NULL, hidden = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{GatingHierarchy},\code{GatingSet},
\code{gatingTemplate} or the name of a gatingTemplate CSV file.}

\item{nodes}{vector of nodes for which the most recent common ancestor should
be returned.}

\item{hidden}{logical indicating whether hidden nodes should be included in
the search for the most recent ancestral node, set to FALSE by default.}

\item{...}{additional arguments passed to \code{\link{cyto_nodes_convert}} to
control the format of the returned ancestral node.}
}
\description{
Get name of most recent common ancestor shared by nodes
}
\examples{
library(CytoExploreRData)

# Activation GatingSet
gs <- GatingSet(Activation)

# Compensation
gs <- cyto_compensate(gs)

# Transformations
gs <- cyto_transform(gs)

# Gating
gs <- cyto_gatingTemplate_apply(gs, Activation_gatingTemplate)

# Ancestral node of CD4 T Cells & CD8 T Cells
cyto_nodes_ancestor(gs, nodes = c("CD4 T Cells", "CD8 T Cells"))

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
