% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_nodes_check}
\alias{cyto_nodes_check}
\title{Check for ambiguous nodes}
\usage{
cyto_nodes_check(x, nodes = NULL, anchor = NULL)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{nodes}{vector of nodes to check.}

\item{anchor}{parental node to anchor to if node is ambiguous.}
}
\description{
\code{cyto_nodes_check} checks if the supplied nodes are ambiguous within the
\code{GatingHierarchy} or \code{GatingSet} and attempts to anchor the node to
a parental node to obtain a unique node path.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
