% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_nodes_convert}
\alias{cyto_nodes_convert}
\title{Convert to unique node paths}
\usage{
cyto_nodes_convert(x, nodes = NULL, anchor = NULL, path = "auto")
}
\arguments{
\item{x}{object of class \code{GatingHierarchy}, \code{GatingSet},
\code{gatingTemplate} or the name of a gatingTemplate CSV file.}

\item{nodes}{vectors of nodes to check and convert.}

\item{anchor}{unique path to parental node to use as an anchor for ambiguous
nodes.}

\item{path}{specifies whether the returned nodes should be in the "full" or
"auto" format, set to "auto" by default}
}
\value{
vector of unique paths for each of the supplied nodes.
}
\description{
\code{cyto_nodes_convert} checks whether the supplied nodes exist in the
supplied \code{GatingHierarchy} or \code{GatingSet} and returns unique node
paths for each of the supplied nodes. In the case of ambiguous nodes,
\code{cyto_nodes_convert} will attempt to anchor the node path to a known
unique parental node.
}
\author{
@author Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
