% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_nodes_fix}
\alias{cyto_nodes_fix}
\title{Rename nodes containing illegal characters in a GatingHierarchy or GatingSet}
\usage{
cyto_nodes_fix(x)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}
}
\value{
a GatingHierarchy or GatingSet with node names containing only legal
  characters.
}
\description{
\code{cyto_nodes_fix()} checks all nodes in a GatingHierarchy or GatingSet to
remove duplicate nodes or rename nodes containing any illegal characters
(including /:|,&) and update any associated logic in boolean gates.
\code{cyto_nodes_fix()} is particularly useful when importing data analysed
using other software platforms to ensure that the data conforms to the
standards enforced within the \code{cytoverse}.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
