% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_nodes_kin}
\alias{cyto_nodes_kin}
\title{Get paths to nodes}
\usage{
cyto_nodes_kin(
  x,
  nodes = NULL,
  type = "children",
  path = "auto",
  hidden = TRUE,
  terminal = FALSE,
  ...
)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}},
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}},
\code{\link[openCyto:gatinGtemplate-class]{gatingTemplate}} or the name of
a gatingTemplate CSV file.}

\item{nodes}{vector of nodes for which the relative nodes should be returned.}

\item{type}{options include \code{"children"}, \code{"parent"},
\code{"grandparent"} or \code{"descendants"} to indicate the type of
relationship required for each of the supplied nodes.}

\item{path}{specifies whether the returned nodes should be in the "full" or
"auto" format, set to "auto" by default.}

\item{hidden}{logical indicating whether hidden nodes should be included in
the search for relatives, set to TRUE by default.}

\item{terminal}{logical to indicate whether only the relative with terminal
nodes should be returned, set to FALSE by default.}

\item{...}{not in use.}
}
\value{
the paths of the nodes of relation \code{type} to each of the
  supplied nodes.
}
\description{
Get paths to nodes
}
\examples{
library(CytoExploreRData)

# Activation GatingSet
gs <- GatingSet(Activation)
gs <- cyto_compensate(gs)
gs <- cyto_transform(gs)
gs <- cyto_gatingTemplate_apply(gs, Activation_gatingTemplate)

# T Cells parent
cyto_nodes_kin(
  "T Cells",
  "parent"
)

# Descendants of Live Cells
cyto_nodes_kin(
  "Live Cells",
  "descendants"
)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
