% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_option}
\alias{cyto_option}
\title{Retrieve or set a global option for CytoExploreR}
\usage{
cyto_option(option, value)
}
\arguments{
\item{option}{name of the option to retrieve or set.}

\item{value}{a value to replace the current option.}
}
\value{
return current value of option if no \code{value} is supplied,
  otherwise update the existing value of the option.
}
\description{
Retrieve or set a global option for CytoExploreR
}
\examples{
# Extract current theme settings
cyto_option("cyto_plot_theme")

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
