% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_panel_design.R
\name{cyto_panel_design}
\alias{cyto_panel_design}
\title{Compute panel design statistics from single colour controls}
\usage{
cyto_panel_design(
  x,
  spectra = NULL,
  type = "CytoDecode",
  channels = NULL,
  details = NULL,
  ...
)
}
\arguments{
\item{x}{object of class \code{GatingSet} or a list of \code{GatingSet}
objects containing a set(s) of pre-gated single colour controls.}

\item{type}{algorithm to use to compute the spectra, if the supplied
GatingSets haven't been through \code{cyto_spillover_compute()} or
\code{cyto_unmix_compute()} and the results supplied through either
\code{spillover} or \code{unmix}.}

\item{channels}{names of channels to computed the statistics over, set to the
column names of \code{spectra} by default.}

\item{...}{additional arguments passed to either
\code{cyto_spillover_compute} or \code{cyto_unmix_compute()}.}

\item{specra}{optional pre-computed spillover or unmixing matrices either as
a vector of CSV file names or a list of matrix objects.}
}
\value{
a list containing spectra, similarity, complexity, brightness,
  interference, spread, total_spread, hotspot and purity and panel complexity
  per panel.
}
\description{
Compute panel design statistics from single colour controls
}
\examples{
\dontrun{
library(CytoExploreRData)

# import single color controls
gs <- GatingSet(Compensation)
gs <- cyto_gatingTemplate_apply(gs)
gs <- cyto_transform(gs)

# compute spillover matrix
spill <- cyto_spillover_compute(gs)

# compute panel design statistics
cyto_panel_design(gs, spillover = spill)
}

}
\seealso{
\code{\link{cyto_spillover_compute}}

\code{\link{cyto_unmix_compute}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
