% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_panels_merge.R
\name{cyto_panels_merge}
\alias{cyto_panels_merge}
\title{Combine multiple overlapping antibody panels}
\usage{
cyto_panels_merge(
  x,
  parent = "root",
  select = NULL,
  channels = NULL,
  merge_by = NULL,
  events = NULL,
  save_as = NULL,
  file = NULL,
  pool = TRUE,
  k = 10,
  ...
)
}
\arguments{
\item{x}{an object of class \code{\link[flowWorkspace:cytoset]{cytoset}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the parent population to extract from each sample prior
to combining the antibody panels, set to the "root" node to use all events
in each sample by default.}

\item{select}{vector or list of sample selection criteria passed to
\code{cyto_select()} to extract the samples that should have their panels
merged.}

\item{channels}{names of the channels or markers over which the panel
integration is to be performed, set to all channels by default.}

\item{merge_by}{a vector of experiment variables to dictate how samples
should be split into groups prior to merging their panels.}

\item{events}{indicates the proportion \code{events < 1} or the number of
events \code{events > 1} to keep in the merged samples per group, set to
the average number of events within each group by default. \code{events}
cannot exceed the maximum number of events within each group. Setting
\code{events = 1} will keep all events within each merged group.}

\item{save_as}{name of a CSV to which the edited panel details should be
saved, defaults to \code{Panel-Details.csv} prefixed with the date unless
\code{file} is specified or a file is found when searching the current
directory for panel details. Setting this argument to NA will prevent
writing of edited panel details to file. Custom file names passed to
\code{save_as} should end in \code{"Panel-Details.csv"} so that
CytoExploreR can automatically find and read in these details when
required.}

\item{file}{name of csv file containing the panel information associated with
each sample in the supplied data. This file must contain a column called
\code{"name"} which contains the names of the samples.}

\item{pool}{logical to indicate whether the integrated data should be
combined into a single cytoset if the groups share the same channels, set
to TRUE by default.}

\item{k}{number of neighbourhoods to consider when performing kNN regression,
set to 10 by default.}

\item{...}{additional arguments passed to \code{FNN::knn.reg()} to offer
control over the algorithm to use when constructing the kNN trees.}
}
\value{
a list of cytosets containing the merged samples for each group in
  \code{merge_by}.
}
\description{
Combine multiple overlapping antibody panels
}
\examples{
\dontrun{
# read in data - add patient_id variable to details 
gs <- cyto_setup(
  "FCS_Files"
)

# merge panels by patient
cs_list <- cyto_panels_merge(
  gs,
  parent = "root",
  channels = c(
  "FSC-A", 
  "SSC-A",
  "PE-A",
  "PE-Cy7-A",
  "FITC-A",
  "APC-A",
  "APC-Cy7-A"
  ),
  merge_by = "patient_id",
  events = 50000,
  pool = TRUE
)

# add merged fdata to new GatingSet
gs <- GatingSet(cs_list[[1]])
}

}
\seealso{
\code{\link{cyto_select}}

\code{\link{cyto_merge_by}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
