% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_parallel}
\alias{cyto_parallel}
\title{Configure parallel processing for CytoExploreR}
\usage{
cyto_parallel(plan, workers = NULL, ...)
}
\arguments{
\item{plan}{future plan to use for parallel processing, options include
\code{"sequential"}, \code{"multisession"}, \code{"multicore"} or
\code{"cluster"}. Only \code{"sequential"} and \code{"multisession"} are
supported on Windows.}

\item{workers}{number of workers to assign to jobs, set to all but one of the
available cores.}
}
\value{
the name of the current future plan.
}
\description{
Configure parallel processing for CytoExploreR
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
