% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-methods.R
\name{cyto_plot.GatingSet}
\alias{cyto_plot.GatingSet}
\title{cyto_plot - GatingSet Method}
\usage{
\method{cyto_plot}{GatingSet}(x, parent, alias = NA, channels,
  axes_trans = NA, group_by = NA, overlay = NA, gate = NA,
  limits = "data", display = 25000, layout, popup = FALSE,
  xlim = NA, ylim = NA, xlab, ylab, title, negate,
  density_modal = TRUE, density_smooth = 0.6, density_stack = 0,
  density_layers = NA, density_cols = NA, density_fill = NA,
  density_fill_alpha = 1, density_line_type = 1,
  density_line_width = 1, density_line_col = "black",
  point_shape = ".", point_size = 2, point_col_scale = NA,
  point_cols = NA, point_col = NA, point_col_alpha = 1,
  contour_lines = 0, contour_line_type = 1, contour_line_width = 1,
  contour_line_col = "black", contour_line_alpha = 1,
  axes_text = c(TRUE, TRUE), axes_text_font = 1, axes_text_size = 1,
  axes_text_col = "black", axes_label_text_font = 1,
  axes_label_text_size = 1.1, axes_label_text_col = "black",
  title_text_font = 2, title_text_size = 1.1,
  title_text_col = "black", legend = FALSE, legend_text = NA,
  legend_text_font = 1, legend_text_size = 1,
  legend_text_col = "black", legend_line_type = NA,
  legend_line_width = NA, legend_line_col = NA, legend_box_fill = NA,
  legend_point_col = NA, gate_line_type = 1, gate_line_width = 2.5,
  gate_line_col = "red", gate_fill = "white", gate_fill_alpha = 0,
  label, label_text = NA, label_stat = "", label_position = "auto",
  label_text_x = NA, label_text_y = NA, label_text_font = 2,
  label_text_size = 0.8, label_text_col = "black",
  label_fill = "white", label_fill_alpha = 0.6, border_line_type = 1,
  border_line_width = 1, border_line_col = "black",
  border_fill = "white", border_fill_alpha = 1, ...)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the population containing the events to plot.}

\item{alias}{name of the gated population for which the gate should be drawn
on the plot. Setting \code{alias} to "" will automatically plot any gates
contructed in the supplied channels.}

\item{channels}{name of the channel(s) or marker(s) to be used to construct
the plot. The length of channels determines the type of plot to be
constructed, either a 1-D density distribution for a single channel or a
2-D scatterplot with blue-red colour scale for two channels.}

\item{axes_trans}{object of class
\code{\link[flowCore:transformList-class]{transformList}} or
\code{\link[flowWorkspace]{transformerList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} which was used to
transform the fluorescent channels of the supplied flowFrame. This
transformation object will be used internally to ensure that the axes
labels of the plot are appropriately transformed. The transformation object
will NOT be applied to the flowFrame internally and should be applied to
the flowFrame prior to plotting.}

\item{group_by}{a vector of pData variables to sort and merge samples into
groups prior to plotting, set to NULL by default to prevent merging. To
merge all samples set this argument to \code{TRUE} or \code{"all"}.}

\item{overlay}{name(s) of the populations to overlay or a \code{flowFrame},
\code{flowSet}, \code{list of flowFrames}, \code{list of flowSets} or
\code{list of flowFrame lists} containing populations to be overlaid onto
the plot(s).}

\item{gate}{gate object(s) to be added to plot.  For \code{cyto_1d_plot} only
gate objects of class
\code{\link[flowCore:rectangleGate-class]{rectangleGate}} in either 1 or 2
dimensions are supported. Multiple gates can be supplied either as a
\code{list} or \code{\link[flowCore:filters-class]{filters}} object.}

\item{limits}{indicates whether the axes limits should be based on the
\code{"data"} or \code{"machine"}, set to "machine" by default to show
complete axes ranges. This argument will only alter the upper axis limits,
to modify the lower limits use \code{xlim} and \code{ylim}.}

\item{display}{numeric [0,1] to control the percentage of events to be
plotted. Specifying a value for \code{display} can substantial improve
plotting speed for less powerful machines.}

\item{layout}{a vector of the length 2 indicating the dimensions of the grid
for plotting \code{c(#rows, #columns)}.}

\item{popup}{logical indicating whether the plot should be constructed in a
pop-up window, set to FALSE by default. \code{popup} will open OS-specific
graphic device prior to plotting. Mac users will need to install
\href{https://www.xquartz.org/}{XQuartz} for this functionality.}

\item{xlim}{lower and upper limits of x axis (e.g. c(0,5)).}

\item{ylim}{lower and upper limits of y axis (e.g. c(0,5)).}

\item{xlab}{x axis label.}

\item{ylab}{y axis label.}

\item{title}{title to use for the plot, set to the name of the sample by
default. Title can be removed by setting this argument to \code{NA}.}

\item{negate}{logical indicating whether a label should be included for the
negated population when gate objects are supplied, set to FALSE by default.}

\item{density_modal}{logical indicating whether density should be normalised
to mode and presented as a percentage for 1-D plots. Set to \code{TRUE} by
default.}

\item{density_smooth}{smoothing parameter passed to
\code{\link[stats:density]{density}} to adjust kernel density for 1-D
plots.}

\item{density_stack}{numeric [0,1] indicating the degree of offset for 1-D
density distributions with overlay, set to 0.5 by default.}

\item{density_layers}{numeric indicating the number of samples to stack in
each plot, set to all samples by default.}

\item{density_cols}{vector colours to draw from when selecting density fill
colours if none are supplied to density_fill.}

\item{density_fill}{fill colour(s) for 1-D density distributions.}

\item{density_fill_alpha}{numeric [0,1] used to control 1-D density fill
colour transparency, set to 1 by default for solid colours.}

\item{density_line_type}{line type(s) to use for 1-D density lines, set to 1
by default to use solid lines. See \code{\link[graphics:par]{lty}} for
alternatives.}

\item{density_line_width}{numeric to control line width(s) for 1-D density
lines, set to 1 by default.}

\item{density_line_col}{colour(s) for 1-D density lines, set to
\code{"black"} by default.}

\item{point_shape}{shape(s) to use for points in 2-D scatterplots, set to
\code{"."} by default to maximise plotting speed.  See
\code{\link[graphics:par]{pch}} for alternatives.}

\item{point_size}{numeric to control the size of points in 2-D scatter plots
set to 2 by default.}

\item{point_col_scale}{vector of ordered colours to use for the density
colour gradient of points.}

\item{point_cols}{vector colours to draw from when selecting colours for
points if none are supplied to point_col.}

\item{point_col}{colour(s) to use for points in 2-D scatter plots, set to NA
by default to use a blue-red density colour scale.}

\item{point_col_alpha}{numeric [0,1] to control point colour transparency in
2-D scatter plots, set to 1 by default to use solid colours.}

\item{contour_lines}{numeric indicating the number of levels to use for
contour lines in 2-D scatter plots, set to 0 by default to turn off contour
lines.}

\item{contour_line_type}{integer [0,6] to control the line type of contour
lines in 2-D scatter plots, set to \code{1} to draw solid lines by default.
See \code{\link[graphics:par]{lty}} for alternatives.}

\item{contour_line_width}{numeric to control line width(s) for contour lines
in 2-D scatter plots, set to 2 by default.}

\item{contour_line_col}{colour(s) to use for contour lines in 2-D scatter
plots, set to \code{"black"} by default.}

\item{axes_text}{logical vector of length 2 indicating whether axis text
should be included for the x and y axes respectively, set to
\code{c(TRUE,TRUE)} by default to display axes text on both axes.}

\item{axes_text_font}{numeric to control the font of axes text, set to 1 for
plain font by default. See \code{\link[graphics:par]{font}} for
alternatives.}

\item{axes_text_size}{numeric to control the size of axes text, set to 1 by
default.}

\item{axes_text_col}{colour to use for axes text, set to \code{"black"} by
default.}

\item{axes_label_text_font}{numeric to control the font axes labels, set to 1
for plain font by default. See \code{\link[graphics:par]{font}} for
alternatives.}

\item{axes_label_text_size}{numeric to control the text size of axes labels,
set to 1.1 by default.}

\item{axes_label_text_col}{colour to use for axes labels text, set to
\code{"black"} by default.}

\item{title_text_font}{numeric to control the font of title text, set to 2
for bold font by default. See \code{\link[graphics:par]{font}} for
alternatives.}

\item{title_text_size}{numeric to control the text size of the plot title,
set to 1.1 by default.}

\item{title_text_col}{colour to use for plot title text, set to
\code{"black"} by default.}

\item{legend}{can be either \code{"line"} or \code{"fill"} to indicate
whether a legend should be constructed based on the density \code{"line"}
or \code{"fill"}, set to FALSE by default to remove the legend.}

\item{legend_text}{vector of labels to use in the legend.}

\item{legend_text_font}{numeric to control the font of legend text, set to 1
for plain font by default. See \code{\link[graphics:par]{font}} for
alternatives.}

\item{legend_text_size}{numeric to control the size of text in the legend,
set to 1 by default.}

\item{legend_text_col}{colour(s) to use for text in legend, set to
\code{"black"} by default.}

\item{legend_line_col}{colour(s) to use for the lines in 1-D plot legends
when legend is set to \code{"line"}.}

\item{legend_box_fill}{fill colour(s) to use for the boxes in 1-D plot
legends when legend is set to \code{"fill"}.}

\item{legend_point_col}{colour(s) to use for points in 2-D scatter plot
legend.}

\item{gate_line_type}{integer [0,6] to control the line type of gates, set to
\code{1} to draw solid lines by default. See
\code{\link[graphics:par]{lty}} for alternatives.}

\item{gate_line_width}{numeric to control the line width(s) of gates, set to
\code{2.5} by default.}

\item{gate_line_col}{colour(s) to use for gates, set to \code{"red"} by
default.}

\item{gate_fill}{fill colour(s) to use for gates, set to "white by default.}

\item{gate_fill_alpha}{numeric to control the fill transparency of gates, set
to 0 by default to remove fill colour(s).}

\item{label}{logical indicating whether gated populations should be labelled.
To include the names of the populations in these labels, supply the
population names to the \code{label_text} argument. The default statistic
is \code{"percent"} for gated data and \code{"count"} for un-gated data.
This argument must be set to TRUE in order to add labels with gates.}

\item{label_text}{vector of population names to use in the labels.The exclude
the population names set this argument to NA.}

\item{label_stat}{indicates the type of statistic to include in the plot
labels, can be \code{"percent"}, \code{"count"}, \code{"mean"},
\code{"median"}, \code{"mode"} or \code{"geo mean"}, set to
\code{"percent"} for gated data or \code{NA} to exclude statistics for
un-gated data. Currently, only \code{"percent"} and \code{"count"} are
supported for 2-D scatter plots.}

\item{label_position}{either "auto" or "manual". The "auto" option (default)
positions labels will be placed in the center of gates and offset if
necessary. The "manual" option will allow label positioning by mouse click.
Label positions are set on a per gate basis, all samples in the same group
will have the same label positions. To individually label plots users must
manually supply the co-ordinates to label_text_x and label_text_y.}

\item{label_text_x}{vector of x co-ordinate(s) to manually adjust the
position plot label(s) on the plot. To interactively position labels set
either \code{label_text_x} or \code{label_text_y} to "select".}

\item{label_text_y}{vector of y co-ordinate(s) to manually adjust the
position plot label(s) on the plot. To interactively position labels set
either \code{label_text_x} or \code{label_text_y} to "select".}

\item{label_text_font}{numeric to control the font of text in plot labels,
set to 2 for bold font by default. See \code{\link[graphics:par]{font}} for
alternatives.}

\item{label_text_size}{numeric to control the size of text in the plot
labels, set to 1 by default.}

\item{label_text_col}{colour(s) to use for text in plot labels, set to
\code{"black"} by default.}

\item{label_fill}{fill colour(s) to use for labels, set to "white" by
default.}

\item{label_fill_alpha}{numeric to control background fill transparency of
label, set to 0.6 by default to introduce some transparency.}

\item{border_line_type}{integer [0,6] to control the line type of plot
border, set to \code{1} by default for a solid border. See
\code{\link[graphics:par]{lty}} for alternatives.}

\item{border_line_width}{numeric to control line width for the plot border,
set to 1 by default.}

\item{border_line_col}{colour to use for the plot border, set to "black" by
default.}

\item{border_fill}{border_fill fill colour to use inside the plot border
(i.e. background colour), set to "white" by default.}

\item{border_fill_alpha}{transparency to use for border_fill colour, set to 1
by default for no transparency.}

\item{...}{additional arguments not currently in use.}
}
\description{
Explore & visualise a GatingSet.
}
\examples{
library(CytoRSuiteData)

# Load samples into GatingSet
fs <- Activation
gs <- GatingSet(fs)

# Apply coompensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(gs))
gs <- transform(gs, trans)

# Apply gatingTemplate
gt <- Activation_gatingTemplate
gating(gt, gs)

# 2-D scatter plot with overlay & Gates
cyto_plot(gs,
  parent = "CD4 T Cells",
  alias = "CD69+ CD4 T Cells",
  channels = c("Alexa Fluor 647-A", "7-AAD-A"),
  overlay = "CD8 T Cells"
)

# 2-D Scatter Plots with Back-Gating & Gates
cyto_plot(gs,
  parent = "T Cells",
  alias = c("CD4 T Cells", "CD8 T Cells"),
  channels = c("Alexa Fluor 488-A", "Alexa Fluor 700-A"),
  overlay = c("CD69+ CD4 T Cells", "CD69+ CD8 T Cells")
)
}
\seealso{
\code{\link{cyto_plot,flowFrame-method}}

\code{\link{cyto_plot,flowSet-method}}

\code{\link{cyto_plot,GatingHierarchy-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
