% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot.R
\name{cyto_plot}
\alias{cyto_plot}
\title{cyto_plot}
\usage{
cyto_plot(
  x,
  parent = "root",
  channels,
  alias = NA,
  axes_trans = NA,
  merge_by = "name",
  overlay = NA,
  gate = NA,
  axes_limits = "auto",
  events = 50000,
  layout,
  margins = c(NA, NA, NA, NA),
  popup = TRUE,
  popup_size = c(NA, NA),
  select = NULL,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  xlab,
  ylab,
  title,
  negate = FALSE,
  hist_stat = "percent",
  hist_bins = 256,
  hist_smooth = 1,
  hist_stack = 0,
  hist_layers = NA,
  hist_cols = NA,
  hist_fill = NA,
  hist_fill_alpha = 1,
  hist_line_type = 1,
  hist_line_width = 1,
  hist_line_col = "black",
  point_shape = ".",
  point_size = 2,
  point_col_scale = NA,
  point_col_smooth = TRUE,
  point_cols = NA,
  point_col = NA,
  point_col_alpha = 1,
  point_fast = FALSE,
  point_bins = 256,
  point_stack = FALSE,
  contour_lines = 0,
  contour_line_type = 1,
  contour_line_width = 1,
  contour_line_col = "black",
  contour_line_alpha = 1,
  axes_limits_buffer = 0.03,
  axes_text = c(TRUE, TRUE),
  axes_text_font = 1,
  axes_text_size = 1,
  axes_text_col = "black",
  axes_label_text_font = 1,
  axes_label_text_size = 1.1,
  axes_label_text_col = "black",
  axes_ticks_line_width = 1,
  axes_ticks_line_col = "black",
  axes_ticks_line_col_alpha = 1,
  title_text_font = 2,
  title_text_size = 1.1,
  title_text_col = "black",
  legend = FALSE,
  legend_text = NA,
  legend_text_font = 1,
  legend_text_size = 1,
  legend_text_col = "black",
  legend_line_type = NA,
  legend_line_width = NA,
  legend_line_col = NA,
  legend_box_fill = NA,
  legend_point_col = NA,
  key = "both",
  key_size = c(1, 1),
  key_scale = "fixed",
  key_text_font = 1,
  key_text_size = 0.9,
  key_text_col = "black",
  key_text_col_alpha = 1,
  key_title = "",
  key_title_text_font = 1,
  key_title_text_size = 1,
  key_title_text_col = "black",
  key_title_text_col_alpha = 1,
  key_hist_line_type = 1,
  key_hist_line_width = 1.5,
  key_hist_line_col = "black",
  key_border_line_width = 1,
  key_border_line_col = "black",
  key_border_line_col_alpha = 1,
  key_ticks_line_width = 1,
  key_ticks_line_col = "black",
  key_ticks_line_col_alpha = 1,
  gate_line_type = 1,
  gate_line_width = 2.5,
  gate_line_col = "red",
  gate_line_col_alpha = 1,
  gate_fill = "white",
  gate_fill_alpha = 0,
  gate_point_shape = NA,
  gate_point_size = 1,
  gate_point_col = "red",
  gate_point_col_alpha = 1,
  label,
  label_text,
  label_stat,
  label_position = "auto",
  label_text_x = NA,
  label_text_y = NA,
  label_text_font = 2,
  label_text_size = 0.8,
  label_text_col = "black",
  label_text_col_alpha = 1,
  label_fill = "white",
  label_fill_alpha = 0.75,
  border_line_type = 1,
  border_line_width = 1,
  border_line_col = "black",
  border_fill = "white",
  border_fill_alpha = 1,
  grid = TRUE,
  grid_line_type = 1,
  grid_line_width = 1,
  grid_line_col = "grey95",
  grid_line_alpha = 1,
  header = NA,
  header_text_font = 2,
  header_text_size = 1,
  header_text_col = "black",
  page = FALSE,
  page_fill = "white",
  page_fill_alpha = 1,
  memory = TRUE,
  seed = 42,
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the population to plot when a \code{GatingHierarchy} or
\code{GatingSet} object is supplied, set to the \code{"root"} node by
default.}

\item{channels}{name of the channel(s) or marker(s) to be used to construct
the plot. The length of channels determines the type of plot to be
constructed, either a histogram for a single channel or a 2-D scatterplot
with density colour scale for two channels.}

\item{alias}{name of the gated population(s) to gated in the plot when a
\code{GatingHierarchy} or \code{GatingSet} object is supplied. Setting
\code{alias} to "" will automatically plot any gates constructed in the
supplied channels. \code{alias} is equivalent to the \code{gate} argument
for \code{cytoset} objects.}

\item{axes_trans}{object of class
\code{\link[flowWorkspace:transformerList]{transformerList}} which was used
to transform the channels of the supplied data. \code{cyto_plot} does not
support in-line transformations and as such the transformations should be
applied to the data prior to plotting. The transformerList is used
internally to ensure that the axes on the constructed plots are
appropriately labelled.}

\item{merge_by}{a vector of experiment variables to sort and merge samples
into groups prior to plotting, set to NA by default to prevent merging. To
merge all samples set this argument to \code{TRUE} or \code{"all"}.}

\item{overlay}{name(s) of the populations to overlay or a \code{cytoset},
\code{list of cytosets} or \code{list of cytoset lists} containing
populations to be overlaid onto the plot(s). This argument can be set to
"children" or "descendants" when a \code{GatingSet} or
\code{GatingHierarchy} to overlay all respective nodes.}

\item{gate}{gate objects to be plotted, can be either objects of class
\code{rectangleGate}, \code{polygonGate}, \code{ellipsoidGate},
\code{quadGate} or \code{filters}. Lists of these supported gate objects
are also supported.}

\item{axes_limits}{options include \code{"auto"}, \code{"trim"},
\code{"data"} or \code{"machine"} to use optimised, lower trimmed (remove
lowest 1% of events), data or machine limits respectively. Set to
\code{"auto"} by default to use optimised axes ranges. Fine control over
axes limits can be obtained by altering the \code{xlim} and \code{ylim}
arguments.}

\item{events}{numeric to control the number or percentage of events to
display. Values [0,1] indicate the percentage of events to display (i.e.
value of 1 will display all events), whilst values larger than 1 indicate
the number of events to display. The default value for \code{events} is set
to 50000 to display 50000 events only. Setting \code{events = NA} will
automatically determine the minimum number of events across groups and use
that number for downsampling.}

\item{layout}{a vector of the length 2 of form \code{c(#rows, #columns)} or a
matrix indicating the dimensions of the grid for plotting.}

\item{margins}{a vector of length 4 to control the margins around the bottom,
left, top and right of the plot, set to c(NA, NA, NA, NA) by default to let
\code{cyto_plot} compute optimal margins.}

\item{popup}{logical indicating whether the plot should be constructed in a
pop-up window, set to TRUE by default. \code{popup} will open OS-specific
graphic device prior to plotting. Mac users will need to install
\href{https://www.xquartz.org/}{XQuartz} for this functionality.}

\item{popup_size}{a vector of length 2 to control the height and width of
pop-up graphics device in inches, set to \code{c(8,8)} by default.}

\item{select}{named list containing experimental variables to be used to
select samples using \code{\link{cyto_select}} when a \code{cytoset} or
\code{GatingSet} is supplied. Refer to \code{\link{cyto_select}} for more
details. Sample selection occurs prior to grouping with \code{merge_by}.}

\item{xlim}{lower and upper limits of x axis (e.g. c(0,250000)).}

\item{ylim}{lower and upper limits of y axis (e.g. c(0,250000)).}

\item{xlab}{x axis label.}

\item{ylab}{y axis label.}

\item{title}{title to use for the plot, set to the name of the sample by
default. Title can be removed by setting this argument to \code{NA}.}

\item{negate}{logical indicating whether a label should be included for the
negated population when gate objects are supplied, set to FALSE by default.
Setting \code{negate = TRUE} will result in the creation of a boolean
filter that contains all the events outside the gates supplied to
\code{alias} or \code{gate}. If such a boolean gate exists in the
GatingSet/GatingHierarchy it will automatically be extracted when
\code{alias = ""}. In order to prevent plotting of these boolean gates,
users will need to explicitly pass the names of the gates they want to
display to \code{alias}.}

\item{hist_stat}{can be either \code{"count"}, \code{"percent"} or
\code{"density"} to indicate the statistic to display on histograms, set to
\code{"count"} by default. The \code{"percent"} option applies modal
normalisation and expresses the result as a percentage.}

\item{hist_bins}{number of bins to use for histograms, set to 256 by default.}

\item{hist_smooth}{smoothing parameter passed to
\code{\link[stats:density]{density}} to adjust the smoothness of the kernel
density for histograms, set to \code{1} by default. Only values greater or
equal to 1 are supported.}

\item{hist_stack}{numeric [0,1] indicating the degree of stacking for
histograms, set to \code{0} by default.}

\item{hist_layers}{numeric indicating the number of histograms to stack in
each plot when there are no overlays, set to 1 histogram per plot by
default. Each plot must contain the same number of histograms.}

\item{hist_cols}{vector colours to draw from when selecting histogram fill
colours if none are supplied to \code{hist_fill}.}

\item{hist_fill}{fill colour(s) for histograms, select from \code{hist_cols}
if not supplied.}

\item{hist_fill_alpha}{numeric [0,1] used to control histogram fill colour
transparency, set to \code{1} by default for solid colours.}

\item{hist_line_type}{line type(s) to use for histogram borders, set to 1 by
default to use solid lines. See \code{\link[graphics:par]{lty}} for
alternatives.}

\item{hist_line_width}{numeric to control line width(s) for histogram borders
lines, set to 1 by default.}

\item{hist_line_col}{colour(s) for histogram borders, set to \code{"black"}
by default.}

\item{point_shape}{shape(s) to use for points in 2-D scatterplots, set to
\code{"."} by default to maximise plotting speed.  See
\code{\link[graphics:par]{pch}} for alternatives. Hex binning is supported
using \code{point_shape = "hex"}.}

\item{point_size}{numeric to control the size of points in 2-D scatter plots
set to 2 by default.}

\item{point_col_scale}{vector of ordered colours to use for the density
colour gradient of points.}

\item{point_col_smooth}{logical indicating whether the 2D binned counts
should be smoothed using kernel density estimates prior to selecting
colours from \code{point_col_scale}, set to TRUE by default. Setting
\code{point_col_smooth} to FALSE will significantly improve plotting speed
on less powerful machines but produce more granular plots.}

\item{point_cols}{vector colours to draw from when selecting colours for
points if none are supplied to point_col.}

\item{point_col}{colour(s) to use for points in 2-D scatter plots, set to NA
by default to use a blue-red density colour scale.}

\item{point_col_alpha}{numeric [0,1] to control point colour transparency in
2-D scatter plots, set to 1 by default to use solid colours.}

\item{point_fast}{logical indicating whether points should be plotted using
the \code{scattermore} package, set to FALSE by default. This is an
optional feature so if you intend to use it, make sure that you have the
scattermore package installed \code{install.packages("scattermore")}.}

\item{point_bins}{number of bins to use for points when \code{point_shape =
"hex"}, set to 256 bins by default.}

\item{point_stack}{logical indicating whether merged data without
\code{overlay} specified should be plotted as multiple layers on the same
plot, set to FALSE by default.}

\item{contour_lines}{numeric indicating the number of levels to use for
contour lines in 2-D scatter plots, set to 0 by default to turn off contour
lines.}

\item{contour_line_type}{integer [0,6] to control the line type of contour
lines in 2-D scatter plots, set to \code{1} to draw solid lines by default.
See \code{\link[graphics:par]{lty}} for alternatives.}

\item{contour_line_width}{numeric to control line width(s) for contour lines
in 2-D scatter plots, set to 2 by default.}

\item{contour_line_col}{colour(s) to use for contour lines in 2-D scatter
plots, set to \code{"black"} by default.}

\item{contour_line_alpha}{numeric [0,1] to control the transparency of
contour lines, set to 1 by default to remove transparency.}

\item{axes_limits_buffer}{decimal indicating the percentage of buffering to
add to either end of the axes limits, set to 0.03 by default.}

\item{axes_text}{logical vector of length 2 indicating whether axis text
should be included for the x and y axes respectively, set to
\code{c(TRUE,TRUE)} by default to display axes text on both axes.
\code{axes_text} can also be set to NA to remove the text only but keep the
axes ticks.}

\item{axes_text_font}{numeric to control the font of axes text, set to 1 for
plain font by default. See \code{\link[graphics:par]{font}} for
alternatives.}

\item{axes_text_size}{numeric to control the size of axes text, set to 1 by
default.}

\item{axes_text_col}{colour to use for axes text, set to \code{"black"} by
default.}

\item{axes_label_text_font}{numeric to control the font axes labels, set to 1
for plain font by default. See \code{\link[graphics:par]{font}} for
alternatives.}

\item{axes_label_text_size}{numeric to control the text size of axes labels,
set to 1.1 by default.}

\item{axes_label_text_col}{colour to use for axes labels text, set to
\code{"black"} by default.}

\item{axes_ticks_line_width}{numeric to control the line width of axes ticks,
set to 1 by default. Refer to \code{lwd} in \code{\link[graphics:par]{par}}
for alternatives.}

\item{axes_ticks_line_col}{colour to use for axes ticks, set to
\code{"black"} by default.}

\item{axes_ticks_line_col_alpha}{numeric [0,1] to control the transparency of
caxes ticks, set to 1 by default to remove transparency.}

\item{title_text_font}{numeric to control the font of title text, set to 2
for bold font by default. See \code{\link[graphics:par]{font}} for
alternatives.}

\item{title_text_size}{numeric to control the text size of the plot title,
set to 1.1 by default.}

\item{title_text_col}{colour to use for plot title text, set to
\code{"black"} by default.}

\item{legend}{can be either \code{"line"} or \code{"fill"} to indicate
whether a legend should be constructed based on the histogram \code{"line"}
or \code{"fill"}, set to FALSE by default to remove the legend.}

\item{legend_text}{vector of labels to use in the legend.}

\item{legend_text_font}{numeric to control the font of legend text, set to 1
for plain font by default. See \code{\link[graphics:par]{font}} for
alternatives.}

\item{legend_text_size}{numeric to control the size of text in the legend,
set to 1 by default.}

\item{legend_text_col}{colour(s) to use for text in legend, set to
\code{"black"} by default.}

\item{legend_line_type}{numeric to control the line type for line legends,
set to 1 by default. Refer to \code{lty} in \code{\link[graphics:par]{par}}
for alternatives.}

\item{legend_line_width}{numeric to control the line width in line legend,
set to 1 by default. Refer to \code{lwd} in \code{\link[graphics:par]{par}}
for alternatives.}

\item{legend_line_col}{colour(s) to use for the lines in 1-D plot legends
when legend is set to \code{"line"}.}

\item{legend_box_fill}{fill colour(s) to use for the boxes in 1-D plot
legends when legend is set to \code{"fill"}.}

\item{legend_point_col}{colour(s) to use for points in 2-D scatter plot
legend.}

\item{key}{options include \code{"scale"} to display only the colour scale,
\code{"both"} to display the colour scale and text to annotate
\code{counts} or \code{fluorescence units}, or \code{"none"} to remove the
key from the plot(s). Set to \code{"both"} by default to display colour
scale and associated text in the key.}

\item{key_size}{a vector of scaling factors between 0 and Inf for the weight
and height of the key, set to \code{c(1,1)} by default.}

\item{key_scale}{can be either \code{"fixed"} to use the same colour scale
for points between plots, or \code{"free"} to set a unique colour scale for
each plot. \code{key_scale} is set to \code{"fixed"} by default create a
shared colour scale which allows comparison of relative counts between
plots. \code{key_scale} also accepts a vector containing the range of
values for the scale, or a list of ranges for the scale in each plot.}

\item{key_text_font}{font to use for text in the key, set to 1 by default for
plain font.}

\item{key_text_size}{numeric to control the size of text in the plot key, set
to 0.9 by default.}

\item{key_text_col}{colour to use for text in the plot key, set to
\code{"black"} by default.}

\item{key_text_col_alpha}{numeric [0, 1] to control the transparency of the
text colour in the plot key, set to 1 by default to remove transparency.}

\item{key_title}{title to place above the key set to \code{"count"} when
\code{point_col = NA} or the name of the channel/marker supplied to
\code{point_col}.}

\item{key_title_text_font}{font to use for the key title, set to 1 by default
for plain font.}

\item{key_title_text_size}{numeric to control the size of title text in teh
key, set to 1 by default.}

\item{key_title_text_col}{colour to use for the key title, set to
\code{"black"} by default.}

\item{key_title_text_col_alpha}{numeric [0, 1] to control the transparency of
the key title text, set to 1 by default to remove transparency.}

\item{key_hist_line_type}{line type(s) to use for histogram borders in the
key, set to 1 by default to use solid lines. See
\code{\link[graphics:par]{lty}} for alternatives.}

\item{key_hist_line_width}{numeric to control line width(s) for histogram
borders lines, set to 1 by default.}

\item{key_hist_line_col}{colour to use for the histogram border in the key,
set to \code{"black"} by default.}

\item{key_border_line_width}{numeric to control the line width in key border,
set to 1 by default. Refer to \code{lwd} in \code{\link[graphics:par]{par}}
for alternatives.}

\item{key_border_line_col}{colour to use for key border, set to
\code{"black"} by default.}

\item{key_border_line_col_alpha}{numeric [0, 1] to control the transparency
of the key border, set to 1 by default to remove transparency.}

\item{key_ticks_line_width}{numeric to control the line width in key ticks,
set to 1 by default. Refer to \code{lwd} in \code{\link[graphics:par]{par}}
for alternatives.}

\item{key_ticks_line_col}{colour to use for key ticks, set to \code{"black"}
by default.}

\item{key_ticks_line_col_alpha}{numeric [0, 1] to control the transparency of
the key ticks, set to 1 by default to remove transparency.}

\item{gate_line_type}{integer [0,6] to control the line type of gates, set to
\code{1} to draw solid lines by default. See
\code{\link[graphics:par]{lty}} for alternatives.}

\item{gate_line_width}{numeric to control the line width(s) of gates, set to
\code{2.5} by default.}

\item{gate_line_col}{colour(s) to use for gates, set to \code{"red"} by
default.}

\item{gate_line_col_alpha}{numeric [0,1] to control the transparency of gate
borders, set to 1 by default for solid colours.}

\item{gate_fill}{fill colour(s) to use for gates, set to "white by default.}

\item{gate_fill_alpha}{numeric to control the fill transparency of gates, set
to 0 by default to remove fill colour(s).}

\item{gate_point_shape}{shape(s) to use for gate vertices, set to \code{NA}
by default to remove gate vertices. See \code{\link[graphics:par]{pch}} for
alternatives.}

\item{gate_point_size}{numeric to control the size of the gate vertices, set
to 1 by default.}

\item{gate_point_col}{colour(s) to use for the gate vertices, set to
\code{"red"} by default.}

\item{gate_point_col_alpha}{numeric [0,1] to control the transparency of gate
vertices, set to 1 by default for solid colours.}

\item{label}{logical indicating whether gated populations should be labelled.
To include the names of the populations in these labels, supply the
population names to the \code{label_text} argument. The default statistic
is \code{"percent"} for gated data. This argument must be set to TRUE in
order to add labels with gates.}

\item{label_text}{vector of population names to use in the labels.To exclude
the population names set this argument to NA.}

\item{label_stat}{indicates the type of statistic to include in the plot
labels, can be \code{"percent"}, \code{"count"}, \code{"mean"},
\code{"median"}, \code{"mode"} or \code{"geo mean"}, set to
\code{"percent"} for gated data or \code{NA} to exclude statistics for
un-gated data. Currently, only \code{"percent"} and \code{"count"} are
supported for 2-D scatter plots.}

\item{label_position}{either "auto" or "manual". The "auto" option (default)
positions labels will be placed in the center of gates and offset if
necessary. The "manual" option will allow label positioning by mouse click.
Label positions are set on a per gate basis, all samples in the same group
will have the same label positions. To individually label plots users must
manually supply the co-ordinates to label_text_x and label_text_y.}

\item{label_text_x}{vector of x co-ordinate(s) to manually adjust the
position plot label(s) on the plot. To interactively position labels set
either \code{label_text_x} or \code{label_text_y} to "select".}

\item{label_text_y}{vector of y co-ordinate(s) to manually adjust the
position plot label(s) on the plot. To interactively position labels set
either \code{label_text_x} or \code{label_text_y} to "select".}

\item{label_text_font}{numeric to control the font of text in plot labels,
set to 2 for bold font by default. See \code{\link[graphics:par]{font}} for
alternatives.}

\item{label_text_size}{numeric to control the size of text in the plot
labels, set to 1 by default.}

\item{label_text_col}{colour(s) to use for text in plot labels, set to
\code{"black"} by default.}

\item{label_text_col_alpha}{numeric [0, 1] to control the transparency of the
text colour, set to 1 by default to remove transparency.}

\item{label_fill}{fill colour(s) to use for labels, set to "white" by
default.}

\item{label_fill_alpha}{numeric to control background fill transparency of
label, set to 0.6 by default to introduce some transparency.}

\item{border_line_type}{integer [0,6] to control the line type of plot
border, set to \code{1} by default for a solid border. See
\code{\link[graphics:par]{lty}} for alternatives.}

\item{border_line_width}{numeric to control line width for the plot border,
set to 1 by default.}

\item{border_line_col}{colour to use for the plot border, set to "black" by
default.}

\item{border_fill}{border_fill fill colour to use inside the plot border
(i.e. background colour), set to "white" by default.}

\item{border_fill_alpha}{transparency to use for border_fill colour, set to 1
by default for no transparency.}

\item{grid}{logical indicating whether to include grid lines in the plot
background, set to TRUE by default. Alternatively, users can supply a
integer to indicate the number of equally spaced quantiles to used for the
grid lines.}

\item{grid_line_type}{integer [0,6] to control the line type of grid lines,
set to \code{1} to draw solid lines by default. See
\code{\link[graphics:par]{lty}} for alternatives.}

\item{grid_line_width}{numeric to control the line width(s) of grid lines,
set to \code{1} by default.}

\item{grid_line_col}{colour to use for grid lines, set to \code{"grey95"} by
default.}

\item{grid_line_alpha}{numeric [0,1] to control the transparency of grid
lines, set to 1 by default to remove transparency.}

\item{header}{text to include above the plots when plotting a \code{cytoset}
or \code{GatingSet}. \code{cyto_plot} expects a single header which will be
added to each full page within the \code{cyto_plot} call. For custom
plotting functions, if the final page has empty panels, we need to set
\code{page = TRUE} in custom plotting functions to ensure that the header
is appropriately added.}

\item{header_text_font}{numeric to control the font of the header text, set
to 2 for bold font by default. See \code{\link[graphics:par]{font}} for
alternatives.}

\item{header_text_size}{numeric to control the size of the header text, set
to 1 by default.}

\item{header_text_col}{colour to use for the header text, set to "black" by
default.}

\item{page}{used internally within cyto_plot() wrapper functions to signal if
a new page is required after constructing the plot(s), set to FALSE by
default. This argument is only required for custom layouts which result in
pages with empty panels that require a header, because cyto_plot() only
adds headers to complete pages.}

\item{page_fill}{colour to use to fill the page prior to adding plot panels,
set to \code{"white"} by default.}

\item{page_fill_alpha}{numeric [0,1] to control the transparency of the page
colour, set to 1 by default to remove transparency.}

\item{memory}{logical indicating whether \code{cyto_plot()} should remember
label co-ordinates when \code{label_position = "manual"} and use those
co-ordinates when \code{cyto_plot_save()} is called, set to TRUE by
default.}

\item{seed}{numeric passed to \code{\link{set.seed}} to ensure that the same
sampling is applied with each \code{\link{cyto_plot}} call, set to an
arbitrary numeric by default. This behaviour can be turned off by setting
this argument to NULL.}

\item{...}{not currently in use.}
}
\value{
a list containing the recorded plots.
}
\description{
Explore and visualise cytometry data.
}
\examples{
library(CytoExploreRData)

# Activation GatingSet
gs <- cyto_load(
  system.file(
    "extdata/Activation-GatingSet",
    package = "CytoExploreRData"
  )
)

# 2-D scatter plot with overlay & Gates
cyto_plot(gs[1:9],
  parent = "CD4 T Cells",
  alias = "CD69+ CD4 T Cells",
  channels = c("Alexa Fluor 647-A", "7-AAD-A"),
  overlay = "CD8 T Cells"
)

# 2-D Scatter Plots with Back-Gating & Gates
cyto_plot(gs[1:9],
  parent = "T Cells",
  alias = c("CD4 T Cells", "CD8 T Cells"),
  channels = c("Alexa Fluor 488-A", "Alexa Fluor 700-A"),
  overlay = c("CD69+ CD4 T Cells", "CD69+ CD8 T Cells")
)
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
