% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot4.R
\name{cyto_plot4}
\alias{cyto_plot4}
\alias{cyto_plot4.flowFrame}
\alias{cyto_plot4.flowSet}
\title{cyto_plot4}
\usage{
\method{cyto_plot4}{flowFrame}(x, channels, axes_trans = NA,
  overlay = NA, gate = NA, limits = "data", display = 25000,
  popup = FALSE, xlim = NA, ylim = NA, xlab, ylab, title,
  negate = FALSE, density_modal = TRUE, density_smooth = 0.6,
  density_stack = 0, density_cols = NA, density_fill = NA,
  density_fill_alpha = 1, density_line_type = 1,
  density_line_width = 1, density_line_col = "black",
  point_shape = ".", point_size = 2, point_col_scale = NA,
  point_cols = NA, point_col = NA, point_col_alpha = 1,
  contour_lines = 0, contour_line_type = 1, contour_line_width = 1,
  contour_line_col = "black", contour_line_alpha = 1,
  axes_text = c(TRUE, TRUE), axes_text_font = 1, axes_text_size = 1,
  axes_text_col = "black", axes_label_text_font = 1,
  axes_label_text_size = 1.1, axes_label_text_col = "black",
  title_text_font = 2, title_text_size = 1.1,
  title_text_col = "black", legend = FALSE, legend_text = NA,
  legend_text_font = 1, legend_text_size = 1,
  legend_text_col = "black", legend_line_type = NA,
  legend_line_width = NA, legend_line_col = NA, legend_box_fill = NA,
  legend_point_col = NA, gate_line_type = 1, gate_line_width = 2.5,
  gate_line_col = "red", gate_fill = "white", gate_fill_alpha = 0,
  label, label_text, label_stat, label_position = "auto",
  label_text_x = NA, label_text_y = NA, label_text_font = 2,
  label_text_size = 1, label_text_col = "black",
  label_fill = "white", label_fill_alpha = 0.6, border_line_type = 1,
  border_line_width = 1, border_line_col = "black",
  border_fill = "white", border_fill_alpha = 1, ...)

\method{cyto_plot4}{flowSet}(x, channels, axes_trans = NA,
  group_by = NA, overlay = NA, gate = NA, limits = "data",
  display = 1, layout, popup = FALSE, xlim = NA, ylim = NA, xlab,
  ylab, title, negate = FALSE, density_modal = TRUE,
  density_smooth = 0.6, density_stack = 0, density_layers = NA,
  density_cols = NA, density_fill = NA, density_fill_alpha = 1,
  density_line_type = 1, density_line_width = 1,
  density_line_col = "black", point_shape = ".", point_size = 2,
  point_col_scale = NA, point_cols = NA, point_col = NA,
  point_col_alpha = 1, contour_lines = 0, contour_line_type = 1,
  contour_line_width = 1, contour_line_col = "black",
  contour_line_alpha = 1, axes_text = c(TRUE, TRUE),
  axes_text_font = 1, axes_text_size = 1, axes_text_col = "black",
  axes_label_text_font = 1, axes_label_text_size = 1.1,
  axes_label_text_col = "black", title_text_font = 2,
  title_text_size = 1.1, title_text_col = "black", legend = FALSE,
  legend_text = NA, legend_text_font = 1, legend_text_size = 1,
  legend_text_col = "black", legend_line_type = NA,
  legend_line_width = NA, legend_line_col = NA, legend_box_fill = NA,
  legend_point_col = NA, gate_line_type = 1, gate_line_width = 2.5,
  gate_line_col = "red", gate_fill = "white", gate_fill_alpha = 0,
  label, label_text = NA, label_stat, label_position = "auto",
  label_text_x = NA, label_text_y = NA, label_text_font = 2,
  label_text_size = 0.8, label_text_col = "black",
  label_fill = "white", label_fill_alpha = 0.6, border_line_type = 1,
  border_line_width = 1, border_line_col = "black",
  border_fill = "white", border_fill_alpha = 1, ...)
}
\description{
cyto_plot4
}
