% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_extensions.R
\name{cyto_plot_animate}
\alias{cyto_plot_animate}
\title{Create cyto_plot animations}
\usage{
cyto_plot_animate(x, save_as = NULL, height = 800, width = 800, delay = 1, ...)
}
\arguments{
\item{x}{a list of recorded plots returned by a call to \code{cyto_plot()} or
the names of a series of PNG files generated by \code{cyto_plot_save()}.}

\item{save_as}{name of a GIF file to which the animated plot should be saved.}

\item{height}{GIF height in pixels, set to 800 pixels by default.}

\item{width}{GIF width in pixels, set to 800 pixels by default.}

\item{delay}{time in seconds to display each image in the animation, set to 1
second by default.}

\item{...}{additional arguments passed to \code{gifski::gifski()}.}
}
\value{
create a GIF file in the current working directory containing the
  animated plot.
}
\description{
Create cyto_plot animations
}
\examples{
\dontrun{
library(CytoExploreRData)

gs <- GatingSet(Activation)

gs <- cyto_compensate(gs)

gs <- cyto_transform(gs)

gs <- cyto_gatingTemplate_apply(gs, Activation_gatingTemplate)

cyto_plot_save("cyto_plot-1.png")
cyto_plot(
gs[29], 
parent = "CD4 T Cells", 
channels = c("CD44", "CD69"), 
title = "0 nM OVA"
)

cyto_plot_save("cyto_plot-2.png")
cyto_plot(
gs[30], 
parent = "CD4 T Cells", 
channels = c("CD44", "CD69"), 
title = "5 nM OVA"
)

cyto_plot_save("cyto_plot-3.png")
cyto_plot(
gs[31], 
parent = "CD4 T Cells", 
channels = c("CD44", "CD69"), 
title = "50 nM OVA"
)

cyto_plot_save("cyto_plot-4.png")
cyto_plot(
gs[32], 
parent = "CD4 T Cells", 
channels = c("CD44", "CD69"), 
title = "50 nM OVA"
)

cyto_plot_animate(
c("cyto_plot-1.png", 
"cyto_plot-2.png", 
"cyto_plot-3.png", 
"cyto_plot-4.png"),
save_as = "cyto_plot_animation.gif"
)

}

}
\seealso{
\code{\link{cyto_plot}}

\code{\link{cyto_plot_save}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
