% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_calibrate}
\alias{cyto_plot_calibrate}
\title{Calibrate channel ranges for cyto_plot}
\usage{
cyto_plot_calibrate(
  x,
  parent = "root",
  channels = NULL,
  select = NULL,
  type = "quantile",
  probs = c(0.01, 0.95),
  limits = c(NA, NA),
  axes_trans = NA,
  anchor = TRUE,
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} to use for the
calibration. For the best calibration is recommended that users supply
samples containing both negative and positive events in each channel.}

\item{parent}{name of the parent population to use for channel calibration
when a \code{GatingHierarchy} or \code{GatingSet} is supplied, set to the
\code{"root"} node by default.}

\item{channels}{names of the channels or markers for which the calibration.
settings should be updated, set to all channels by default. All channels
within the same \code{cyto_plot_calibrate} call will use the same settings
for \code{type} and \code{probs}. Multiple calls to
\code{cyto_plot_calibrate} allow for channel-wise control over these
settings which are stored globally.}

\item{select}{sample selection criteria passed to \code{cyto_select()} to
extract samples that should be used to calibrate the channels, set to NULL
by default to calibrate channels using data from all samples.}

\item{type}{indicates the type of calibration to perform, options include
\code{"range"} or \code{"quantile"}, set to \code{"quantile"} by default.
Range calibration simply uses the full range of values across samples for
the calibration. Quantile calibration computes an lower and upper quantile
for each channel, values falling outside the calibration range are assigned
the bottom or top colour. Alternatively, users can also supply their own
custom calibration function to \code{type} that accepts a matrix.}

\item{probs}{vector of lower and upper probabilities passed to
\code{stats:quantile} to compute quantiles, set to \code{c(0.01, 0.99)} by
default.}

\item{limits}{manually supply channel limits in a named list or matrix,
values should be supplied on the linear scale as per
\code{\link{cyto_plot}}. Any missing values in \code{limits} will be
computed using the method indicated by \code{type}.}

\item{axes_trans}{object of class \code{transformerList} containing the
definitions of the transformers applied to the supplied data. These
transformers are only required when a cytoset to transform supplied channel
limits.}

\item{anchor}{logical indicating whether the lower limit of the scale should
be anchored to zero, set to TRUE by default.}

\item{...}{not in use.}
}
\value{
saves calibration settings for use by \code{\link{cyto_plot}}.
}
\description{
The density colour scale for points in \code{cyto_plot} can now be made to
represent fluorescent intensity of a third channel, instead of local density
in the 2D plotting space. In order for this feature to work properly, the
full range for the channel/marker of interest is required to appropriately
set the colour scale. \code{cyto_plot_calibrate} performs this range
calibration, so that this information can be used in all downstream
\code{cyto_plot} calls.
}
\examples{
library(CytoExploreRData)

# Activation Gatingset
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                          package = "CytoExploreRData"))

# Calibration
cyto_plot_calibrate(gs)

# Colour based on Hoechst-405 staining
cyto_plot(gs[1],
parent = "root",
channels = c("FSC-A", "SSC-A"),
point_col = "Hoechst-405")

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
