% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_calibrate_reset}
\alias{cyto_plot_calibrate_reset}
\title{Remove current calibration settings}
\usage{
cyto_plot_calibrate_reset()
}
\value{
remove current calibration settings created by
  \code{cyto_plot_calibrate}.
}
\description{
Remove current calibration settings
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
