% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_compensation-methods.R
\name{cyto_plot_compensation.flowFrame}
\alias{cyto_plot_compensation.flowFrame}
\title{Plot Compensation in All Fluorescent Channels - flowFrame Method}
\usage{
\method{cyto_plot_compensation}{flowFrame}(x, channel_match = NULL,
  compensate = FALSE, spillover = NULL, axes_trans = NA, layout,
  popup = FALSE, title = NA, header = NULL, header_text_font = 2,
  header_text_size = 1, header_text_col = "black", ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}}
containing gated compensation controls and an unstained control.}

\item{channel_match}{name of the fluorescent channel associated with the
\code{\link[flowCore:flowFrame-class]{flowFrame}}. If not supplied users
will need to select the channel from a dropdown menu.}

\item{compensate}{logical indicating whether the samples should be
compensated prior to plotting, set to FALSE by default. If no spillover
matrix is supplied to the spillover_file argument the spillover matrix will
extracted from the samples.}

\item{spillover}{name of spillover matrix csv file including .csv file
extension to apply to sample when \code{compensate} is TRUE. If no
\code{spillover} is supplied the spillover matrix will be extracted
directly from the \code{flowFrame} and applied to the sample when
\code{compensate} is TRUE.}

\item{axes_trans}{object of class
\code{\link[flowCore:transformList-class]{transformList}} or
\code{\link[flowWorkspace]{transformerList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} which was used to
transform the fluorescent channels of the supplied flowFrame. This
transform object will be used internally to ensure axes labels of the plot
are appropriately transformed. The transform object will NOT be applied to
the flowFrame internally and should be applied to the flowFrame prior to
plotting.}

\item{layout}{vector of grid dimensions \code{c(#rows,#columns)} for each
plot.}

\item{popup}{logical indicating whether plots should be constructed in a
pop-up window.}

\item{title}{text to include above each plot, set to NA by default to remove
titles.}

\item{header}{title to use for the plots, set to the name of the sample by
default. Turn off the header by setting this argument to NA.}

\item{header_text_font}{font to use for header text, set to 2 by default.}

\item{header_text_size}{text size for header, set to 1 by default.}

\item{header_text_col}{colour for header text, set to "black" by default.}

\item{...}{additional arguments passed to
\code{\link{cyto_plot,flowFrame-method}}.}
}
\description{
Plot each compensation control in all fluorescent channels to identify any
potential compensation issues. The unstained control is overlaid in black as
a reference.
}
\examples{
library(CytoRSuiteData)

# Don't run - bypass directory check for external files
options("CytoRSuite_wd_check" = FALSE)

# Load in compensation controls
gs <- GatingSet(Compensation)

# Gate single cells using gate_draw
gt <- Compensation_gatingTemplate
gating(gt, gs)

# Compensation plots
cyto_plot_compensation(getData(gs, "Single Cells")[[1]],
  channel_match = "7-AAD-A",
  overlay = getData(gs, "Single Cells")[[4]]
)

# Don't run - return "CytoRSuite_wd_check" to default
options("CytoRSuite_wd_check" = TRUE)
}
\seealso{
\code{\link{cyto_plot,flowFrame-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
