% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_comp.R
\name{cyto_plot_compensation}
\alias{cyto_plot_compensation}
\title{Visualise compensation of spillover in all channels}
\usage{
cyto_plot_compensation(
  x,
  parent = "root",
  channel_match = NULL,
  spillover = NULL,
  compensate = TRUE,
  overlay = "both",
  axes_trans = NA,
  axes_limits = "machine",
  layout,
  header,
  point_col = c("magenta", "blue", "grey40"),
  point_col_alpha = 0.5,
  hist_fill = c("magenta", "blue", "grey40"),
  hist_fill_alpha = 0.5,
  lines = TRUE,
  line_type = 1,
  line_width = 2,
  line_col = "red",
  line_col_alpha = 1,
  text = TRUE,
  text_font = 1,
  text_size = 1.2,
  text_col = "black",
  text_col_alpha = 1,
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} containing data for
single colour compensation controls.}

\item{parent}{names of the parent populations to extract from each control
for plotting.}

\item{channel_match}{a vector of channels one per control or the name of a
CSV file which indicates the name of the channel associated with each
control. If \code{channel_match} is not supplied or the channels are not
annotated in \code{cyto_details(x)} the user will be asked to interactively
enter this information prior to plotting.}

\item{spillover}{a matrix or name of a CSV file containing the spillover
coefficients that should be applied to the data when \code{compensate} is
TRUE.}

\item{compensate}{logical required for \code{cytoset} objects to indicate
whether the data should be compensated prior to plotting, set to TRUE by
default. This argument should be set to FALSE if the \code{spillover} has
already been applied to the data.}

\item{overlay}{can be either \code{"unstained"}, \code{"compensated"},
\code{"both"} or \code{"none"} to allow control over which data is overlaid
onto the plots, set to \code{"both"} by default.}

\item{axes_trans}{object of class \code{transformerList} containing the
definitions of the transformations that have been applied to the data.
Transformers will be automatically extracted from \code{GatingSet} objects
but must be supplied manually for \code{cytoset} objects. If no transformer
definitions are extracted or supplied, new transformers of the logicle type
will be generated internally and applied to the data for better
visualisation.}

\item{axes_limits}{options include \code{"auto"}, \code{"trim"},
\code{"data"} or \code{"machine"} to use optimised, lower trimmed (remove
lowest 1% of events), data or machine limits respectively. Set to
\code{"auto"} by default to use optimised axes ranges. Fine control over
axes limits can be obtained by altering the \code{xlim} and \code{ylim}
arguments.}

\item{layout}{a vector of the length 2 of form \code{c(#rows, #columns)} or a
matrix indicating the dimensions of the grid for plotting.}

\item{header}{a vector of text to include at the top of each plot layout, set
to the names of each single colour control by default.}

\item{point_col}{vector of colours to use for layers in scatter plots, set to
\code{c("magenta", "blue", "grey40")} by default.}

\item{point_col_alpha}{numeric [0,1] to control point colour transparency in
2-D scatter plots, set to \code{0.5} by default.}

\item{hist_fill}{vector of colours to use for layers in 1-D histograms, by
default matches \code{point_col} for consistency.}

\item{hist_fill_alpha}{numeric [0,1] used to control histogram fill colour
transparency, set to \code{0.5} by default.}

\item{lines}{logical indicating whether a robust linear model should be
fitted to the data and added to the plots, set to TRUE by default.}

\item{line_type}{type of line to use for fitted robust linear models, set to
1 by default for solid lines.}

\item{line_width}{width of lines to use for fitted robust linear models, set
to 2 by default.}

\item{line_col}{colour to use for fitted robust linear models, set to
\code{"red"} by default.}

\item{line_col_alpha}{numeric [0,1] to control transparency of fitted robust
linear models, set to 1 by default to use solid colours.}

\item{text}{logical indicating whether the plots should be annotated with
their respective spillover coefficients in \code{spillover}, set to TRUE by
default.}

\item{text_font}{font to use for spillover coefficient text, set to 1 by
default for plain font.}

\item{text_size}{numeric to control the size of the spillover coefficient
text, set to 1.2 by default.}

\item{text_col}{colour to use for spillover coefficient text, set to
\code{"black"} by default.}

\item{text_col_alpha}{numeric [0,1] to control transparency of spillover
coefficient text, set to 1 by default to use solid colours.}

\item{...}{additional arguments passed to \code{\link{cyto_plot}}.}
}
\value{
a list of recorded plots for each single colour control.
}
\description{
Visualise compensation of spillover in all channels
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
