% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_complete}
\alias{cyto_plot_complete}
\title{Save custom plot and reset associated settings}
\usage{
cyto_plot_complete(...)
}
\arguments{
\item{...}{graphical parameters supplied by name to be passed to
\code{\link{cyto_plot_par}} to reset the parameters of the current graphics
device.}
}
\description{
\code{cyto_plot_custom()} signals that a custom plot is ready for saving with
\code{cyto_plot_save()} and also resets any settings that were set by
\code{cyto_plot_custom()}. For this reason it is recommended that
\code{cyto_plot_complete()} be called after any plotting on a graphics device
that has been set up using \code{cyto_plot_custom()}.
}
\examples{
library(CytoExploreRData)

# Load samples into GatingSet
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(gs))
gs <- transform(gs, trans)

# Apply gatingTemplate
gt <- Activation_gatingTemplate
gt_gating(gt, gs)

# Save custom plot
cyto_plot_save("Custom.png",
  height = 8,
  width = 16
)

# Set out plot layout
cyto_plot_layout(c(1, 2))

# Add 2D plot
cyto_plot(gs[[4]],
  parent = "CD4 T Cells",
  alias = "",
  channels = c("Alexa Fluor 647-A", "7-AAD-A"),
  layout = FALSE
)

# Add 1D plot
cyto_plot(gs,
  parent = "CD4 T Cells",
  alias = "",
  channels = "7-AAD-A",
  hist_stack = 0.6,
  layout = FALSE
)

# Signal that the plot is complete
cyto_plot_complete()
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
