% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_contour-methods.R
\docType{methods}
\name{cyto_plot_contour,list-method}
\alias{cyto_plot_contour,list-method}
\title{Contour Lines for cyto_plot - list Method}
\usage{
\S4method{cyto_plot_contour}{list}(x, channels, contour_lines = 15,
  contour_line_type = 1, contour_line_width = 1,
  contour_line_col = "black")
}
\arguments{
\item{x}{a list of \code{flowFrame} objects to use for contour lines to
overlay onto an existing cyto_plot.}

\item{channels}{channels used to construct the existing cyto_plot.}

\item{contour_lines}{numeric indicating the number of levels to use for
contour lines, set to 15 by default.}

\item{contour_line_type}{type of line to use for contour lines, set to 1 by
default.}

\item{contour_line_width}{line width for contour lines, set to 2 by default.}

\item{contour_line_col}{colour to use for contour lines, set to "black" by
default.}
}
\description{
Contour Lines for cyto_plot - list Method
}
\examples{
library(CytoRSuiteData)

# Load in Samples
fs <- Activation

# Apply compensation
fs <- compensate(fs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(fs[[4]], cyto_fluor_channels(fs))
fs <- transform(fs, trans)

# Plot
cyto_plot(fs,
  channels = c("Alexa Fluor 488-A", "Alexa Fluor 700-A"),
  group_by = "all", axes_trans = trans
)

# Contour lines for control sample
cyto_plot_contour(list(fs[[1]], fs[[2]], fs[[3]], fs[[4]]),
  channels = c("Alexa Fluor 488-A", "Alexa Fluor 700-A"),
  contour_lines = c(20, 0, 0, 0)
)
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
