% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_contour.R
\name{cyto_plot_contour}
\alias{cyto_plot_contour}
\title{Add contour lines to cyto_plot}
\usage{
cyto_plot_contour(
  x,
  parent = "root",
  channels,
  overlay = NA,
  select = NULL,
  events = 1,
  contour_bins = 256,
  contour_smooth = 0,
  contour_lines = 15,
  contour_line_type = 1,
  contour_line_width = 1,
  contour_line_col = "black",
  contour_line_alpha = 1,
  seed = 42,
  ...
)
}
\arguments{
\item{x}{either an object of class
\code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the parent population to extract from GatingHierarchy
or GatingSt objects, set to the "root" node by default.}

\item{channels}{channels used to construct the existing cyto_plot.}

\item{overlay}{name(s) of the populations to overlay or a \code{cytoset} or
\code{list of cytosets} containing populations to be overlaid onto the
plot(s). This argument can be set to "children" or "descendants" when a
\code{GatingSet} or \code{GatingHierarchy} to overlay all respective nodes.}

\item{select}{named list containing experimental variables to be used to
select samples using \code{\link{cyto_select}} when a \code{flowSet} or
\code{GatingSet} is supplied. Refer to \code{\link{cyto_select}} for more
details. Sample selection occurs prior to grouping with \code{merge_by}.}

\item{events}{controls the number or percentage of events to display, set to
1 by default to display all events.}

\item{contour_bins}{number of bins to use for contour lines, set to 256 bins
by default.}

\item{contour_lines}{numeric indicating the number of levels to use for
contour lines, set to 15 by default.}

\item{contour_line_type}{type of line to use for contour lines, set to 1 by
default.}

\item{contour_line_width}{line width for contour lines, set to 1 by default.}

\item{contour_line_col}{colour to use for contour lines, set to "black" by
default.}

\item{contour_line_alpha}{numeric [0,1] to control transparency of contour
lines, set to 1 by default to remove transparency.}

\item{seed}{numeric passed to \code{\link{set.seed}} to ensure that the same
sampling is applied with each \code{\link{cyto_plot_contour}} call, set to
an arbitrary numeric by default. This behaviour can be turned off by
setting this argument to NULL.}

\item{...}{not in use.}
}
\description{
Add contour lines to cyto_plot
}
\examples{
library(CytoExploreRData)

# Activation GatingSet
gs <- cyto_load(
  system.file(
    "extdata/Activation-GatingSet",
    package = "CytoExploreRData"
  )
)

# Visualise T Cells
cyto_plot(gs[[1]],
parent = "T Cells",
channels = c("CD4", "CD8"))

# Add contour lines
cyto_plot_contour(gs[[1]],
parent = "T Cells",
channels = c("CD4", "CD8"),
contour_lines = 15)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
