% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_density.R
\name{cyto_plot_density}
\alias{cyto_plot_density}
\alias{cyto_plot_density.flowFrame}
\alias{cyto_plot_density.list}
\title{Add density distributions to empty plots}
\usage{
cyto_plot_density(x, ...)

\method{cyto_plot_density}{flowFrame}(x, channel, overlay = NA,
  display = 1, density_modal = TRUE, density_smooth = 1.5,
  density_stack = 0.5, density_cols = NA, density_fill = NA,
  density_fill_alpha = 1, density_line_type = 1,
  density_line_width = 1, density_line_col = "black")

\method{cyto_plot_density}{list}(x, density_cols = NA,
  density_fill = NA, density_fill_alpha = 1, density_line_type = 1,
  density_line_width = 1, density_line_col = "black", ...)
}
\arguments{
\item{x}{object of class flowFrame or a list of density objects.}

\item{channel}{name of the channels to be used to construct the plot.}

\item{overlay}{list of flowFRame objects to overlay.}

\item{display}{controls the number or percentage of events to display, set to
1 by default to display all events.}

\item{density_modal}{logical indicating whether density should be normalised
to mode and presented as a percentage. Set to \code{TRUE} by default.}

\item{density_smooth}{smoothing parameter passed to
\code{\link[stats:density]{density}} to adjust kernel density.}

\item{density_stack}{numeric [0,1] indicating the degree of offset for
overlaid populations, set to 0.5 by default.}

\item{density_cols}{vector of colls to use to select density_fill colours}

\item{density_fill}{colour(s) used to fill polygons.}

\item{density_fill_alpha}{numeric [0,1] used to control fill transparency,
set to 1 by default to remove transparency.}

\item{density_line_type}{line type(s) to use for border(s), set to solid
lines by default.}

\item{density_line_width}{line width for border.}

\item{density_line_col}{colour(s) for border line, set to "black" by default.}

\item{x}{list of density objects for plotting}
}
\description{
Add density distributions to empty plots
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
