% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_explore.R
\name{cyto_plot_explore}
\alias{cyto_plot_explore}
\title{Explore cytometry data in a series of bivariate scatterplots}
\usage{
cyto_plot_explore(
  x,
  parent = "root",
  channels_x = NULL,
  channels_y = NULL,
  select = NULL,
  merge_by = "name",
  overlay = NA,
  order = "channels",
  layout,
  axes_trans = NA,
  header,
  title,
  page_fill = "white",
  page_fill_alpha = 1,
  ylim = c(NA, NA),
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the population to plot when a \code{GatingHierarchy} or
\code{GatingSet} object is supplied, set to the \code{"root"} node by
default.}

\item{channels_x}{vector of channels or markers to plot on the x axis of the
plots, by default we use channels returned by
\code{\link{cyto_fluor_channels}}. If \code{channel} metadata variable is
present each sample with be plotted with the specified channel on the x
axis.}

\item{channels_y}{vector of channels or markers to plot on the y axis of the
plots, by default we use all channels as returned by
\code{\link{cyto_channels}}. \code{channels_y} overlapping with
\code{channels_x} will be plotted as histograms.}

\item{select}{named list containing experimental variables to be used to
select samples using \code{\link{cyto_select}} when a \code{flowSet} or
\code{GatingSet} is supplied. Refer to \code{\link{cyto_select}} for more
details. Sample selection occurs prior to grouping with \code{merge_by}.}

\item{merge_by}{a vector of pData variables to sort and merge samples into
groups prior to plotting, set to NA by default to prevent merging. To merge
all samples set this argument to \code{TRUE} or \code{"all"}.}

\item{overlay}{name(s) of the populations to overlay or a \code{cytoset},
\code{list of cytosets} or \code{list of cytoset lists} containing
populations to be overlaid onto the plot(s). This argument can be set to
"children" or "descendants" when a \code{GatingSet} or
\code{GatingHierarchy} to overlay all respective nodes.}

\item{order}{can be either \code{"channels"} or \code{"groups"} to control
the order in which the data is plotted. Setting \code{order = "channels"}
will plot each group on a single page with a plot for each channel in
\code{channels_y}. On the other hand, setting \code{order = "groups"} will
plot each channel on a separate page with a plot for each group in
\code{merge_by}.}

\item{layout}{a vector of the length 2 of form \code{c(#rows, #columns)} or a
matrix indicating the dimensions of the grid for plotting.}

\item{axes_trans}{object of class
\code{\link[flowWorkspace:transformerList]{transformerList}} which was used
to transform the channels of the supplied data. \code{cyto_plot} does not
support in-line transformations and as such the transformations should be
applied to the data prior to plotting. The transformerList is used
internally to ensure that the axes on the constructed plots are
appropriately labelled.}

\item{header}{can be set to NA to remove headers from plots. Alternatively,
users can supply a vector of headers for each page, this can be tricky to
get right so it is recommended for users to use the default headers where
possible.}

\item{title}{text to display above each plot, set to the name of the group or
channels displayed in the plot by default.}

\item{page_fill}{colour to use to fill the page prior to adding plot panels,
set to \code{"white"} by default.}

\item{page_fill_alpha}{numeric [0,1] to control the transparency of the page
colour, set to 1 by default to remove transparency.}

\item{ylim}{lower and upper limits of y axes (e.g. c(0,250000)).}

\item{...}{additional arguments passed to \code{\link{cyto_plot}}.}
}
\value{
list of recorded plots.
}
\description{
Explore cytometry data in a series of bivariate scatterplots
}
\examples{
\dontrun{
library(CytoExploreRData)

# Activation GatingSet
gs <- cyto_load(
  system.file(
    "extdata/Activation-GatingSet",
    package = "CytoExploreRData"
  )
)

# Plot a sample in all relevant channels
cyto_plot_explore(
  gs,
  parent = "Live Cells",
  select = 32,
  order = "channels",
  channels_x = c("Va2", "CD4", "CD8", "CD44", "CD69", "CD11c"),
  channels_y = c("Va2", "CD4", "CD8", "CD44", "CD69", "CD11c"),
)

# Compare differences between samples in all relevant channels
cyto_plot_explore(
  gs,
  parent = "Live Cells",
  select = 32,
  order = "groups",
  channels_x = c("Va2", "CD4", "CD8", "CD44", "CD69", "CD11c"),
  channels_y = c("Va2", "CD4", "CD8", "CD44", "CD69", "CD11c"),
)
}

}
\seealso{
\code{\link{cyto_plot}}

\code{\link{cyto_plot_profile}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
