% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_gate-methods.R
\docType{methods}
\name{cyto_plot_gate,polygonGate-method}
\alias{cyto_plot_gate,polygonGate-method}
\title{Plot polygonGate Objects onto an Existing Plot}
\usage{
\S4method{cyto_plot_gate}{polygonGate}(x, channels, gate_line_type = 1,
  gate_line_width = 2.5, gate_line_col = "red", gate_point = FALSE,
  gate_point_shape = 16, gate_point_size = 1)
}
\arguments{
\item{x}{an object of class
\code{\link[flowCore:polygonGate-class]{polygonGate}}.}

\item{channels}{fluorescent channels to used to construct the plot.}

\item{gate_line_type}{integer [0,6] which controls the line type, set to
\code{1} to draw solid lines by default.}

\item{gate_line_width}{numeric to adjust line thickness of gates, set to
\code{2.5} by default.}

\item{gate_line_col}{indicates the colour of the gate to be constructed, set
to \code{"red"} by default.}

\item{gate_point}{logical indicating whether points should be included when
plotting the gates, set to \code{FALSE} by default.}

\item{gate_point_shape}{integer [0,25] passed to pch to control the shape of
the points, set to \code{16} to draw filled circles by default. For other
shapes refer to \code{\link[graphics:points]{?pch}}.}

\item{gate_point_size}{numeric character expansion to control the size of the
points in the drawn gate, set to \code{1} by default.}
}
\value{
gate object with modified co-ordinates for plotting.
}
\description{
Plot polygonGate Objects onto an Existing Plot
}
\examples{
library(CytoRSuiteData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(gs))
gs <- transform(gs, trans)

# Gate using gate_draw
gating(Activation_gatingTemplate, gs)

# Plot
cyto_plot(gs[[4]],
  parent = "root",
  channels = c("FSC-A", "SSC-A")
)

# Cells gate
cyto_plot_gate(getGate(gs, "Cells")[[1]],
  channels = c("FSC-A", "SSC-A"),
  gate_line_col = "purple"
)
}
\seealso{
\code{\link{cyto_plot_gate,rectangleGate-method}}

\code{\link{cyto_plot_gate,ellipsoidGate-method}}

\code{\link{cyto_plot_gate,list-method}}

\code{\link{cyto_plot_gate,filters-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
