% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_gate3.R
\name{cyto_plot_gate3}
\alias{cyto_plot_gate3}
\alias{cyto_plot_gate3.default}
\alias{cyto_plot_gate3.rectangleGate}
\alias{cyto_plot_gate3.polygonGate}
\alias{cyto_plot_gate3.ellipsoidGate}
\alias{cyto_plot_gate3.quadGate}
\alias{cyto_plot_gate3.filters}
\alias{cyto_plot_gate3.list}
\title{Plot Gate Objects onto an Existing cyto_plot}
\usage{
\method{cyto_plot_gate3}{default}(x, ...)

\method{cyto_plot_gate3}{rectangleGate}(gate, channels,
  gate_line_type = 1, gate_line_width = 2.5, gate_line_col = "red",
  gate_fill = "white", gate_fill_alpha = 0)

\method{cyto_plot_gate3}{polygonGate}(gate, channels, gate_line_type = 1,
  gate_line_width = 2.5, gate_line_col = "red", gate_fill = "white",
  gate_fill_alpha = 0)

\method{cyto_plot_gate3}{ellipsoidGate}(gate, channels,
  gate_line_type = 1, gate_line_width = 2.5, gate_line_col = "red",
  gate_fill = "white", gate_fill_alpha = 0)

\method{cyto_plot_gate3}{quadGate}(gate, channels, gate_line_type = 1,
  gate_line_width = 2.5, gate_line_col = "red", gate_fill = "white",
  gate_fill_alpha = 0)

\method{cyto_plot_gate3}{filters}(gate, channels, gate_line_type = 1,
  gate_line_width = 2.5, gate_line_col = "red", gate_fill = "white",
  gate_fill_alpha = 0)

\method{cyto_plot_gate3}{list}(gate, channels, gate_line_type = 1,
  gate_line_width = 2.5, gate_line_col = "red", gate_fill = "white",
  gate_fill_alpha = 0)
}
\arguments{
\item{x}{gate object of class
\code{\link[flowCore:rectangleGate-class]{rectangleGate}},
\code{\link[flowCore:polygonGate-class]{polygonGate}},
\code{\link[flowCore:ellipsoidGate-class]{ellipsoidGate}}, \code{list} or
\code{\link[flowCore:filters-class]{filters}}.}

\item{channels}{fluorescent channels to used to construct the plot.}

\item{gate_line_type}{integer [0,6] which controls the line type, set to
\code{1} to draw solid lines by default.}

\item{gate_line_width}{numeric to adjust line thickness of gates, set to
\code{2.5} by default.}

\item{gate_line_col}{indicates the colour of the gate to be constructed, set
to \code{"red"} by default.}

\item{gate_fill}{fill colour(s) to use for gates, set to "white" by default.}

\item{gate_fill_alpha}{numeric [0,1] to control gate fill colour
transparency, set to 0 by default to make gate fill colour completely
transparent.}
}
\value{
invisibly return modified gate objects with dimensions appropriate
  for the constructed plot.
}
\description{
Plot Gate Objects onto an Existing cyto_plot
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
