% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_gating_scheme-methods.R
\docType{methods}
\name{cyto_plot_gating_scheme,GatingHierarchy-method}
\alias{cyto_plot_gating_scheme,GatingHierarchy-method}
\title{Plot Cytometry Gating Strategies - GatingHierarchy Method}
\usage{
\S4method{cyto_plot_gating_scheme}{GatingHierarchy}(x,
  gatingTemplate = NULL, back_gate = FALSE, gate_track = FALSE,
  show_all = FALSE, header, title, popup = FALSE, layout, point_col,
  density_stack = 0.5, density_fill, gate_line_col,
  border_line_col = NA, border_line_width, legend = TRUE, legend_text,
  legend_text_size = 1.2, title_text_col = NA, label_text_size = 0.8,
  ...)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}}.}

\item{gatingTemplate}{name of the gatingTemplate csv file used to gate
\code{x}. If not supplied the gating scheme will be obtained directly from
the GatingSet.}

\item{back_gate}{names of the population(s) to back-gate, set to \code{FALSE}
by default to turn off back-gating. To back-gate all populations set this
argument to \code{"all"}.}

\item{gate_track}{logical indicating whether gate colour should be tracked
throughout gating scheme, set to TRUE by default.}

\item{show_all}{logical indicating whether every population should be
included in every plot in the gating scheme, set to \code{FALSE} by
default.}

\item{header}{character string to use as the header for the plot layout, set
to "Gating Scheme" by default.}

\item{title}{vector of titles to use above each plot.}

\item{popup}{logical indicating whether the gating scheme should be plotted
in a pop-up window, set to FALSE by default.}

\item{layout}{a vector of the length 2 indicating the dimensions of the grid
for plotting \code{c(#rows, #columns)}.}

\item{point_col}{colour of points in 2D plots set to NA to use default
red-blue colour scale. Control the colour of overlays by supplying multiple
colours to this argument (e.g. c("blue","red")).}

\item{density_stack}{numeric [0,1] indicating the degree of offset for 1-D
density distributions with overlay, set to 0.5 by default.}

\item{density_fill}{fill colour for 1D density distributions. Control the
colour of overlays by supplying multiple colours to this argument (e.g.
c(NA,"red")).}

\item{gate_line_col}{vector of colours to use for gates. Individual gate
colours can only be controlled when \code{gate_track} is set to TRUE.}

\item{border_line_col}{line colour for plot border, set to "black" by
default.}

\item{border_line_width}{line width for plot border, set to 3 when gate_track
is TRUE.}

\item{legend}{logical indicating whether a legend should be included when an
overlay is supplied.}

\item{legend_text}{vector of character strings to use for legend when an
overlay is supplied.}

\item{legend_text_size}{character expansion for legend text, set to 1.2 by
default.}

\item{title_text_col}{colour for plot title.}

\item{label_text_size}{numeric to control the size of text in the plot
labels, set to 0.8 by default.}

\item{...}{extra arguments passed to cyto_plot, see
\code{\link{cyto_plot,flowFrame-method}} for more details.}
}
\description{
\code{cyto_plot_gating_scheme} automatically plots the entire gating scheme
and has full support for gate tracking and back-gating through
\code{gate_track} and \code{back_gate}.
}
\examples{
library(CytoRSuiteData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(gs))
gs <- transform(gs, trans)

# Gate using gate_draw
gt <- Activation_gatingTemplate
gating(gt, gs)

# Gating scheme
cyto_plot_gating_scheme(gs[[4]])

# Back-gating
cyto_plot_gating_scheme(gs[[4]],
  back_gate = TRUE
)

# Gate-tracking
cyto_plot_gating_scheme(gs[[4]],
  gate_track = TRUE
)
}
\seealso{
\code{\link{cyto_plot_gating_scheme,GatingHierarchy-method}}

\code{\link{cyto_plot,flowFrame-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
