% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_gating_scheme-methods.R
\name{cyto_plot_gating_scheme}
\alias{cyto_plot_gating_scheme}
\title{Plot Cytometry Gating Strategies}
\usage{
cyto_plot_gating_scheme(x, ...)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{...}{additional method specific arguments.}
}
\description{
\code{cyto_plot_gating_scheme} automatically plots the entire gating scheme
and has full support for gate tracking and back-gating through
\code{gate_track} and \code{back_gate}.
}
\seealso{
\code{\link{cyto_plot_gating_scheme,GatingHierarchy-method}}

\code{\link{cyto_plot_gating_scheme,GatingSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
