% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_gating_tree.R
\name{cyto_plot_gating_tree}
\alias{cyto_plot_gating_tree}
\alias{cyto_plot_gating_tree.GatingSet}
\alias{cyto_plot_gating_tree.GatingHierarchy}
\alias{cyto_plot_gating_tree.gatingTemplate}
\title{Plot Gating Trees}
\usage{
\method{cyto_plot_gating_tree}{GatingSet}(x, ...)

\method{cyto_plot_gating_tree}{GatingHierarchy}(x, stat = NULL, ...)

\method{cyto_plot_gating_tree}{gatingTemplate}(x, ...)
}
\description{
Plot Gating Trees
}
