% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_grid-methods.R
\docType{methods}
\name{cyto_plot_grid,flowSet-method}
\alias{cyto_plot_grid,flowSet-method}
\title{Grid layout for cyto_plot - flowSet Method}
\usage{
\S4method{cyto_plot_grid}{flowSet}(x, channels = NULL,
  axes_trans = NULL, group_by = FALSE, overlay = NULL, gate = NA,
  limits = "machine", display = NULL, layout = NULL, popup = FALSE,
  xlim = NULL, ylim = NULL, title = c(NA, NA), xlab, ylab,
  format = "row", header = NULL, panel_x_label = NULL,
  panel_y_label = NULL, density_modal = TRUE, density_smooth = 1.5,
  density_stack = 0, density_fill = NA, density_fill_alpha = 1,
  density_line_type = 1, density_line_width = 1,
  density_line_col = "black", point_shape = ".", point_col = NA,
  point_alpha = NA, contour_lines = 15, contour_line_type = 1,
  contour_line_width = 1, contour_line_col = "black",
  axes_text_font = 1, axes_text_size = 1, axes_text_col = "black",
  axes_label_text_font = 1, axes_label_text_size = 1.1,
  axes_label_text_col = "black", title_text_font = c(2, 2),
  title_text_size = c(1.2, 1.2), title_text_col = c("black", "black"),
  legend = FALSE, gate_line_type = 1, gate_line_width = 2.5,
  gate_line_col = "red", label = TRUE, label_text = NA,
  label_stat = "percent", label_text_font = 2, label_text_size = 1,
  label_text_col = "black", label_box_x = NA, label_box_y = NA,
  label_box_alpha = 0.6, border_line_type = 1, border_line_width = 1,
  border_line_col = "black", header_text_font = 2,
  header_text_size = 1.5, header_text_col = "black",
  panel_label_font = c(2, 2), panel_label_size = c(1, 1),
  panel_label_col = c("black", "black"), ...)
}
\description{
\code{cyto_plot_grid} is wrapper for cyto_plot which provides a grid layout
 without excess white space. \code{cyto_plot_grid} is particularly useful for
 visualising and facetting data grouped using \code{group_by}.
}
\details{
@param x object of class \code{flowSet}.
 @param ... additional method-specific arguments for \code{cyto_plot}.
 @param format indicates whether to plot in a "row" or "column" when a single
   variable is supplied.

@importFrom flowCore parameters flowFrame
 @importFrom flowWorkspace pData
 @importfrom graphics par mtext text

@author Dillon Hammill (Dillon.Hammill@anu.edu.au)

@export
}
