% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_grid-methods.R
\docType{methods}
\name{cyto_plot_grid,GatingSet-method}
\alias{cyto_plot_grid,GatingSet-method}
\title{Grid layout for cyto_plot - GatingSet Method}
\usage{
\S4method{cyto_plot_grid}{GatingSet}(x, parent, alias, channels, group_by,
  header, title_text_x, title_text_y, ...)
}
\description{
\code{cyto_plot_grid} is wrapper for cyto_plot which provides a grid layout
 without excess white space. \code{cyto_plot_grid} is particularly useful for
 visualising data grouped using \code{group_by}.
 
 @param x object of class \code{GatingSet}.
 @param ... additional method-specific arguments for cyto_plot.
 
 @author Dillon Hammill (Dillon.Hammill@anu.edu.au)
 
 @export
}
