% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_heatmap.R
\name{cyto_plot_heatmap}
\alias{cyto_plot_heatmap}
\title{Plot heatmaps using HeatmapR}
\usage{
cyto_plot_heatmap(x, ...)
}
\arguments{
\item{x}{object of class \code{"matrix"} or \code{"data.frame"}.}

\item{...}{additional arguments passed to \code{HeatmapR::heat_map()}.}
}
\value{
recorded heatmap.
}
\description{
A CytoExploreR wrapper for \code{HeatmapR::heat_map()} to allow visualisation
of matrices or data.frames as heatmaps.
}
\examples{
\dontrun{
  library(CytoExploreRData)
  gs <- GatingSet(Activation)
  spill <- cyto_spillover_extract(gs)[[1]]
  cyto_plot_heatmap(
    spill,
    cell_text = TRUE,
    title = "Spillover Matrix"
  )
}

}
\seealso{
\code{\link[HeatmapR]{heat_map}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
