% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_hist.R
\name{cyto_plot_hist}
\alias{cyto_plot_hist}
\alias{cyto_plot_hist.flowFrame}
\alias{cyto_plot_hist.list}
\title{Add histograms to empty plots}
\usage{
cyto_plot_hist(x, ...)

\method{cyto_plot_hist}{flowFrame}(
  x,
  channel,
  overlay = NA,
  display = 1,
  hist_stat = "percent",
  hist_smooth = 0.6,
  hist_stack = 0.5,
  hist_cols = NA,
  hist_fill = NA,
  hist_fill_alpha = 1,
  hist_line_type = 1,
  hist_line_width = 1,
  hist_line_col = "black",
  ...
)

\method{cyto_plot_hist}{list}(
  x,
  hist_cols = NA,
  hist_fill = NA,
  hist_fill_alpha = 1,
  hist_line_type = 1,
  hist_line_width = 1,
  hist_line_col = "black",
  ...
)
}
\arguments{
\item{x}{list of density objects for plotting}

\item{...}{not in use.}

\item{channel}{name of the channels to be used to construct the plot.}

\item{overlay}{list of flowFRame objects to overlay.}

\item{display}{controls the number or percentage of events to display, set to
1 by default to display all events.}

\item{hist_stat}{can be either \code{"count"}, \code{"percent"} or
\code{"density"} to indicate the statistic to display on histograms, set to
\code{"percent"} by default. The \code{"percent"} option applies modal
normalisation and expresses the result as a percentage.}

\item{hist_smooth}{smoothing parameter passed to
\code{\link[stats:density]{density}} to adjust the smoothness kernel
density, set to \code{0.6} by default.}

\item{hist_stack}{numeric [0,1] indicating the degree of stacking for
histograms, set to \code{0.5} by default.}

\item{hist_cols}{vector colours to draw from when selecting histogram fill
colours if none are supplied to \code{hist_fill}.}

\item{hist_fill}{fill colour(s) for histograms, select from \code{hist_cols}
if not supplied.}

\item{hist_fill_alpha}{numeric [0,1] used to control histogram fill colour
transparency, set to \code{1} by default for solid colours.}

\item{hist_line_type}{line type(s) to use for histogram borders, set to 1 by
default to use solid lines. See \code{\link[graphics:par]{lty}} for
alternatives.}

\item{hist_line_width}{numeric to control line width(s) for histogram borders
lines, set to 1 by default.}

\item{hist_line_col}{colour(s) for histogram borders, set to \code{"black"}
by default.}
}
\description{
Add histograms to empty plots
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
