% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_label-methods.R
\docType{methods}
\name{cyto_plot_label,flowFrame,polygonGate-method}
\alias{cyto_plot_label,flowFrame,polygonGate-method}
\title{Add boxed text labels to cyto_plot - polygonGate Method}
\usage{
\S4method{cyto_plot_label}{flowFrame,polygonGate}(x, gates, channels,
  trans = NULL, text = NA, stat = NA, text_x = NA, text_y = NA,
  text_font = 2, text_size = 0.8, text_col = "black",
  box_alpha = 0.6)
}
\arguments{
\item{x}{a \code{\link[flowCore:flowFrame-class]{flowFrame}} gated in the
existing plot.}

\item{gates}{an object of class
\code{\link[flowCore:polygonGate-class]{polygonGate}}.}

\item{channels}{a vector indicating the fluorescent channel(s) to be used for
gating.}

\item{trans}{object of class
\code{\link[flowCore:transformList-class]{transformList}} or
\code{\link[flowWorkspace]{transformerList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} which was used to
transform the fluorescent channels of the supplied flowFrame.}

\item{text}{the name of the gated population, set to NA by default to only
include percent in labels.}

\item{stat}{indicates the type of statistic to include in the label, can be
either \code{"percent"} or \code{"count"}. \code{stat} is set to
\code{"percent"} by default.}

\item{text_x}{vector containing the x co-ordiantes for the plot labels. Set
to \code{NULL} by default to place labels in the center of the gates.}

\item{text_y}{vector containing the x co-ordiantes for the plot labels. Set
to \code{NULL} by default to place labels in the center of the gates.}

\item{text_font}{integer [1,2,3,4] passed to \code{text} to alter the font,
set to \code{2} by default for a bold font.}

\item{text_size}{numeric character expansion used to control the size of the
text in the labels, set to \code{0.8} by default. See \code{?text} for
details.}

\item{text_col}{specify text colour in label for each gate, defaults to
\code{"black"} for all gates.}

\item{box_alpha}{numeric [0,1] controls the transparency of the background,
set to \code{0.6} by default.}
}
\value{
add a boxed text label to cyto_plot.
}
\description{
\code{cyto_plot_label} takes on a \code{flowFrame} object, population name
\code{text}, \code{channels} and a gate object to construct a text label for
the plot with the population name and frequency.
}
\examples{
library(CytoRSuiteData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(fs))
gs <- transform(gs, trans)

# Gate using gate_draw
gating(Activation_gatingTemplate, gs)

# Plot
cyto_plot(gs[[4]],
  parent = "root",
  channels = c("FSC-A", "SSC-A")
)

# Cells gate
gt <- getGate(gs, "Cells")[[1]]
cyto_plot_gate(gt,
  channels = c("FSC-A", "SSC-A")
)

# Labels
cyto_plot_label(getData(gs, "root")[[4]],
  gates = gt,
  trans = trans,
  channels = c("FSC-A", "SSC-A"),
  text = "Cells",
  stat = "percent",
  text_col = "magenta",
  text_size = 1.2
)
}
\seealso{
\code{\link{cyto_plot_label,flowFrame,rectangleGate-method}}

\code{\link{cyto_plot_label,flowFrame,ellipsoidGate-method}}

\code{\link{cyto_plot_label,flowFrame,list-method}}

\code{\link{cyto_plot_label,flowFrame,filters-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
