% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_label.R
\name{cyto_plot_labeller}
\alias{cyto_plot_labeller}
\title{Add labels to existing cyto_plot}
\usage{
cyto_plot_labeller(label_text = NA, label_text_x = NA,
  label_text_y = NA, label_text_font = 2, label_text_size = 0.8,
  label_text_col = "black", label_fill = "white",
  label_fill_alpha = 0.6)
}
\arguments{
\item{label_text}{character string to include in the label.}

\item{label_text_x}{vector containing the x co-ordinates for the plot labels.
Label positions can be interactively selected if no co-ordinates are
manually supplied.}

\item{label_text_y}{vector containing the x co-ordinates for the plot labels.
Label positions can be interactively selected if no co-ordinates are
manually supplied.}

\item{label_text_font}{integer [1,2,3,4] passed to \code{text} to alter the
font, set to \code{2} by default for a bold font.}

\item{label_text_size}{numeric character expansion used to control the size
of the text in the labels, set to \code{0.8} by default. See \code{?text}
for details.}

\item{label_text_col}{specify text colour in label for each gate, defaults to
\code{"black"} for all gates.}

\item{label_fill}{fill colour to use for labels, set to "white" by default.}

\item{label_fill_alpha}{numeric [0,1] controls the transparency of the fill
colour, set to \code{0.6} by default.}
}
\description{
Convenient labelling function to add prepared text labels to an existing
cyto_plot.
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
