% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_line.R
\name{cyto_plot_line}
\alias{cyto_plot_line}
\title{Add lines with points to empty cyto_plot}
\usage{
cyto_plot_line(
  x,
  line_type = 1,
  line_width = 1,
  line_col = "red",
  line_col_alpha = 1,
  line_fill = NA,
  line_fill_alpha = 1,
  line_point_shape = NA,
  line_point_size = 1,
  line_point_col = NA,
  line_point_col_alpha = NA,
  ...
)
}
\arguments{
\item{x}{a list of matrices or data.frames containing the \code{"x"} and
\code{"y"} co-ordinates of the lines.}

\item{line_type}{integer [1,4] to control the type of each line, set to 1 by
default for solid lines.}

\item{line_width}{numeric to control the width of each line, set to 1 by
default.}

\item{line_col}{colour(s) to use for each line, set to "red" by default.}

\item{line_col_alpha}{numeric [0, 1] to control the transparency of line
colours, set to 1 by default for solid colours.}

\item{line_fill}{colour(s) to use to fill the polygons below each line, set
to NA by default to remove the polygons.}

\item{line_fill_alpha}{numeric [0, 1] to control the transparency of the
polygon fill colour(s), set to 1 by default for solid colour(s).}

\item{line_point_shape}{shape(s) to use for points, set to \code{NA} by
default to remove points from the plot. See \code{\link[graphics:par]{pch}}
for alternatives.}

\item{line_point_size}{numeric to control the size of points per line, set to
1 by default.}

\item{line_point_col}{colour(s) to use for the points in each line, set to
"red" by default.}

\item{line_point_col_alpha}{numeric [0, 1] to control the transparency of
point colours, set to 1 by default for solid colours.}

\item{...}{not in use.}
}
\description{
Add lines with points to empty cyto_plot
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
