% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_new_page}
\alias{cyto_plot_new_page}
\title{Add plots to a new page}
\usage{
cyto_plot_new_page()
}
\description{
\code{cyto_plot_fill()} is a handy function designed to fill incomplete
\code{cyto_plot()} layouts so that the next call to \code{cyto_plot()}
generates a plot on a new page.
}
\examples{
library(CytoExploreRData)
library(flowWorkspace)

cs <- flowSet_to_cytoset(Activation)

# create custom plot layout
cyto_plot_custom(layout = c(2,2))

# add some plots - leave some panels empty
cyto_plot(
  cs[1:3],
  channels = c("FSC-A", "SSC-A")
)

# signal page is complete - use new page for next plots
cyto_plot_new_page()

# add these plots to new page
cyto_plot(
  cs[4:7],
  channels = c("FSC-A", "SSC-A")
)

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
