% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_outer_margins}
\alias{cyto_plot_outer_margins}
\title{Set outer margins for cyto_plot layout}
\usage{
cyto_plot_outer_margins(outer_margins = NULL)
}
\arguments{
\item{outer_margins}{vector of length 4 indicating the number of lines of
text to place around the \code{c(bottom, left, top, right)} outer borders
of the plot, set to NULL by default to use \code{par("oma")}.}
}
\description{
Set outer margins for cyto_plot layout
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
