% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_overlay-methods.R
\docType{methods}
\name{cyto_plot_overlay,list-method}
\alias{cyto_plot_overlay,list-method}
\title{Overlay Points on cyto_plot - list Method}
\usage{
\S4method{cyto_plot_overlay}{list}(x, channels, point_shape = ".",
  point_size = 2, point_col = "black", point_alpha = 1)
}
\arguments{
\item{x}{list of \code{flowFrame} objects to overlay onto an existing
cyto_plot.}

\item{channels}{channels used to construct the existing cyto_plot.}

\item{point_shape}{point character to use for points, set to "." by default
to maximise plotting speed.}

\item{point_size}{numeric specifying the degree of character expansion for
points, set to 2 by default.}

\item{point_col}{colours to use for points, set to \code{"black"} by default.}

\item{point_alpha}{numeric [0,1] used to control colour transparency, set to
1 by default to remove transparency.}
}
\description{
Overlay Points on cyto_plot - list Method
}
\examples{
library(CytoRSuiteData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(gs))
gs <- transform(gs, trans)

# Gate using gate_draw
gating(Activation_gatingTemplate, gs)

# Plot
cyto_plot(gs[[4]],
  parent = "Live Cells",
  channels = c("APC-Cy7-A", "PE-A"),
  axes_trans = trans
)

# Overlay CD4 T Cells & Dendritic Cells
cyto_plot_overlay(list(getData(gs, "CD4 T Cells")[[4]], getData(gs, "Dendritic Cells")[[4]]),
  channels = c("APC-Cy7-A", "PE-A"),
  point_col = c("magenta", "red")
)
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
