% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_overlay-methods.R
\name{cyto_plot_overlay}
\alias{cyto_plot_overlay}
\title{Overlay Points on cyto_plot}
\usage{
cyto_plot_overlay(x, ...)
}
\arguments{
\item{x}{object to overlay onto an existing cyto_plot.}

\item{...}{additional method-specific arguments}
}
\description{
Overlay Points on cyto_plot
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
