% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-internal-helpers.R
\name{.cyto_plot_overlay_format}
\alias{.cyto_plot_overlay_format}
\title{Format overlay for use in cyto_plot}
\usage{
.cyto_plot_overlay_format(x, ...)
}
\arguments{
\item{x}{object of class \code{flowFrame}, \code{flowSet},
\code{GatingHierarchy} or \code{GatingSet}.}

\item{overlay}{object to overlay.}

\item{display}{numeric [0,1] to control the percentage of events to be
plotted. Specifying a value for \code{display} can substantial improve
plotting speed for less powerful machines.}
}
\description{
\code{cyto_plot_overlay_format} will check whether the supplied overlay is
supported and convert it into an appropriate format for use in
\code{\link{cyto_plot}}.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
