% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_par}
\alias{cyto_plot_par}
\title{Extract a list of graphical parameters set by cyto_plot}
\usage{
cyto_plot_par(..., reset = FALSE)
}
\arguments{
\item{...}{additional graphical parameters to be set by the current
\code{cyto_plot()} call (e.g. mfrow = c(1,2)).}

\item{reset}{logical indicating whether these svaed graphical parameters
should be reset, set to FALSE by default.}
}
\value{
invisibly return a list of set graphical parameters.
}
\description{
This function is designed to used within \code{cyto_plot()} to return a list
of graphical parameters that have been set by \code{cyto_plot()}. This is not
intended to be used by the user, but can be used to check if there are any
graphical parameters that have not been properly reset after a call to
\code{cyto_plot()}. If required, these graphical parameters can be reset by
setting \code{reset = TRUE}.
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
