% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_point.R
\name{cyto_plot_point}
\alias{cyto_plot_point}
\alias{cyto_plot_point.flowFrame}
\alias{cyto_plot_point.list}
\title{Add points and contour lines to empty cyto_plot}
\usage{
cyto_plot_point(x, ...)

\method{cyto_plot_point}{flowFrame}(x, channels, overlay = NA,
  display = 1, point_shape = ".", point_size = 2,
  point_col_scale = NA, point_cols = NA, point_col = NA,
  point_col_alpha = 1, contour_lines = 0, contour_line_type = 1,
  contour_line_width = 1, contour_line_col = "black",
  contour_line_alpha = 1, ...)

\method{cyto_plot_point}{list}(x, channels, display = 1,
  point_shape = ".", point_size = 2, point_col_scale = NA,
  point_cols = NA, point_col = NA, point_col_alpha = 1,
  contour_lines = 0, contour_line_type = 1, contour_line_width = 1,
  contour_line_col = "black", contour_line_alpha = 1, ...)
}
\arguments{
\item{x}{object of class \code{flowFrame} or a list of \code{flowFrame}
onjects.}

\item{channels}{names of the channels used to construct the plot.}

\item{overlay}{optional argument if x is a flowFrame to overlay a list of
flowFrame objects.}

\item{display}{controls the number or percentage of events to display, set to
1 by default to display all events.}

\item{point_shape}{shape(s) to use for points in 2-D scatterplots, set to
\code{"."} by default to maximise plotting speed.  See
\code{\link[graphics:par]{pch}} for alternatives.}

\item{point_size}{numeric to control the size of points in 2-D scatter plots
set to 2 by default.}

\item{point_col_scale}{vector of ordered colours to use for the density
colour gradient of points.}

\item{point_cols}{vector colours to draw from when selecting colours for
points if none are supplied to point_col.}

\item{point_col}{colour(s) to use for points in 2-D scatter plots, set to NA
by default to use a blue-red density colour scale.}

\item{point_col_alpha}{numeric [0,1] to control point colour transparency in
2-D scatter plots, set to 1 by default to use solid colours.}

\item{contour_lines}{numeric indicating the number of levels to use for
contour lines, set to 0 by default to exclude contour lines.}

\item{contour_line_type}{type of line to use for contour lines, set to 1 by
default.}

\item{contour_line_width}{line width for contour lines, set to 2 by default.}

\item{contour_line_col}{colour to use for contour lines, set to "black" by
default.}

\item{contour_line_alpha}{numeric [0,1] to control transparency of contour
lines, set to 1 by default to remove transparency.}
}
\description{
Add points and contour lines to empty cyto_plot
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
