% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_profile.R
\docType{methods}
\name{cyto_plot_profile,flowSet-method}
\alias{cyto_plot_profile,flowSet-method}
\title{Plot Expression Profile in All Fluorescent Channels - flowSet Method}
\usage{
\S4method{cyto_plot_profile}{flowSet}(x, channels = NULL,
  group_by = FALSE, axes_trans = NULL, title = NA,
  density_stack = 0.5, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowSet-class]{flowSet}}.}

\item{channels}{a vector channels to use to construct the plots, set to all
channels by default.}

\item{group_by}{a vector of pData variables to sort and merge samples into
groups, set to FALSE by default to prevent merging. To merge all samples
set this argument to "all".}

\item{axes_trans}{object of class
\code{\link[flowCore:transformList-class]{transformList}} or
\code{\link[flowWorkspace]{transformerList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} which was used to
transform the fluorescent channels of the supplied flowFrame. This
transform object will be used internally to ensure axes labels of the plot
are appropriately transformed. The transform object will NOT be applied to
the flowFrame internally and should be applied to the flowFrame prior to
plotting.}

\item{title}{a title for the plots, set to the file name of x by default.}

\item{density_stack}{numeric [0,1] indicating the degree of offset for
overlaid populations, set to 0.5 by default.}

\item{...}{additional arguments passed to
\code{\link{cyto_plot,flowFrame-method}}.}
}
\description{
Plot Expression Profile in All Fluorescent Channels - flowSet Method
}
\examples{
library(CytoRSuiteData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(gs))
gs <- transform(gs, trans)

# Gate using gate_draw
gt <- Activation_gatingTemplate
gating(gt, gs)

# Plot expression profile in all channels
cyto_plot_profile(getData(gs, "T Cells"),
  axes_trans = trans
)
}
\seealso{
\code{\link{cyto_plot,flowFrame-method}}

\code{\link{cyto_plot_profile,flowFrame-method}}

\code{\link{cyto_plot_profile,GatingSet-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
