% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_profile-methods.R
\docType{methods}
\name{cyto_plot_profile,GatingSet-method}
\alias{cyto_plot_profile,GatingSet-method}
\title{Plot Expression Profile in All Fluorescent Channels - GatingSet Method}
\usage{
\S4method{cyto_plot_profile}{GatingSet}(x, parent = NULL,
  channels = NULL, axes_trans = NULL, ...)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the population used to construct the plot.}

\item{channels}{a vector channels to use to construct the plots, set to all
channels by default.}

\item{axes_trans}{object of class
\code{\link[flowCore:transformList-class]{transformList}} or
\code{\link[flowWorkspace]{transformerList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} which was used to
transform the fluorescent channels of the supplied flowFrame. This
transform object will be used internally to ensure axes labels of the plot
are appropriately transformed. The transform object will NOT be applied to
the flowFrame internally and should be applied to the flowFrame prior to
plotting.}

\item{...}{additional arguments passed to
\code{\link{cyto_plot,flowFrame-method}}.}
}
\description{
Plot Expression Profile in All Fluorescent Channels - GatingSet Method
}
\examples{
library(CytoRSuiteData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(gs))
gs <- transform(gs, trans)

# Gate using gate_draw
gt <- Activation_gatingTemplate
gating(gt, gs)

# Plot expression profile in all channels
cyto_plot_profile(gs,
  parent = "T Cells"
)
}
\seealso{
\code{\link{cyto_plot,flowFrame-method}}

\code{\link{cyto_plot_profile,flowFrame-method}}

\code{\link{cyto_plot_profile,flowSet-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
