% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_profile.R
\name{cyto_plot_profile}
\alias{cyto_plot_profile}
\title{Plot expression profile in multiple channels}
\usage{
cyto_plot_profile(
  x,
  parent = "root",
  channels = NULL,
  overlay = NA,
  select = NULL,
  merge_by = "name",
  order = "channels",
  layout,
  hist_layers = NA,
  hist_stack = 0,
  axes_trans = NA,
  header,
  page_fill = "white",
  page_fill_alpha = 1,
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the population to plot when a \code{GatingHierarchy} or
\code{GatingSet} object is supplied, set to the \code{"root"} node by
default.}

\item{channels}{names of the channels or markers in which the data should be
plotted, set to all channels except \code{"Time"} and \code{"Event-ID"} by
default.}

\item{overlay}{name(s) of the populations to overlay or a \code{cytoset},
\code{list of cytosets} or \code{list of cytoset lists} containing
populations to be overlaid onto the plot(s). This argument can be set to
"children" or "descendants" when a \code{GatingSet} or
\code{GatingHierarchy} to overlay all respective nodes.}

\item{select}{named list containing experimental variables to be used to
select samples using \code{\link{cyto_select}} when a \code{flowSet} or
\code{GatingSet} is supplied. Refer to \code{\link{cyto_select}} for more
details. Sample selection occurs prior to grouping with \code{merge_by}.}

\item{merge_by}{a vector of pData variables to sort and merge samples into
groups prior to plotting, set to NA by default to prevent merging. To merge
all samples set this argument to \code{TRUE} or \code{"all"}.}

\item{order}{can be either \code{"channels"} or \code{"groups"} to control
the order in which the data is plotted. Setting \code{order = "channels"}
will plot each group on a single page in all the supplied channels. On the
other hand, setting \code{order = "groups"} will plot each channel on a
separate page for all groups.}

\item{layout}{a vector of the length 2 of form \code{c(#rows, #columns)} or a
matrix indicating the dimensions of the grid for plotting.}

\item{hist_layers}{numeric indicating the number of histograms to stack in
each plot when there are no overlays, set to 1 histogram per plot by
default. Each plot must contain the same number of histograms.}

\item{hist_stack}{numeric [0,1] indicating the degree of stacking for
histograms, set to \code{0} by default.}

\item{axes_trans}{object of class
\code{\link[flowWorkspace:transformerList]{transformerList}} which was used
to transform the channels of the supplied data. \code{cyto_plot} does not
support in-line transformations and as such the transformations should be
applied to the data prior to plotting. The transformerList is used
internally to ensure that the axes on the constructed plots are
appropriately labelled.}

\item{header}{can be set to NA to remove headers from plots. Alternatively,
users can supply a vector of headers for each page, this can be tricky to
get right so it is recommended for users to use the default headers where
possible.}

\item{page_fill}{colour to use to fill the page prior to adding plot panels,
set to \code{"white"} by default.}

\item{page_fill_alpha}{numeric [0,1] to control the transparency of the page
colour, set to 1 by default to remove transparency.}

\item{...}{additional arguments passed to \code{\link{cyto_plot}}.}
}
\value{
list of recorded plots.
}
\description{
Plot expression profile in multiple channels
}
\examples{
\dontrun{
library(CytoExploreRData)

# Activation GatingSet
gs <- cyto_load(
  system.file(
    "extdata/Activation-GatingSet",
    package = "CytoExploreRData"
  )
)

# Expression profile in all channels per sample
cyto_plot_profile(
  gs,
  parent = "Live Cells",
  order = "channels"
)

# Compare expression between samples in each channel
cyto_plot_profile(
  gs,
  parent = "Live Cells",
  order = "groups"
)
}

}
\seealso{
\code{\link{cyto_plot}}

\code{\link{cyto_plot_explore}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
