% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_record}
\alias{cyto_plot_record}
\title{Record an existing cyto_plot}
\usage{
cyto_plot_record()
}
\description{
\code{cyto_plot_record} will record an existing plot such that it can be
saved to an R object for future reference.
}
\examples{
\dontrun{

# Load CytoExploreRData to acces data
library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Construct cyto_plot
cyto_plot(fs[[1]],
  channels = c("FSC-A", "SSC-A")
)

# Record plot and save to object called p
p <- cyto_plot_record()

# Calling p will bring back the recorded plot
p
}

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
