% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_save}
\alias{cyto_plot_save}
\title{Save High Resolution cyto_plot Images}
\usage{
cyto_plot_save(file, width = 7, height = 7, units = "in",
  res = 300, multiple = FALSE, ...)
}
\arguments{
\item{file}{name of the file to which the plot should be saved (including the
file extension). Supported file formats include png, tiff jpeg and pdf.}

\item{width}{numeric indicating the width of exported plot in \code{units},
set to 7 by default for image with width of 7 inches.}

\item{height}{numeric indicating the height of the exported plot in
\code{units}, set to 7 by default for image with height of 7 inches.}

\item{units}{units to be used to set plot size, can be either pixels
(\code{px}), inches (\code{inches}), centimetres (\code{cm}) or millimetres
(\code{mm}). Set to \code{"in"} by default.}

\item{res}{resolution in ppi, set to 300 by default.}

\item{multiple}{logical indicating whether multiple pages should be saved to
separate numbered files, set to \code{TRUE} by default.}

\item{...}{additional arguments for the appropriate \code{png()},
\code{tiff()}, \code{jpeg()} or \code{pdf} graphics devices.}
}
\description{
Save High Resolution cyto_plot Images
}
\examples{
library(CytoExploreRData)

# Load samples into GatingSet
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(gs))
gs <- transform(gs, trans)

# Apply gatingTemplate
gt <- Activation_gatingTemplate
gating(gt, gs)

# Save png image of gating scheme after plotting
cyto_plot_save("Gating-Scheme.png",
  width = 20,
  height = 16
)
cyto_plot_gating_scheme(gs[[1]])

# Save multiple pages to the same pdf file
cyto_plot_save("CD4-T-Cells.pdf",
  height = 8,
  width = 16,
  multiple = TRUE
)
cyto_plot(gs,
  parent = "CD4 T Cells",
  alias = "",
  channels = c("Alexa Fluor 647-A", "7-AAD-A"),
  layout = c(1, 2)
)
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
