% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_theme}
\alias{cyto_plot_theme}
\title{Create custom themes for cyto_plot}
\usage{
cyto_plot_theme(..., reset = FALSE)
}
\arguments{
\item{...}{arguments supported by cyto_plot_theme.}

\item{reset}{logical indicating whether \code{cyto_plot_theme()} settings
should be reset to use default settings, set to FALSE by default.}
}
\description{
\code{cyto_plot_theme} provides an easy way to alter the theme used by
\code{cyto_plot}. By calling \code{cyto_plot_theme} prior to plotting,
subsequent plots will inherit these arguments so there is no need to supply
them manually each time. For a complete list of supported arguments see
\code{cyto_plot_theme_args}.
}
\examples{
# Make all plots have a black background
cyto_plot_theme(border_fill = "black")

# Black ground with custom colour scale for points and purple gates
cyto_plot_theme(
  border_fill = "black",
  point_col_scale = c(
    "cyan",
    "green",
    "yellow",
    "orange",
    "red",
    "darkred"
  ),
  gate_line_col = "magenta"
)

# Reset to default setting
cyto_plot_theme_reset()

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
