% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_theme_args}
\alias{cyto_plot_theme_args}
\title{Get supported cyto_plot_theme arguments}
\usage{
cyto_plot_theme_args()
}
\value{
vector of argument names supported by cyto_plot_theme.
}
\description{
Get supported cyto_plot_theme arguments
}
\examples{
cyto_plot_theme_args()

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
