% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{cyto_plot_theme_colorblind}
\alias{cyto_plot_theme_colorblind}
\title{Colorblind-friendly theme}
\usage{
cyto_plot_theme_colorblind(
  point_shape = "hex",
  point_col_scale = c("#440154", "#31688E", "#35B779", "#FDE725"),
  grid_line_col = "grey85",
  grid_line_alpha = 0.5,
  border_fill = "white",
  border_line_col = "grey40",
  border_line_width = 0.8,
  axes_text_col = "grey20",
  axes_label_text_col = "black",
  axes_ticks_line_col = "grey40",
  title_text_col = "black",
  title_text_font = 2,
  key_text_col = "grey20",
  key_title_text_col = "black",
  key_ticks_line_col = "grey40",
  key_border_line_col = "grey40",
  page_fill = "white",
  contour_line_col = "grey30",
  header_text_col = "black",
  legend_text_col = "black",
  gate_line_col = "#D55E00",
  gate_line_width = 1.2,
  gate_fill_alpha = 0.08,
  gate_fill = "#D55E00",
  label_fill_alpha = 0,
  label_text_col = "#D55E00",
  label_text_font = 2,
  ...
)
}
\arguments{
\item{...}{additional arguments to customize the \code{cyto_plot()} theme,
see \code{cyto_plot_theme_args()} for details.}
}
\description{
A theme using a colorblind-safe palette based on the Viridis color scale.
Ensures plots are accessible to viewers with color vision deficiencies while
maintaining clear visual distinction between density levels.
}
\examples{
\dontrun{
cyto_plot_theme_colorblind()
cyto_plot(gs[[1]], parent = "root", channels = c("FSC-A", "SSC-A"))
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
