% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{cyto_plot_theme_publication}
\alias{cyto_plot_theme_publication}
\title{Publication-ready grayscale theme}
\usage{
cyto_plot_theme_publication(
  point_shape = "hex",
  point_col_scale = c("grey95", "grey70", "grey40", "grey10"),
  grid = FALSE,
  border_fill = "white",
  border_line_col = "black",
  border_line_width = 1,
  axes_text_col = "black",
  axes_text_size = 1,
  axes_text_font = 1,
  axes_label_text_col = "black",
  axes_label_text_size = 1.1,
  axes_label_text_font = 2,
  axes_ticks_line_col = "black",
  axes_ticks_line_width = 1,
  title_text_col = "black",
  title_text_size = 1.2,
  title_text_font = 2,
  key_text_col = "black",
  key_title_text_col = "black",
  key_ticks_line_col = "black",
  key_border_line_col = "black",
  page_fill = "white",
  contour_line_col = "grey30",
  contour_line_width = 0.8,
  header_text_col = "black",
  legend_text_col = "black",
  gate_line_col = "black",
  gate_line_width = 1,
  gate_line_type = 2,
  gate_fill_alpha = 0,
  gate_fill = "black",
  label_fill_alpha = 0,
  label_text_col = "black",
  label_text_font = 2,
  label_text_size = 1,
  ...
)
}
\arguments{
\item{...}{additional arguments to customize the \code{cyto_plot()} theme,
see \code{cyto_plot_theme_args()} for details.}
}
\description{
A grayscale theme optimized for journal publications and print. Uses only
black, white, and grey tones to ensure plots look good in both color and
black-and-white print.
}
\examples{
\dontrun{
cyto_plot_theme_publication()
cyto_plot(gs[[1]], parent = "root", channels = c("FSC-A", "SSC-A"))
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
