% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_theme_reset}
\alias{cyto_plot_theme_reset}
\title{Reset cyto_plot_theme to default settings}
\usage{
cyto_plot_theme_reset()
}
\description{
Reset cyto_plot_theme to default settings
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
