% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_progress}
\alias{cyto_progress}
\title{CytoExploreR progress bar}
\usage{
cyto_progress(pb = NULL, label = NULL, total = 100, clear = FALSE, ...)
}
\arguments{
\item{pb}{an existing progress bar to increment.}

\item{label}{text to include to the left of the progress bar.}

\item{total}{the maximum number of iterations for the progress bar.}

\item{clear}{logical indicating whether the progress bar should be removed
from the console once complete, set to FALSE by default.}

\item{...}{additional arguments passed to \code{progress_bar}.}
}
\value{
progress bar object.
}
\description{
CytoExploreR progress bar
}
\examples{
\dontrun{
  pb <- cyto_progress(total = 10)
  lapply(1:10, function(z) {
    cyto_progress(pb)
    Sys.sleep(0.2)
  })
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
