% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_warp.R
\name{cyto_quantile_norm}
\alias{cyto_quantile_norm}
\title{Apply quantile normalisation to cytometry data}
\usage{
cyto_quantile_norm(
  x,
  parent = "root",
  channels = NULL,
  select = NULL,
  probs = seq(0, 1, 0.01),
  ...
)
}
\arguments{
\item{x}{object of class cytoset, GatingHierachy or GatingSet.}

\item{parent}{name of the parent population to normalise when a
GatingHierarchy or GatingSet is supplied. A new GatingSet will be created
post normalisation if the parent population is not set to \code{"root"}.}

\item{channels}{a vector of channels or markers over which quantile
normalisation should be performed, set to all channels excluding
\code{"Time"}, \code{"Sample|Event-ID"} or any dimension-reduced
parameters.}

\item{select}{sample selection criteria passed to \code{cyto_select()} to
identify the samples containing the reference quantiles to which the
remaining samples should be normalised. Set to Null by default to use the
average quantiles across all samples as the refernce quantiles for every
sample.}

\item{probs}{vector of quantiles over which the data is to be normalised, set
to \code{seq(0, 1, 0.01)} by default to obtain quantiles for every percent
of the data.}

\item{...}{not in use.}
}
\value{
a \code{cytoset}, \code{GatingHierarchy} or \code{GatingSet}
  containing the quantile normalised data.
}
\description{
Apply quantile normalisation to cytometry data
}
\examples{
library(CytoExploreRData)

gs <- GatingSet(Activation)

gs <- cyto_compensate(gs)

gs <- cyto_transform(gs)

gs <- cyto_quantile_norm(gs, channels = c("PE-A", "CD8"))

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
