% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_sample_n}
\alias{cyto_sample_n}
\title{Compute sample number for each cytoframe in a cytoset}
\usage{
cyto_sample_n(x, display = 1, parent = NULL, seed = NULL)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{display}{can be either a numeric [0,1] or integer to indicate the
percentage or number of events to keep respectively.}

\item{parent}{name of the parental population to extract from GatingHierarchy
or GatingSet objects.}

\item{seed}{value used to \code{set.seed()} internally. Setting a value for
seed will return the same result with each run.}
}
\value{
named vector containing the number of events to extract from each
  cytoframe within the cytoset.
}
\description{
This function is used within \code{cyto_coerce} and \code{cyto_plot} to
compute the number of events to extract from each cytoframe in cytoset prior
to coercion.
}
\examples{
# Activation GatingSet
gs <- cyto_load(
 system.file(
   "extdata/Activation-GatingSet",
   package = "CytoExploreRData"
   )
 )

 # SAMPLE SIZE
 cyto_sample_n(gs,
 parent = "T Cells",
 display = 50000)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
