% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_sample_to_node}
\alias{cyto_sample_to_node}
\title{Sample GatingSet to node}
\usage{
cyto_sample_to_node(x, node = NULL, count = NULL)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{node}{name of the gated bead population to use for the calculation. If
not supplied internal checks will be made for populations named "Single
Beads" or "Beads".}

\item{count}{minimum number of events to downsample to, set to the minimum
count for the specified node across samples by default. \code{count} must
be less than or equal to the minimum count across the samples.}
}
\description{
\code{cyto_sample_to_node} can be used to downsample the root node of a
GatingSet based on the number of events in a particular node. A classic
example is downsampling to a bead population to give an indication of the
relative number of events per sample. For example, if Sample A contains 100
beads and 75000 events, and Sample B contains 50 beads and 5000 events.
\code{cyto_sample_to_node} will downsample each sample to contain 50 beads
and therefore 5000 events for Sample A and 3750 events for Sample B.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
