% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_select}
\alias{cyto_select}
\title{Select samples based on experiment variables}
\usage{
cyto_select(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{...}{named list containing experimental variables to be used to select
samples or named arguments containing the levels of the variables to
select. See below examples for use cases. Selected samples can be excluded
by setting \code{exclude} to TRUE.}
}
\value{
\code{cytoset} or \code{GatingSet} restricted to samples which meet
  the designated selection criteria.
}
\description{
Select samples based on experiment variables
}
\examples{
library(CytoExploreRData)

# Activation Gatingset
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                          package = "CytoExploreRData"))

# Look at experiment details
cyto_details(gs)

# Select Stim-C samples with 0 and 500 nM OVA concentrations
gs_stim_c <- cyto_select(gs,
  Treatment = "Stim-C",
  OVAConc = c(0, 500)
)

# Select Stim-A and Stim-C treatment groups
gs_stim_ac <- cyto_select(
  gs,
  list("Treatment" = c("Stim-A", "Stim-C"))
)

# Exclude Stim-D treatment group
gs_stim_abc <- cyto_select(gs,
  Treatment = "Stim-D",
  exclude = TRUE
)
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
