% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_setup}
\alias{cyto_setup}
\title{Load.fcs files into GatingSet and annotate with experiment details}
\usage{
cyto_setup(path = ".", gatingTemplate = NULL, ...)
}
\arguments{
\item{path}{points to the location of the .fcs files to read in (e.g. name of
a folder in current working directory).}

\item{gatingTemplate}{name of a gatingTemplate csv file to be used for gate
saving.}
}
\value{
object of class
  \code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.
}
\description{
Load.fcs files into GatingSet and annotate with experiment details
}
\examples{
# Load in CytoExploreRData to acces data
library(CytoExploreRData)

# Get path to Activation .fcs files in CytoExploreRData
datadir <- system.file("extdata", package = "CytoExploreRData")
path <- paste0(datadir, "/Activation")

# Load in .fcs files into an annotated GatingSet
fs <- cyto_setup(path)

# Markers have been assigned
cyto_extract(gs, "root")[[1]]

# Experiment details have been updated
cyto_details(gs)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
