% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_slot}
\alias{cyto_slot}
\title{Accessory function to extract elements from objects}
\usage{
cyto_slot(x, slot = NULL)
}
\arguments{
\item{x}{an object from which element(s) are to be extracted.}

\item{slot}{can be either the name of a slot to extract or a function that
converts \code{x} to the desired output.}
}
\value{
the element extracted or formatted from \code{x}.
}
\description{
Accessory function to extract elements from objects
}
\examples{
\dontrun{
library(CytoExploreRData)

# Create named list
res <- list("A" = 3, "B" = 4)

# Extract B from list
cyto_slot(res, "B")
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
