% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_som.R
\name{cyto_som}
\alias{cyto_som}
\title{Train a self-organising map (SOM) on cytometry data}
\usage{
cyto_som(
  x,
  select = NULL,
  parent = "root",
  channels = NULL,
  grid = c(14, 14),
  rlen = 10,
  alpha = c(0.05, 0.01),
  events = 1,
  map = "t-SNE",
  trans = NA,
  som = NULL,
  plot = TRUE,
  gatingTemplate = NULL,
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{select}{passed to \code{cyto_select()} to control which samples are
passed to the SOM, set to all samples by default.}

\item{parent}{name of the parent population upon which the SOM should be
trained, set to the \code{"root"} node by default.}

\item{channels}{name(s) of the channel(s) or marker(s) to use when training
the SOM, no default set so channels must be supplied manually.}

\item{grid}{vector of length 2 indicating the dimensions of the SOM grid, set
to \code{c(14,14)} by default.}

\item{rlen}{number of time the data should be presented to the network, set
to 10 by default.}

\item{alpha}{learning rate, default is to decline linearly from 0.05 to 0.01
over \code{rlen} updates.}

\item{events}{indicates the number or proportion of events to use when
training the SOM, set to 1 by default to use all available events.}

\item{map}{passed to the \code{type} argument of cyto_map() to perform
dimensionality reduction on the SOM code vectors, set to \code{"t-SNE"} by
default. Alternatively, a vector of channel names indicating the
dimension-reduced channels in a supplied \code{som} that should be
inherited by the new SOM. Refer to \code{\link{cyto_map}} or details.}

\item{trans}{object of class \code{transformerList} containing the
transformers applied to the supplied data, only required for objects of
class \code{cytoset}.}

\item{som}{a \code{cytoset}, \code{GatingHierarchy} or \code{GatingSet}
containing a pre-computed SOM upon which the supplied data should be
mapped. If the parameters defined by \code{map} are located, the existing
dimension-reduced co-ordinates will be transferred to the new SOM. The data
will be mapped to the existing SOM by computing the distance to the nearest
node using the parameters defined by \code{channels}.}

\item{plot}{logical indicating whether the constructed SOM should be plotted
in the channels produced by \code{map}, set to TRUE by default.}

\item{gatingTemplate}{name of a gatingTemplate csv file to set as the active
\code{gatingTemplate} to store downstream gates on the constructed SOM, set
to NULL by default to bypass \code{gatingTemplate} assignment.}

\item{...}{additional arguments passed to \code{cyto_map()} for
dimensionality reduction.}

\item{radius}{radius of the neighbourhood for SOM training, set to start with
a value that covers two thirds of all unit-to-unit distances.}
}
\value{
a cytoset or GatingSet containing the code vectors for every sample
  with associated counts and dimension-reduced co-ordinates.
}
\description{
\code{cyto_som()} provides an intuitive way to map data stored in a
\code{cytoset}, \code{GatingHierarchy} or \code{GatingSet} into a compressed
self-organising map (SOM). The SOM is trained on the combined data from the
population specified by \code{parent} in the required \code{channels} using
the requested proportion or number of \code{events}. The codebook vectors
generated from the trained SOM are then passed to \code{cyto_map()} to obtain
dimension-reduced co-ordinates using the method specified by \code{map}. The
codes and dimension-reduced co-ordinates from the the trained SOM are stored
in a new cytoframe for each sample with the number of events assigned to each
node being stored in a new channel called \code{"SOM_counts"}. Data can also
be mapped to an existing SOM trained by \code{cyto_som()} if a SOM
\code{cytoset} or \code{GatingSet} is passed to \code{som}.
}
\examples{
\dontrun{
library(CytoExploreRData)

# Activation GatingSet
gs <- GatingSet(Activation)
gs <- cyto_compensate(gs)
gs <- cyto_transform(gs)
gs <- cyto_gatingTemplate_apply(gs, Activation_gatingTemplate)

# Build SOM
gs_som <- cyto_som(
  gs,
  select = 1:32,
  parent = "Live Cells",
  channels = c(
    "FSC-A",
    "SSC-A",
    cyto_markers(gs)
  ),
  grid = c(8,8)
)

# Visualise SOMs
cyto_plot_som(
  gs,
  parent = "Live Cells",
  channels = c("t-SNE-1", "t-SNE-2")
)

# Visualise SOM maps
cyto_plot_som_map(
  gs,
  select = c(1, 32),
  parent = "Live Cells",
  channels = c("t-SNE-1", "t-SNE-2"),
  point_col = c(
    "FSC-A",
    "SSC-A",
    cyto_markers(gs)
  )
)
}

}
\seealso{
\code{\link{cyto_map}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
