% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_som.R
\name{cyto_som_check}
\alias{cyto_som_check}
\title{Check if an object contains a Self-Organising Map (SOM)}
\usage{
cyto_som_check(x)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}
}
\value{
logical indicating whether the supplied object contains a SOM.
}
\description{
Check if an object contains a Self-Organising Map (SOM)
}
\examples{
\dontrun{
  library(CytoExploreRData)
  cyto_som_check(Activation)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
