% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_sort}
\alias{cyto_sort}
\title{Sort a flowSet or GatingSet by experiment variables}
\usage{
cyto_sort(x, sort_by = NULL)
}
\arguments{
\item{x}{object of class \code{flowSet} or \code{GatingSet}.}

\item{sort_by}{names of cyto_details variables to use for sorting. For more
control over the sorting order, specify the factor levels for each variable
in a list (e.g. list(Treatment = c("Stim-A","Stim-C","Stim-B", "Stim-D"))).}
}
\value{
\code{flowSet} or \code{GatingSet} object sorted based on experiment
  variables.
}
\description{
Sort a flowSet or GatingSet by experiment variables
}
\examples{
library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Sort by Treatment and OVAConc
fs <- cyto_sort(fs,
list(Treatment = c("Stim-A", "Stim-B", "Stim-C", "Stim-D"),
OVAConc = c(0, 5, 50, 500)))

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
