% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_spectra_compare.R
\name{cyto_spectra_compare}
\alias{cyto_spectra_compare}
\title{Compute spectral similarity between spectra}
\usage{
cyto_spectra_compare(x, select = NULL, save_as = NULL, heatmap = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{"matrix"} or list of matrix objects with each
row representing a uniquely named spectrum and each column representing a
detector or channel.}

\item{select}{vector of row to select from each matrix by name, set to NULL
by default to compare all rows.}

\item{save_as}{name of a CSV file to which the similarity matrix should be
saved, set to NULL by default to bypass saving.}

\item{heatmap}{logical indicating whether the computed similarity scores
should be displayed in a heatmap, set to to TRUE by default.}

\item{...}{additional arguments passed to \code{HeatmapR::heat_map()} to
cusomise the displayed heatmap.}
}
\value{
computed similarity matrix, heatmap and exported CSV file if
  requested.
}
\description{
Compute spectral similarity between spectra
}
\seealso{
\code{\link{cyto_plot_spectra}}

\code{\link{cyto_unmix_compute}}

\code{\link{cyto_spillover_compute}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
