% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_spillover_edit2.R
\name{cyto_spillover_edit2}
\alias{cyto_spillover_edit2}
\alias{cyto_spillover_edit2.flowSet}
\alias{cyto_spillover_edit2.GatingSet}
\title{Interactively Edit Spillover Matrices in Real-Time}
\usage{
\method{cyto_spillover_edit2}{flowSet}(x, channel_match = NULL,
  spillover = NULL, axes_trans = NULL, point_size = 3,
  axes_text_size = 1.7, axes_label_text_size = 2,
  title_text_size = 2, header_text_size = 1.5, ...)

\method{cyto_spillover_edit2}{GatingSet}(x, parent = NULL,
  channel_match = NULL, spillover = NULL, axes_trans = NULL,
  point_size = 3, axes_text_size = 1.7, axes_label_text_size = 2,
  title_text_size = 2, header_text_size = 1.5, ...)
}
\description{
#' @return save edited spillover matrix to .csv file named
  "Spillover-matrix.csv" or spillover.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
