% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_spillover_extract}
\alias{cyto_spillover_extract}
\title{Extract spillover matrix from cytometry object}
\usage{
cyto_spillover_extract(x)
}
\arguments{
\item{x}{object of class flowFrame, flowSet, GatingHierachy or GatingSet.}
}
\value{
list of spillover matrices or NULL.
}
\description{
Extract spillover matrix from cytometry object
}
\examples{
library(CytoExploreRData)

# Activation GatingSet
gs <- GatingSet(Activation)

# Apply compensation
gs <- cyto_compensate(gs)

# Extract spillover matrices
spill <- cyto_spillover_extract(gs)
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
