% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_spillover_spread_compute.R
\name{cyto_spillover_spread_compute}
\alias{cyto_spillover_spread_compute}
\title{Compute Spillover Spreading Matrix}
\usage{
cyto_spillover_spread_compute(
  x,
  axes_trans = NULL,
  channel_match = NULL,
  compensated = FALSE,
  spillover = NULL,
  spillover_spread = NULL,
  axes_limits = "machine",
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} containing
compensated, transformed and gated compensation controls.}

\item{axes_trans}{object of class
\code{\link[flowWorkspace:transformerList]{transformerList}} generated by a
\code{cyto_transformer} which contains the transformer definitions to be
used internally to transform the fluorescent channels of the samples as
required.}

\item{channel_match}{name of csv file to associate a fluorescent channel with
each of the compensation controls. The \code{channel_match} file should
contain a "name" column with the names of the compensation controls and a
"channel" column to associate a fluorescent channel with each compensation
control. Users need not generate this file by hand as it will be created
following the channel selection process.}

\item{compensated}{logical indicating whether the supplied compensation
controls have been compensated. Compensation controls must be compensated
prior tro calculation of spillover spreading matrix. If \code{compensated}
is set to FALSE, the compensation controls will be compensated internally
using the supplied \code{spillover} matrix.}

\item{spillover}{name of the output spillover matrix csv file to be used
internally to compensate the compensation controls. If no spillover matrix
is supplied \code{spillover_spread_compute}, the spillover matrix attached
to compensation controls will be applied.}

\item{spillover_spread}{name of the csv file to which the spillover spreading
matrix will be saved, set to \code{"date-Spillover-Spread-Matrix.csv"} by
default.}

\item{axes_limits}{options include \code{"auto"}, \code{"data"} or
\code{"machine"} to use optimised, data or machine limits respectively. Set
to \code{"machine"} by default to use entire axes ranges. Fine control over
axes limits can be obtained by altering the \code{xlim} and \code{ylim}
arguments.}

\item{...}{additional arguments passed to \code{\link{cyto_plot}}.}

\item{parent}{name of the population to use for downstream calculations when
a \code{GatingSet} object is supplied, set to the last gated population by
default (e.g. "Single Cells").}
}
\description{
\code{spillover_spread} computes the spillover spreading matrix as described
by Nguyen et al. (2013) using gated and compensated compensation controls.
}
\references{
Nguyen R, Perfetto S, Mahnke YD, Chattopadhyay P & Roederer M,
  (2013). Quantifying spillover spreading for comparing instrument
  performance and aiding in multicolor panel design. Cytometry A,
  83(3):306-15.
}
\seealso{
\code{\link{cyto_spillover_compute}}

\code{\link{cyto_spillover_edit}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
