% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_split}
\alias{cyto_split}
\title{Split samples merged with cyto_merge}
\usage{
cyto_split(x, names = NULL)
}
\arguments{
\item{x}{object of class \code{flowFrame}.}

\item{names}{vector of names to assign to each of the extracted flowFrames
hwne saving the split files.}
}
\value{
list of split flowFrames.
}
\description{
Extract individual samples merged using \code{cyto_merge()} based on
\code{"Sample ID"} column created by \code{cyto_barcode()}.
}
\examples{

# Load CytoExploreRData to access data
library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Merge samples
fr <- cyto_merge_by(fs, "all")[[1]]

# Split merged samples
fr_list <- cyto_split(fr, names = cyto_names(fs))

}
\seealso{
\code{\link{cyto_merge_by}}

\code{\link{cyto_barcode}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
