% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_stats_compute-methods.R
\docType{methods}
\name{cyto_stats_compute,GatingSet-method}
\alias{cyto_stats_compute,GatingSet-method}
\title{Compute Statistics - GatingSet Method}
\usage{
\S4method{cyto_stats_compute}{GatingSet}(x, alias = NULL,
  parent = NULL, channels = NULL, trans = NULL, stat = "median",
  density_smooth = 1.5, save = TRUE)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{alias}{name(s) of the population(s) for which the statistic should be
calculated.}

\item{parent}{name(s) of the parent population(s) used calculate population
frequencies. The frequency of alias in each parent will be returned as a
percentage.}

\item{channels}{names of of channels for which statistic should be
calculated, set to all channels by default.}

\item{trans}{object of class
\code{\link[flowCore:transformList-class]{transformList}} or
\code{\link[flowWorkspace:transformerList]{transformerList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} used to transform
the fluorescent channels of x. The transformation list is required to apply
the inverse transformation such that the required statistics are returned
on the original linear scale.}

\item{stat}{name of the statistic to calculate, options include
\code{"count"}, \code{"freq"}, \code{"median"}, \code{"mode"},
\code{"mean"}, \code{"geo mean"}, \code{"CV"}, or \code{"freq"}.}

\item{density_smooth}{smoothing parameter passed to
\code{\link[stats:density]{density}} when calculating mode, set to 1.5 by
default.}

\item{save}{logical indicating whether statistical results should be saved to
a csv file.}
}
\description{
Calculate and export flow cytometry statistics for a GatingSet.
}
\examples{
library(CytoRSuiteData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(gs))
gs <- transform(gs, trans)

# Gate using gate_draw
gt <- Activation_gatingTemplate
gating(gt, gs)

# Compute statistics
cyto_stats_compute(gs,
  alias = "T Cells",
  channels = c("Alexa Fluor 488-A", "PE-A"),
  stat = "median",
  save = FALSE
)

cyto_stats_compute(gs,
  alias = "T Cells",
  stat = "geo mean",
  save = FALSE
)

cyto_stats_compute(gs,
  alias = c("CD4 T Cells", "CD8 T Cells"),
  stat = "count",
  save = FALSE
)

cyto_stats_compute(gs,
  alias = c("CD4 T Cells", "CD8 T Cells"),
  parent = c("Live Cells", "T Cells"),
  stat = "freq",
  save = FALSE
)
}
\seealso{
\code{\link{cyto_stats_compute,flowFrame-method}}

\code{\link{cyto_stats_compute,flowSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
