% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_transform.R
\name{cyto_transform_arcsinh}
\alias{cyto_transform_arcsinh}
\alias{cyto_transform_arcsinh.flowFrame}
\alias{cyto_transform_arcsinh.flowSet}
\alias{cyto_transform_arcsinh.GatingHierarchy}
\alias{cyto_transform_arcsinh.GatingSet}
\title{Definition(s) of ArcSinh Transformation(s)}
\usage{
cyto_transform_arcsinh(x, ...)

\method{cyto_transform_arcsinh}{flowFrame}(x, channels = NULL,
  raw_max = 262144, width = 4.5, trans_min = 0,
  equal_space = FALSE, breaks = 5, plot = TRUE, popup = FALSE, ...)

\method{cyto_transform_arcsinh}{flowSet}(x, channels = NULL,
  select = NULL, raw_max = 262144, width = 4.5, trans_min = 0,
  equal_space = FALSE, breaks = 5, plot = TRUE, popup = FALSE, ...)

\method{cyto_transform_arcsinh}{GatingHierarchy}(x, channels = NULL,
  parent = "root", raw_max = 262144, width = 4.5, trans_min = 0,
  equal_space = FALSE, breaks = 5, plot = TRUE, popup = FALSE)

\method{cyto_transform_arcsinh}{GatingSet}(x, channels = NULL,
  parent = "root", select = NULL, raw_max = 262144, width = 4.5,
  trans_min = 0, equal_space = FALSE, breaks = 5, plot = TRUE,
  popup = FALSE)
}
\arguments{
\item{x}{an object of class \code{flowFrame}, \code{flowSet},
\code{GatingHierarchy} or \code{GatingSet}.}

\item{...}{not in use.}

\item{channels}{name(s) of channel(s)/marker(s) for which transformation
functions must be generated. Set to all the fluorescent channels by default
if no channels/markers are supplied.}

\item{raw_max}{maximum value of the raw data.}

\item{width}{width of the transformed display in asymptotic decades.}

\item{trans_min}{minimum of transformed data in the display in asymptotic
decades, set to 0 by default.}

\item{equal_space}{logical indicating whether breaks should be equally
spaced, set to FALSE by default.}

\item{breaks}{number of required breaks.}

\item{plot}{logical indicating whether the results of transformations should
be plotted using \code{cyto_plot}.}

\item{popup}{logical indicating whether the plots should be constructed in a
pop-up window.}

\item{select}{list of selection criteria passed to \code{cyto_select} to
select a subset of samples for \code{cyto_plot}.}

\item{parent}{name of the parent population to use for generating the
transformation(s).}
}
\description{
CytoRSuite implementation of \code{flowWorkspace}
\code{\link[flowWorkspace:asinh_Gml2]{asinh_Gml2}} transformation which
always returns a \code{transformerList} object and displays the result of the
transformation(s) using \code{cyto_plot}. To combine different types of
transformations have a look at \code{cyto_transform_combine}.
\code{cyto_transform} should be used to apply the transformations to the
data.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
