% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_transform.R
\name{cyto_transform_asinh}
\alias{cyto_transform_asinh}
\alias{cyto_transform_asinh.flowFrame}
\alias{cyto_transform_asinh.flowSet}
\alias{cyto_transform_asinh.GatingHierarchy}
\alias{cyto_transform_asinh.GatingSet}
\title{Definition(s) of Inverse Hyperbolic Sine Transformation(s)}
\usage{
cyto_transform_asinh(x, ...)

\method{cyto_transform_asinh}{flowFrame}()

\method{cyto_transform_asinh}{flowSet}()

\method{cyto_transform_asinh}{GatingHierarchy}()

\method{cyto_transform_asinh}{GatingSet}()
}
\arguments{
\item{x}{an object of class \code{flowFrame}, \code{flowSet},
\code{GatingHierarchy} or \code{GatingSet}.}

\item{channels}{name(s) of channel(s)/marker(s) for which transformation
functions must be generated. Set to all the fluorescent channels by default
if no channels/markers are supplied.}

\item{parent}{name of the parent population to use for generating the
transformation(s).}

\item{select}{list of selection criteria passed to \code{cyto_select} to
select a subset of samples for generating the transformation(s).}

\item{raw_max}{}

\item{type}{}

\item{breaks}{}
}
\description{
CytoRSuite implementation of \code{flowWorkspace}
\code{\link[flowWorkspace:asinh_Gml2]{asinh_Gml2}} transformation which
always returns a \code{transformerList} object and displays the result of the
transformation(s) using \code{cyto_plot}. To combine different types of
transformations have a look at \code{cyto_transform_combine}.
\code{cyto_transform} should be used to apply the transformations to the
data.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
