% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_transform.R
\name{cyto_transform_combine}
\alias{cyto_transform_combine}
\title{Combine Transformation Definitions}
\usage{
cyto_transform_combine(...)
}
\arguments{
\item{...}{objects of class \code{transformerList} to be combined into a
single \code{transformerList} object.}
}
\description{
\code{cyto_transform_combine} makes it easy to combine transformation
definitions obtained from \code{cyto_transform_arcsinh},
\code{cyto_transform_biex} and/or \code{cyto_transform_logicle} prior to
applying these transformations to the data using \code{cyto_transform}.
}
\seealso{
\code{\link{cyto_transform_arcsinh}}

\code{\link{cyto_transform_biex}}

\code{\link{cyto_transform_logicle}}

\code{\link{cyto_transform}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
