% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_transform_convert}
\alias{cyto_transform_convert}
\alias{cyto_transform_convert.transformList}
\alias{cyto_transform_convert.transformerList}
\title{Convert transformerLists to transformLists and/or extract inverse
transformations}
\usage{
cyto_transform_convert(x, inverse = FALSE)

\method{cyto_transform_convert}{transformList}(x, inverse = FALSE)

\method{cyto_transform_convert}{transformerList}(x, inverse = FALSE)
}
\arguments{
\item{x}{object of class
\code{\link[flowCore:transformList-class]{transformList}} or
\code{\link[flowWorkspace]{transformerList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} which was used to
transform the fluorescent channels of the samples.}

\item{inverse}{logical indicating whether the returned transformList should
contain the inverse transformations.}
}
\value{
A \code{\link[flowCore:transformList-class]{transformList}} containing
  the desired transformations.
}
\description{
Convert transformerLists to transformLists and/or extract inverse
transformations
}
\examples{
library(CytoRSuiteData)

# Load in samples to flowSet
fs <- Activation

# Add fs to GatingSet
gs <- GatingSet(fs)

# Get transformList using estimateLogicle
trans <- estimateLogicle(fs[[32]], cyto_fluor_channels(fs))

# Get transformList containing inverse transformarions
inv <- cyto_transform_convert(trans, inverse = TRUE)

# Convert transformerList into transformList
trans <- estimateLogicle(gs[[32]], cyto_fluor_channels(gs))

# Convert transformerList into inverse transformList
inv <- cyto_transform_convert(trans, inverse = FALSE)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
