% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_transform.R
\name{cyto_transform_hyperlog}
\alias{cyto_transform_hyperlog}
\alias{cyto_transform_hyperlog.flowFrame}
\alias{cyto_transform_hyperlog.flowSet}
\alias{cyto_transform_hyperlog.GatingHierarchy}
\alias{cyto_transform_hyperlog.GatingSet}
\title{Definition(s) of Hyperlog Transformation(s)}
\usage{
cyto_transform_hyperlog(x, ...)

\method{cyto_transform_hyperlog}{flowFrame}()

\method{cyto_transform_hyperlog}{flowSet}()

\method{cyto_transform_hyperlog}{GatingHierarchy}()

\method{cyto_transform_hyperlog}{GatingSet}()
}
\description{
Definition(s) of Hyperlog Transformation(s)
}
