% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_transform.R
\name{cyto_transform_log}
\alias{cyto_transform_log}
\alias{cyto_transform_log.flowFrame}
\alias{cyto_transform_log.flowSet}
\alias{cyto_transform_log.GatingHierarchy}
\alias{cyto_transform_log.GatingSet}
\title{Definition(s) of Log Transformation(s)}
\usage{
cyto_transform_log(x, ...)

\method{cyto_transform_log}{flowFrame}(x, ...)

\method{cyto_transform_log}{flowSet}(x, ...)

\method{cyto_transform_log}{GatingHierarchy}(x, ...)

\method{cyto_transform_log}{GatingSet}(x, ...)
}
\description{
CytoRSuite implementation of \code{flowWorkspace}
\code{\link[flowWorkspace:flowjo_flog]{asinh_Gml2}} transformation which
always returns a \code{transformerList} object and displays the result of the
transformation(s) using \code{cyto_plot}. To combine different types of
transformations have a look at \code{cyto_transform_combine}.
\code{cyto_transform} should be used to apply the transformations to the
data.
}
\seealso{
\code{\link[flowWorkspace:flowjo_flog]{flowjo_flog}}

\code{\link{cyto_transform_arcsinh}}

\code{\link{cyto_transform_biex}}

\code{\link{cyto_transform_logicle}}

\code{\link{cyto_transform_combine}}

\code{\link{cyto_transform}}
}
