% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_transform.R
\name{cyto_transform_logicle}
\alias{cyto_transform_logicle}
\alias{cyto_transform_logicle.flowFrame}
\alias{cyto_transform_logicle.flowSet}
\alias{cyto_transform_logicle.GatingHierarchy}
\alias{cyto_transform_logicle.GatingSet}
\title{Definition(s) of Logicle Transformation(s)}
\usage{
cyto_transform_logicle(x, ...)

\method{cyto_transform_logicle}{flowFrame}(x, channels, width = 4.5,
  raw_max = 262144, trans_min = 0, quantile = 0.05,
  type = "instrument", popup = FALSE, equal_space = FALSE,
  breaks = 6, ...)

\method{cyto_transform_logicle}{flowSet}(x, channels, select = NULL,
  trans_max = 4.5, raw_max = 262144, trans_min = 0,
  quantile = 0.05, type = "instrument", popup = FALSE,
  equal_space = FALSE, breaks = 6, ...)

\method{cyto_transform_logicle}{GatingHierarchy}(x, channels,
  parent = "root", trans_max = 4.5, raw_max = 262144,
  trans_min = 0, quantile = 0.05, type = "instrument",
  popup = FALSE, equal_space = FALSE, breaks = 6, ...)

\method{cyto_transform_logicle}{GatingSet}(x, channels, parent = "root",
  select = NULL, trans_max = 4.5, raw_max = 262144, trans_min = 0,
  quantile = 0.05, type = "instrument", popup = FALSE,
  equal_space = FALSE, breaks = 6, ...)
}
\arguments{
\item{x}{object of class \code{flowFrame}, \code{flowSet},
\code{GatingHierarchy} or \code{GatingSet}.}

\item{...}{additional arguments passed to \code{cyto_plot}.}

\item{channels}{name(s) of channel(s)/marker(s) for which transformation
functions must be generated. Set to all the fluorescent channels by default
if no channels/markers are supplied.}

\item{width}{width of transformed display in asymptotic decades.}

\item{raw_max}{maximum value of the raw data.}

\item{trans_min}{minimum of the transformed data in symptotic decades, set to
0 by default.}

\item{quantile}{quantile of the negative data value (used to calculate
linearisation width (w)).}

\item{type}{indicates whether the "instrument" or "data" limits should be
used as the range for the data.}

\item{popup}{logical indicating whether the plots should be constructed in a
pop-up window.}

\item{equal_space}{logical indicating whether breaks should be equally
spaced, set to FALSE by default.}

\item{breaks}{number of required breaks.}

\item{select}{list of selection criteria passed to \code{cyto_select} to
select a subset of samples for \code{cyto_plot}.}

\item{parent}{name of the parent population to use for generating the
transformation(s).}
}
\value{
transformerList object.
}
\description{
\code{CytoRSuite} implementation of \code{flowCore} logicle transformation
which allows the use of multiple samples to estimate transformation
parameters, always returns a \code{transformerList} object and displays the
result of the transformation(s) using \code{cyto_plot}. To combine different
types of transformations have a look at \code{cyto_transform_combine}.
\code{cyto_transform} should be used to apply the transformations to the
data.
}
\seealso{
\code{\link[flowCore:logicleTransform]{estimateLogicle}}

\code{\link[flowWorkspace:estimateLogicle]{estimateLogicle}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
