% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_transformers.R
\name{cyto_transformer_biex}
\alias{cyto_transformer_biex}
\alias{cyto_transformer_biex.GatingSet}
\alias{cyto_transformer_biex.GatingHierarchy}
\alias{cyto_transformer_biex.flowSet}
\alias{cyto_transformer_biex.flowFrame}
\title{Definition(s) of biexponential transformation(s)}
\usage{
cyto_transformer_biex(x, ...)

\method{cyto_transformer_biex}{GatingSet}(x, channels = NULL,
  parent = "root", select = NULL, plot = TRUE, popup = FALSE, ...)

\method{cyto_transformer_biex}{GatingHierarchy}(x, channels = NULL,
  parent = "root", select = NULL, plot = TRUE, popup = FALSE)

\method{cyto_transformer_biex}{flowSet}(x, channels = NULL,
  select = NULL, plot = TRUE, popup = FALSE, ...)

\method{cyto_transformer_biex}{flowFrame}(x, channels = NULL,
  select = NULL, plot = TRUE, popup = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{flowFrame}, \code{flowSet},
\code{GatingHierarchy} or \code{GatingSet}.}

\item{...}{additional arguments passed to
\code{\link[flowWorkspace:flowJoTrans]{flowJoTrans}}.}

\item{channels}{name(s) of channel(s)/marker(s) for which transformation
functions must be generated. Set to all the fluorescent channels by default
if no channels/markers are supplied.}

\item{parent}{name of the parent population to use for generating the
transformation(s).}

\item{select}{list of selection criteria passed to \code{cyto_select} to
select a subset of samples for \code{cyto_plot}.}

\item{plot}{logical indicating whether the results of transformations should
be plotted using \code{cyto_plot}.}

\item{popup}{logical indicating whether the plots should be constructed in a
pop-up window.}
}
\value{
a \code{transformerList} object.
}
\description{
Definition(s) of biexponential transformation(s)
}
\seealso{
\code{\link[flowWorkspace:flowJoTrans]{flowJoTrans}}

\code{\link[flowWorkspace:flowJo_biexp_trans]{flowJo_biexp_trans}}

\code{\link{cyto_transformer_log}}

\code{\link{cyto_transformer_arcsinh}}

\code{\link{cyto_transformer_logicle}}

\code{\link{cyto_transformer_combine}}

\code{\link{cyto_transform}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
