% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_transformers.R
\name{cyto_transformers_combine}
\alias{cyto_transformers_combine}
\title{Combine cyto_transformer definitions}
\usage{
cyto_transformers_combine(...)
}
\arguments{
\item{...}{objects of class
\code{\link[flowWorkspace:transformerList]{transformerList}} to be combined
into a single \code{transformerList} object.}
}
\description{
\code{cyto_transformers_combine} makes it easy to combine transformation
definitions obtained from \code{\link{cyto_transformers_define}} prior to
applying these transformations to the data using
\code{\link{cyto_transform}}.
}
\seealso{
\code{\link{cyto_transformers_define}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
