% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_transformers.R
\name{cyto_transformers_define}
\alias{cyto_transformers_define}
\title{Define transformers for channels}
\usage{
cyto_transformers_define(
  x,
  parent = "root",
  channels = NULL,
  type = "logicle",
  select = NULL,
  events = 50000,
  probs = 0.05,
  plot = TRUE,
  progress = TRUE,
  ...
)
}
\arguments{
\item{x}{an object of class \code{flowFrame}, \code{flowSet},
\code{GatingHierarchy} or \code{GatingSet}.}

\item{parent}{name of the parent population to use for generating the
transformation(s).}

\item{channels}{name(s) of channel(s)/marker(s) for which transformation
functions must be generated. Set to all the fluorescent channels by default
if no channels/markers are supplied.}

\item{type}{a vector indicating the type of transformations to apply to each
channel, options include \code{"log"}, \code{"asinh"}, \code{"asinh_Gml2"},
\code{"biexponential"} or \code{"logicle"}. To set different transformation
types for each channel, supply a vector of transformation types named with
each of the channels. Users can also define their own custom transformation
functions by supplying a list of transformation and inverse transformation
functions for each channel (e.g. list("PE-A" = list(transform = "asinh",
inverse = "sinh"))). All channels will have the same settings for each
transformation type, multiple calls to \code{cyto_transformers_define()}
can be made to modify these settings per channel. Multiple transformerLists
can be joined together using \code{cyto_transformers_combine()}.}

\item{select}{list of selection criteria passed to \code{cyto_select} to
select a subset of samples for \code{cyto_plot}.}

\item{events}{number or frequency of events to display in plots, set to 50000
events by default. See \code{\link{cyto_plot}} for details.}

\item{probs}{quantile to use when estimating cofactor or w using negative
events, set to 0.05 by default.}

\item{plot}{logical indicating whether the results of transformations should
be plotted using \code{cyto_plot}.}

\item{progress}{logical indicating whether to display the progress bar(s)
when computing transformer definitions, set to TRUE by default.}

\item{...}{additional arguments passed to
\code{\link[flowWorkspace:flowjo_log_trans]{flowjo_log_trans}},
\code{\link[flowWorkspace:asinh_Gml2]{asinh_Gml2}},
\code{\link[flowWorkspace:flowjo_biexp]{flowjo_biexp}} or
\code{\link[flowWorkspace:estimateLogicle]{estimateLogicle}} and
\code{\link{cyto_plot}}.}
}
\value{
a \code{transformerList} object containing the transformation
  definitions for each channel.
}
\description{
\code{cyto_transformers_define()} is a convenient function that allows users
to optimise cytoverse transformer definitions for channels, as
transformations are not applied directly to the supplied data and data
transformations are automatically visualised. The type of transformation to
apply can be controlled through the \code{type} argument and links to
relevant documentation for additional arguments are supplied below. Separate
transformer definitions can be combined into the same transformerList using
\code{\link{cyto_transformers_combine}}. Once optimisied, transformers can be
applied to the data using \code{\link{cyto_transform}}.
}
\seealso{
\code{\link[flowWorkspace:flowjo_log_trans]{flowjo_log_trans}}

\code{\link[flowWorkspace:asinh_Gml2]{asinh_Gml2}}

\code{\link[flowWorkspace:flowjo_biexp]{flowjo_biexp}}

\code{\link[flowCore:logicleTransform]{estimateLogicle}}

\code{\link[flowWorkspace:estimateLogicle]{estimateLogicle}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
