% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_transformers.R
\name{cyto_transformers_extract}
\alias{cyto_transformers_extract}
\title{Extract transformers from a GatingHierarchy or GatingSet}
\usage{
cyto_transformers_extract(x)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}}
or \code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}
}
\value{
transformerList or NA.
}
\description{
Extract transformers from a GatingHierarchy or GatingSet
}
\examples{

library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Activation GatingSet
gs <- GatingSet(fs)

# Apply transformations
gs <- cyto_transform(gs, channels = c("CD4", "CD8"))

# Extract transformations
trans <- cyto_transformers_extract(gs)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
