% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channels-helpers.R
\name{cyto_unmix_channels}
\alias{cyto_unmix_channels}
\title{Extract unmixed channels}
\usage{
cyto_unmix_channels(x)
}
\arguments{
\item{x}{object of class cytoframe, cytoset, GatingHierrachy or GatingSet 
potentially containing unmixed parameters.}
}
\value{
the names of the unmixed parameters or NULL.
}
\description{
Extract unmixed channels
}
\examples{
\dontrun{
cyto_unmix_channels(gs)
}


}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
