% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_vars}
\alias{cyto_vars}
\title{Retrieve the names of the experiment variables}
\usage{
cyto_vars(x)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}
}
\value{
a vector containing the names of the experiment variables.
}
\description{
Retrieve the names of the experiment variables
}
\examples{
library(CytoExploreRData)

# Activation Gatingset
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                          package = "CytoExploreRData"))

# experiment variables
cyto_vars(gs)

}
\seealso{
\code{\link{cyto_details}}

\code{\link{cyto_select}}

\code{\link{cyto_groups}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
