% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_warp.R
\name{cyto_warp}
\alias{cyto_warp}
\title{Apply normalisation algorithms to remove batch effects in cytometry data}
\usage{
cyto_warp(
  x,
  parent = "root",
  channels = NULL,
  type = "quantile",
  select = NULL,
  group_by = "all",
  input = "cytoset",
  target = NULL,
  probs = seq(0, 1, 0.01),
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the parent population to extract for normalisation, set
to the \code{"root"} node by default.}

\item{channels}{names of the channels or markers to be normalised set to all
channels excluding \code{"Time"}, \code{"Event-ID"} and \code{"Sample-ID"}.}

\item{type}{indicates the type of normalisation to perform, native options
include \code{"fdaNorm"} or \code{"warpSet"}, \code{"GaussNorm"},
\code{"quantile"} and \code{"CytoNorm"}. Custom normalisation functions can
also be supplied either by name or as a function.}

\item{select}{selection criteria to restrict the data to a subset of the
samples prior to normalisation, set to \code{NULL} by default to use all
samples.}

\item{group_by}{grouping arguments passed to \code{cyto_group_by()} to split
the samples into groups prior to normalisation, set to \code{"all"} by
default to normalise all samples together.}

\item{input}{indicates how the data should be formatted prior to passing it a
custom normalisation function passed through \code{type}.}

\item{target}{sample selection criteria passed to \code{cyto_select()} within
each group specified by \code{group_by} to identify the reference sample(s)
against which the rest of the samples should be aligned.}

\item{probs}{vector of quantiles over which quantile normalisation should be
performed, set to every percent by default.}

\item{...}{additional arguments passed to the normalisation function
specified by \code{type}.}
}
\value{
a \code{cytoset} or \code{GatingSet} containing the normalised data.
}
\description{
Apply normalisation algorithms to remove batch effects in cytometry data
}
\references{
Finak, G et al. (2013). High throughput flow cytometry data
  normalisation for clinical trials. Cytometry A 85(3).

Van Gassen, S. et al. (2019). CytoNorm: A normalisation algorithm
  for cytometry data. Cytometry A 97(3).
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
