% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoExploreR-package.R
\docType{package}
\name{CytoExploreR-package}
\alias{CytoExploreR}
\alias{CytoExploreR-package}
\title{CytoExploreR: Interactive Analysis of Cytometry Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An intuitive and interactive approach to analysing cytometry data in R. This package has been developed under ROpenSci gudelines to integrate conventional and cutting edge cytometry analysis tools under a unified open-source framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DillonHammill/CytoExploreR}
  \item Report bugs at \url{https://github.com/DillonHammill/CytoExploreR/issues}
}

}
\author{
\strong{Maintainer}: Dillon Hammill \email{Dillon.Hammill@anu.edu.au}

Other contributors:
\itemize{
  \item Greg Finak \email{gfinak@fredhutch.org} [contributor]
  \item Mike Jiang \email{wjiang2@fhcrc.org} [contributor]
}

}
\keyword{internal}
