% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot2-internal.R
\name{.cyto_plot_1d_v2.flowFrame}
\alias{.cyto_plot_1d_v2.flowFrame}
\title{cyto_plot_1d - flowFrame Method}
\usage{
\method{.cyto_plot_1d_v2}{flowFrame}(x, channel = NA, axes_trans = NA,
  overlay = NA, gate = NA, display = 1, limits = "machine",
  popup = FALSE, xlim = NA, ylim = NA, title, xlab, ylab,
  density_modal = TRUE, density_smooth = 1.5, density_stack = 0.5,
  density_cols = NA, density_fill, density_fill_alpha = 1,
  density_line_type = 1, density_line_width = 1,
  density_line_col = "black", axes_text_font = 1, axes_text_size = 1,
  axes_text_col = "black", axes_label_text_font = 1,
  axes_label_text_size = 1.1, axes_label_text_col = "black",
  title_text_font = 2, title_text_size = 1.1,
  title_text_col = "black", legend = FALSE, legend_text,
  legend_text_font = 1, legend_text_size = 1,
  legend_text_col = "black", legend_line_col = NA,
  legend_box_fill = NA, gate_line_type = 1, gate_line_width = 2.5,
  gate_line_col = "red", label = TRUE, label_text = NA,
  label_stat = "percent", label_text_font = 2, label_text_size = 1,
  label_text_col = "black", label_box_x = NA, label_box_y = NA,
  label_box_alpha = 0.6, border_line_type = 1, border_line_width = 1,
  border_line_col = "black", border_fill = "white", ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}}.}

\item{channel}{name of the channel or marker to be used to construct the
plot.}

\item{axes_trans}{object of class
\code{\link[flowCore:transformList-class]{transformList}} or
\code{\link[flowWorkspace]{transformerList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} which was used to
transform the fluorescent channels of the supplied flowFrame. This
transformation object will be used internally to ensure that the axes
labels of the plot are appropriately transformed. The transformation object
will NOT be applied to the flowFrame internally and should be applied to
the flowFrame prior to plotting.}

\item{overlay}{a \code{flowFrame}, \code{flowSet} or list of
\code{flowFrames} to be overlaid onto the plot.}

\item{gate}{gate object(s) to be added to plot.  For \code{cyto_plot_1d} only
gate objects of class
\code{\link[flowCore:rectangleGate-class]{rectangleGate}} in either 1 or 2
dimensions are supported. Mulitple gates can be supplied either as a
\code{list} or \code{\link[flowCore:filters-class]{filters}} object.}

\item{display}{numeric indicating the number of events to plot, set to all
events by default.}

\item{limits}{indicates whether the axes limits should be based on the
\code{"data"} or \code{"machine"}, set to "machine" by default to show
complete axes ranges. This argument will only alter the upper axis limits,
to modify the lower limits use \code{xlim} and \code{ylim}.}

\item{popup}{logical indicating whether the plot should be constructed in a
pop-up window, set to FALSE by default. \code{popup} will open OS-specific
graphic device prior to plotting. Mac users will need to install
\href{https://www.xquartz.org/}{XQuartz} for this functionality.}

\item{xlim}{lower and upper limits of x axis (e.g. c(0,5)).}

\item{ylim}{lower and upper limits of y axis (e.g. c(0,5)).}

\item{title}{title to use for the plot, set to the name of the sample by
default. Title can be removed by setting this argument to \code{NA}.}

\item{xlab}{x axis label.}

\item{ylab}{y axis label.}

\item{density_modal}{logical indicating whether density should be normalised
to mode and presented as a percentage. Set to \code{TRUE} by default.}

\item{density_smooth}{smoothing parameter passed to
\code{\link[stats:density]{density}} to adjust kernel density.}

\item{density_stack}{numeric [0,1] indicating the degree of offset for
overlaid populations, set to 0.5 by default.}

\item{density_cols}{vector colours to draw from when selecting density fill
colours when none are supplied.}

\item{density_fill}{colour(s) used to fill polygons.}

\item{density_fill_alpha}{numeric [0,1] used to control fill transparency,
set to 1 by default to remove transparency.}

\item{density_line_type}{line type(s) to use for border(s), set to solid
lines by default.}

\item{density_line_width}{line width for border.}

\item{density_line_col}{colour(s) for border line, set to "black" by default.}

\item{axes_text_font}{numeric indicating the font to use for axes, set to 1
for plain font by default. See \code{\link[graphics:par]{?par}} font for
details.}

\item{axes_text_size}{character expansion for axis text.}

\item{axes_text_col}{colour of axis text.}

\item{axes_label_text_font}{numeric indicating the font to use for title, set
to 1 for plain font by default. See \code{\link[graphics:par]{?par}} font
for details.}

\item{axes_label_text_size}{character expansion for axis labels.}

\item{axes_label_text_col}{colour of axis labels.}

\item{title_text_font}{numeric indicating the font to use for title, set to 2
for bold font by default. See \code{\link[graphics:par]{?par}} font for
details.}

\item{title_text_size}{character expansion for plot title.}

\item{title_text_col}{colour for plot title.}

\item{legend}{can be either \code{"line"} or \code{"fill"} to indicate
whether a legend should be constructed based on the density \code{"line"}
or \code{"fill"}, set to FALSE by default to remove the legend.}

\item{legend_text}{vector of labels to use for the legend.}

\item{legend_text_font}{numeric indicating the font to use for labels, set to
1 for plain font by default. See \code{\link[graphics:par]{?par}} font for
details.}

\item{legend_text_size}{character expansion for legend text, set to 1 by
default.}

\item{legend_text_col}{colour of text used in legend, set to \code{"black"}
by default.}

\item{legend_line_col}{vector of line colours to use for legend.}

\item{legend_box_fill}{vector of fill colours to use for legend.}

\item{gate_line_type}{integer [0,6] which controls the line type, set to
\code{1} to draw solid lines by default.}

\item{gate_line_width}{numeric to adjust line thickness of gates, set to
\code{2.5} by default.}

\item{gate_line_col}{indicates the colour of the gate to be constructed, set
to \code{"red"} by default.}

\item{label}{logical indicating whether gated populations should be labelled.
If the names of the populations are supplied as the text.labels argument,
the population name and frequency will be included in the labels, otherwise
only the population frequencies will be included in the labels.}

\item{label_text}{vector of population names to use in labels. Set to NA to
exclude population names.}

\item{label_stat}{indicates the type of statistic to include in the label,
can be \code{"percent"}, \code{"count"}, \code{"mean"}, \code{"median"},
\code{"mode"} or \code{"geo mean"}, set to \code{"percent"} for gated data
or \code{NA} to exclude statistics for un-gated data.}

\item{label_text_font}{numeric indicating the font to use for labels, set to
2 for bold font by default. See \code{\link[graphics:par]{?par}} font for
details.}

\item{label_text_size}{character expansion for label text, set to 0.8 by
default.}

\item{label_text_col}{colour of text used in labels, set to \code{"black"} by
default.}

\item{label_box_x}{x co-ordinate to manually adjust the position of the label
on the plot.}

\item{label_box_y}{y co-ordinate(s) to manually adjust the position of the
label(s) on the plot.}

\item{label_box_alpha}{numeric controlling backgropund fill transparency of
label boxes, set to 0.6 by default to introduce some transparency.}

\item{border_line_type}{line type to use for plot border, set to 1 by default
for a sold border.}

\item{border_line_width}{line width for plot border, set to 1 by default.}

\item{border_line_col}{line colour for plot border, set to "black" by
default.}

\item{border_fill}{border_fill fill colour to use inside the plot border
(i.e. background colour), set to "white" by default.}

\item{...}{additional arguments passed to \code{\link[graphics:plot]{plot}}.}
}
\description{
Visualise 1-D flow cytometry density distributions for a flowFrame.
}
\examples{
\dontrun{
library(CytoRSuiteData)
fs <- Activation

cyto_plot_1d(fs[[1]],
  channel = "FSC-A",
  overlay = fs[[2]],
  density_fill = c("red", "blue")
)
}

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
