% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot2-internal.R
\name{.cyto_plot_1d_v2}
\alias{.cyto_plot_1d_v2}
\title{cyto_plot_1d}
\usage{
.cyto_plot_1d_v2(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}} or
\code{\link[flowCore:flowSet-class]{flowSet}}.}

\item{...}{additional method-specific arguments for cyto_plot_1d.}
}
\description{
Visualise 1-D flow cytometry density distributions.
}
\seealso{
\code{\link{cyto_plot_1d,flowFrame-method}}

\code{\link{cyto_plot_1d,flowSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
