% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawGate-methods.R
\docType{methods}
\name{drawGate,flowFrame-method}
\alias{drawGate,flowFrame-method}
\title{drawGate flowFrame Method.}
\usage{
\S4method{drawGate}{flowFrame}(x, channels = NULL, alias = NULL,
  subSample = 250000, gate_type = "polygon", axis = "x", adjust = 1.5,
  plot = TRUE, labs = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{flowFrame}.}

\item{channels}{a vector indicating the fluorescent channel(s) to be used for gating. If a single channel is supplied, a histogram of
of the kernel density will be constructed.}

\item{alias}{the name(s) of the populations to be gated. If multiple population names are supplied (e.g. \code{c("CD3,"CD4)}) multiple 
gates will be returned. \code{alias} is \code{NULL} by default which will halt the gating routine.}

\item{subSample}{numeric indicating the number of events to plot to speed up plotting. If \code{subSample} is greater than the total
number of events, all events will be plotted. \code{subSample} is set to 250000 events by default.}

\item{gate_type}{vector of gate type names used to construct the gates. Multiple \code{gate_types} are supported but should be accompanied with
an \code{alias} argument of the same length (i.e. one \code{gate_type} per \code{alias}). Supported \code{gate_types} are \code{polygon, rectangle,
ellipse, threshold, boundary, interval, quadrant and web} which can be abbreviated as upper or lower case first letters as well. Default \code{gate_type}
is \code{"polygon"}.}

\item{axis}{indicates whether the \code{"x"} or \code{"y"} axis should be gated for 2-D interval gates.}

\item{adjust}{smooothing factor passed to \code{density()} for 1D plots (defaults to 1.5).}

\item{plot}{logical indicating whether the data should be plotted. This feature allows for constructing gates of different types over 
existing plots which may already contain a different gate type. For demonstration of this feature refer to the package vignette.}

\item{labs}{logical indicating whether to include \code{plotLabels} for the gated population(s), \code{TRUE} by default.}

\item{...}{additional arguments for \code{plotDens}.}
}
\description{
drawGate flowFrame Method.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
