% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawGate-methods.R
\docType{methods}
\name{drawGate,flowSet-method}
\alias{drawGate,flowSet-method}
\title{drawGate flowSet Method}
\usage{
\S4method{drawGate}{flowSet}(x, select = NULL, channels = NULL,
  alias = NULL, subSample = 250000, type = "polygon", axis = "x",
  adjust = 1.5, labels = TRUE, plot = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowSet-class]{flowSet}}.}

\item{select}{vector containing the indicies of samples within gs to use for
plotting.}

\item{channels}{vector of channel names to use for plotting, can be of length
1 for 1-D density histogram or length 2 for 2-D scatter plot.}

\item{alias}{the name(s) of the populations to be gated. If multiple
population names are supplied (e.g. \code{c("CD3,"CD4)}) multiple gates
will be returned. \code{alias} is \code{NULL} by default which will halt
the gating routine.}

\item{subSample}{numeric indicating the number of events to plot, set to all
events by default. Reducing the sample size can significantly increase
plotting speed on less powerful machines.}

\item{type}{vector of gate type names used to construct the gates. Multiple
gate types are supported but should be accompanied with an \code{alias}
argument of the same length (i.e. one \code{type} per \code{alias}).
Supported gate types are \code{polygon, rectangle, ellipse, threshold,
boundary, interval, quadrant and web} which can be abbreviated as upper or
lower case first letters as well. Default \code{type} is \code{"interval"}
for 1D gates and \code{"polygon"} for 2D gates.}

\item{axis}{indicates whether the \code{"x"} or \code{"y"} axis should be
gated for 2-D interval gates.}

\item{adjust}{smooothing factor passed to
\code{\link[stats:density]{density}} for 1-D plots (defaults to 1.5).}

\item{labels}{logical indicating whether to include \code{\link{plotLabels}}
for the gated population(s), \code{TRUE} by default.}

\item{plot}{logical indicating whether a plot should be drawn, set to
\code{TRUE} by default.}

\item{...}{additional arguments for \code{\link{plotCyto,flowSet-method}}.}
}
\value{
a \code{\link[flowCore:filters-class]{filters}} list containing the
  drawn gate objects.
}
\description{
Manually draw gates around populations for analysis of flow cytometry data.
}
\seealso{
\code{\link{plotCyto1d,flowSet-method}}

\code{\link{plotCyto2d,flowSet-method}}

\code{\link{drawGate,flowFrame-method}}

\code{\link{drawGate,GatingSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
