% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating-functions.R
\name{drawQuadrants}
\alias{drawQuadrants}
\title{Draw Quadrants to Gate Multiple Flow Cytometry Populations.}
\usage{
drawQuadrants(fr, channels, alias = NULL, plot = TRUE, labels = TRUE,
  ...)
}
\arguments{
\item{fr}{a \code{\link[flowCore:flowFrame-class]{flowFrame}} object
containing the flow cytometry data for plotting and gating.}

\item{channels}{vector of channel names to use for plotting, can be of length
1 for 1-D density histogram or length 2 for 2-D scatter plot.}

\item{alias}{the name(s) of the 4 populations to be gated. \code{alias} is
\code{NULL} by default which will halt the gating routine.}

\item{plot}{logical indicating whether the data should be plotted. This
feature allows for constructing gates of different types over existing
plots which may already contain a different gate type.}

\item{labels}{logical indicating whether to include \code{\link{plotLabels}}
for the gated population(s), \code{TRUE} by default.}

\item{...}{additional arguments for \code{\link{plotCyto,flowFrame-method}}.}
}
\value{
a\code{\link[flowCore:filters-class]{filters}} list containing the 4
  constructed \code{\link[flowCore:rectangleGate]{rectangleGate}}
  objects.
}
\description{
\code{drawQuadrants} constructs an interactive plotting window for user to
select the crosshair center of 4 populations which is used to construct 4
\code{\link[flowCore:rectangleGate]{rectangleGate}} objects which are
stored in a\code{\link[flowCore:filters-class]{filters}}  list. Populations
are assigned in the following order: bottom left, bottom right, top right and
top left.
}
\seealso{
\code{\link{plotCyto1d,flowFrame-method}}

\code{\link{plotCyto2d,flowFrame-method}}

\code{\link{drawGate}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
\keyword{FlowJo,}
\keyword{draw,}
\keyword{gating,}
\keyword{manual,}
\keyword{openCyto,}
\keyword{quadrants}
\keyword{rectangleGate,}
