% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{extractGate}
\alias{extractGate}
\title{Extract saved gate(s) from gatingTemplate.}
\usage{
extractGate(parent, alias, gtfile)
}
\arguments{
\item{parent}{name of the parental population.}

\item{alias}{name of the population for which the gate must be extracted.}

\item{gtfile}{name of the \code{gatingTemplate} csv file (e.g. "gatingTemplate.csv") where the gate(s) are saved.}
}
\description{
Extract saved gate(s) from gatingTemplate.
}
