% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIt-SNE.R
\name{fftRtsne}
\alias{fftRtsne}
\title{FIt-SNE Based on Kluger Lab FIt-SNE}
\usage{
fftRtsne(
  X,
  dims = 2,
  perplexity = 30,
  theta = 0.5,
  max_iter = 1000,
  fft_not_bh = TRUE,
  ann_not_vptree = TRUE,
  stop_early_exag_iter = 250,
  exaggeration_factor = 12,
  no_momentum_during_exag = FALSE,
  start_late_exag_iter = -1,
  late_exag_coeff = 1,
  mom_switch_iter = 250,
  momentum = 0.5,
  final_momentum = 0.8,
  learning_rate = 200,
  n_trees = 50,
  search_k = -1,
  rand_seed = -1,
  nterms = 3,
  intervals_per_integer = 1,
  min_num_intervals = 50,
  K = -1,
  sigma = -30,
  initialization = NULL,
  data_path = NULL,
  result_path = NULL,
  load_affinities = NULL,
  fast_tsne_path = NULL,
  nthreads = 0,
  perplexity_list = NULL,
  get_costs = FALSE,
  df = 1
)
}
\arguments{
\item{X}{a matrix contain ing the data to be mapped.}

\item{dims}{dimensionality of the embedding. Default 2.}

\item{perplexity}{perplexity is used to determine the bandwidth of the
Gaussian kernel in the input space.  Default 30.}

\item{theta}{set to 0 for exact.  If non-zero, then will use either Barnes
Hut or FIt-SNE based on nbody_algo.  If Barnes Hut, then this determins the
accuracy of BH approximation. Default 0.5.}

\item{max_iter}{number of iterations of t-SNE to run. Default 1000.}

\item{fft_not_bh}{if theta is nonzero, this determins whether to use FIt-SNE
or Barnes Hut approximation. Default is FIt-SNE. set to be True for FIt-SNE}

\item{ann_not_vptree}{use vp-trees (as in bhtsne) or approximate nearest
neighbors (default). Set to be TRUE for approximate nearest neighbors.}

\item{stop_early_exag_iter}{when to switch off early exaggeration. Default
250.}

\item{exaggeration_factor}{coefficient for early exaggeration (>1). Default
12.}

\item{no_momentum_during_exag}{set to 0 to use momentum and other
optimization tricks. 1 to do plain,vanilla gradient descent (useful for
testing large exaggeration coefficients)}

\item{start_late_exag_iter}{when to start late exaggeration. set to -1 to not
use late exaggeration. Default -1.}

\item{late_exag_coeff}{late exaggeration coefficient. Set to -1 to not use
late exaggeration. Default -1.}

\item{mom_switch_iter}{unknown.}

\item{momentum}{unknown.}

\item{final_momentum}{unkown.}

\item{learning_rate}{unknown.}

\item{n_trees}{unknown.}

\item{search_k}{unknown.}

\item{rand_seed}{unknown.}

\item{nterms}{- If using FIt-SNE, this is the number of interpolation points
per sub-interval.}

\item{intervals_per_integer}{see min_num_intervals}

\item{min_num_intervals}{let maxloc = ceil(max(max(X))) and minloc =
floor(min(min(X))). i.e. the points are in a [minloc]^no_dims by
[maxloc]^no_dims interval/square. The number of intervals in each dimension
is either min_num_intervals or ceil((maxloc -
minloc)/intervals_per_integer), whichever is larger. min_num_intervals must
be an integer >0, and intervals_per_integer must be >0. Defaults are
min_num_intervals=50 and intervals_per_integer = 1.}

\item{K}{number of nearest neighbours to get when using fixed sigma. Default
-30 (None).}

\item{sigma}{fixed sigma value to use when perplexity==-1. Default -1 (None).}

\item{initialization}{N x no_dims array to intialize the solution. Default:
None.}

\item{data_path}{unknown}

\item{result_path}{unknown}

\item{load_affinities}{- If 1, input similarities are loaded from a file and
not computed If 2, input similarities are saved into a file. If 0,
affinities are neither saved nor loaded.}

\item{fast_tsne_path}{unknown}

\item{nthreads}{unknown}

\item{perplexity_list}{if perplexity==0 then perplexity combination will be
used with values taken from perplexity_list. Default: NULL df - Degree of
freedom of t-distribution, must be greater than 0. Values smaller than 1
correspond to heavier tails, which can often resolve substructure in the
embedding. See Kobak et al. (2019) for details. Default is 1.0.}

\item{get_costs}{unknown}

\item{df}{unknown}
}
\description{
Modified version of \url{https://github.com/KlugerLab/FIt-SNE/}
implementation to expose argument names and defaults within CytoExploreR.
This function should not be used directly, data should instead be mapped
using \code{\link{cyto_map}}.
}
\seealso{
\code{\link{cyto_map}}
}
