% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{file_ext_append}
\alias{file_ext_append}
\title{Append file name with an extension}
\usage{
file_ext_append(x, ext = "csv")
}
\arguments{
\item{x}{vector of filenames.}

\item{ext}{vector of extensions to append.}
}
\description{
Append file name with an extension
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
