% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{file_wd_check}
\alias{file_wd_check}
\title{Check if a file exists in the current working directory}
\usage{
file_wd_check(name)
}
\arguments{
\item{name}{filename including file extension to be checked.}
}
\value{
TRUE/FALSE if file exists in the current working directory.
}
\description{
Check if a file exists in the current working directory
}
\examples{
file_wd_check("gatingTemplate.csv")

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
