% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_draw-methods.R
\docType{methods}
\name{gate_draw,flowSet-method}
\alias{gate_draw,flowSet-method}
\title{gate_draw flowSet Method}
\usage{
\S4method{gate_draw}{flowSet}(x, select = NULL, channels = NULL,
  alias = NULL, display = 1/length(x), type = "polygon",
  axis = "x", density_smooth = 1.5, label = TRUE, plot = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowSet-class]{flowSet}}.}

\item{select}{vector containing the indicies of samples within gs to use for
plotting.}

\item{channels}{vector of channel names to use for plotting, can be of length
1 for 1-D density histogram or length 2 for 2-D scatter plot.}

\item{alias}{the name(s) of the populations to be gated. If multiple
population names are supplied (e.g. \code{c("CD3,"CD4)}) multiple gates
will be returned. \code{alias} is \code{NULL} by default which will halt
the gating routine.}

\item{display}{numeric [0,1] to control the percentage of events to be
plotted. Specifying a value for \code{display} can substantial improve
plotting speed for less powerful machines.}

\item{type}{vector of gate type names used to construct the gates. Multiple
gate types are supported but should be accompanied with an \code{alias}
argument of the same length (i.e. one \code{type} per \code{alias}).
Supported gate types are \code{polygon, rectangle, ellipse, threshold,
boundary, interval, quadrant and web} which can be abbreviated as upper or
lower case first letters as well. Default \code{type} is \code{"interval"}
for 1D gates and \code{"polygon"} for 2D gates.}

\item{axis}{indicates whether the \code{"x"} or \code{"y"} axis should be
gated for 2-D interval gates.}

\item{density_smooth}{smooothing factor passed to
\code{\link[stats:density]{density}} for 1-D plots (defaults to 1.5).}

\item{label}{logical indicating whether to include \code{\link{cyto_plot_label}}
for the gated population(s), \code{TRUE} by default.}

\item{plot}{logical indicating whether a plot should be drawn, set to
\code{TRUE} by default.}

\item{...}{additional arguments for \code{\link{cyto_plot,flowSet-method}}.}
}
\value{
a \code{\link[flowCore:filters-class]{filters}} list containing the
  drawn gate objects.
}
\description{
Manually draw gates around populations for analysis of flow cytometry data.
}
\examples{
\dontrun{
library(CytoRSuiteData)

# Load in samples
fs <- Activation

# draw gates using gate_draw - add contour lines & overlay control
gate_draw(fs,
  channels = c("FSC-A", "SSC-A"),
  alias = "Cells",
  type = "polygon",
  contour_lines = 15,
  overlay = fs[[1]]
)

# gt is a filters object containing the contructed polygonGate
gt[[1]]
}

}
\seealso{
\code{\link{cyto_plot,flowSet-method}}

\code{\link{gate_draw,flowFrame-method}}

\code{\link{gate_draw,GatingSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
