% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCyto-plugins.R
\name{gate_draw}
\alias{gate_draw}
\title{drawGate Plugin for openCyto}
\usage{
gate_draw(fr, pp_res, channels, alias, subSample, gate_range = NULL,
  min = NULL, max = NULL, type = "polygon", axis = "x",
  adjust = 1.5, plot = TRUE, transList = NULL, ...)
}
\arguments{
\item{fr}{a \code{\link[flowCore:flowFrame-class]{flowFrame}} object
containing the flow cytometry data for gating.}

\item{pp_res}{output of preprocessing function.}

\item{channels}{name(s) of the channel(s) to be used for plotting.}

\item{alias}{used within gating_args to supply the names of the populations
to drawGate (e.g. gating_args = alias='Lymphocytes').}

\item{subSample}{number of events to plot.}

\item{gate_range}{range in which gate should be constructed (only needed for
autogating functions).}

\item{min}{argument passed to \code{truncate_flowFrame} to restrict data to
values > \code{min}.}

\item{max}{argument passed to \code{truncate_flowFrame} to restrict data to
values < \code{max}.}

\item{type}{type of gate to be constructed, supported types include
\code{c("polygon", "rectangle", "interval", "threshold", "ellipse",
"quadrant")}.}

\item{axis}{indicates the axis to use for gating for \code{type="interval"}
when 2 fluorescent channel are supplied.}

\item{adjust}{numeric smoothing factor used for 1D density plots.}

\item{plot}{logical indicating whether a plot should be constructed.}

\item{transList}{object of class
\code{\link[flowCore:transformList-class]{transformerList}} generated by
\code{estimateLogicle} to transform fluorescent channels for gating.
\code{transList} is required if logicle transformation has already been
applied to \code{x} using estimateLogicle. \code{computeSpillover} will
automatically call \code{\link[flowCore:logicleTransform]{estimateLogicle}}
internally to transform channels prior to gating, if \code{transList} is
supplied it will be used for the transformation instead.}

\item{...}{additional arguments passsed to \code{\link{plotCyto}}.}
}
\value{
a \code{polygonGate} constructed from coordinates supplied by
  \code{DrawGate}.
}
\description{
\code{drawGate} allows the user to draw polygon gates directly onto plots of
flow cytometry data. Simply left click to gate and right click to close the
gate.
}
\examples{
\dontrun{
fs <- Activation # load in .fcs files

gs <- GatingSet(fs) # add flowSet to GatingSet

template <- add_pop(
gs, alias = "Lymphocytes", pop = "+", parent = "root",
dims = "FSC-A,SSC-A", gating_method = "drawGate",
gating_args = "subSample=1000,alias='Lymphocytes',type='ellipse'",
collapseDataForGating = TRUE, groupBy = 2
)

# gating window will open to construct gate left click vertices on plot
# and close gate by right click and selecting "stop".
plotCyto(gs[[1]],
         parent = "root",
         alias = "Lymphocytes",
         channels = c("FSC-A","SSC-A"))
}

}
\seealso{
\code{\link{plotCyto1d,flowFrame-method}}

\code{\link{plotCyto1d,flowSet-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
\keyword{gating,}
\keyword{manual,}
\keyword{polygon,}
\keyword{polygonGate}
