% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating-functions.R
\name{gate_ellipse_draw}
\alias{gate_ellipse_draw}
\title{Draw Ellipsoid Gate(s) Around Populations.}
\usage{
gate_ellipse_draw(fr, alias = NULL, channels, plot = TRUE,
  label = TRUE, ...)
}
\arguments{
\item{fr}{a \code{\link[flowCore:flowFrame-class]{flowFrame}} object
containing the flow cytometry data for plotting and gating.}

\item{alias}{the name(s) of the populations to be gated. If multiple
population names are supplied (e.g. \code{c("CD3","CD4")}) multiple gates
will be returned. \code{alias} is \code{NULL} by default which will halt
the gating routine.}

\item{channels}{vector of channel names to use for plotting, can be of length
1 for 1-D density histogram or length 2 for 2-D scatter plot.}

\item{plot}{logical indicating whether the data should be plotted. This
feature allows for constructing gates of different types over existing
plots which may already contain a different gate type.}

\item{label}{logical indicating whether to include \code{\link{cyto_plot_label}}
for the gated population(s), \code{TRUE} by default.}

\item{...}{additional arguments for \code{\link{cyto_plot,flowFrame-method}}.}
}
\value{
a\code{\link[flowCore:filters-class]{filters}} list containing the
  constructed \code{\link[flowCore:ellipsoidGate-class]{ellipsoidGate}}
  object(s).
}
\description{
\code{gate_ellipse_draw} constructs an interactive plotting window for user to
select the limits of a population in 2 dimensions (4 points) which is
constructed into \code{\link[flowCore:ellipsoidGate-class]{ellipsoidGate}}
object and stored in a \code{\link[flowCore:filters-class]{filters}} list.
}
\examples{
\dontrun{

# Copy and paste into console to interactively draw gates

library(CytoRSuiteData)

# Load in samples to flowSet
fs <- Activation

# Transform fluorescent channels
fs <- transform(fs, estimateLogicle(fs[[4]], cyto_fluor_channels(fs)))

# Get ellipsoidGate using gate_ellipse_draw
eg <- gate_ellipse_draw(fs[[4]],
  alias = "Cells",
  channels = c("PE-A", "Alexa Fluor 700-A"),
  overlay = fs[[1]]
)

# eg is a filters object - extract ellipsoidGate using `[[`
eg[[1]]
}

}
\seealso{
\code{\link{cyto_plot,flowFrame-method}}

\code{\link{gate_draw}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
\keyword{draw,}
\keyword{ellipse}
\keyword{ellipsoidGate,}
\keyword{gating,}
\keyword{manual,}
\keyword{openCyto,}
