% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating-functions.R
\name{gate_interval_draw}
\alias{gate_interval_draw}
\title{Draw Interval Gate(s) Around Populations.}
\usage{
gate_interval_draw(fr, alias = NULL, channels, plot = TRUE,
  axis = "x", label = TRUE, ...)
}
\arguments{
\item{fr}{a \code{\link[flowCore:flowFrame-class]{flowFrame}} object
containing the flow cytometry data for plotting and gating.}

\item{alias}{the name(s) of the populations to be gated. If multiple
population names are supplied (e.g. \code{c("CD3,"CD4)}) multiple gates
will be returned. \code{alias} is \code{NULL} by default which will halt
the gating routine.}

\item{channels}{vector of channel names to use for plotting, can be of length
1 for 1-D density histogram or length 2 for 2-D scatter plot.}

\item{plot}{logical indicating whether the data should be plotted. This
feature allows for constructing gates of different types over existing
plots which may already contain a different gate type.}

\item{axis}{indicates whether the \code{"x"} or \code{"y"} axis should be
gated for 2-D interval gates.}

\item{label}{logical indicating whether to include
\code{\link{cyto_plot_label}} for the gated population(s), \code{TRUE} by
default.}

\item{...}{additional arguments for \code{\link{cyto_plot,flowFrame-method}}.}
}
\value{
a\code{\link[flowCore:filters-class]{filters}} list containing the
  constructed \code{\link[flowCore:rectangleGate-class]{rectangleGate}}
  object(s).
}
\description{
\code{gate_interval_draw} constructs an interactive plotting window for user
to select the lower and upper bounds of a population (through mouse click)
which is constructed into a
\code{\link[flowCore:rectangleGate-class]{rectangleGate}} object and stored
in a \code{\link[flowCore:filters-class]{filters}} list. Both 1-D and 2-D
interval gates are supported, for 2-D interval gates an additional argument
\code{axis} must be supplied to indicate which axis should be gated.
}
\examples{
\dontrun{

# Copy and paste into console to interactively draw gates

library(CytoRSuiteData)

# Load in samples to flowSet
fs <- Activation

# Transform fluorescent channels
fs <- transform(fs, estimateLogicle(fs[[4]], cyto_fluor_channels(fs)))

# Get 1-D interval gate using gate_interval_draw - overlay control
ig <- gate_interval_draw(fs[[4]],
  alias = "Cells",
  channels = "PE-A",
  overlay = fs[[1]],
  density_stack = 0.5
)

# ig is a filters object - extract rectangleGate using `[[`
ig[[1]]

# Get 2-D interval gate on y axis using gate_interval_draw
ig <- gate_interval_draw(fs[[4]],
  alias = "Cells",
  channels = c("PE-A", "Alexa Fluor 488-A"),
  axis = "y"
)

# ig is a filters object - extract rectangleGate using `[[`
ig[[1]]
}

}
\seealso{
\code{\link{cyto_plot,flowFrame-method}}

\code{\link{gate_draw}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
\keyword{draw,}
\keyword{gating,}
\keyword{interval}
\keyword{manual,}
\keyword{openCyto,}
\keyword{rectangleGate,}
