% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCyto-plugins.R
\name{gate_manual}
\alias{gate_manual}
\title{Apply Stored Gates to GatingSet}
\usage{
gate_manual(fr, pp_res, channels, gate)
}
\arguments{
\item{fr}{a \code{\link[flowCore:flowFrame-class]{flowFrame}} object
containing the flow cytometry data for gating.}

\item{pp_res}{output of preprocessing function.}

\item{channels}{fluorescent channel(s) to use for gating.}

\item{gate}{stored \code{drawGate} gate in csv file for passing to openCyto.}
}
\value{
pass saved gate to openCyto to apply to all samples.
}
\description{
\code{drawGate} allows the user to draw gates directly onto plots of flow
cytometry data. This plugin provides a way of passing manually drawn gates to
\code{openCyto}.
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
\keyword{gating,}
\keyword{manual,}
\keyword{polygon,}
\keyword{polygonGate}
