% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating-functions.R
\name{gate_quadrant_draw}
\alias{gate_quadrant_draw}
\title{Draw Quadrant Gates Around Populations.}
\usage{
gate_quadrant_draw(fr, alias = NULL, channels, plot = TRUE,
  label = TRUE, ...)
}
\arguments{
\item{fr}{a \code{\link[flowCore:flowFrame-class]{flowFrame}} object
containing the flow cytometry data for plotting and gating.}

\item{alias}{the name(s) of the 4 populations to be gated. \code{alias} is
\code{NULL} by default which will halt the gating routine.}

\item{channels}{vector of channel names to use for plotting, can be of length
1 for 1-D density histogram or length 2 for 2-D scatter plot.}

\item{plot}{logical indicating whether the data should be plotted. This
feature allows for constructing gates of different types over existing
plots which may already contain a different gate type.}

\item{label}{logical indicating whether to include \code{\link{cyto_plot_label}}
for the gated population(s), \code{TRUE} by default.}

\item{...}{additional arguments for \code{\link{cyto_plot,flowFrame-method}}.}
}
\value{
a\code{\link[flowCore:filters-class]{filters}} list containing the 4
  constructed \code{\link[flowCore:rectangleGate-class]{rectangleGate}}
  objects.
}
\description{
\code{gate_quadrant_draw} constructs an interactive plotting window for user to
select the crosshair center of 4 populations which is used to construct 4
\code{\link[flowCore:rectangleGate-class]{rectangleGate}} objects which are
stored in a\code{\link[flowCore:filters-class]{filters}}  list. Populations
are assigned in the following order: bottom left, bottom right, top right and
top left.
}
\examples{
\dontrun{

# Copy and paste into console to interactively draw gates

library(CytoRSuiteData)

# Load in samples to flowSet
fs <- Activation

# Transform fluorescent channels
fs <- transform(fs, estimateLogicle(fs[[4]], cyto_fluor_channels(fs)))

# Get quadrant gates using gate_quadrant_draw
qg <- gate_quadrant_draw(fs[[4]],
  alias = c("DN", "CD4", "DP", "CD8"),
  channels = c("Alexa Fluor 700-A", "Alexa Fluor 488-A")
)

# qg is a filters object - extract each rectangleGate using `[[`
qg[[4]]
}

}
\seealso{
\code{\link{cyto_plot,flowFrame-method}}

\code{\link{gate_draw}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
\keyword{FlowJo,}
\keyword{draw,}
\keyword{gating,}
\keyword{manual,}
\keyword{openCyto,}
\keyword{quadrants}
\keyword{rectangleGate,}
