% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating-functions.R
\name{gate_rectangle_draw}
\alias{gate_rectangle_draw}
\title{Draw Rectangle Gate(s) Around Populations.}
\usage{
gate_rectangle_draw(fr, alias = NULL, channels, plot = TRUE,
  label = TRUE, ...)
}
\arguments{
\item{fr}{a \code{\link[flowCore:flowFrame-class]{flowFrame}} object
containing the flow cytometry data for plotting and gating.}

\item{alias}{the name(s) of the populations to be gated. If multiple
population names are supplied (e.g. \code{c("CD3,"CD4)}) multiple gates
will be returned. \code{alias} is \code{NULL} by default which will halt
the gating routine.}

\item{channels}{vector of channel names to use for plotting, can be of length
1 for 1-D density histogram or length 2 for 2-D scatter plot.}

\item{plot}{logical indicating whether the data should be plotted. This
feature allows for constructing gates of different types over existing
plots which may already contain a different gate type.}

\item{label}{logical indicating whether to include
\code{\link{cyto_plot_label}} for the gated population(s), \code{TRUE} by
default.}

\item{...}{additional arguments for \code{\link{cyto_plot,flowFrame-method}}.}
}
\value{
a\code{\link[flowCore:filters-class]{filters}} list containing the
  constructed \code{\link[flowCore:rectangleGate-class]{rectangleGate}}
  object(s).
}
\description{
\code{gate_rectangle_draw} constructs an interactive plotting window to allow
manual selection of the co-ordinates of a rectangle gate(s) (through mouse
click) which are constructed into
\code{\link[flowCore:rectangleGate-class]{rectangleGate}} objects and stored
in a \code{\link[flowCore:filters-class]{filters}} list. Simply select 2
diagonal co-ordinates to construct the rectangleGate(s).
}
\examples{
\dontrun{

# Copy and paste into console to interactively draw gates

library(CytoRSuiteData)

# Load in samples to flowSet
fs <- Activation

# Transform fluorescent channels
fs <- transform(fs, estimateLogicle(fs[[4]], cyto_fluor_channels(fs)))

# Get polygonGate using gate_rectangle_draw - add contour lines
rg <- gate_rectangle_draw(fs[[4]],
  alias = "Cells",
  channels = c("FSC-A", "SSC-A"),
  contour_lines = 15
)

# rg is a filters object - extract rectangleGate using `[[`
rg[[1]]
}

}
\seealso{
\code{\link{cyto_plot,flowFrame-method}}

\code{\link{gate_draw}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
\keyword{draw,}
\keyword{gating,}
\keyword{manual,}
\keyword{openCyto}
\keyword{rectangleGate,}
