% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating-functions.R
\name{gate_threshold_draw}
\alias{gate_threshold_draw}
\title{Draw Threshold Gate(s) Around Populations.}
\usage{
gate_threshold_draw(fr, alias = NULL, channels, plot = TRUE,
  label = TRUE, ...)
}
\arguments{
\item{fr}{a \code{\link[flowCore:flowFrame-class]{flowFrame}} object
containing the flow cytometry data for plotting and gating.}

\item{alias}{the name(s) of the populations to be gated. Multiple
\code{threshold} gates are not currently supported. \code{alias} is
\code{NULL} by default which will halt the gating routine.}

\item{channels}{vector of channel names to use for plotting, can be of length
1 for 1-D density histogram or length 2 for 2-D scatter plot.}

\item{plot}{logical indicating whether the data should be plotted. This
feature allows for constructing gates of different types over existing
plots which may already contain a different gate type.}

\item{label}{logical indicating whether to include
\code{\link{cyto_plot_label}} for the gated population(s), \code{TRUE} by
default.}

\item{...}{additional arguments for \code{\link{cyto_plot,flowFrame-method}}.}
}
\value{
a\code{\link[flowCore:filters-class]{filters}} list containing the
  constructed \code{\link[flowCore:rectangleGate-class]{rectangleGate}}
  object.
}
\description{
\code{gate_threshold_draw} constructs an interactive plotting window for user
to select the lower bound of a population which is constructed into a
\code{\link[flowCore:rectangleGate-class]{rectangleGate}} object and stored
in a \code{\link[flowCore:filters-class]{filters}} list. Both 1-D and 2-D
threshold gates are supported, for 2-D threshold gates all events above the
select x and y coordinates are included in the gate. Multiple threshold gates
are not currently supported.
}
\examples{
\dontrun{

# Copy and paste into console to interactively draw gates

library(CytoRSuiteData)

# Load in samples to flowSet
fs <- Activation

# Transform fluorescent channels
fs <- transform(fs, estimateLogicle(fs[[4]], cyto_fluor_channels(fs)))

# Get 1-D threshold gate using gate_threshold_draw
tg <- gate_threshold_draw(fs[[4]],
  alias = "Cells",
  channels = c("PE-A")
)

# tg is a filters object - extract rectangleGate using `[[`
tg[[1]]

#' # Get 2-D threshold gate using gate_threshold_draw - overlay control
tg <- gate_threshold_draw(fs[[4]],
  alias = "Cells",
  channels = c("Alexa Fluor 647-A", "7-AAD-A"),
  overlay = fs[[1]]
)

# tg is a filters object - extract rectangleGate using `[[`
tg[[1]]
}

}
\seealso{
\code{\link{cyto_plot,flowFrame-method}}

\code{\link{gate_draw}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
\keyword{draw,}
\keyword{gating,}
\keyword{manual,}
\keyword{openCyto,}
\keyword{rectangleGate,}
\keyword{threshold}
