% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating-functions.R
\name{gate_web_draw}
\alias{gate_web_draw}
\title{Draw Web Gates Around Populations.}
\usage{
gate_web_draw(fr, alias = NULL, channels, plot = TRUE, label = TRUE,
  ...)
}
\arguments{
\item{fr}{a \code{\link[flowCore:flowFrame-class]{flowFrame}} object
containing the flow cytometry data for plotting and gating.}

\item{alias}{the name(s) of the populations to be gated. If multiple
population names are supplied (e.g. \code{c("CD3","CD4")}) multiple gates
will be returned. \code{alias} is \code{NULL} by default which will halt
the gating routine. Recommended for 3 or more populations.}

\item{channels}{vector of channel names to use for plotting, can be of length
1 for 1-D density histogram or length 2 for 2-D scatter plot.}

\item{plot}{logical indicating whether the data should be plotted. This
feature allows for constructing gates of different types over existing
plots which may already contain a different gate type.}

\item{label}{logical indicating whether to include \code{\link{cyto_plot_label}}
for the gated population(s), \code{TRUE} by default.}

\item{...}{additional arguments for \code{\link{cyto_plot,flowFrame-method}}.}
}
\value{
a\code{\link[flowCore:filters-class]{filters}} list containing the
  constructed \code{\link[flowCore:polygonGate-class]{polygonGate}}
  object(s).
}
\description{
\code{gate_web_draw} is a variation of drawQuadrant which allows more flexibility
with gate co-ordinates (angled lines) and supports any number of gates as
indicated by the \code{alias} argument. To construct the gate simply select
the center point and surrounding divider points on plot edge. \code{gate_web_draw}
will contruct the \code{\link[flowCore:polygonGate-class]{polygonGate}}
objects and store them in a \code{\link[flowCore:filters-class]{filters}}
list.
}
\examples{
\dontrun{

# Copy and paste into console to interactively draw gates

library(CytoRSuiteData)

# Load in samples to flowSet
fs <- Activation

# Transform fluorescent channels
fs <- transform(fs, estimateLogicle(fs[[4]], cyto_fluor_channels(fs)))

# Get web gates using gate_web_draw
wg <- gate_web_draw(fs[[4]],
  alias = c("DN", "CD4", "CD8"),
  channels = c("Alexa Fluor 700-A", "Alexa Fluor 488-A")
)

# wg is a filters object - extract each polygonGate using `[[`
wg[[4]]
}

}
\seealso{
\code{\link{cyto_plot,flowFrame-method}}

\code{\link{gate_draw}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
\keyword{draw,}
\keyword{gate_web_draw}
\keyword{gating,}
\keyword{manual,}
\keyword{openCyto,}
\keyword{polygonGate,}
