% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingTemplate-helpers.R
\name{gatingTemplate_edit}
\alias{gatingTemplate_edit}
\title{Interactively Edit a gatingTemplate}
\usage{
gatingTemplate_edit(x, gatingTemplate = NULL)
}
\arguments{
\item{x}{object of class \code{GatingSet} which has been gated with the 
supplied gatingTemplate.}

\item{gatingTemplate}{name of the gatingTemplate csv file to be edited.}
}
\value{
update gatingTemplate csv file and update the GatingSet accordingly.
}
\description{
\code{gatingTemplate_edit} provides an interactive interface to editing the
openCyto gatingTemplate. This function is intended to aid in adding boolean
and reference gates to the gatingTemplate. Users should NOT modify existing
entries in the gatingTemplate.
}
\examples{
\dontrun{
library(CytoRSuite)

# gs is a GatingSet object
gatingTemplate_edit(gs, "gatingTemplate.csv")
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
