% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\docType{methods}
\name{getChannels,flowFrame-method}
\alias{getChannels,flowFrame-method}
\title{Extract fluorescent parameters from a flowFrame}
\usage{
\S4method{getChannels}{flowFrame}(x)
}
\arguments{
\item{x}{object \code{flowFrame}}
}
\value{
vector of fluorescent channels
}
\description{
Extract fluorescent parameters from a flowFrame
}
