% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{getChannels}
\alias{getChannels}
\title{Extract Fluorescent Channels}
\usage{
getChannels(x)
}
\description{
Extract Fluorescent Channels
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
