% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-statistics.R
\name{getFreq}
\alias{getFreq}
\title{Get Population Frequency}
\usage{
getFreq(gs, alias, parent, type = "percent")
}
\arguments{
\item{gs}{an object of class \code{GatingSet}.}

\item{alias}{name of the population.}

\item{parent}{name of the parent population to use for calculation}
}
\description{
Get Population Frequency
}
