% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{getGateType}
\alias{getGateType}
\title{Get gate_type from Saved Gate.}
\usage{
getGateType(gates)
}
\arguments{
\item{gates}{an object of class \code{filters} containing the gates from which the \code{gate_type(s)} must be obtained.}
}
\value{
vector of gate_type names for supplied gates.
}
\description{
Get gate_type from Saved Gate.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
