% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{ozette_light_theme}
\alias{ozette_light_theme}
\title{Light theme to match Ozette's platform}
\usage{
ozette_light_theme(
  point_shape = "hex",
  point_col_scale = c("#131D2A", "#1853C9", "#EFF412", "#FF8000", "#FF1200"),
  axes_text_col = "grey40",
  key_text_col = "grey40",
  ...
)
}
\arguments{
\item{...}{additional arguments to customize the \code{cyto_plot()} theme,
see \code{cyto_plot_theme_args()} for details.}
}
\description{
Light theme to match Ozette's platform
}
\author{
@author Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
