% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{ozette_theme}
\alias{ozette_theme}
\title{Combination of light and dark themes for Ozette's slide decks}
\usage{
ozette_theme(
  point_shape = "hex",
  point_col_scale = c("#131D2A", "#1853C9", "#EFF412", "#FF8000", "#FF1200"),
  grid_line_col = "grey40",
  grid_line_alpha = 0.1,
  border_fill = "white",
  border_line_col = "grey95",
  axes_text_col = "grey70",
  axes_label_text_col = "white",
  axes_ticks_line_col = "grey95",
  title_text_col = "white",
  key_text_col = "grey70",
  key_title_text_col = "white",
  key_ticks_line_col = "grey95",
  key_border_line_col = "grey95",
  page_fill = "#0B1117",
  contour_line_col = "black",
  header_text_col = "white",
  legend_text_col = "white",
  gate_line_col = "black",
  gate_line_width = 1,
  gate_fill_alpha = 0.1,
  gate_fill = "grey40",
  label_fill_alpha = 0,
  label_text_col = "black",
  label_text_font = 1,
  ...
)
}
\arguments{
\item{...}{additional arguments to customize the \code{cyto_plot()} theme,
see \code{cyto_plot_theme_args()} for details.}
}
\description{
Combination of light and dark themes for Ozette's slide decks
}
\author{
@author Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
