% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCyto-methods.R
\docType{methods}
\name{plotCyto,flowFrame-method}
\alias{plotCyto,flowFrame-method}
\title{plotCyto - flowFrame Method}
\usage{
\S4method{plotCyto}{flowFrame}(x, channels, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}}.}

\item{channels}{vector of channel names to use for plotting, can be of length
1 for 1D density distribution or length 2 for 2D scatter plot.}

\item{...}{additional arguments passed to
\code{\link{plotCyto1d,flowFrame-method}} or
\code{\link{plotCyto2d,flowFrame-method}}.}
}
\description{
Explore & visualise flow cytometry data for a flowFrame.
}
\examples{
\dontrun{
  fs <- Activation
  trans <- estimateLogicle(fs[[2]],"PE-A")
  fs <- transform(fs, trans)
  plotCyto(fs[[2]], channels = c("PE-A"), transList = trans, overlay = fs[[1]])
  plotCyto(fs[[2]], channels = c("FSC-A","PE-A"), transList = trans, overlay = fs[[1]])
}

}
\seealso{
\code{\link{plotCyto1d,flowFrame-method}}

\code{\link{plotCyto2d,flowFrame-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
