% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCyto-methods.R
\docType{methods}
\name{plotCyto,flowSet-method}
\alias{plotCyto,flowSet-method}
\title{plotCyto - flowSet Method}
\usage{
\S4method{plotCyto}{flowSet}(x, channels, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowSet-class]{flowSet}}.}

\item{channels}{vector of channel names to use for plotting, can be of length
1 for 1D density distribution or length 2 for 2D scatter plot.}

\item{...}{additional arguments passed to
\code{\link{plotCyto1d,flowSet-method}} or
\code{\link{plotCyto2d,flowSet-method}}.}
}
\description{
Explore & visualise flow cytometry data for a flowSet.
}
\examples{
\dontrun{
  fs <- Activation
  trans <- estimateLogicle(fs[[2]],"PE-A")
  fs <- transform(fs, trans)
  plotCyto(fs, channels = c("PE-A"), transList = trans, overlay = fs[[1]])
  plotCyto(fs, channels = c("FSC-A","PE-A"), transList = trans, overlay = fs[[1]])
}

}
\seealso{
\code{\link{plotCyto1d,flowSet-method}}

\code{\link{plotCyto2d,flowSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
