% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCyto-methods.R
\name{plotCyto}
\alias{plotCyto}
\title{plotCyto}
\usage{
plotCyto(x, ...)
}
\arguments{
\item{x}{object of class \code{flowFrame}, \code{flowSet} or
\code{GatingSet}.}

\item{...}{additional method-specific arguments for plotCyto.}
}
\description{
Explore & visualise flow cytometry data.
}
\details{
\code{plotCyto} is a versatile wrapper function built on base graphics
capable of plotting a variety of existing flow cytometry classes including
\code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} and
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}. The type of plot
constructed by \code{plotCyto} is determined by the number of channels
supplied to the \code{channels} argument. If a single channel is supplied, a
1-D density distribution will be plotted in the specified channel using
\code{\link{plotCyto1d,flowFrame-method}}. Supplying two channels to this
argument will result in construction of a 2-D scatterplot with a blue-red
density colour scale for points using
\code{\link{plotCyto2d,flowFrame-method}}.

Some key features of \code{plotCyto} include: \itemize{ \item minimal coding
\item stacked density distributions \item 2-D contour lines \item back-gating
\item overlays \item population labelling \item grid to visualise multiple
samples \item plot gate objects \item plot in pop-up windows \item sampling
\item fully customisable \item copy and paste arguments from drawGate }

For more method-specific information refer to
\code{\link{plotCyto1d,flowFrame-method}} and
\code{\link{plotCyto2d,flowFrame-method}}.
}
\seealso{
\code{\link{plotCyto,flowFrame-method}}

\code{\link{plotCyto,flowSet-method}}

\code{\link{plotCyto,GatingHierarchy-method}}

\code{\link{plotCyto,GatingSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
