% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCyto-internal.R
\docType{methods}
\name{plotCyto1d,flowSet-method}
\alias{plotCyto1d,flowSet-method}
\title{plotCyto1d - flowSet Method}
\usage{
\S4method{plotCyto1d}{flowSet}(x, channel, transList = NULL,
  merge = FALSE, overlay = NULL, stack = c(0, 1), offset = 0.5,
  limits = "machine", xlim = NULL, mfrow, popup = FALSE, main, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowSet-class]{flowSet}}.}

\item{channel}{name of the channel or marker to be used to construct the
plot.}

\item{transList}{object of class
\code{\link[flowCore:transformList-class]{transformList}} or
\code{\link[flowWorkspace]{transformerList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} which was used to
transform the fluorescent channels of the supplied flowFrame. This
transList object will be used internally to ensure axes labels of the plot
are appropriately transformed. The transList will NOT be applied to the
flowFrame internally and should be applied to the flowFrame prior to
plotting.}

\item{merge}{logical indicating whether the flowFrames should be merged prior
to plotting to yield a single plot, set to FALSE by default.}

\item{overlay}{a \code{flowFrame}, \code{flowSet}, \code{list of flowFrames},
\code{list of flowSets} or \code{list of flowFrame lists} containing
populations to be overlaid onto the plot(s).}

\item{stack}{vector of length 2 indicating offset for samples and number of
samples per plot. Set to \code{c(0,1)} to plot each sample in a separate
panel.}

\item{offset}{numeric [0,1] indicating the degree of offset for overlayed
populations, set to 0.5 by default.}

\item{limits}{indicates whether the axes limits should be based on the
\code{"data"} or \code{"machine"}, set to "machine" by default to show
complete axes ranges. This argument will only alter the upper axis limits,
to modify the lower limits use \code{xlim} and \code{ylim}.}

\item{xlim}{lower and upper limits of x axis (e.g. c(0,5)).}

\item{mfrow}{a vector of the length 2 indicating the dimensions of the grid
for plotting \code{c(#rows, #columns)}.}

\item{popup}{logical indicating whether the plot should be constructed in a
pop-up window, set to FALSE by default. \code{popup} will open OS-specific
graphic device prior to plotting. Mac users will need to install
\href{https://www.xquartz.org/}{XQuartz} for this functionality.}

\item{main}{vector of titles to use for each plot, set to name of the sample
by default.}

\item{...}{additional arguments passed to
\code{\link{plotCyto1d,flowFrame-method}}.}
}
\description{
Visualise 1-D flow cytometry density distributions for a flowSet.
}
\examples{
\dontrun{
  fs <- Activation
  plotCyto1d(fs, channel = "FSC-A", legend = TRUE)
}

}
\seealso{
\code{\link{plotCyto1d,flowFrame-method}}

\code{\link{plotCyto,flowSet-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
