% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCyto-internal.R
\name{plotCyto1d}
\alias{plotCyto1d}
\title{plotCyto1d}
\usage{
plotCyto1d(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}} or
\code{\link[flowCore:flowSet-class]{flowSet}}.}

\item{...}{additional method-specific arguments for plotCyto1d.}
}
\description{
Visualise 1-D flow cytometry density distributions.
}
\details{
For a complete list of customisation arguments see
\code{\link{plotCyto1d,flowFrame-method}}.
}
\seealso{
\code{\link{plotCyto1d,flowFrame-method}}

\code{\link{plotCyto1d,flowSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
