% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCyto-internal.R
\docType{methods}
\name{plotCyto2d,flowFrame-method}
\alias{plotCyto2d,flowFrame-method}
\title{plotCyto2d - flowFrame Method}
\usage{
\S4method{plotCyto2d}{flowFrame}(x, channels, transList = NULL,
  overlay = NULL, contours = 0, gates = NULL, popup = FALSE,
  subSample = NULL, limits = "machine", xlim = NULL, ylim = NULL,
  main, xlab, ylab, pch = ".", cex.pts = 2, col, alpha = 1,
  col.contour = "black", lwd.contour = 1, cex.axis = 1,
  col.axis = "black", cex.lab = 1.1, col.lab = "black",
  cex.main = 1.1, col.main = "black", legend = FALSE, text.legend,
  col.legend, fill.legend, cex.legend = 1, col.gate = "red",
  lwd.gate = 2.5, lty.gate = 1, labels = TRUE, text.labels,
  format.labels = c("alias", "percent"), cex.labels = 0.8,
  font.labels = 2, col.labels = "black", alpha.labels = 0.6, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}}.}

\item{channels}{name of the channels or markers to be used to construct the
plot.}

\item{transList}{object of class
\code{\link[flowCore:transformList-class]{transformList}} or
\code{\link[flowWorkspace]{transformerList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} which was used to
transform the fluorescent channels of the supplied flowFrame. This
transList object will be used internally to ensure axes labels of the plot
are appropriately transformed. The transList will NOT be applied to the
flowFrame internally and should be applied to the flowFrame prior to
plotting.}

\item{overlay}{a \code{flowFrame} object to be overlayed onto the plot.}

\item{contours}{numeric indicating the number of levels to use for contour
lines, set to 0 by default to turn off contour lines.}

\item{gates}{gate object(s) to be added to plot. Gates can be of class
\code{\link[flowCore:rectangleGate]{rectangleGate}},
\code{\link[flowCore:polygonGate]{polygonGate}} or
\code{\link[flowCore:ellipsoidGate]{ellipsoidGate}}. Multiple gates of
mixed classes can also be supplied in \code{list} or
\code{\link[flowCore:filters-class]{filters}} objects.}

\item{popup}{logical indicating whether the plot should be constructed in a
pop-up window, set to FALSE by default. \code{popup} will open OS-specific
graphic device prior to plotting. Mac users will need to install
\href{https://www.xquartz.org/}{XQuartz} for this functionality.}

\item{subSample}{numeric indicating the number of events to plot, set to all
events by default. Reducing the sample size can significantly increase
plotting speed on less powerful machines.}

\item{limits}{indicates whether the axes limits should be based on the
\code{"data"} or \code{"machine"}, set to "machine" by default to show
complete axes ranges. This argument will only alter the upper axis limits,
to modify the lower limits use \code{xlim} and \code{ylim}.}

\item{xlim}{lower and upper limits of x axis (e.g. c(0,5)).}

\item{ylim}{lower and upper limits of x axis (e.g. c(0,5)).}

\item{main}{title to use for the plot, set to the name of the sample by
default.}

\item{xlab}{x axis label.}

\item{ylab}{y axis label.}

\item{pch}{point character to use for points, set to "." by default to
maximise plotting speed.}

\item{cex.pts}{numeric specifying the degree of character expansion for
points, set to 2 by default.}

\item{col}{colours to use for points, set to NA by default to blue-red
density colour scale.}

\item{alpha}{numeric [0,1] used to control col transparency, set to 1 by
default to remove transparency.}

\item{col.contour}{colour to use for contour lines, set to "black" by
default.}

\item{lwd.contour}{line width for contour lines, set to 2 by default.}

\item{cex.axis}{character expansion for axis text.}

\item{col.axis}{colour of axis text.}

\item{cex.lab}{character expansion for axis labels.}

\item{col.lab}{colour of axis labels.}

\item{cex.main}{character expansion for plot title.}

\item{col.main}{colour for plot title.}

\item{legend}{logical indicating whether a legend should be included for
plots including overlays, set to FALSE by default.}

\item{text.legend}{vector of labels to use for the legend.}

\item{col.legend}{vector of line colours to use for legend.}

\item{fill.legend}{vector of fill colours to use for legend.}

\item{cex.legend}{character expansion for legend text, set to 1 by default.}

\item{col.gate}{indicates the colour of the gate to be constructed, set to
\code{"red"} by default.}

\item{lwd.gate}{numeric to adjust line thickness of gates, set to \code{2.5}
by default.}

\item{lty.gate}{integer [0,6] which controls the line type, set to \code{1}
to draw solid lines by default.}

\item{labels}{logical indicating whether gated populations should be
labelled. If the names of the populations are supplied as the text.labels
argument, the population name and frequency will be included in the labels,
otherwise only the population frequencies will be included in the labels.}

\item{text.labels}{vector of population names to use in labels.}

\item{format.labels}{indicates the type of text to include in the label, can
be either \code{"alias"}, \code{"percent"}, \code{"count"},
\code{c("alias","percent")} or \code{c("alias","count")}. Set to
\code{c("alias","percent")} by default.}

\item{cex.labels}{character expansion for label text, set to 0.8 by default.}

\item{font.labels}{numeric indicating the font to use for labels, set to 2
for bold font by default. See \code{\link[graphics:par]{?par}} font for
details.}

\item{col.labels}{colour of text used in labels, set to \code{"black"} by
default.}

\item{alpha.labels}{numeric controlling backgropund fill transparency of
labels, set to 0.6 by default to introduce some transparency.}

\item{...}{additional arguments passed to \code{\link[graphics:plot]{plot}}.}
}
\description{
Visualise 2-D flow cytometry scatterplots with blue-red density colour scale
for a flowFrame.
}
\examples{
\dontrun{
  fs <- Activation
  plotCyto2d(fs[[1]], channel = c("FSC-A","SSC-A"), overlay = fs[[2]], col = c(NA,"purple"))
  plotCyto2d(fs[[1]], channel = c("FSC-A","SSC-A"), overlay = fs[[2]], col = c("black","red"))

}

}
\seealso{
\code{\link{plotCyto2d,flowSet-method}}

\code{\link{plotCyto,flowFrame-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
