% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCyto-internal.R
\docType{methods}
\name{plotCyto2d,flowSet-method}
\alias{plotCyto2d,flowSet-method}
\title{plotCyto2d - flowSet Method}
\usage{
\S4method{plotCyto2d}{flowSet}(x, channels, transList = NULL,
  merge = FALSE, overlay = NULL, subSample = NULL, mfrow,
  popup = FALSE, limits = "machine", xlim = NULL, ylim = NULL,
  main, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowSet-class]{flowSet}}.}

\item{channels}{names of channels or markers to be used to construct the
plot.}

\item{transList}{object of class
\code{\link[flowCore:transformList-class]{transformList}} or
\code{\link[flowWorkspace]{transformerList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} which was used to
transform the fluorescent channels of the supplied flowFrame. This
transList object will be used internally to ensure axes labels of the plot
are appropriately transformed. The transList will NOT be applied to the
flowFrame internally and should be applied to the flowFrame prior to
plotting.}

\item{merge}{logical indicating whether the flowFrames should be merged prior
to plotting to yield a single plot, set to FALSE by default.}

\item{overlay}{a \code{flowFrame}, \code{flowSet}, \code{list of flowFrames},
\code{list of flowSets} or \code{list of flowFrame lists} containing
populations to be overlayed onto the plot(s).}

\item{subSample}{numeric indicating the number of events to plot, set to all
events by default. Reducing the sample size can significantly increase
plotting speed on less powerful machines.}

\item{mfrow}{a vector of the length 2 indicating the dimensions of the grid
for plotting \code{c(#rows, #columns)}.}

\item{popup}{logical indicating whether the plot should be constructed in a
pop-up window, set to FALSE by default. \code{popup} will open OS-specific
graphic device prior to plotting. Mac users will need to install
\href{https://www.xquartz.org/}{XQuartz} for this functionality.}

\item{limits}{indicates whether the axes limits should be based on the
\code{"data"} or \code{"machine"}, set to "machine" by default to show
complete axes ranges. This argument will only alter the upper axis limits,
to modify the lower limits use \code{xlim} and \code{ylim}.}

\item{xlim}{lower and upper limits of x axis (e.g. c(0,5)).}

\item{ylim}{lower and upper limits of x axis (e.g. c(0,5)).}

\item{main}{title for the plot, set to the name of the sample by default.}

\item{...}{additional arguments passed to \code{\link{plotCyto2d,flowFrame-method}}.}
}
\description{
Visualise 2-D flow cytometry scatterplots with blue-red density colour scale
for a flowSet.
}
\examples{
\dontrun{
  fs <- Activation
  plotCyto2d(fs, channel = c("FSC-A","SSC-A"), overlay = fs[[2]], col = c(NA,"purple"))
  plotCyto2d(fs, channel = c("FSC-A","SSC-A"), overlay = fs[[2]], col = c("black","red"))

}

}
\seealso{
\code{\link{plotCyto2d,flowFrame-method}}

\code{\link{plotCyto,flowSet-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
