% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCyto-internal.R
\name{plotCyto2d}
\alias{plotCyto2d}
\title{plotCyto2d}
\usage{
plotCyto2d(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}} or
\code{\link[flowCore:flowSet-class]{flowSet}}.}

\item{...}{additional method-specific arguments for plotCyto2d.}
}
\description{
Visualise 2-D flow cytometry scatterplots with blue-red density colour scale.
}
\details{
For a complete list of customisation arguments see
\code{\link{plotCyto2d,flowFrame-method}}.
}
\seealso{
\code{\link{plotCyto2d,flowFrame-method}}

\code{\link{plotCyto2d,flowSet-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
