% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCytoComp-methods.R
\docType{methods}
\name{plotCytoComp,flowFrame-method}
\alias{plotCytoComp,flowFrame-method}
\title{Plot Compensation - flowFrame Method}
\usage{
\S4method{plotCytoComp}{flowFrame}(x, compensate = FALSE,
  spfile = NULL, transList = NULL, mfrow, popup = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}}
containing gated compensation controls and an unstained control.}

\item{compensate}{logical indicating whether the samples should be
compensated prior to plotting, set to FALSE by default. If no spillover
matrix is supplied to the spfile argument the spillover matrix will
extracted from the samples.}

\item{spfile}{name of spillover matrix csv file including .csv file extension
to use as a starting point for editing. If \code{spfile} is not supplied
the spillover matrix will be extracted directly from the \code{flowFrame}.}

\item{transList}{object of class
\code{\link[flowCore:transformList-class]{transformList}} or
\code{\link[flowWorkspace]{transformerList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} which was used to
transform the fluorescent channels of the supplied flowFrame. This
transList object will be used internally to ensure axes labels of the plot
are appropriately transformed. The transList will NOT be applied to the
flowFrame internally and should be applied to the flowFrame prior to
plotting.}

\item{mfrow}{vector of grid dimensions \code{c(#rows,#columns)} for each
plot.}

\item{popup}{logical indicating whether plots should be constructed in a
pop-up window.}

\item{...}{additional arguments passed to \code{\link{plotCyto2d,flowFrame-method}}.}
}
\description{
Plot each compensation control in all fluorescent channels to identify any
potential compensation issues. The unstained control is overlayed in black as
a reference.
}
\examples{
\dontrun{
plotCytoComp(fr, transList = trans, compensate = TRUE, spfile = "Spillover Matrix.csv")
}

}
\seealso{
\code{\link{plotCyto2d,flowFrame-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
