% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCytoComp-methods.R
\name{plotCytoComp}
\alias{plotCytoComp}
\title{Plot Compensation}
\usage{
plotCytoComp(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}} or
\code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} containing gated
compensation controls and an unstained control.}

\item{...}{additional method-specific arguments.}
}
\description{
Plot each compensation control in all fluorescent channels to identify any
potential compensation issues. The unstained control is overlayed in black as
a reference.
}
\seealso{
\code{\link{plotCytoComp,flowFrame-method}}

\code{\link{plotCytoComp,flowSet-method}}

\code{\link{plotCytoComp,GatingSet-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
