% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCytoExprs-methods.R
\docType{methods}
\name{plotCytoExprs,flowSet-method}
\alias{plotCytoExprs,flowSet-method}
\title{Plot Population Density Distribution in All Channels - flowSet Method}
\usage{
\S4method{plotCytoExprs}{flowSet}(x, channels = NULL, transList = NULL,
  merge = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowSet-class]{flowSet}}.}

\item{channels}{a vector channels to use to construct the plots, set to all
channels by default.}

\item{transList}{object of class
\code{\link[flowCore:transformList-class]{transformList}} or
\code{\link[flowWorkspace]{transformerList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} which was used to
transform the fluorescent channels of the supplied flowFrame. This
transList object will be used internally to ensure axes labels of the plot
are appropriately transformed. The transList will NOT be applied to the
flowFrame internally and should be applied to the flowFrame prior to
plotting.}

\item{merge}{logical indicating whether flowFrames should be merged prior to
plotting, set to TRUE by default.}

\item{...}{additional arguments passed to \code{\link{plotCyto1d,flowFrame-method}}.}
}
\description{
Plot Population Density Distribution in All Channels - flowSet Method
}
\examples{
\dontrun{
plotCytoExprs(fs, transList = trans)
}

}
\seealso{
\code{\link{plotCyto1d,flowFrame-method}}

\code{\link{plotCytoExprs,flowFrame-method}}

\code{\link{plotCytoExprs,GatingHierarchy-method}}

\code{\link{plotCytoExprs,GatingSet-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
