% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCytoExprs-methods.R
\name{plotCytoExprs}
\alias{plotCytoExprs}
\title{Plot Population Density Distribution in All Channels}
\usage{
plotCytoExprs(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{...}{additional method-specific arguments.}
}
\description{
Plot Population Density Distribution in All Channels
}
\seealso{
\code{\link{plotCytoExprs,flowFrame-method}}

\code{\link{plotCytoExprs,flowSet-method}}

\code{\link{plotCytoExprs,GatingHierarchy-method}}

\code{\link{plotCytoExprs,GatingSet-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
