% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCytoGates-methods.R
\docType{methods}
\name{plotCytoGates,GatingSet-method}
\alias{plotCytoGates,GatingSet-method}
\title{Plot Flow Cytometry Gating Strategies - GatingSet Method}
\usage{
\S4method{plotCytoGates}{GatingSet}(x, gtfile = NULL, merge = TRUE,
  overlay = NULL, title = "Gating Strategy", main, mfrow,
  popup = FALSE, col = NA, fill = "blue", legend = TRUE,
  text.legend, cex.legend = 1.2, ...)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{gtfile}{name of the gatingTemplate csv file used to gate \code{x}. If
not supplied the gating strategy will be obtained directly from the
GatingSet.}

\item{merge}{logical indicating whether the flowFrames should be merged prior
to plotting to yield a single plot, set to FALSE by default.}

\item{overlay}{name(s) of population(s) to overlay onto gating strategy for
back-gating.}

\item{title}{character string to use as the title for the plot layout, set to
"Gating Strategy" by default.}

\item{main}{vector of titles to use above each plot.}

\item{mfrow}{a vector of the length 2 indicating the dimensions of the grid
for plotting \code{c(#rows, #columns)}.}

\item{popup}{logical indicating whether the gating strategy should be plotted
in a pop-up window, set to FALSE by default.}

\item{col}{colour of points in 2D plots set to NA to use default red-blue
colour scale. Control the fill colour of overlays by supplying multiple
colours to this argument (e.g. c("blue","red")).}

\item{fill}{fill colour for 1D density distributions. Control the colour of
overlays by supplying multiple colours to this argument (e.g. c(NA,"red")).}

\item{legend}{logical indicating whether a legend should be included when an
overlay is supplied.}

\item{text.legend}{vector of character strings to use for legend when an
overlay is supplied.}

\item{cex.legend}{character expansion for legend text, set to 1.2 by default.}

\item{...}{extra arguments passed to plotCyto, see
\code{\link{plotCyto1d,flowFrame-method}} and
\code{\link{plotCyto2d,flowFrame-method}} for more details.}
}
\description{
\code{plotCytoGates} automatically plots the entire gating strategy and has
full support for back-gating through \code{overlay}.
}
\seealso{
\code{\link{plotCytoGates,GatingHierarchy-method}}

\code{\link{plotCyto1d,flowFrame-method}}

\code{\link{plotCyto2d,flowFrame-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
