% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCytoGates-methods.R
\name{plotCytoGates}
\alias{plotCytoGates}
\title{Plot Flow Cytometry Gating Strategies}
\usage{
plotCytoGates(x, ...)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{...}{additional method specific arguments.}
}
\description{
\code{plotCytoGates} automatically plots the entire gating strategy and has
full support for back-gating through \code{overlay}.
}
\seealso{
\code{\link{plotCytoGates,GatingHierarchy-method}}

\code{\link{plotCytoGates,GatingSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
