% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGates-methods.R
\docType{methods}
\name{plotGates,rectangleGate-method}
\alias{plotGates,rectangleGate-method}
\title{Plot rectangleGate Objects onto an Existing plot}
\usage{
\S4method{plotGates}{rectangleGate}(x, channels, col.gate = "red",
  lwd.gate = 2.5, lty.gate = 1, pts.gate = FALSE, pch.gate = 16,
  cex.gate = 1)
}
\arguments{
\item{x}{an object of class
\code{\link[flowCore:rectangleGate]{rectangleGate}}.}

\item{channels}{fluorescent channels to used to construct the plot.}

\item{col.gate}{indicates the colour of the gate to be constructed, set to
\code{"red"} by default.}

\item{lwd.gate}{numeric to adjust line thickness of gates, set to \code{2.5}
by default.}

\item{lty.gate}{integer [0,6] which controls the line type, set to \code{1}
to draw solid lines by default.}

\item{pts.gate}{logical indicating whether points should be included when
plotting the gates, set to \code{FALSE} by default.}

\item{pch.gate}{integer [0,25] passed to pch to control the shape of the
points, set to \code{16} to draw filled circles by default. For other
shapes refer to \code{\link[graphics:points]{?pch}}.}

\item{cex.gate}{numeric character expansion to control the size of the points
in the drawn gate, set to \code{1} by default.}
}
\description{
Plot rectangleGate Objects onto an Existing plot
}
\seealso{
\code{\link{plotGates,polygonGate-method}}

\code{\link{plotGates,ellipsoidGate-method}}

\code{\link{plotGates,list-method}}

\code{\link{plotGates,filters-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
