% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGates-methods.R
\name{plotGates}
\alias{plotGates}
\title{Plot Gate Objects onto an Existing Plot}
\usage{
plotGates(x, ...)
}
\arguments{
\item{x}{gate object of class
\code{\link[flowCore:rectangleGate]{rectangleGate}},
\code{\link[flowCore:polygonGate]{polygonGate}},
\code{\link[flowCore:ellipsoidGate]{ellipsoidGate}}, \code{list} or
\code{\link[flowCore:filters-class]{filters}}.}

\item{...}{additional method-specific arguments.}
}
\description{
Plot Gate Objects onto an Existing Plot
}
\seealso{
\code{\link{plotGates,rectangleGate-method}}

\code{\link{plotGates,polygonGate-method}}

\code{\link{plotGates,ellipsoidGate-method}}

\code{\link{plotGates,list-method}}

\code{\link{plotGates,filters-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
