% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLabels-methods.R
\docType{methods}
\name{plotLabels,flowFrame,filters-method}
\alias{plotLabels,flowFrame,filters-method}
\title{Plot Labels for Gated Populations - filters Method}
\usage{
\S4method{plotLabels}{flowFrame,filters}(x, gates, channels,
  alias = NULL, format.text = NULL, x.text = NULL, y.text = NULL,
  font.text = 2, col.text = "black", cex.text = 0.8, alpha = 0.6)
}
\arguments{
\item{x}{a \code{\link[flowCore:flowFrame-class]{flowFrame}} gated in the
existing plot.}

\item{gates}{an object of class \code{\link[flowCore:filters-class]{filters}}
containing objects of class
\code{\link[flowCore:rectangleGate]{rectangleGate}},
\code{\link[flowCore:polygonGate]{polygonGate}} or
\code{\link[flowCore:ellipsoidGate]{ellipsoidGate}}.}

\item{channels}{a vector indicating the fluorescent channel(s) to be used for
gating.}

\item{alias}{the name of the gated population.}

\item{format.text}{indicates the type of text to include in the label, can be
either \code{"alias"}, \code{"percent"}, \code{"count"},
\code{c("alias","percent")} or \code{c("alias","count")}. Set to
\code{c("alias","percent")} by default.}

\item{x.text}{vector containing the x co-ordiantes for the plot labels. Set
to \code{NULL} by default to place labels in the center of the gates.}

\item{y.text}{vector containing the x co-ordiantes for the plot labels. Set
to \code{NULL} by default to place labels in the center of the gates.}

\item{font.text}{integer [1,2,3,4] passed to \code{text} to alter the font,
set to \code{2} by default for a bold font.}

\item{col.text}{specify text colour in label for each gate, defaults to
\code{"black"} for all gates.}

\item{cex.text}{numeric character expansion used to control the size of the
text in the labels, set to \code{0.8} by default. See \code{?text} for
details.}

\item{alpha}{numeric [0,1] controls the transparency of the background, set
to \code{0.6} by default.}
}
\value{
add label to existing plot with information about gated population.
}
\description{
\code{plotLabels} takes on a \code{flowFrame} object, population name
\code{alias}, \code{channels} and a gate object to construct a text label for
the plot with the population name and frequency.
}
\seealso{
\code{\link{plotLabels,flowFrame,rectangleGate-method}}

\code{\link{plotLabels,flowFrame,polygonGate-method}}

\code{\link{plotLabels,flowFrame,ellipsoidGate-method}}

\code{\link{plotLabels,flowFrame,list-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
