% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plotLabels}
\alias{plotLabels}
\title{Label Plots with Population Name and Frequency.}
\usage{
plotLabels(fr, channels, alias, gate)
}
\arguments{
\item{fr}{a \code{flowFrame} object containing the flow cytometry data for plotting and gating.}

\item{channels}{a vector indicating the fluorescent channel(s) to be used for gating.}

\item{alias}{the name of the gated population.}

\item{gate}{an object of class \code{polygonGate}, \code{rectangleGate} or \code{ellipsoidGate}.}

\item{...}{additional arguments for plotDens.}
}
\value{
add text box label to existing plot with population name and frequency.
}
\description{
\code{plotLabels} takes on a \code{flowFrame} object, population name \code{alias}, \code{channels} and a gate object to construct
a text label for the plot with the population name and frequency.
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
\keyword{cytometry,}
\keyword{draw,}
\keyword{drawPlot,}
\keyword{flow}
\keyword{gates,}
\keyword{gating,}
\keyword{labels}
\keyword{manual,}
\keyword{openCyto,}
