% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{removeGate}
\alias{removeGate}
\title{Remove gate and edit gatingTemplate .csv file}
\usage{
removeGate(gs, alias = NULL, gtfile = NULL, children = TRUE)
}
\arguments{
\item{gs}{an object of class \code{GatingSet}.}

\item{alias}{name(s) of the gate to remove (e.g. "Single Cells").}

\item{gtfile}{name of the \code{gatingTemplate} csv file (e.g. "gatingTemplate.csv").}

\item{children}{logical indicating whether descendant populations should also be removed from the gtfile, set to \code{TRUE} by default.}
}
\value{
an object of class \code{gatingSet} with gate and children removed, as well as gatingTemplate file with population removed.
}
\description{
Remove gate and edit gatingTemplate .csv file
}
