% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{restrictFrame}
\alias{restrictFrame}
\title{Restrict flowFrame}
\usage{
restrictFrame(fr, size = 5000)
}
\arguments{
\item{fr}{object of class \code{flowFrame}.}

\item{size}{numeric indicating the number of events to keep. If \code{size > nrow(fr)}, size is set to \code{nrow(fr)}.}
}
\value{
restricted \code{flowFrame}.
}
\description{
Restrict flowFrame
}
