% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{sampleFrame}
\alias{sampleFrame}
\title{Sample a flowFrame}
\usage{
sampleFrame(fr, size = 250000)
}
\arguments{
\item{fr}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}}.}

\item{size}{numeric indicating the number of events to keep. If \code{size >
nrow(fr)}, size is set to \code{nrow(fr)}.}
}
\value{
\code{\link[flowCore:flowFrame-class]{flowFrame}} restricted to
  \code{size} events.
}
\description{
Sample a flowFrame
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
