% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\docType{methods}
\name{selectChannels,flowFrame-method}
\alias{selectChannels,flowFrame-method}
\title{Select Fluorescent Channel for Compensation Controls - flowFrame Method}
\usage{
\S4method{selectChannels}{flowFrame}(x)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}}.}
}
\value{
selected channel associated with the supplied flowFrame.
}
\description{
Select Fluorescent Channel for Compensation Controls - flowFrame Method
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
