% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\docType{methods}
\name{selectChannels,flowSet-method}
\alias{selectChannels,flowSet-method}
\title{Select Fluorescent Channel for Compensation Controls - flowSet Method}
\usage{
\S4method{selectChannels}{flowSet}(x)
}
\arguments{
\item{x}{object of class
\code{\link[flowCore:flowSet-class]{flowSet}} containing compensation
controls.}
}
\value{
vector of channels in order of flowSet.
}
\description{
Select Fluorescent Channel for Compensation Controls - flowSet Method
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
