% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\docType{methods}
\name{selectChannels,GatingSet-method}
\alias{selectChannels,GatingSet-method}
\title{Select Fluorescent Channel for Compensation Controls - GatingSet Method}
\usage{
\S4method{selectChannels}{GatingSet}(x)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} containing
compensation controls.}
}
\value{
vector of channels in order of GatingSet.
}
\description{
Select Fluorescent Channel for Compensation Controls - GatingSet Method
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
