% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{selectChannels}
\alias{selectChannels}
\title{Select flurescent channels for compensation controls}
\usage{
selectChannels(fs)
}
\arguments{
\item{fs}{object of class \code{flowSet} containing compensation controls}
}
\value{
add selected channels to \code{pData(fs)$channel}
}
\description{
Select flurescent channels for compensation controls
}
