% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{selectFrames}
\alias{selectFrames}
\title{Select flowFrames Based on pData}
\usage{
selectFrames(fs, pData)
}
\arguments{
\item{fs}{an object of class \code{\link[flowCore:flowSet-class]{flowSet}}.}

\item{pData}{vector of form \code{c("column","row")} indicating the rows of
\code{pData(fs)} to extract.}
}
\value{
an object of calss \code{\link[flowCore:flowSet-class]{flowSet}}
  containing the extracted \code{\link[flowCore:flowFrame-class]{flowFrame}}
  objects.
}
\description{
Select flowFrames Based on pData
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
