% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillover_compute-methods.R
\docType{methods}
\name{spillover_compute,GatingSet-method}
\alias{spillover_compute,GatingSet-method}
\title{Compute Spillover Matrix - GatingSet Method}
\usage{
\S4method{spillover_compute}{GatingSet}(x, parent = NULL,
  axes_trans = NULL, channel_match = NULL,
  spillover = "Spillover-Matrix.csv", ...)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} containing pre-gated
single stain compensation controls and a universal unstained control.
Currently, spillover_compute does not pre-gate samples to obtain a
homogeneous cell population for downstream calculations. We therefore
recommend pre-gating samples based on FSC and SSC parameters prior to
passing them to spillover_compute and indicate the population of interest
using the \code{parent} argument.}

\item{parent}{name of the pre-gated population to use for downstream
calculations, set to the last node of the GatingSet by default (e.g.
"Single Cells").}

\item{axes_trans}{object of class
\code{\link[flowWorkspace:transformerList]{transformerList}} generated by
\code{estimateLogicle} to transform fluorescent channels for gating.
\code{axes_trans} is required if logicle transformation has already been
applied to \code{x} using estimateLogicle. \code{spillover_compute} will
automatically call \code{\link[flowCore:logicleTransform]{estimateLogicle}}
internally to transform channels prior to gating, if \code{axes_trans} is
supplied it will be used for the transformation instead.}

\item{channel_match}{name of .csv file containing the names of the samples in
a column called "name" and their matching channel in a column called
"channel". \code{spillover_compute} will the guide you through the channel
selection process and generate a channel match file called
"Compensation-Channels.csv" automatically. If you already have a complete
channel_match and would like to bypass the channel selection process,
simply pass the name of the channel_match to this argument (e.g.
"Compensation-Channels.csv").}

\item{spillover}{name of the output spillover csv file, set to
\code{"Spillover-Matrix.csv"} by default.}

\item{...}{additional arguments passed to
\code{\link{cyto_plot,flowFrame-method}}.}
}
\value{
spillover matrix object and \code{"Spillover-Matrix.csv"} file.
}
\description{
\code{spillover_compute} uses the method described by Bagwell & Adams 1993 to
calculate the fluorescent spillover matrix using a reference universal
unstained control and single stain compensation controls.
}
\details{
Calculate spillover matrix using
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} containing gated
single stain compensation controls and an unstained control.
\code{spillover_compute} uses the method described by Bagwell & Adams 1993 to
calculate fluorescent spillover values using single stain compensation
controls and a universal unstained control. \code{spillover_compute} begins
by the user selecting which fluorescent channel is associated with each
control from a dropdown menu. Following channel selection,
\code{spillover_compute} runs through each control and plots the density
distribution of the unstained control in red and the compensation control in
blue. Users can then gate the positive signal for spillover calculation using
an interval gate. The percentage spillover is calculated based on the median
fluorescent intensities of the stained populations and the universal
unstained sample. The computed spillover matrix is returned as an R object
and written to a named .csv file for future use.
}
\examples{
library(CytoRSuiteData)

# Bypass directory check for external files
options("CytoRSuite_wd_check" = FALSE)

# Don't run - skips the gating process
options("CytoRSuite_interact" = FALSE)

# Load in compensation controls
fs <- Compensation
gs <- GatingSet(Compensation)

# Gate using gate_draw
gt <- Compensation_gatingTemplate
gating(gt, gs)

# Channel match fille
cmfile <- system.file("extdata",
  "Compensation-Channels.csv",
  package = "CytoRSuiteData"
)

# Compute fluorescent spillover matrix
spill <- spillover_compute(gs,
  parent = "Single Cells",
  channel_match = cmfile,
  spillover = "Example-spillover.csv"
)

# Compensate samples
gs <- compensate(gs, spill)

# Return CytoRSuite_wd_check to default
options("CytoRSuite_wd_check" = TRUE)

# Return CytoRSuite_interact to default
options("CytoRSuite_interact" = TRUE)
}
\references{
C. B. Bagwell \& E. G. Adams (1993). Fluorescence spectral
  overlap compensation for any number of flow cytometry parameters. in:
  Annals of the New York Academy of Sciences, 677:167-184.
}
\seealso{
\code{\link{cyto_plot,flowFrame-method}}.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
