% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillover_compute-methods.R
\name{spillover_compute}
\alias{spillover_compute}
\title{Compute Spillover Matrix}
\usage{
spillover_compute(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{...}{additional method-specific arguments for spillover_compute.}
}
\description{
\code{spillover_compute} uses the method described by Bagwell & Adams 1993 to
calculate the fluorescent spillover matrix using a reference universal
unstained control and single stain compensation controls.
}
\details{
\code{spillover_compute} begins by the user selecting which fluorescent
channel is associated with each control from a dropdown menu. Following
channel selection, \code{spillover_compute} runs through each control and
plots the density distribution of the unstained control in red and the
compensation control in blue. Users can then gate the positive signal for
spillover calculation using an interval gate. The percentage spillover is
calculated based on the median fluorescent intensities of the stained
populations and the universal unstained sample. The computed spillover matrix
is returned as an R object and written to a named .csv file for future use.
\code{spillover_compute} has methods for both
\code{\link[flowCore:flowSet-class]{flowSet}} and
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} objects so refer to
their respective help pages for more information.
}
\references{
C. B. Bagwell \& E. G. Adams (1993). Fluorescence spectral
  overlap compensation for any number of flow cytometry parameters. in:
  Annals of the New York Academy of Sciences, 677:167-184.
}
\seealso{
\code{\link{spillover_compute,flowSet-method}}

\code{\link{spillover_compute,GatingSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
