% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillover_edit-methods.R
\docType{methods}
\name{spillover_edit,GatingSet-method}
\alias{spillover_edit,GatingSet-method}
\title{Edit Spillover Matrix GatingSet Method}
\usage{
\S4method{spillover_edit}{GatingSet}(x, parent = NULL,
  channel_match = NULL, spillover = NULL, display = 1,
  axes_trans = NULL, ...)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the pre-gated population to be plotted (e.g. "Single
Cells").}

\item{channel_match}{name of .csv file containing the names of the samples in
a column called "name" and their matching channel in a column called
"channel". \code{spillover_edit} will the guide you through the channel
selection process and generate a channel match file called
"Compensation-Channels.csv" automatically. If you already have a complete
channel_match and would like to bypass the channel selection process,
simply pass the name of the channel_match to this argument (e.g.
"Compensation-Channels.csv").}

\item{spillover}{name of spillover matrix csv file including .csv file
extension to use as a starting point for editing. If \code{spillover} is
not supplied the spillover matrix will be extracted directly from the
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} and the edited
matrix saved to \code{"Spillover-Matrix.csv"}.}

\item{display}{numeric indicating the number of events to plot, set to 5000
events by default.}

\item{axes_trans}{object of class
\code{\link[flowCore:transformList-class]{transformList}} or
\code{\link[flowWorkspace:transformerList]{transformerList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} which was used to
transform the fluorescent channels of the flowSet.}

\item{...}{additional arguments passed to
\code{\link{cyto_plot,flowFrame-method}}.}
}
\value{
save edited spillover matrix to .csv file named
  "Spillover-matrix.csv" or spillover.
}
\description{
Edit spillover matrices in real-time using a shiny interface.
}
\details{
\code{spillover_edit} provides an interactive shiny interface for editing
fluorescent spillover matrices. \code{spillover_edit} takes on either a
\code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} containing
untransformed single stain compensation controls and a universal unstained
control. It is recommended that samples be pre-gated based on FSC and SSC
parameters to obtain a homogeneous population for calculation of fluorescent
spillover. Users begin by selecting the unstained control and a stained
control from dropdown menus of sample names. \code{spillover_edit} leverages
\code{cyto_plot} to plot the stained sample and overlay the unstained control
in black. Users should then select the channel associated with the selected
control on the \code{x axis} and go through all other channels on the \code{y
axis}. The displayed spillover matrix is extracted directly from the
\code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} unless another
spillover matrix is supplied through the spillover argument. To edit the
spillover matrix simply modify the appropriate cell in the the table. The new
spillover matrix will be re-applied to the samples with each edit and
automatically re-plotted so you can track changes in real-time. To aid in
selection of an appropriate spillover value, the median fluorescent intensity
of the unstained control is indicated by a red line and median fluorescent
intensity of the stained control is tracked with a purple line. These
features can be turned off by de-selecting the check boxes. Changes to the
spillover matrix are automatically saved to a csv file called
\code{"Spillover-Matrix.csv"} in the case where the \code{spillover} is not
specified or to the same name as the specified \code{spillover}.
}
\examples{
\dontrun{
library(CytoRSuiteData)

# Load in compensation controls
fs <- Compensation
gs <- GatingSet(fs)

# Gate single cells using gate_draw
gt <- Compensation_gatingTemplate
gating(gt, gs)

# Channel match file
cmfile <- system.file("extdata", "Compensation-Channels.csv", package = "CytoRSuiteData")

# Edit spillover matrix
spillover_edit(gs,
  parent = "Single Cells",
  channel_match = cmfile
)

}
}
\seealso{
\code{\link{spillover_edit,flowSet-method}}

\code{\link{cyto_plot,flowFrame-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
