% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillover_spread-methods.R
\docType{methods}
\name{spillover_spread_compute,flowSet-method}
\alias{spillover_spread_compute,flowSet-method}
\title{Compute Spillover Spreading Matrix}
\usage{
\S4method{spillover_spread_compute}{flowSet}(x, axes_trans = NULL,
  channel_match = NULL, compensated = FALSE, spillover = NULL,
  spillover_spread = "Spillover-Spread-Matrix.csv", ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowSet-class]{flowSet}}
containing pre-gated single stain compensation controls and a universal
unstained control.}

\item{axes_trans}{object of class
\code{\link[flowCore:transformList-class]{transformList}} generated by
\code{estimateLogicle} to transform fluorescent channels for gating.
\code{axes_trans} is required if logicle transformation has already been
applied to \code{x} using estimateLogicle. \code{spillover_cspread} will
automatically call \code{\link[flowCore:logicleTransform]{estimateLogicle}}
internally to transform channels prior to gating, if \code{axes_trans} is
supplied it will be used for the transformation instead.}

\item{channel_match}{name of .csv file containing the names of the samples in
a column called "name" and their matching channel in a column called
"channel". \code{spillover_spread} will the guide you through the channel
selection process and generate a channel match file called
"Compensation-Channels.csv" automatically. If you already have a complete
channel_match and would like to bypass the channel selection process,
simply pass the name of the channel_match to this argument (e.g.
"Compensation-Channels.csv").}

\item{compensated}{logical indicating whether the supplied compensation
controls have been compensated. Compensation controls must be compensated
prior tro calculation of spillover spreading matrix. If \code{compensated}
is set to FALSE, the compensation controls will be compensated internally
using the supplied \code{spillover}. If no spillover matrix is supplied
\code{spillover_spread_compute} will look for a "Spillover-Matrix.csv" file
or as a last resort use the spillover matrix attached to compensation
controls.}

\item{spillover}{name of the output spillover matrix csv file to be used
internally to compensate the compensation controls. If no spillover matrix
is supplied \code{spillover_spread_compute} will look for a
"Spillover-Matrix.csv" file or as a last resort use the spillover matrix
attached to compensation controls.}

\item{spillover_spread}{name of the csv file to which the spillover spreading
matrix will be saved, set to \code{"Spillover-Spread-Matrix.csv"} by
default.}

\item{...}{additional arguments passed to
\code{\link{cyto_plot,flowFrame-method}}.}
}
\value{
spillover spreading matrix object, write spillover spreading matrix
  to csv file called \code{"Spillover-Spread-Matrix.csv"} and plot the
  resulting matrix as a heatmap.
}
\description{
\code{spillover_spread} computes the spillover spreading matrix as described
by Nguyen et al. (2013) using gated and compensated compensation controls
which include a reference universal unstained control.
}
\references{
Nguyen R, Perfetto S, Mahnke YD, Chattopadhyay P & Roederer M,
  (2013). Quantifying spillover spreading for comparing instrument
  performance and aiding in multicolor panel design. Cytometry A,
  83(3):306-15.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
