% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spillover_spread-methods.R
\name{spillover_spread_compute}
\alias{spillover_spread_compute}
\title{Compute Spillover Spreading Matrix}
\usage{
spillover_spread_compute(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{...}{additional method-specific arguments for spillover_spread.}
}
\description{
\code{spillover_spread} computes the spillover spreading matrix as described
by Nguyen et al. (2013) using gated and compensated compensation controls
which include a reference universal unstained control.
}
\references{
Nguyen R, Perfetto S, Mahnke YD, Chattopadhyay P & Roederer M,
  (2013). Quantifying spillover spreading for comparing instrument
  performance and aiding in multicolor panel design. Cytometry A,
  83(3):306-15.
}
\seealso{
\code{\link{spillover_spread_compute,flowSet-method}}

\code{\link{spillover_spread_compute,GatingSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
