% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_spillover_compute.R
\name{watershed_1d}
\alias{watershed_1d}
\title{Automated gating using watershed algorithm}
\usage{
watershed_1d(
  x,
  x_control = NULL,
  channel,
  smooth = 3,
  bins = 128,
  limits = c(NA, NA),
  bandwidth = NA,
  alpha = 0.05,
  beta = 0.3,
  peak = FALSE,
  plot = FALSE,
  ...
)
}
\arguments{
\item{x}{a cytoset object.}

\item{x_control}{a cytoset object representing background.}

\item{channel}{name of the channel or marker to gate.}

\item{smooth}{numeric to control the smoothness of the kernel density
estimate, set to 3 by default.}

\item{bins}{number of bins to used for the kernel density estimate, set to
\code{128} by default.}

\item{limits}{numeric limits over which the density is to be estimated, set
to the range of values in \code{x} by default.}

\item{bandwidth}{optional bandwidth for the density estimate.}

\item{alpha}{density cutoff as a proportion of the maximum density to
eliminate small peaks, set to \code{0.05} by default.}

\item{beta}{numeric to control the proportion of the x axis to be allocated
to negative expression values, set to 0.3 by default.}

\item{peak}{logical to indicate whether the last gate should have the peak
density as its lower bound, set to FALSE by default.}

\item{plot}{logical to plot the kernel density estimate, density peaks, beta
and segmentation boundaries, set to FALSE by default.}
}
\value{
a set of \code{filters} containing \code{rectangleGate} objects to
  segment the specified \code{channel}.
}
\description{
Automated gating using watershed algorithm
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
