% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Activation}
\alias{Activation}
\title{Activation}
\format{A flowSet containing two flowFrame objects.}
\usage{
Activation
}
\description{
An experiment performed to assess the in vitro activation of OT-I and OT-II transgenic T cells co-cultured with
OVA-presenting dendritic cells (DCs). This data set is comprised of two samples saved as a \code{flowSet} object.
The first sample is an unactivated control whilst the second is an activated sample.
}
\details{
Pooled DCs & T cells were labelled as follows:
\itemize{
\item Hoechst 33258, a DNA intercalating dye which fluoresces in Alexa Fluor 405-A and Alexa Fluor 430-A channels
\item CD11c-APC-Cy7 to identify DCs within the samples.
\item Va2-PE to identify cells which express the OVA-specific TCR (i.e. OT-I and OT-II T cells).
\item CD8-FITC to identify the CD8+ OT-I transgenic T cells (Alexa Fluor 488-A).
\item CD4-Alexa Fluor 700 to identify the CD4+ OT-II transgenic T cells (Alexa Fluor 700-A)
\item CD44-APC to assess T cell activation (Alexa Fluor 647-A)
\item CD69-PerCP Cy5.5 to assess T cell activation (7-AAD-A)
}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
\keyword{datasets}
