% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{checkChannels}
\alias{checkChannels}
\title{Check Supplied Channels}
\usage{
checkChannels(x, channels, plot)
}
\arguments{
\item{x}{an object of class
\code{\link[flowCore:flowFrame-class]{flowFrame}}.}

\item{channels}{vector of channel names (e.g. c("PE-A","APC-A")) or marker
names.}

\item{plot}{logical indicating whether the channels will be used to construct
a plot, set to TRUE by default to accept 1 or 2 channels only.}
}
\value{
vector of valid channel names.
}
\description{
\code{checkChannels} will check whether the supplied channels are valid for
the \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowset}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} . \code{checkChannels}
will also return the channels if the marker names are supplied.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
