% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{checkChannels}
\alias{checkChannels}
\title{Check Supplied Channels.}
\usage{
checkChannels(x, channels)
}
\arguments{
\item{x}{an object of class \code{flowFrame}, \code{flowSet} or \code{GatingSet}.}

\item{channels}{vector of length 1 or 2 indicating the names of the channels to use for plotting and gating.}
}
\value{
Stops gating process if channels are incorrect or returns nothing.
}
\description{
Check Supplied Channels.
}
