% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{checkGateType}
\alias{checkGateType}
\title{Check Supplied Gate Types(s).}
\usage{
checkGateType(gate_type)
}
\arguments{
\item{gate_type}{vector indicating the types of gates to construct using \code{drawGate}.}
}
\value{
Stops gating process if gate_type is incorrect or returns \code{gate_type} as full lower case name(s).
}
\description{
Check Supplied Gate Types(s).
}
