% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{checkGateType}
\alias{checkGateType}
\title{Check Gate Types(s) Supplied drawGate.}
\usage{
checkGateType(gate_type, alias)
}
\arguments{
\item{gate_type}{vector indicating the types of gates to construct using \code{drawGate}.}

\item{alias}{names of the populations to be gated.}
}
\value{
Stop gating process if gate_type is incorrect or returns \code{gate_type} as full lower case name(s). If a single gate_type is supplied for multiple populations,
the same gate_type will be used for all populations.
}
\description{
Check Gate Types(s) Supplied drawGate.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
