% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-functions.R
\name{checkOSGD}
\alias{checkOSGD}
\title{Check Operating System & Open New Graphics Device}
\usage{
checkOSGD()
}
\description{
\code{checkOSGD} is used internally by plotCyto to open an OS-specific
interactive garphics device to facilitate gate drawing. Mac users will need
to install \href{https://www.xquartz.org/}{XQuartz} for this functionality.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
