% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSpillover-methods.R
\docType{methods}
\name{computeSpillover,flowSet-method}
\alias{computeSpillover,flowSet-method}
\title{Compute Spillover Matrix - flowSet Method}
\usage{
\S4method{computeSpillover}{flowSet}(x, trans = NULL, cmfile = NULL,
  spfile = "Spillover Matrix.csv", ...)
}
\arguments{
\item{x}{object of class \code{flowSet} containing pre-gated single stain compensation controls and a universal unstained control. Currently,
computeSpillover does not pre-gate samples to obtain a homogeneous cell population for downstream calculations. We therefore recommmend pre-gating
samples based on FSC and SSC parameters prior to passing them to computeSpillover (i.e. \code{x} should contain events for single cells only).
Passing raw files to computeSpillover will result in inaccurate calculations of fluorescent spillover.}

\item{trans}{object of class \code{"transformList"} generated by \code{estimateLogicle} to transform fluorescent channels for gating. \code{trans}
is required if logicle transformation has already been applied to \code{x} using estimateLogicle. \code{computeSpillover} will automatically call
estimateLogicle internally to transform channels prior to gating, if \code{trans} is supplied it will be used for the transformation instead.}

\item{cmfile}{name of .csv file containing the names of the samples in a column called "name" and their matching channel in a column called "channel".
\code{computeSpillover} will the guide you through the channel selection process and generate a channel match file called "Compensation Channels.csv" 
automatically. If you already have a complete cmfile and would like to bypass the channel selection process, simply pass the name of the cmfile to 
this argument (e.g. "Compensation Channels.csv").}

\item{spfile}{name of the output spillover csv file, set to \code{"Spillover Matrix.csv"} by default.}

\item{...}{additional arguments passed to estimateLogicle to transform fluorescent channels prior to gating.}
}
\value{
spillover matrix object and \code{"Spillover Matrix.csv"} file.
}
\description{
Calculate spillover matrix using \code{GatingSet} containing gated single stain compensation controls and an unstained control.
\code{computeSpillover} uses the method described by Bagwell & Adams 1993 to calculate fluorescent spillover values using single 
stain compensation controls and a universal unstained control. \code{computeSpillover} begins by the user selecting which fluorescent
channel is associated with each control from a dropdown menu. Following channel selection, \code{computeSpillover} runs through each control
and plots a histogram which is gated by the user using an interval gate (refer to cytoSuite drawInterval documentation) to 
select the positive population to use for the spillover calculation. The percentage spillover is calculated based on the median fluorescent
intensities of the stained populations and the universal unstained sample. The computed spillover matrix is returned as an r object and written
to a named .csv file for future use.
}
\examples{
\dontrun{
fs <- Activation
gate1 <- drawGate(fs, alias = "Cells", channels = c("FSC-A","SSC-A"))
Cells <- Subset(fs, g1[[1]])
gate2 <- drawGate(Cells, alias = "Singlets", channels = c("FSC-A","FSC-H"))
Singlets <- Subset(Cells, g2[[1]])
spill <- computeSpillover(Singlets, spfile = "Example Spillover Matrix.csv")
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
