% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSpillover-methods.R
\docType{methods}
\name{computeSpillover,flowSet-method}
\alias{computeSpillover,flowSet-method}
\title{Caluclate Spillover Matrix Using Pre-gated Compensation Controls}
\usage{
\S4method{computeSpillover}{flowSet}(x, pdfile = NULL, spfile = NULL, ...)
}
\arguments{
\item{x}{object of class \code{flowSet} containing pre-gated compensation single stain controls as well as an unstained control.}

\item{pdfile}{\code{pData} csv file containing additional column \code{"channel"} indicating the fluorescent channel associated
with each sample. This channel should be set to \code{"Unstained"} for unstained controls.}

\item{spfile}{name of the output spillover csv file, set to \code{"Spillover Matrix.csv"} by default.}

\item{alias}{name of the gated population to use for downstream calculations, set to the last node of the GatingSet by default.}
}
\value{
spillover matrix object and \code{"Spillover Matrix.csv"} file as side effect.
}
\description{
Caluclate Spillover Matrix Using Pre-gated Compensation Controls
}
