% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSpillover-methods.R
\docType{methods}
\name{computeSpillover,GatingSet-method}
\alias{computeSpillover,GatingSet-method}
\title{Caluclate Spillover Matrix - flowSet Method}
\usage{
\S4method{computeSpillover}{GatingSet}(x, gtfile = NULL, alias = NULL,
  pdfile = NULL, spfile = NULL, ...)
}
\arguments{
\item{x}{object of class \code{GatingSet} containing gated compensation single stain controls as well as an unstained control.}

\item{gtfile}{openCyto \code{gatingTemplate} csv file to be used to gate samples prior to spillover calculation. If no
\code{gatingTemplate} is supplied samples will be gated on FSC-A/SSC-A and SSC-W/SSC-H using \code{drawGate}. The gating
strategy will be saved as a side effect in a gatingTemplate csv file called \code{"Compensation gatingTemplate.csv"}.}

\item{alias}{name of the gated population to use for downstream calculations, set to the last node of the GatingSet by default.}

\item{pdfile}{\code{pData} csv file containing additional column \code{"channel"} indicating the fluorescent channel associated
with each sample. This channel should be set to \code{"Unstained"} for unstained controls.}

\item{spfile}{name of the output spillover csv file, set to \code{"Spillover Matrix.csv"} by default.}
}
\value{
spillover matrix and \code{"Spillover Matrix.csv"} file.
}
\description{
Caluclate Spillover Matrix - flowSet Method
}
