% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compensation-functions.R
\name{computeSpillover}
\alias{computeSpillover}
\title{Caluclate Spillover Matrix}
\usage{
computeSpillover(fs, file = NULL)
}
\arguments{
\item{fs}{object of class \code{flowSet} containing compensation single stain controls as well as an unstained control.}
}
\value{
spillover matrix and \code{"Spillover Matrix.csv"} file.
}
\description{
Caluclate Spillover Matrix
}
