% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSpillover-methods.R
\name{computeSpillover}
\alias{computeSpillover}
\title{Compute Spillover Matrix}
\usage{
computeSpillover(x, ...)
}
\arguments{
\item{x}{object of class \code{flowSet} or \code{GatingSet}.}
}
\description{
Compute spillover matrix using single stain compensation controls and an unstained control.
}
\references{
C. B. Bagwell \& E. G. Adams (1993). Fluorescence spectral overlap compensation for any number of flow cytometry parameters. in: Annals of the New York Academy of Sciences, 677:167-184.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
