% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_apply}
\alias{cyto_apply}
\title{Apply a function over cytometry objects}
\usage{
cyto_apply(x, FUN, ..., simplify = TRUE, use.exprs = FALSE)
}
\arguments{
\item{x}{object of class \code{flowSet} or \code{cytoset}.}

\item{FUN}{function to apply to each \code{flowFrame} or \code{cytoframe} in
the supplied \code{flowSet} or \code{cytoset}.}

\item{...}{optional arguments to \code{FUN}.}

\item{simplify}{logical indicating whether the resulting list should be
simplified into a flowSet or bound into a matrix or data.frame using
\code{rbind}, set to TRUE by default.}

\item{use.exprs}{logical indicating whether the function should be applied to
the raw data in each flowFrame, set to FALSE by default.}
}
\description{
Apply a function over cytometry objects
}
\examples{
library(CytoExploreRData)

# Activation dataset
fs <- Activation

# Sample each flowFrame
fs <- cyto_apply(fs, cyto_sample)

}
\seealso{
\code{\link[flowCore:fsApply]{fsApply}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
