% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_barcode}
\alias{cyto_barcode}
\title{Barcode each file in a flowSet with a sample ID}
\usage{
cyto_barcode(x)
}
\arguments{
\item{x}{object of class \code{flowSet} to be barcoded.}
}
\value{
barcoded flowSet with \code{"Sample ID"} column added and annotated.
}
\description{
Adds a new parameter to each of the flowFrames in the flowSet called
\code{"Sample ID"} to barcode events from each flowFrame.
}
\examples{

# Load in CytoExploreRData to access files
library(CytoExploreRData)

# Load in samples
fs <- Activation

# Barcode
cyto_barcode(fs)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
