% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_barcode}
\alias{cyto_barcode}
\title{Barcode each file in a cytoset with a sample ID}
\usage{
cyto_barcode(x, type = "samples")
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}} to
be barcoded.}

\item{type}{indicates whether the \code{"samples"}, \code{"events"} or
\code{"both"} should be barcoded, set \code{"samples"} by default.}
}
\value{
barcoded cytoset with \code{"Sample ID"} and/or \code{"Event ID"}
  column added and annotated.
}
\description{
Adds a new parameter to each of the cytoframes in the cytoset called
\code{"Sample ID"} to barcode events from each cytoframe
}
\examples{

# Load in CytoExploreRData to access files
library(CytoExploreRData)

# Load in samples
fs <- Activation

# Barcode
cyto_barcode(fs)
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
