% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_calibrate}
\alias{cyto_calibrate}
\title{Calibrate channel ranges for cyto_plot}
\usage{
cyto_calibrate(x, parent = "root", type = "range", probs = c(0.01, 0.99), ...)
}
\arguments{
\item{x}{object of class \code{cytoframe}, \code{cytoset},
\code{GatingHierarchy} or \code{GatingSet} to use for the calibration. For
the best calibration is recommended that users supply samples containing
both negative and positive events in each channel.}

\item{type}{indicates the type of calibration to perform, options include
\code{"range"} or \code{"quantile"}. Range calibration simply uses the full
range of values across samples for the calibration. Quantile calibration
computes an lower and upper quantile for each channel, values falling
outside the calibration range are assigned the bottom or top colour.}

\item{probs}{vector of lower and upper probabilities passed to
\code{stats:quantile} to compute quantiles, set to \code{c(0.01, 0.99)} by
default.}

\item{...}{not in use.}

\item{name}{of the parent population to use for channel calibration when a
\code{GatingHierarchy} or \code{GatingSet} is supplied, set to the
\code{"root"} node by default.}
}
\value{
saves calibration settings for use by \code{\link{cyto_plot}}.
}
\description{
The density colour scale for points in \code{cyto_plot} can now be made to
represent fluorescent intensity of a third channel, instead of local density
in the 2D plotting space. In order for this feature to work properly, the
full range for the channel/marker of interest is required to appropriately
set the colour scale. \code{cyto_calibrate} performs this range calibration,
so that this information can be used in all downstream \code{cyto_plot}
calls.
}
\examples{
library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Calibration
cyto_calibrate(fs)

# Colour based on Hoechst-405 staining
cyto_plot(fs[1],
channels = c("FSC-A", "SSC-A"),
point_col = "Hoechst-405")

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
