% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_cbind}
\alias{cyto_cbind}
\alias{cyto_cbind.flowFrame}
\alias{cyto_cbind.flowSet}
\title{Bind new columns to cytoframe or cytoset}
\usage{
\method{cyto_cbind}{flowFrame}(x, cols = NULL)

\method{cyto_cbind}{flowSet}(x, cols = NULL)
}
\arguments{
\item{x}{object of class \code{cytoframe}, \code{cytoset} or a list of
\code{cytoframes}.}

\item{cols}{matrix of columns to be added to \code{x} can be supplied in a
list for \code{cytoset} and \code{list} methods.}
}
\description{
Bind new columns to cytoframe or cytoset
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
