% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channels-helpers.R
\name{cyto_channel_match}
\alias{cyto_channel_match}
\title{Create a csv file assigning a channel to each compensation control}
\usage{
cyto_channel_match(x, save_as = NULL)
}
\arguments{
\item{x}{object of \code{\link[flowWorkspace:cytoset]{cytoset}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{save_as}{name to use for the saved channel match csv file, set to
\code{"date-Channel-Match.csv"}.}
}
\value{
update \code{cyto_details} of \code{cytoset} or \code{GatingSet},
  write channel matching to csv file and return channel matching as a
  data.frame.
}
\description{
Create a csv file assigning a channel to each compensation control
}
\examples{
if(interactive()) {
library(CytoExploreRData)

# Compensation Gatingset
gs <- load_gs(system.file("extdata/Compensation-GatingSet",
                          package = "CytoExploreRData"))

# Select a channel for each control from dropdown menu
cyto_channel_match(gs)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
