% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_channel_match}
\alias{cyto_channel_match}
\title{Table Editor for Channel Match File Construction}
\usage{
cyto_channel_match(x, save_as = NULL)
}
\arguments{
\item{x}{object of \code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{save_as}{name to use for the saved channel match csv file, set to
\code{"date-Channel-Match.csv"}.}
}
\value{
save constructed channel_match csv file.
}
\description{
Table Editor for Channel Match File Construction
}
\examples{
\dontrun{
# Load in CytoExploreRData to access data
library(CytoExploreR)

# Generate channel match file for compensation controls
cyto_channel_match(Compensation)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
