% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channels-helpers.R
\name{cyto_channel_match}
\alias{cyto_channel_match}
\title{Table Editor for Channel Match File Construction}
\usage{
cyto_channel_match(x, save_as = NULL, menu = TRUE)
}
\arguments{
\item{x}{object of \code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{save_as}{name to use for the saved channel match csv file, set to
\code{"date-Channel-Match.csv"}.}

\item{menu}{logical indicating whether channels should be selected from a
drop down menu instead of manually typing them in.}
}
\value{
update \code{cyto_details} of \code{flowSet} or \code{GatingSet},
  write channel matching to csv file and return channel matching as a
  data.frame.
}
\description{
Table Editor for Channel Match File Construction
}
\examples{
\dontrun{
# Load in CytoExploreRData to access data
library(CytoExploreR)

# Generate channel match file for compensation controls
cyto_channel_match(Compensation)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
