% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_channel_match}
\alias{cyto_channel_match}
\title{Table Editor for Channel Match File Construction}
\usage{
cyto_channel_match(x, channel_match = NULL)
}
\arguments{
\item{x}{object of \code{flowSet} or \code{GatingSet}.}

\item{channel_match}{name to use for the saved channel match csv file, set to
\code{"date-Channel-Match.csv"}.}
}
\value{
save constructed channel_match csv file.
}
\description{
Table Editor for Channel Match File Construction
}
\examples{
\dontrun{
library(CytoRSuite)

# Generate channel match file for compensation controls
cyto_channel_match(Compensation)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
