% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channels-helpers.R
\name{cyto_channel_select}
\alias{cyto_channel_select}
\title{Select Fluorescent Channel for Compensation Controls}
\usage{
cyto_channel_select(x)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} containing
compensation controls.}
}
\value{
vector of channels in order of compensation Control samples.
}
\description{
Select Fluorescent Channel for Compensation Controls
}
\examples{
\dontrun{
library(CytoExploreRData)

# Load in samples
fs <- Compensation

# Select a channel for each control from dropdown menu
cyto_channel_select(fs)
}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
