% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channels-helpers.R
\name{cyto_channels_extract}
\alias{cyto_channels_extract}
\title{Extract channels associated with certain markers}
\usage{
cyto_channels_extract(x, channels, plot = FALSE, ...)
}
\arguments{
\item{x}{an object of class
\code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{channels}{vector of channel and/or marker names (e.g. c("Alexa Fluor
700-A","CD8")).}

\item{plot}{logical indicating whether the channels will be used to construct
a plot, set to FALSE by default. If set to TRUE an additional check will be
performed to ensure that only 1 or 2 \code{channels} are supplied.}

\item{...}{additional arguments passed to \code{\link[base:grep]{grepl}} for
character matching. For exact character string matching to override the
default which ignores character case, set \code{fixed} to TRUE.}
}
\value{
A vector of valid channel names.
}
\description{
\code{cyto_channels_extract} will check whether the supplied channels or
marker names are valid for the
\code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} and return a vector of
valid channel names. \code{cyto_channels_extract} is particularly useful for
determining which channel(s) are associated with particular marker(s).
}
\examples{
library(CytoExploreRData)

# Activation GatingSet
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                          package = "CytoExploreRData"))

# Extract channels for CD4 & CD8
cyto_channels_extract(gs, c("CD4", "CD8"))

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
