% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channels-helpers.R
\name{cyto_channels_restrict}
\alias{cyto_channels_restrict}
\alias{cyto_channels_restrict.flowFrame}
\alias{cyto_channels_restrict.flowSet}
\alias{cyto_channels_restrict.GatingSet}
\title{Restrict the channels of a cytometry object}
\usage{
\method{cyto_channels_restrict}{flowFrame}(x, exclude = NULL, ...)

\method{cyto_channels_restrict}{flowSet}(x, exclude = NULL, ...)

\method{cyto_channels_restrict}{GatingSet}(x, exclude = NULL, ...)
}
\arguments{
\item{x}{object of class \code{flowFrame}, \code{flowSet}or \code{GatingSet}.}

\item{exclude}{vector of privileged channels or markers to remove in addition
to the channels removed by default, set to NULL by default.}

\item{...}{additional arguments passed to \code{\link{cyto_channels}} to
control character matching for the exclude argument.}
}
\value{
an object of class \code{flowFrame}, \code{flowSet},
  \code{GatingHierarchy} or \code{GatingSet} with unused channels removed.
}
\description{
\code{cyto_channels_restrict} removes any unused channels (channels lacking
marker assignments) from a \code{flowFrame}, \code{flowSet},
\code{GatingHierarchy} or \code{GatingSet}. By default,
\code{cyto_channels_restrict} will always retain any FSC, SSC or Time
channels irrespective of marker assignment. Removal of channels that contain
marker assignments or channels that are privileged channels (FSC/SSC/Time)
can be forced through use of the \code{exclude} argument.
}
\examples{

library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Channels
cyto_channels(fs)

# Remove unused channels
fs <- cyto_channels_restrict(fs)

# Channels removed
cyto_channels(fs)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
