% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channels-helpers.R
\name{cyto_channels_restrict}
\alias{cyto_channels_restrict}
\title{Restrict the channels of a cytometry object}
\usage{
cyto_channels_restrict(x, exclude = NULL, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[flowWorkspace:cytoset]{cytoset}}or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{exclude}{vector of privileged channels or markers to remove in addition
to the channels removed by default, set to NULL by default.}

\item{...}{additional arguments passed to \code{\link{cyto_channels}} to
control character matching for the exclude argument.}
}
\value{
an object of class \code{flowFrame}, \code{flowSet},
  \code{GatingHierarchy} or \code{GatingSet} with unused channels removed.
}
\description{
\code{cyto_channels_restrict} removes any unused channels (channels lacking
marker assignments) from a \code{cytoframe}, \code{cytoset},
\code{GatingHierarchy} or \code{GatingSet}. By default,
\code{cyto_channels_restrict} will always retain any FSC, SSC or Time
channels irrespective of marker assignment. Removal of channels that contain
marker assignments or channels that are privileged channels (FSC/SSC/Time)
can be forced through use of the \code{exclude} argument.
}
\examples{

library(CytoExploreRData)

# Activation Gatingset
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                          package = "CytoExploreRData"))
# Channels
cyto_channels(gs)

# Remove unused channels
gs <- cyto_channels_restrict(gs)

# Channels removed
cyto_channels(gs)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
