% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_clean}
\alias{cyto_clean}
\title{Apply flowAI anomaly detection to clean cytometry data}
\usage{
cyto_clean(x, ...)
}
\arguments{
\item{x}{object of class \code{flowFrame}, \code{flowSet},
\code{GatingHierarchy} or \code{GatingSet}. The \code{root} node extracted
when a \code{GatingSet} or \code{GatingHierachy} is supplied.}

\item{...}{additional arguments passed to
\code{\link[flowAI:flow_auto_qc]{flow_auto_qc}}.}
}
\description{
Apply flowAI anomaly detection to clean cytometry data
}
\examples{
\dontrun{
library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Clean Activation flowSet
fs <- cyto_clean(fs)

# Activation GatingSet
gs <- GatingSet(fs)

# Clean Activation GatingSet
gs <- cyto_clean(gs)
}

}
\references{
Monaco,G. et al. (2016) flowAI: automatic and interactive anomaly
  discerning tools for flow cytometry data. Bioinformatics. 2016 Aug
  15;32(16):2473-80.
  \url{https://academic.oup.com/bioinformatics/article/32/16/2473/2240408}
}
\seealso{
\code{\link[flowAI:flow_auto_qc]{flow_auto_qc}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
