% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_clean}
\alias{cyto_clean}
\title{Apply anomaly detection algorithms to obtain clean cytometry data}
\usage{
cyto_clean(x, type = "flowAI", channels = NULL, trans = NA, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{CatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}. The \code{root}
node extracted when a \code{GatingSet} or \code{GatingHierachy} is
supplied.}

\item{type}{the method to use when cleaning the data, options include
\code{"flowAI"}, \code{"flowClean"}, \code{"flowCut"} or \code{"PeacoQC"},
set to \code{"flowAI"} by default.}

\item{channels}{names of the channel(s) or marker(s) to clean, must include
the \code{Time} parameter for flow rate checks.}

\item{trans}{object of class
transformerList containing the definitions of the transformations already
applied to the supplied \code{cytoset}.}

\item{...}{additional arguments passed to \code{flowAI::flow_auto_qc},
\code{flowClean::flowClean}, \code{flowCut::flowCut} or
\code{PeacoQC::PeacoQC}.}
}
\description{
Apply anomaly detection algorithms to obtain clean cytometry data
}
\examples{
library(CytoExploreRData)

# Activation GatingSet
gs <- cyto_load(
 system.file(
   "extdata/Activation-GatingSet",
   package = "CytoExploreRData"
   )
 )

# Clean Activation GatingSet
gs <- cyto_clean(gs)

}
\references{
Monaco G, et al. (2016) flowAI: automatic and interactive anomaly
  discerning tools for flow cytometry data. Bioinformatics. 2016 Aug
  15;32(16):2473-80.
  \url{https://academic.oup.com/bioinformatics/article/32/16/2473/2240408}

Fletez-Brant K, Spidlen J, Brinkman R, Roederer M, Chattopadhyay
  P (2016). flowClean: Automated identification and removal of fluorescence
  anaomalies in flow cytometry data. Cytometry A 89(5).
  \url{https://onlinelibrary.wiley.com/doi/full/10.1002/cyto.a.22837}

Meskas J, Wang S, Brinkman R (2021). flowCut --- An R Package for
  precise and accurate automated removal of outlier events and flagging of
  files based on time versus fluorescence analysis. bioRxiv.
  \url{https://doi.org/10.1101/2020.04.23.058545}

Emmaneel A, et al. (2021) PeacoQC: peak-based selection of high
  quality cytometry data. Cytometry A.
  \url{https://onlinelibrary.wiley.com/doi/10.1002/cyto.a.24501}
}
\seealso{
\code{\link[flowAI:flow_auto_qc]{flow_auto_qc}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
