% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_coerce}
\alias{cyto_coerce}
\title{Coerce cytoframes in a cytoset}
\usage{
cyto_coerce(
  x,
  sample = 1,
  seed = NULL,
  barcode = FALSE,
  format = "cytoset",
  name = "merge",
  overwrite = NULL,
  ...
)
}
\arguments{
\item{x}{\code{\link[flowWorkspace:cytoset]{cytoset}}.}

\item{sample}{numeric to control how events should be sampled when coercing
samples. If a single value is supplied it is passed to \code{cyto_sample_n}
to compute the number of events to extract from each sample to obtain
\code{sample} total events in the coerced data. Alternatively, users can
manually supply a \code{sample} value for each file for more control over
the way in which samples are sampled prior to merging.}

\item{seed}{numeric passed to \code{\link{set.seed}} to return the same
sample with each run, set to NULL by default for random sampling.}

\item{barcode}{logical indicating whether the samples should be barcoded
prior to coercion, set to FALSE by default.}

\item{format}{can be either 1 - matrix, 2 - cytoframe or 3 - cytoset to
control the format of the returned data, set to \code{"cytoset"} by
default.}

\item{name}{character string indicating the name to use for the coerced
object, set to \code{"merge"} by default.}

\item{overwrite}{logical passed to \code{cyto_barcode} to control whether
existing barcodes should be overwritten, the user will be asked
interactively if not manually supplied.}

\item{...}{additional arguments passed to \code{\link{cyto_sample}}.}
}
\value{
a sampled and coerced \code{matrix}, \code{cytoframe} or
  \code{cytoset}.
}
\description{
\code{cyto_coerce} is an efficient method for simultaneously coercing and
downsampling elements of a \code{cytoset}. Basically, the individual
\code{cytoframe} elements are sampled prior to coercion using a combination
of \code{cyto_sample_n} and \code{cyto_sample}. \code{format} provides
control over the format in which the coerced data should be returned.
}
\examples{
library(CytoExploreRData)

# Activation GatingSet
gs <- cyto_load(
 system.file(
   "extdata/Activation-GatingSet",
   package = "CytoExploreRData"
   )
 )

# Activation cytoset
cs <- cyto_data_extract(gs, "root")[["root"]]

# Coerce & Sample 50000 total events
cyto_coerce(cs, sample = 50000, seed = 56)

# Sample 5000 events each & coerce
cyto_coerce(cs, sample = rep(5000, 33))

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
