% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_convert}
\alias{cyto_convert}
\title{Convert between cytometry objects}
\usage{
cyto_convert(x, return = "flowFrame", parent = "root", display = 1)
}
\arguments{
\item{x}{flowFrame, flowSet, GatingHierarchy, GatingSet or flowFrame list or
flowSet list.}

\item{return}{either 'flowFrame', 'flowSet', 'flowFrame list' or 'flowSet
list'.}

\item{parent}{name of parent population to extract from GatingHierarchy and
GatingSet objects.}

\item{display}{percentage of events to include.}
}
\value{
object specified by 'return' argument.
}
\description{
Automatically removes 'Original' column for coerced objects.
}
\examples{
library(CytoRSuiteData)

# Convert flowSet to list of flowFrames
cyto_convert(Activation, "flowFrame list")

# Convert flowSet to flowFrame
cyto_convert(Activation, "flowFrame")

# Convert GatingSet to flowFrame
cyto_convert(GatingSet(Activation), "flowFrame", parent = "T Cells")

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
