% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_convert}
\alias{cyto_convert}
\title{Convert flowFrame/flowSet objects to cytoframe/cytoset objects}
\usage{
cyto_convert(x, ...)
}
\arguments{
\item{x}{object of class \code{flowFrame}, \code{flowSet}, \code{cytoframe}
or \code{cytoset}.}

\item{...}{additional arguments passed to}
}
\value{
a cytoframe or cytoset.
}
\description{
Convert flowFrame/flowSet objects to cytoframe/cytoset objects
}
\examples{
library(CytoExploreRData)

# flowSet to cytoset
cs <- cyto_convert(Activation)

}
\seealso{
\code{\link[flowWorkspace:convert]{flowFrame_to_cytoframe}}

\code{\link[flowWorkspace:convert]{flowSet_to_cytoset}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
