% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_copy}
\alias{cyto_copy}
\title{Copy a cytoset or cytoframe}
\usage{
cyto_copy(x)
}
\arguments{
\item{x}{a \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}},
\code{\link[flowWorkspace:cytoset]{cytoset}} or
\code{\link[flowWorkspace:cytoframe]{cytoframe}} object.}
}
\value{
flowFrame or flowSet object.
}
\description{
A wrapper around
\code{\link[flowWorkspace:cytoframe_to_flowFrame]{cytoframe_to_flowFrame}}
and \code{\link[flowWorkspace:cytoset_to_flowSet]{cytoset_to_flowSet}}.
}
\examples{

library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Activation cytoset
cs <- flowSet_to_cytoset(fs)

# Activation flowSet
fs <- cyto_copy(fs)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
