% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_copy}
\alias{cyto_copy}
\title{Copy a cytoset, cytoframe or GatingSet}
\usage{
cyto_copy(x)
}
\arguments{
\item{x}{cytoframe, cytoset or GatingSet to be copied.}
}
\value{
copied cytoframe, cytoset or GatingSet.
}
\description{
Copy a cytoset, cytoframe or GatingSet
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
