% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_details}
\alias{cyto_details}
\title{Extract experiment details}
\usage{
cyto_details(x, convert = FALSE, drop = FALSE, factor = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace-cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{convert}{logical indicating whether
\code{\link[utils:type.convert]{type.convert}} should be called on each
column.}

\item{drop}{logical indicating whether row names should be dropped, set to
FALSE by default.}

\item{factor}{logical indicating whether characters should be converted to
factors when convert is TRUE, set to FALSE by default.}

\item{...}{additional arguments passed to
\code{\link[utils:type.convert]{type.convert}}.}
}
\value{
experiment details as data.frame.
}
\description{
Simply an autocomplete-friendly wrapper around
\code{\link[flowWorkspace:pData-methods]{pData}} with some additional
formatting options.
}
\examples{
library(CytoExploreRData)

# Activation GatingSet
gs <- cyto_load(
 system.file(
   "extdata/Activation-GatingSet",
   package = "CytoExploreRData"
   )
 )
                          
# Experiment variables
cyto_details(gs)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
