% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_details}
\alias{cyto_details}
\title{Extract experiment details}
\usage{
cyto_details(x)
}
\arguments{
\item{object}{of class \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}
}
\value{
experiment details as data.frame.
}
\description{
Simply an autocomplete-friendly wrapper around
\code{\link[flowWorkspace:pData-methods]{pData}}. A call is made to
\code{\link{cyto_names}} if a
\code{\link[flowCore:flowFrame-class]{flowFrame}} is supplied.
}
