% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_details}
\alias{cyto_details}
\title{Extract Experiment Details}
\usage{
cyto_details(x)
}
\arguments{
\item{object}{of class \code{flowFrame}, \code{flowSet},
\code{GatingHierarchy} or \code{GatingSet}.}
}
\value{
experiment details as data.frame.
}
\description{
Simply an autocomplete-friendly wrapper around
\code{\link[flowWorkspace:pData]{pData}}.
}
