% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_details_edit}
\alias{cyto_details_edit}
\title{Interactively edit cyto_details for a cytoset or GatingSet}
\usage{
cyto_details_edit(x, file = NULL, save_as = NULL, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{file}{name of csv file containing the experimental information
associated with each sample in the supplied data. This file must contain a
column called \code{"name"} which contains the names of the samples.}

\item{save_as}{name of a CSV to which the edited experimental details should
be saved, defaults to \code{Experiment-Details.csv} prefixed with the date
unless \code{file} is specified or a file is found when searching the
current directory for experimental details. Setting this argument to NA
will prevent writing of edited experimental details to file.}

\item{...}{additional arguments passed to \code{DataEditR::data_edit()} when
\code{cyto_details_edit()} is run in interactive mode.}
}
\value{
NULL and return \code{flowSet} or \code{GatingSet} with updated
  experimental details.
}
\description{
Interactively edit cyto_details for a cytoset or GatingSet
}
\examples{
\dontrun{
# Load in CytoExploreRData to access data
library(CytoExploreRData)

# Load in samples
fs <- Activation

# Edit cyto_details in table editor
cyto_details_edit(fs)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
