% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_details_edit}
\alias{cyto_details_edit}
\title{Interactively edit cyto_details for a flowSet or GatingSet}
\usage{
cyto_details_edit(x, file = NULL)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{file}{name of csv file containing experimental information.}
}
\value{
NULL and return \code{flowSet} or \code{GatingSet} with updated
  experimental details.
}
\description{
Interactively edit cyto_details for a flowSet or GatingSet
}
\examples{
\dontrun{
# Load in CytoExploreRData to access data
library(CytoExploreRData)

# Load in samples
fs <- Activation

# Edit cyto_details in table editor
cyto_details_edit(fs)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
