% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_extract}
\alias{cyto_extract}
\title{Extract a valid flowFrame or flowSet}
\usage{
cyto_extract(x, parent = "root", ...)
}
\arguments{
\item{x}{object of class flowFrame, flowSet, GatingHierarchy or GatingSet.}

\item{parent}{name of the parent population to extract from GatingHierachy or
GatingSet objects.}
}
\value{
either a flowFrame or flowSet object
}
\description{
\code{cyto_extract} is essentially a wrapper for
\code{flowWorkspace::getData()} which also accepts flowFrame or flowSet
objects. The \code{parent} population is extracted from GatingHierarchy or
GatingSet objects whilst flowFrame or flowSet objects are returned as is.
}
\examples{
library(CytoRSuiteData)

# GatingSet
gs <- GatingSet(Activation)

# Extract root population
cyto_extract(gs, "root")
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
