% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_extract}
\alias{cyto_extract}
\title{Extract a valid flowFrame or flowSet}
\usage{
cyto_extract(x, parent = "root", ...)
}
\arguments{
\item{x}{object of class \code{flowFrame}, \code{flowSet},
\code{GatingHierarchy} or \code{GatingSet}.}

\item{parent}{name of the parent population to extract from
\code{GatingHierachy} or \code{GatingSet} objects.}
}
\value{
either a \code{flowFrame} or a \code{flowSet}.
}
\description{
\code{cyto_extract} is essentially a wrapper for
\code{\link[flowWorkspace:gs_pop_get_data]{gs_pop_get_data}} which also
accepts \code{\link[flowCore:flowFrame-class]{flowFrame}} or
\code{\link[flowCore:flowSet-class]{flowSet}} objects. The \code{parent}
population is extracted from
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} objects whilst
\code{flowFrame} or \code{flowSet} objects are returned as is.
}
\examples{

# Load in CytoExploreRData to access data
library(CytoExploreRData)

# GatingSet
gs <- GatingSet(Activation)

# Extract root population
cyto_extract(gs, "root")

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
