% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_file_search.R
\name{cyto_file_search}
\alias{cyto_file_search}
\title{Search for CSV file in current directory}
\usage{
cyto_file_search(x, dir = ".", colnames = NULL, rownames = NULL, ...)
}
\arguments{
\item{x}{regular expression to use when searching for files by name in
specified directory.}

\item{dir}{location where the search should be performed.}

\item{colnames}{minimal requirement for column names in the file.}

\item{rownames}{minimal requirement for row names in the file.}

\item{...}{additional named arguments indicating values to find in particular
columns of the file (e.g. name = c("Activation_1.fcs")).}
}
\value{
NULL or data read in from file matching search criteria.
}
\description{
Search for CSV file in current directory
}
\examples{
\dontrun{ 
  # search for experiment details of samples
  cyto_file_search( 
    "Experiment-Details.csv$", 
    rownames = c("Activation_1.fcs", "Activation_2.fcs")
  ) 
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
