% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_filter}
\alias{cyto_filter}
\title{Filter samples based on experiment variables}
\usage{
cyto_filter(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{...}{tidyverse-style subsetting using comma separated logical
predicates based on experimental variables stored in
\code{cyto_details(x)}. See examples below for demonstration.}
}
\value{
\code{cytoset} or \code{GatingSet} restricted to samples which meet
  the filtering criteria.
}
\description{
Filter samples based on experiment variables
}
\examples{
library(CytoExploreRData)

# Activation Gatingset
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                          package = "CytoExploreRData"))

# Load in CytoExploreRData to access data
library(CytoExploreRData)

# Look at experiment details
cyto_details(Activation)

# Select Stim-C samples with 0 and 500 nM OVA concentrations
fs <- cyto_filter(
  Activation,
  Treatment == "Stim-C",
  OVAConc \%in\% c(0, 500)
)

# Select Stim-A and Stim-C treatment groups
fs <- cyto_filter(
  Activation,
  Treatment \%in\% c("Stim-A", "Stim-C")
)
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
