% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channels-helpers.R
\name{cyto_fluor_channels}
\alias{cyto_fluor_channels}
\title{Extract Fluorescent Channels}
\usage{
cyto_fluor_channels(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{...}{additional arguments passed to \code{\link{cyto_channels}}.}
}
\description{
Extract Fluorescent Channels
}
\examples{
library(CytoExploreRData)

# Activation GatingSet
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                          package = "CytoExploreRData"))

# GatingSet
cyto_fluor_channels(gs)

# GatingHierarchy
cyto_fluor_channels(gs[[1]])

# cytoset
cs <- cyto_data_extract(gs, "root")[["root"]]
cyto_fluor_channels(cs)

# cytoframe
cyto_fluor_channels(cs[[1]])

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
