% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channels-helpers.R
\name{cyto_fluor_channels}
\alias{cyto_fluor_channels}
\alias{cyto_fluor_channels.flowFrame}
\alias{cyto_fluor_channels.flowSet}
\alias{cyto_fluor_channels.GatingHierarchy}
\alias{cyto_fluor_channels.GatingSet}
\title{Extract Fluorescent Channels}
\usage{
cyto_fluor_channels(x)

\method{cyto_fluor_channels}{flowFrame}(x)

\method{cyto_fluor_channels}{flowSet}(x)

\method{cyto_fluor_channels}{GatingHierarchy}(x)

\method{cyto_fluor_channels}{GatingSet}(x)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}
}
\description{
Extract Fluorescent Channels
}
\examples{
library(CytoExploreRData)

# Load in samples
fs <- Activation

# Add samples to GatingSet
gs <- GtaingSet(fs)

# Fluorescent channels of flowFrame
cyto_fluor_channels(fs[[1]])

# Fluorescent channels for a flowSet
cyto_fluor_channels(fs)

# Fluorescent channels for GatingHierarchy
cyto_fluor_channels(gs[[1]])

# Fluorescent channels for GatingSet
cyto_fluor_channels(gs)
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
