% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gate-helpers.R
\name{cyto_gate_copy}
\alias{cyto_gate_copy}
\title{Copy gates to other nodes in GatingSet and update gatingTemplate}
\usage{
cyto_gate_copy(
  x,
  parent = NULL,
  alias = NULL,
  copy = NULL,
  gatingTemplate = NULL
)
}
\arguments{
\item{x}{object of class \code{GatingSet}.}

\item{parent}{name of the parental node to which the gates should be copied.}

\item{alias}{vector of names for the copied gates, must be the same length as
copy.}

\item{copy}{vector of names to copy to new parent.}

\item{gatingTemplate}{name of the \code{gatingTemplate} csv file (e.g.
"gatingTemplate.csv") where the new entries should be saved.}
}
\value{
object of class \code{GatingSet} with copied gates applied and update
  gatingTemplate csv file with appropriate entries.
}
\description{
Copy gates to other nodes in GatingSet and update gatingTemplate
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
