% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gate_draw.R
\name{cyto_gate_draw}
\alias{cyto_gate_draw}
\alias{cyto_gate_draw.GatingSet}
\alias{cyto_gate_draw.flowSet}
\alias{cyto_gate_draw.flowFrame}
\title{Interactively draw and construct gate objects}
\usage{
\method{cyto_gate_draw}{GatingSet}(x, parent = "root", alias = NULL,
  channels = NULL, type = NULL, gatingTemplate = NULL,
  overlay = NA, group_by = "all", select = NULL, negate = FALSE,
  axis = "x", label = TRUE, plot = TRUE, ...)

\method{cyto_gate_draw}{flowSet}(x, alias = NULL, channels = NULL,
  type = NULL, overlay = NA, group_by = "all", select = NULL,
  axis = "x", label = TRUE, plot = TRUE, ...)

\method{cyto_gate_draw}{flowFrame}(x, alias = NULL, channels = NULL,
  type = NULL, overlay = NA, axis = "x", label = TRUE,
  plot = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the \code{parent} population to extract for gating when
a \code{GatingSet} object is supplied.}

\item{alias}{the name(s) of the populations to be gated. If multiple
population names are supplied (e.g. \code{c("CD3,"CD4)}) multiple gates
will be returned. \code{alias} is \code{NULL} by default which will halt
the gating routine.}

\item{channels}{vector of channel names to use for plotting, can be of length
1 for 1-D density histogram or length 2 for 2-D scatter plot.}

\item{type}{vector of gate type names used to construct the gates. Multiple
gate types are supported but should be accompanied with an \code{alias}
argument of the same length (i.e. one \code{type} per \code{alias}).
Supported gate types are \code{polygon, rectangle, ellipse, threshold,
boundary, interval, quadrant and web} which can be abbreviated as upper or
lower case first letters as well. Default \code{type} is \code{"interval"}
for 1D gates and \code{"polygon"} for 2D gates.}

\item{gatingTemplate}{name of \code{gatingTemplate} csv file to which the
\code{gatingTemplate} entries for the \code{GatingSet} method should be
saved.}

\item{overlay}{name(s) of the populations to overlay or a \code{flowFrame},
\code{flowSet}, \code{list of flowFrames} or \code{list of flowSets}
containing populations to be overlaid onto the plot(s).}

\item{group_by}{vector of \code{\link{cyto_details}} column names (e.g.
c("Treatment","Concentration") indicating how the samples should be grouped
prior to gating, set to the length of x by default to construct a single
gate for all samples. If group_by is supplied a different gate will be
constructed for each group. This argument only applies to \code{flowSet} or
\code{GatingSet} objects.}

\item{select}{designates which samples will be plotted and used for
determining the best location to set the drawn gate(s). Filtering steps
should be comma separated and wrapped in a list (e.g. list(Treatment ==
"Stim-A", OVAConc %in% c(0,0.5))). Refer to \code{\link{cyto_select}}.}

\item{negate}{logical indicating whether a gatingTemplate entry should be
made for the negated population (i.e. all events outside the constructed
gates), set to FALSE by default. If negate is set to TRUE, a name for the
negated population MUST be supplied at the end of the alias argument.}

\item{axis}{indicates whether the \code{"x"} or \code{"y"} axis should be
gated for 2-D interval gates.}

\item{label}{logical indicating whether to include
\code{\link{cyto_plot_label}} for the gated population(s), \code{TRUE} by
default.}

\item{plot}{logical indicating whether a plot should be drawn, set to
\code{TRUE} by default.}

\item{...}{additional arguments for \code{\link{cyto_plot}}.}
}
\value{
\code{flowFrame} and \code{flowSet} methods return a list of
  \code{\link[flowCore:filters-class]{filters}} objects, whilst the
  \code{GatingSet} applies the constructed gates directly to the
  \code{GatingSet} and adds appropriate entries into the specified
  \code{gatingTemplate}. The \code{GatingSet} method does return the
  constructed gates but instead invisibly returns the \code{gatingTemplate}
  entries.
}
\description{
\code{cyto_gate_draw} allows to users to interactive draw gates around
populations which are returned as \code{flowCore} gate objects. The flowFrame
and flowSet methods simply return the constructed gates as a list of
\code{\link[flowCore:filters-class]{filters}}, whilst the GatingSet method
automatically applies the constructed gates to the GatingSet and saves the
constructed gates in an \code{openCyto}
\code{\link[openCyto:gatingTemplate-class]{gatingTemplate}} for future use.
See \code{\link{cyto_gate_edit}}, \code{\link{cyto_gate_remove}} and
\code{\link{cyto_gate_rename}} to manipulate constructed gates and modify
their entries in the gatingTemplate.
}
\examples{
\dontrun{
# Gate drawing requires an interactive R session
library(CytoRSuiteData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(fs))
gs <- transform(gs, trans)

# Gate using cyto_gate_draw
gating(Activation_gatingTemplate, gs)

# draw gates using cyto_gate_draw
cyto_gate_draw(gs,
  parent = "Dendritic Cells",
  channels = c("Alexa Fluor 488-A", "Alexa Fluor 700-A"),
  alias = c("CD8+ DC", "CD4+ DC"),
  gatingTemplate = "Example-gatingTemplate.csv",
  type = "rectangle",
  contour_lines = 15
)

# Constructed gate applied directly to GatingSet
cyto_nodes(gs)
}

}
\seealso{
\code{\link{cyto_plot}}

\code{\link{cyto_gate_edit}}

\code{\link{cyto_gate_remove}}

\code{\link{cyto_gate_rename}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
