% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingTemplate-modifiers.R
\name{cyto_gate_edit}
\alias{cyto_gate_edit}
\title{Edit Existing Gate(s).}
\usage{
cyto_gate_edit(x, parent = NULL, alias = NULL, channels = NULL,
  type = NULL, gatingTemplate = NULL, display = NULL,
  select = NULL, overlay = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{GatingSet}.}

\item{parent}{name of the parental population.}

\item{alias}{name(s) of the gate to edit (e.g. "Single Cells").}

\item{channels}{name(s) of the channel(s) used to construct the gate(s). This
argument is not necessary but is included to allow conversion of
\code{cyto_gate_draw} code to \code{cyto_gate_remove} code by simply changing
\code{"draw"} to \code{"remove"}.}

\item{type}{vector of gate type names used to construct the gates. Multiple
\code{types} are supported but should be accompanied with an \code{alias}
argument of the same length (i.e. one \code{type} per \code{alias}).
Supported \code{gate_types} are \code{polygon, rectangle, ellipse,
threshold, boundary, interval, quadrant and web} which can be abbreviated
as upper or lower case first letters as well. Default \code{type} is
\code{"polygon"}.}

\item{gatingTemplate}{name of the \code{gatingTemplate} csv file (e.g.
"gatingTemplate.csv") where the gate is saved.}

\item{display}{numeric [0,1] to control the percentage of events to be
plotted. Specifying a value for \code{display} can substantial improve
plotting speed for less powerful machines.}

\item{select}{vector containing the indicies of samples within gs to use for
plotting.}

\item{overlay}{name(s) of the population(s) to overlay onto the plot.}

\item{...}{additional arguments passed to cyto_plot, see ?cyto_plot for
details.}
}
\value{
an object of class \code{GatingSet} with edited gate applied, as well
  as gatingTemplate file with edited gate saved.
}
\description{
Edit Existing Gate(s).
}
\examples{
\dontrun{
library(CytoRSuiteData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(gs))
gs <- transform(gs, trans)

# Gate using cyto_gate_draw
gt <- Activation_gatingTemplate
gating(gt, gs)

# Edit CD4 T Cells Gate - replace gatingTemplate name
gate_edit(gs,
  parent = "T Cells",
  alias = "CD4 T Cells",
  gatingTemplate = "gatingTemplate.csv"
)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
