% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gate-helpers.R
\name{cyto_gate_edit}
\alias{cyto_gate_edit}
\title{Edit Existing Gate(s).}
\usage{
cyto_gate_edit(
  x,
  parent = NULL,
  alias = NULL,
  channels = NULL,
  type = NULL,
  gatingTemplate = NULL,
  overlay = NA,
  group_by = "all",
  select = NULL,
  negate = FALSE,
  display = 25000,
  axis = "x",
  label = TRUE,
  ...
)
}
\arguments{
\item{x}{an object of class \code{GatingSet}.}

\item{parent}{name of the parental population.}

\item{alias}{name(s) of the gate to edit (e.g. "Single Cells").}

\item{channels}{name(s) of the channel(s) used to construct the gate(s). This
argument is not necessary but is included to allow conversion of
\code{cyto_gate_draw} code to \code{cyto_gate_remove} code by simply changing
\code{"draw"} to \code{"remove"}.}

\item{type}{vector of gate type names used to construct the gates. Multiple
\code{types} are supported but should be accompanied with an \code{alias}
argument of the same length (i.e. one \code{type} per \code{alias}).
Supported \code{gate_types} are \code{polygon, rectangle, ellipse,
threshold, boundary, interval, quadrant and web} which can be abbreviated
as upper or lower case first letters as well. Default \code{type} is
\code{"polygon"}.}

\item{gatingTemplate}{name of the \code{gatingTemplate} csv file (e.g.
"gatingTemplate.csv") where the gate is saved.}

\item{overlay}{name(s) of the populations to overlay or a \code{flowFrame},
\code{flowSet}, \code{list of flowFrames} or \code{list of flowSets}
containing populations to be overlaid onto the plot(s). Only overlaid
flowSet objects are subjected to sampling by \code{display}.}

\item{group_by}{vector of pData column names (e.g.
c("Treatment","Concentration") indicating how the samples should be grouped
prior to gating, set to the length of x by default to construct a single
gate for all samples. If group_by is supplied a different gate will be
constructed for each group.}

\item{select}{vector containing the indices of samples within gs to use for
plotting.}

\item{negate}{logical indicating whether a gatingTemplate entry should be
made for the negated population (i.e. all events outside the constructed
gates), set to FALSE by default. If negate is set to TRUE, a name for the
negated population MUST be supplied at the end of the alias argument.}

\item{display}{fraction or number of events to display in the plot during the
gating process, set to 25 000 events by default.}

\item{axis}{indicates whether the \code{"x"} or \code{"y"} axis should be
gated for 2-D interval gates.}

\item{label}{logical indicating whether to include
\code{\link{cyto_plot_label}} for the gated population(s), \code{TRUE} by
default.}

\item{...}{additional arguments for \code{\link{cyto_plot.flowFrame}}.}
}
\value{
an object of class \code{GatingSet} with edited gate applied, as well
  as gatingTemplate file with edited gate saved.
}
\description{
Edit Existing Gate(s).
}
\examples{
\dontrun{
library(CytoExploreRData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- cyto_compensate(gs)

# Transform fluorescent channels
gs <- cyto_transform(gs)

# Gate using cyto_gate_draw
gt <- Activation_gatingTemplate
gt_gating(gt, gs)

# Edit CD4 T Cells Gate - replace gatingTemplate name
cyto_gate_edit(gs,
  parent = "T Cells",
  alias = "CD4 T Cells",
  gatingTemplate = "gatingTemplate.csv"
)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
