% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gate-helpers.R
\name{cyto_gate_extract}
\alias{cyto_gate_extract}
\title{Extract Saved Gate(s) from gatingTemplate.}
\usage{
cyto_gate_extract(x, parent, alias, gatingTemplate = NULL, ...)
}
\arguments{
\item{x}{object of class GatingHierarchy or GatingSet.}

\item{parent}{name of the parental population.}

\item{alias}{name of the population for which the gate must be extracted.}

\item{gatingTemplate}{name of the \code{gatingTemplate} csv file (e.g.
"gatingTemplate.csv") where the gate(s) are saved.}

\item{...}{not used.}
}
\description{
Extract Saved Gate(s) from gatingTemplate.
}
\examples{
library(CytoExploreRData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- cyto_compensate(gs)

# Transform fluorescent channels
gs <- cyto_transform(gs)

# Gate using cyto_gate_draw
gt <- Activation_gatingTemplate
gt_gating(gt, gs)

# gatingTemplate
gtfile <- system.file("extdata",
  "Activation-gatingTemplate.csv",
  package = "CytoExploreRData"
)

# Extract T Cells gate
cyto_gate_extract("Live Cells", "T Cells", gatingTemplate = gtfile)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
