% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingTemplate-modifiers.R
\name{cyto_gate_extract}
\alias{cyto_gate_extract}
\title{Extract Saved Gate(s) from gatingTemplate.}
\usage{
cyto_gate_extract(parent, alias, gatingTemplate = NULL, ...)
}
\arguments{
\item{parent}{name of the parental population.}

\item{alias}{name of the population for which the gate must be extracted.}

\item{gatingTemplate}{name of the \code{gatingTemplate} csv file (e.g.
"gatingTemplate.csv") where the gate(s) are saved.}

\item{...}{not used.}
}
\description{
Extract Saved Gate(s) from gatingTemplate.
}
\examples{
library(CytoRSuiteData)

# Bypass working directory check for external files
options("CytoRSuite_wd_check" = FALSE)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(gs))
gs <- transform(gs, trans)

# Gate using cyto_gate_draw
gt <- Activation_gatingTemplate
gating(gt, gs)

# gatingTemplate
gtfile <- system.file("extdata",
  "Activation-gatingTemplate.csv",
  package = "CytoRSuiteData"
)

# Extract T Cells gate
gate_extract("Live Cells", "T Cells", gatingTemplate = gtfile)

# Reset CytoRSuite_wd_check to default
options("CytoRSuite_wd_check" = TRUE)
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
