% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gatingTemplate-helpers.R
\name{cyto_gatingTemplate_active}
\alias{cyto_gatingTemplate_active}
\title{Retrieve or set active gatingTemplate}
\usage{
cyto_gatingTemplate_active(x = NULL, reset = FALSE, ask = FALSE)
}
\arguments{
\item{x}{name of the gatingTemplate to set as the active gatingTemplate. If
not supplied, the name of the current active gatingTemplate will be
returned.}

\item{reset}{logical indicating whether the active gatingTemplate setting
should be reset so that no active gatingTemplate is set, set to FALSE by
default.}

\item{ask}{logical indicating whether the user should be asked to supply a
name for the active gatingTemplate if one has not been set, set to FALSE
by default.}
}
\value{
retrieve or set current active gatingTemplate.
}
\description{
Retrieve or set active gatingTemplate
}
\seealso{
\code{\link{cyto_gatingTemplate_create}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
