% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingTemplate-helpers.R
\name{cyto_gatingTemplate_apply}
\alias{cyto_gatingTemplate_apply}
\title{Apply gates saved in gatingTemplate to a GatingSet}
\usage{
cyto_gatingTemplate_apply(x, gatingTemplate = NULL, ...)
}
\arguments{
\item{x}{object of class \code{GatingSet}.}

\item{gatingTemplate}{name of the gatingTemplate csv file which contains the
gates to be applied to the GatingSet.}

\item{...}{additional arguments passed to gating.}
}
\value{
NULL and update the GatingSet in the global environment with the
  gates in the gatingTemplate.
}
\description{
A convenient wrapper for
\code{\link[openCyto:gatingTemplate]{gatingTemplate}} and
\code{\link[openCyto:gating]{gating}} to apply a gatingTemplate csv file to a
GatingSet.
}
