% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_gatingTemplate-helpers.R
\name{cyto_gatingTemplate_generate}
\alias{cyto_gatingTemplate_generate}
\alias{cyto_gatingTemplate_generate.GatingHierarchy}
\alias{cyto_gatingTemplate_generate.GatingSet}
\title{Generate a CytoExploreR gatingTemplate for a GatingHierarchy or GatingSet}
\usage{
\method{cyto_gatingTemplate_generate}{GatingHierarchy}(x, gatingTemplate = NULL, ...)

\method{cyto_gatingTemplate_generate}{GatingSet}(x, gatingTemplate = NULL, ...)
}
\arguments{
\item{x}{object of class GatingHierarchy or GatingSet.}

\item{gatingTemplate}{name of a CSV file to which the generated
gatingTemplate should be saved.}

\item{...}{additional arguments passed to \code{gatingTemplate}.}
}
\value{
write generated gatingTemplate to designated CSV file and return the
  gatingTemplate object.
}
\description{
Generate a CytoExploreR gatingTemplate for a GatingHierarchy or GatingSet
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
