% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_group_by}
\alias{cyto_group_by}
\title{Group a flowSet or GatingSet by experiment variables}
\usage{
cyto_group_by(x, group_by = "all", select = NULL)
}
\arguments{
\item{x}{an object of class \code{\link[flowWorkspace:cytoset]{cytoset}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{group_by}{names of cyto_details variables to use for merging, set to
"all" to group all samples in \code{x}. The order of the grouping can be
controlled by specifying the factor levels in a list (e.g. list(Treatment =
c("Stim-A","Stim-C","Stim-B", "Stim-D"))). Set this argument to
\code{"all"} or \code{NA} to place all samples within the same group.}

\item{select}{vector of indices or named list containing experimental
variables passed to \code{\link{cyto_select}} to be used to select samples
prior to grouping.}
}
\value{
a named list of \code{cytoset} or \code{GatingSet} objects
  respectively.
}
\description{
Group a flowSet or GatingSet by experiment variables
}
\examples{
library(CytoExploreRData)

# Activation Gatingset
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                          package = "CytoExploreRData"))

# Group by Treatment
cyto_group_by(gs, "Treatment")

# Group GatingSet by Treatment and OVAConc
cyto_group_by(gs, c("Treatment", "OVAConc"))

}
\seealso{
\code{\link{cyto_select}}

\code{\link{cyto_groups}}

\code{\link{cyto_sort_by}}

\code{\link{cyto_merge_by}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
