% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_group_by}
\alias{cyto_group_by}
\title{Group a flowSet or GatingSet by experiment variables}
\usage{
cyto_group_by(x, group_by = "all")
}
\arguments{
\item{x}{an object of class \code{flowSet} or \code{GatingSet}.}

\item{group_by}{names of cyto_details variables to use for merging. Set to "all" to
merge all samples in \code{x}.}

\item{parent}{name of the parent population to extract from GatingSet object.}
}
\value{
a named list of \code{flowSet} or \code{GatingSet} objects
  respectively.
}
\description{
Group a flowSet or GatingSet by experiment variables
}
\examples{
library(CytoRSuiteData)

# Group flowSet by Treatment
cyto_group_by(Activation, "Treatment")

# Group GatingSet by Treatment and OVAConc
gs <- GatingSet(Activation)
cyto_group_by(gs, c("Treatment", "OVAConc"))
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
