% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_list}
\alias{cyto_list}
\title{Convert cytoframe or cytoset to named cytoframe list}
\usage{
cyto_list(x)
}
\arguments{
\item{x}{object of class \code{cytoframe} or \code{cytoset} to be listed.}
}
\value{
list of \code{cytoframe} objects.
}
\description{
Convert cytoframe or cytoset to named cytoframe list
}
\examples{
library(CytoExploreRData)

# Activation dataset
fs <- Activation

# Convert to list
cf_list <- cyto_list(fs)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
