% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_load}
\alias{cyto_load}
\title{Load FCS files into cytoset}
\usage{
cyto_load(
  path = ".",
  select = NULL,
  exclude = NULL,
  sort = TRUE,
  barcode = FALSE,
  restrict = FALSE,
  copy = FALSE,
  split = NULL,
  events = 1,
  fixed = FALSE,
  ignore.case = TRUE,
  seed = NULL,
  ...
)
}
\arguments{
\item{path}{points to the location of the .fcs files to read in. Preferably
the name of folder in current working directory, but paths to individual
files are also allowed. Alternatively, users can supply a named list of
matrices to be converted into a cytoset for use within CytoExploreR.}

\item{select}{vector of file names to select when loading files, set to NULL
be default to select all files in the specified directory.}

\item{exclude}{vector of file names to exclude when loading files, set to
NULL by default to load all files in the specified directory.}

\item{sort}{logical indicating whether attempts should be made to sort the
files by name prior to loading, set to \code{TRUE} by default.}

\item{barcode}{logical indicating whether the cytoframes should be barcoded
using \code{cyto_barcode}, set to FALSE by default.}

\item{restrict}{logical indicating whether unassigned channels should be
dropped from the returned cytoset, set to FALSE by default. See
\code{\link{cyto_channels_restrict}}.}

\item{copy}{logical indicating whether \code{GatingSet} archives should be
copied after loading, set to FALSE by default.}

\item{split}{indicates the name of a channel by which barcoded samples should
be split using \code{cyto_split()}.}

\item{events}{numeric passed to \code{cyto_sample()} to indicate the
proportion or number of events to import from each sample, set to 1 by
default to import all events. Setting \code{events} to NA will result in
downsampling to the minimum number of events across samples.}

\item{fixed}{logical passed to \code{grepl()} during file name matching in
\code{select} and \code{exclude} to control whether an exact match is
required, set to FALSE by default for more flexible matching.}

\item{ignore.case}{logical passed to \code{grepl()} during file name matching
in \code{select} and \code{exclude} to control whether case insensitive
matching is required, set to TRUE by default.}

\item{seed}{integer passed tp \code{set.seed()} prior to sampling with
\code{cyto_sample()} to ensure reproducility of sampling, set to NULL by
default to use a random sample.}

\item{...}{additional arguments passed to
\code{\link[flowWorkspace:load_cytoset_from_fcs]{load_cytoset_from_fcs()}}.}
}
\value{
object of class \code{\link[flowWorkspace:cytoset]{cytoset}} or
  \code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.
}
\description{
\code{cyto_load} is a convenient wrapper around
\code{\link[base:list.files]{list.files}} and
\code{\link[flowWorkspace:load_cytoset_from_fcs]{load_cytoset_from_fcs}}
which makes it easy to load .fcs files into a cytoset. \code{cyto_load} is
also a wrapper around \code{\link[flowWorkspace:save_gs]{load_gs}} to load
saved GatingSet objects.
}
\examples{
# Load in CytoExploreRData to access data
library(CytoExploreRData)

# Get path to Activation .fcs files in CytoExploreRData
datadir <- system.file("extdata", package = "CytoExploreRData")
path <- paste0(datadir, "/Activation")

# Load in .fcs files into cytoset
cs <- cyto_load(path)

# cs is a cytoset
class(cs)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
