% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_map.R
\name{cyto_map}
\alias{cyto_map}
\alias{cyto_map.flowFrame}
\alias{cyto_map.flowSet}
\title{Create reduced dimension maps of cytometry data}
\usage{
cyto_map(x, ...)

\method{cyto_map}{flowFrame}(x, method = "tSNE", save = FALSE,
  save_as = "cyto_map", plot = TRUE, seed, ...)

\method{cyto_map}{flowSet}(x, method = "tSNE", save = FALSE,
  save_as = "cyto_map", plot = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{flowFrame} or \code{flowSet}.}

\item{...}{additional arguments passed to the called dimension reduction
function. Links to the documentation for these functions can be found
below.}

\item{method}{direction reduction method to use to generate the map,
supported options include "PCA", "tSNE" and "UMAP".}

\item{save}{logical indicating whether the mapped \code{flowFrame} or
\code{flowSet} should be saved as .fcs file(s) in a folder in the  current
working directory.}

\item{save_as}{name of the folder to save the .fcs files to when save is
TRUE, set to "cyto_map" by default.}

\item{plot}{logical indicating whether the constructed map should be plotted
using \code{cyto_plot}.}

\item{seed}{integer to set seed prior to mapping to ensure consistent results
between runs.}
}
\description{
Create reduced dimension maps of cytometry data
}
\seealso{
\code{\link[stats:prcomp]{PCA}}

\code{\link[Rtsne:Rtsne]{tSNE}}

\code{\link[umap:umap]{UMAP}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
