% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_map.R
\name{cyto_map}
\alias{cyto_map}
\alias{cyto_map.GatingSet}
\alias{cyto_map.flowSet}
\alias{cyto_map.flowFrame}
\title{Create dimension-reduced maps of cytometry data}
\usage{
cyto_map(x, ...)

\method{cyto_map}{GatingSet}(
  x,
  parent = "root",
  select = NULL,
  channels = NULL,
  display = 1,
  type = "UMAP",
  merge = TRUE,
  split = TRUE,
  names = NULL,
  save_as = NULL,
  inverse = FALSE,
  trans = NULL,
  plot = TRUE,
  seed = NULL,
  ...
)

\method{cyto_map}{flowSet}(
  x,
  select = NULL,
  channels = NULL,
  display = 1,
  type = "UMAP",
  merge = TRUE,
  split = TRUE,
  names = NULL,
  save_as = NULL,
  inverse = FALSE,
  trans = NULL,
  plot = TRUE,
  seed = NULL,
  ...
)

\method{cyto_map}{flowFrame}(
  x,
  channels = NULL,
  display = 1,
  type = "UMAP",
  split = TRUE,
  names = NULL,
  save_as = NULL,
  inverse = FALSE,
  trans = NULL,
  plot = TRUE,
  seed = NULL,
  ...
)
}
\arguments{
\item{x}{object of class \code{flowFrame} or \code{flowSet}.}

\item{...}{additional arguments passed to the called dimension reduction
function. Links to the documentation for these functions can be found
below.}

\item{parent}{name of the parent population to extract from
\code{GatingHierarchy} or \code{GatingSet} objects for mapping, set to the
\code{"root"} node by default.}

\item{select}{designates which samples should be used for mapping when a
\code{flowSet} or \code{GatingSet} object is supplied. Filtering steps
should be comma separated and wrapped in a list. Refer to
\code{\link{cyto_select}}.}

\item{channels}{vector of channels names indicating the channels that should
be used by the dimension reduction algorithm to compute the 2-dimensional
map, set to all channels with assigned markers by default. Restricting the
number of channels can greatly improve processing speed and resolution.}

\item{display}{total number of events to map, all events in the combined data
are mapped by default.}

\item{type}{dimension reduction type to use to generate the map, supported
options include "PCA", "tSNE", "FIt-SNE", "UMAP" and "EmbedSOM".}

\item{merge}{logical indicating whether the samples should be merged prior to
mapping to generate a consensus map, set to TRUE by default.}

\item{split}{logical indicating whether samples merged using
\code{cyto_merge_by} should be split prior to writing fcs files, set to
FALSE by default.}

\item{names}{original names of the samples prior to merging using
\code{cyto_merge_by}, only required when split is TRUE. These names will be
re-assigned to each of split flowFrames and included in the file names.}

\item{save_as}{passed to \code{cyto_save} to indicate a folder where the
mapped FCS files should be saved, set to NULL by default to turn off saving
of FCS files.}

\item{inverse}{logical indicating whether the data should be inverse
transformed prior to writing FCS files, set to FALSE by default. Inverse
transformations of \code{flowFrame} or \code{flowSet} objects requires
passing of transformers through the \code{trans} argument.}

\item{trans}{object of class \code{transformerList} containing the
transformation definitions applied to the supplied data. Used internally
when \code{inverse_transform} is TRUE, to inverse the transformations prior
to writing FCS files.}

\item{plot}{logical indicating whether the constructed map should be plotted
using \code{cyto_plot}.}

\item{seed}{integer to set seed prior to mapping to ensure more consistent
results between runs.}
}
\value{
flowFrame, flowSet, GatingHierarchy or GatingSet containing the
  mapped projection parameters.
}
\description{
\code{cyto_map} is a convenient wrapper to produce dimension-reduced maps of
cytometry data using PCA, tSNE, FIt-SNE, UMAP and EmbedSOM. These
dimensionality reduction functions are called using the default settings, but
can be altered by passing relvant arguments through \code{cyto_map}. To see a
full list of customisable parameters refer to the documentation for each of
these functions by clicking on the links below.
}
\details{
If you use \code{cyto_map} to map your cytometry data, be sure to cite the
publication that describes the dimensionality reduction algorithm that you
have chosen to use. References to these publications can be found in the
references section of this document.
}
\references{
Gabriel K. (1971). The biplot graphical display of matrices with
  application to principal component analysis. Biometrika 58, 453–467.
  \url{doi:10.1093/biomet/58.3.453}.

Maaten, L. van der, & Hinton, G. (2008). Visualizing Data using
  t-SNE. Journal of Machine Learning Research 9, 2579–2605.
  \url{http://www.jmlr.org/papers/volume9/vandermaaten08a/vandermaaten08a.pdf}.

Linderman, G., Rachh, M., Hoskins, J., Steinerberger, S.,
  Kluger., Y. (2019). Fast interpolation-based t-SNE for improved
  visualization of single-cell RNA-seq data. Nature Methods.
  \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6402590/}.

McInnes, L., & Healy, J. (2018). UMAP: uniform manifold
  approximation and projection for dimension reduction. Preprint at
  \url{https://arxiv.org/abs/1802.03426}.

Kratochvíl, M., Koladiya, A., Balounova, J., Novosadova, V.,
  Fišer, K., Sedlacek, R., Vondrášek, J., and Drbal, K. (2018). Rapid
  single-cell cytometry data visualization with EmbedSOM. Preprint at
  \url{https://www.biorxiv.org/content/10.1101/496869v1}.
}
\seealso{
\code{\link[stats:prcomp]{PCA}}

\code{\link[Rtsne:Rtsne]{tSNE}}

\code{\link{fftRtsne}}

\code{\link[umap:umap]{UMAP}}

\code{\link[EmbedSOM:SOM]{SOM}}

\code{\link[EmbedSOM:EmbedSOM]{EmbedSOM}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
