% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channels-helpers.R
\name{cyto_markers}
\alias{cyto_markers}
\title{Extract marker names}
\usage{
cyto_markers(x, select = NULL, exclude = NULL, append = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} or a list of these
objects.}

\item{select}{vector of channels or markers for which the channel/marker
combinations should be returned.}

\item{exclude}{vector of channels or markers for which the channel/marker
combinations should not be returned.}

\item{append}{logical indicating whether the name of the channel should be
appended to the marker names in the form \code{<marker> channel}, set to
FALSE by default.}

\item{...}{additional arguments passed to \code{\link[base:grep]{grepl}} for
character matching. For exact character string matching to override the
default which ignores character case, set \code{fixed} to TRUE.}
}
\value{
vector of marker names or NULL if no markers have been assigned.
}
\description{
Extract marker names
}
\examples{

# Load in CytoExploreRData to access data
library(CytoExploreRData)

# Activation GatingSet
gs <- cyto_load(system.file("extdata/Activation-GatingSet", 
                package = "CytoExploreRData"))

# GatingHierarchy
cyto_markers(gs[[1]])

# GatingSet
cyto_markers(gs)

# cytoset
cs <- cyto_data_extract(gs, "root")[["root"]]
cyto_markers(cs)

# cytoframe
cyto_markers(cs[[1]])

}
\seealso{
\code{\link{cyto_channels}}

\code{\link{cyto_fluor_channels}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
