% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_markers}
\alias{cyto_markers}
\title{Assign marker names to flowFrame or flowSet}
\usage{
cyto_markers(x, file = NULL)
}
\arguments{
\item{x}{object of class \code{flowFrame}, \code{flowSet},
\code{GatingHierarchy} or \code{GatingSet}.}

\item{file}{name of csv file containing columns 'Channel' and 'Marker'.}
}
\value{
save inputs to "Experiment-Markers.csv" and returns updated samples.
}
\description{
\code{cyto_markers} opens an editable table containing a list of channels and
markers for a \code{flowFrame}, \code{flowSet}, \code{GatingHierarchy} or
\code{GatingSet}. Users can edit the \code{name} or \code{desc} columns with
updated channel names or marker names respectively. These entries will be
updated in the \code{x} upon closing the window and saved to a
"Experiment-markers.csv" file for future use.
}
\examples{
\dontrun{
library(CytoRSuiteData)

# Load in samples
fs <- Activation

# Add marker names to channels - edit table
fs <- cyto_markers(fs)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
