% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{cyto_markers}
\alias{cyto_markers}
\title{Assign marker names to flowFrame or flowSet}
\usage{
cyto_markers(x, file = NULL)
}
\arguments{
\item{x}{object of class \code{flowFrame} or \code{flowSet}.}

\item{file}{name of csv file containing columns 'Channel' and 'Marker'.}
}
\value{
NULL and update marker names of \code{x}.
}
\description{
\code{markers_assign} opens an editable table containing a list of channels
and markers for a \code{flowFrame} or \code{flowSet}. Users can edit the
\code{markers} column as required and these entries will be updated in the
\code{flowFrame} or \code{flowSet} upon closing the window.
}
\examples{
\dontrun{
library(CytoRSuiteData)

# Load in samples
fs <- Activation

# Add marker names to channels - edit table
cyto_markers(fs)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
