% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_markers_edit}
\alias{cyto_markers_edit}
\title{Assign marker names to flowFrame or flowSet}
\usage{
cyto_markers_edit(x, file = NULL)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{flowSet},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingSet}} or
\code{GatingSet}.}

\item{file}{name of csv file containing columns 'Channel' and 'Marker'.}
}
\value{
save inputs to "Experiment-Markers.csv" and returns updated samples.
}
\description{
\code{cyto_markers_edit} opens an editable table containing a list of
channels and markers for a \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}},
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}. Users can edit the
\code{name} or \code{desc} columns with updated channel names or marker names
respectively. These entries will be updated in the \code{x} upon closing the
window and saved to a "Experiment-markers.csv" file for future use.
}
\examples{

\dontrun{
# Load in CytoExploreRData to access data
library(CytoExploreRData)

# Load in samples
fs <- Activation

# Add marker names to channels - edit table
fs <- cyto_markers_edit(fs)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
