% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_channels-helpers.R
\name{cyto_markers_extract}
\alias{cyto_markers_extract}
\title{Extract marker names for certain channels}
\usage{
cyto_markers_extract(
  x,
  channels,
  skip = NULL,
  append = FALSE,
  plot = FALSE,
  escape = TRUE,
  ignore.case = TRUE,
  ...
)
}
\arguments{
\item{x}{an object of class \code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{channels}{vector of channel and/or marker names (e.g. c("Alexa Fluor
700-A","CD8")) or channel indices.}

\item{skip}{vector of markers/channels in \code{channels} to bypass when
converting to valid markers, for example \code{"Unstained"} is bypassed
when checking markers in the channel match file.}

\item{append}{logical indicating whether the name of the channel should be
appended to the marker names in the form \code{<marker> channel}, set to
FALSE by default.}

\item{plot}{logical indicating whether the channels will be used to construct
a plot, set to FALSE by default. If set to TRUE an additional check will be
performed to ensure that only 1 or 2 \code{channels} are supplied.}

\item{escape}{logical to indicate whether special characters should be
escaped prior to performing partial marker matching, set to TRUE by
default.}

\item{ignore.case}{logical indicating whether case insensitive channel
matches should be returned, set to TRUE bu default.}

\item{...}{additional arguments passed to \code{\link[base:grep]{grepl}} for
character matching. For exact character string matching to override the
default which ignores character case, set \code{fixed} to TRUE.}
}
\value{
A vector of marker names.
}
\description{
\code{cyto_markers_extract} will check whether the supplied channels or
marker names are valid for the
\code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} and return a vector of
marker names. The name of the channel will be returned if there is no
associated marker found.
}
\examples{
library(CytoExploreRData)

# Activation GatingSet
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                          package = "CytoExploreRData"))

# Extract channels for CD4 & CD8
cyto_markers_extract(gs, c("Alexa Fluor 700-A", "CD8"))

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
