% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_merge_by}
\alias{cyto_merge_by}
\title{Merge a cytoset by experiment variables}
\usage{
cyto_merge_by(
  x,
  parent = "root",
  merge_by = "all",
  select = NULL,
  barcode = TRUE,
  format = "cytoset",
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the parent population to merge when a
\code{GatingHierarchy} or \code{GatingSet} object is supplied, set to the
\code{"root"} node by default.}

\item{merge_by}{vector of \code{\link{cyto_details}} column names (e.g.
c("Treatment","Concentration") indicating how the samples should be grouped
prior to merging. Set this argument to \code{"all"} or \code{NA} to merge
all samples.}

\item{select}{selection criteria passed to \code{\link{cyto_select}} which
indicates which samples in each group to retain prior to merging, set to
NULL by default to merge all samples in each group. Filtering steps should
be comma separated and wrapped in a list. Refer to
\code{\link{cyto_select}} for more details.}

\item{barcode}{logical indicating whether a call should be made to
\code{\link{cyto_barcode}} prior to grouping and merging samples, set to
TRUE by default. Barcoding helps \code{\link{cyto_sample}} to appropriately
sample events based on the number of merged samples.}

\item{format}{either \code{"cytoframe"} or \code{"cytoset"} to indicate the
desired format for merged groups, set to \code{"cytoset"} by default.}

\item{...}{additional arguments passed to \code{\link{cyto_barcode}}.}
}
\value{
list of flowFrames merged by the grouping variables specified by
  \code{merge_by}.
}
\description{
\code{cyto_merge_by} makes a call to \code{cyto_group_by} to split samples
into groups based on experiment variables. The resulting groups are then
coerced in a \code{cytoset} containing a single \code{cytoframe} with the
merged data. \code{cyto_merge_by} is the preferred way to merge samples in
CytoExploreR as it will ensure appropriate sampling in \code{cyto_plot}.
}
\examples{
library(CytoExploreRData)

# Activation Gatingset
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                          package = "CytoExploreRData"))

# Experiment details
cyto_details(gs)

# Merge samples by 'Treatment'
cs_list <- cyto_merge_by(gs, "Treatment")

# Merge samples by 'OVAConc'
cs_list <- cyto_merge_by(gs, "OVAConc")

}
\seealso{
\code{\link{cyto_groups}}

\code{\link{cyto_group_by}}

\code{\link{cyto_select}}

\code{\link{cyto_barcode}}

\code{\link{cyto_sample}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
