% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_merge_by}
\alias{cyto_merge_by}
\title{Merge a flowSet by experiment variables}
\usage{
cyto_merge_by(x, merge_by = "all")
}
\arguments{
\item{x}{object of class \code{flowSet}.}
}
\value{
list of flowFrames merged by the grouping variables specified by
  \code{merge_by}.
}
\description{
\code{cyto_merge_by} makes a call to \code{cyto_group_by} to split samples
into groups based on experiment variables. The resulting groups are then
converted to flowFrames using \code{cyto_convert}. \code{cyto_merge_by} is
the preferred way to merge samples in CytoExploreR as it will ensure
appropriate sampling in \code{cyto_plot}.
}
\examples{

# Load CytoExploreRData to acces data
library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Experiment details
cyto_details(fs)

# Merge samples by 'Treatment'
fr_list <- cyto_merge_by(fs, "Treatment")

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
