% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_merge_by}
\alias{cyto_merge_by}
\title{Merge a flowSet by experiment variables}
\usage{
cyto_merge_by(x, merge_by = "all", select = NULL, barcode = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{flowSet}.}

\item{merge_by}{vector of \code{\link{cyto_details}} column names (e.g.
c("Treatment","Concentration") indicating how the samples should be grouped
prior to merging.}

\item{select}{selection critieria passed to \code{\link{cyto_select}} which
indicates which samples in each group to retain prior to merging, set to
NULL by default to merge all samples in each group. Filtering steps should
be comma separated and wrapped in a list. Refer to
\code{\link{cyto_select}} for more details.}

\item{barcode}{logical indicating whether a call should be made to
\code{\link{cyto_barcode}} prior to grouping and merging samples, set to
TRUE by default. Barcoding helps \code{\link{cyto_sample}} to appropriately
sample events based on the number of merged samples.}

\item{...}{additional arguments passed to \code{\link{cyto_barcode}}.}
}
\value{
list of flowFrames merged by the grouping variables specified by
  \code{merge_by}.
}
\description{
\code{cyto_merge_by} makes a call to \code{cyto_group_by} to split samples
into groups based on experiment variables. The resulting groups are then
converted to flowFrames using \code{cyto_convert}. \code{cyto_merge_by} is
the preferred way to merge samples in CytoExploreR as it will ensure
appropriate sampling in \code{cyto_plot}.
}
\examples{

# Load CytoExploreRData to acces data
library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Experiment details
cyto_details(fs)

# Merge samples by 'Treatment'
fr_list <- cyto_merge_by(fs, "Treatment")
}
\seealso{
\code{\link{cyto_group_by}}

\code{\link{cyto_select}}

\code{\link{cyto_barcode}}

\code{\link{cyto_sample}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
