% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_names<-}
\alias{cyto_names<-}
\title{Replacement method for cyto_names}
\usage{
cyto_names(x) <- value
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{value}{vector of replacement names.}
}
\description{
Replacement method for cyto_names
}
\examples{
library(CytoExploreRData)

# Activation Gatingset
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                         package = "CytoExploreRData"))

# GatingSet
cyto_names(gs)[1] <- "Activation_001.fcs"
cyto_names(gs)

# cytoset
cs <- cyto_data_extract(gs, "root")[["root"]]
cyto_names(cs)[1:2] <- c("Activation_1.fcs", "Activation_2.fcs")
cyto_names(cs)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
