% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_names}
\alias{cyto_names}
\alias{cyto_names.flowFrame}
\alias{cyto_names.flowSet}
\alias{cyto_names.GatingHierarchy}
\alias{cyto_names.GatingSet}
\alias{cyto_names.list}
\title{Extract sample names}
\usage{
cyto_names(x)

\method{cyto_names}{flowFrame}(x)

\method{cyto_names}{flowSet}(x)

\method{cyto_names}{GatingHierarchy}(x)

\method{cyto_names}{GatingSet}(x)

\method{cyto_names}{list}(x)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}
}
\value{
names associated with the supplied object.
}
\description{
Simply a convenient and autocomplete-friendly wrapper around
\code{\link[flowCore:identifier-methods]{identifier}}
\code{\link[flowWorkspace:sampleNames]{sampleNames}} to extract the sample
names from flowFrame, flowSet GatingHierarchy or GatingSet. Anonymous
\code{\link[flowCore:flowFrame-class]{flowFrame}} identifiers will be
converted to \code{"Combined Events"}.
}
\examples{

#' # Load in CytoExploreRData to access data
library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Activation GatingSet
gs <- GatingSet(fs)

# flowFrame
cyto_names(fs[[1]])

# flowSet
cyto_names(fs)

# GatingHierarchy
cyto_names(gs[[1]])

# GatingSet
cyto_names(gs)
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
