% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_names}
\alias{cyto_names}
\alias{cyto_names.flowFrame}
\alias{cyto_names.flowSet}
\alias{cyto_names.GatingHierarchy}
\alias{cyto_names.GatingSet}
\title{Extract sample names}
\usage{
cyto_names(x)

\method{cyto_names}{flowFrame}(x)

\method{cyto_names}{flowSet}(x)

\method{cyto_names}{GatingHierarchy}(x)

\method{cyto_names}{GatingSet}(x)
}
\arguments{
\item{object}{of class \code{flowFrame}, \code{flowSet},
\code{GatingHierarchy} or \code{GatingSet}.}
}
\value{
names associated with the supplied object.
}
\description{
Simply a convenient and autocomplete-friendly wrapper around
\code{\link[flowCore:identifier]{identifier}}
\code{\link[flowWorkspace:sampleNames]{sampleNames}} to extract the sample
names from flowFrame, flowSet GatingHierarchy or GatingSet.
}
