% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_names}
\alias{cyto_names}
\title{Extract sample names}
\usage{
cyto_names(x)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}
}
\value{
names associated with the supplied object.
}
\description{
Simply a convenient and autocomplete-friendly wrapper around
\code{\link[flowWorkspace:sampleNames]{sampleNames}} to extract the sample
names from a cytoset, GatingHierarchy or GatingSet.
}
\examples{
library(CytoExploreRData)

# Activation Gatingset
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                         package = "CytoExploreRData"))

# GatingSet
cyto_names(gs)

# GatingHierarchy
cyto_names(gs[[1]])

# cytoset
cs <- cyto_data_extract(gs, "root")[["root"]]
cyto_names(cs)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
