% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_nodes}
\alias{cyto_nodes}
\title{Names of gated populations in gatingTemplate/GatingHierarchy/GatingSet}
\usage{
cyto_nodes(
  x,
  path = "full",
  terminal = FALSE,
  hidden = FALSE,
  select = NULL,
  exclude = NULL,
  bool = NULL,
  ignore.case = TRUE,
  ...
)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingSet}},
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}},
\code{\link[openCyto:gatingTemplate-class]{gatingTemplate}}, or the name of
a gatingTemplate CSV file.}

\item{path}{can be either \code{"full"} or \code{"auto"} to return the
complete path or shortest unique path to each node.}

\item{terminal}{logical to indicate whether only the paths to terminal nodes
should be returned, set to FALSE by default.}

\item{hidden}{logical to control whether the paths to hidden nodes should be
included, set to FALSE by default. \code{hidden} acts in place of the
\code{showHidden} argument for consistency with other CytoExploreR APIs.}

\item{select}{either a vector indices or character strings for partial
matching to the \code{path} nodes to extract the required nodes, set to
NULL by default to select all nodes.}

\item{exclude}{either a vector indices or character strings for partial
matching to the \code{path} nodes to prevent extraction of certain nodes,
set to NULL by default to include all nodes.}

\item{bool}{can be either TRUE to only return boolean nodes, FALSE to exclude
boolean nodes or NULL to include boolean nodes with other nodes.}

\item{ignore.case}{logical indicating whether case should be ignored when
selecting nodes using the \code{select} or \code{exclude} arguments, set to
TRUE by default.}

\item{...}{additional arguments passed to
\code{\link[flowWorkspace:gs_get_pop_paths]{gs_get_pop_paths}} or
\code{\link[openCyto:gt_get_nodes]{gt_get_nodes}}.}
}
\value{
character vector of gated node/population names.
}
\description{
\code{cyto_nodes} is simply an autocomplete-friendly wrapper for
\code{\link[flowWorkspace:gs_get_pop_paths]{gs_get_pop_paths}} with
additional \code{select} and \code{exclude} arguments for node selection.
}
