% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_nodes}
\alias{cyto_nodes}
\title{Extract Names of Gated Populations in GatingHierarchy or GatingSet}
\usage{
cyto_nodes(x, ...)
}
\arguments{
\item{x}{object of class \code{GatingHierarchy} or \code{GatingSet}.}

\item{...}{additional arguments passed to
\code{\link[flowWorkspace:getNodes]{getNodes}}.}
}
\value{
character vector of gated node/population names.
}
\description{
\code{cyto_nodes} is simply an autocomplete-friendly wrapper for
\code{\link[flowWorkspace:getNodes]{getNodes}}
}
