% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_peaks_assign.R
\name{cyto_peaks_assign}
\alias{cyto_peaks_assign}
\title{Assign peak detectors to pre-gated single colour controls}
\usage{
cyto_peaks_assign(x, events = 500, overwrite = FALSE)
}
\arguments{
\item{x}{object of class \code{cytoset}, \code{GatingHierarchy} or
\code{GatingSet}.}

\item{events}{number of events to extract from unstained and stained controls
to identify peak detectors, set to 500 events by default.}

\item{overwrite}{logical indicating whether existing detector assignments in
\code{cyto_details(x)$channel} be overwritten, set to FALSE by default.}
}
\value{
updated \code{"channel"} column in \code{cyto_details(x)}.
}
\description{
Assign peak detectors to pre-gated single colour controls
}
\seealso{
\code{\link{cyto_channel_match}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
