% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-methods.R
\name{cyto_plot}
\alias{cyto_plot}
\title{cyto_plot}
\usage{
cyto_plot(x, ...)
}
\arguments{
\item{x}{object of class \code{flowFrame}, \code{flowSet},
\code{GatingHierarchy} or \code{GatingSet}.}

\item{...}{additional method-specific arguments for cyto_plot.}
}
\description{
Explore & visualise cytometry data.
}
\details{
\code{cyto_plot} is a cytometry visualisation tool built on base graphics
which supports all existing cytometry objects including
\code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} and
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} objects. The type of
plot constructed by \code{cyto_plot} is determined by the number of channels
supplied to the \code{channels} argument. If a single channel is supplied, a
1-D density distribution will be plotted. Supplying two channels to this
argument will result in construction of a 2-D scatterplot with a blue-red
density colour scale for points.

Some key features of \code{cyto_plot} include: \itemize{ \item autocomplete
friendly arguments \item minimal coding \item 1-D density distributions \item
2-D scatterplots \item stacked density distributions \item 2-D contour lines
\item back-gating \item overlays \item population labelling \item grid to
visualise multiple samples \item plot gate objects \item plot in pop-up
windows \item sampling \item fully customisable \item copy and paste
arguments from gate_draw }
}
\seealso{
\code{\link{cyto_plot,flowFrame-method}}

\code{\link{cyto_plot,flowSet-method}}

\code{\link{cyto_plot,GatingHierarchy-method}}

\code{\link{cyto_plot,GatingSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
