% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_compensation-methods.R
\name{cyto_plot_compensation}
\alias{cyto_plot_compensation}
\title{Plot Compensation in All Fluorescent Channels}
\usage{
cyto_plot_compensation(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}} or
\code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} containing gated
compensation controls and an unstained control.}

\item{...}{additional method-specific arguments.}
}
\description{
Plot each compensation control in all fluorescent channels to identify any
potential compensation issues. The unstained control is overlaid in black as
a reference.
}
\seealso{
\code{\link{cyto_plot_compensation,flowFrame-method}}

\code{\link{cyto_plot_compensation,flowSet-method}}

\code{\link{cyto_plot_compensation,GatingSet-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
