% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_complete}
\alias{cyto_plot_complete}
\title{Indicate Completion of Custom cyto_plot Layout for Saving}
\usage{
cyto_plot_complete()
}
\description{
Indicate Completion of Custom cyto_plot Layout for Saving
}
\examples{
library(CytoRSuiteData)

# Load samples into GatingSet
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(gs))
gs <- transform(gs, trans)

# Apply gatingTemplate
gt <- Activation_gatingTemplate
gating(gt, gs)

# Save custom plot
cyto_plot_save("Custom.png",
  height = 8,
  width = 16
)

# Set out plot layout
cyto_plot_layout(1, 2)

# Add 2D plot
cyto_plot(gs[[4]],
  parent = "CD4 T Cells",
  alias = "",
  channels = c("Alexa Fluor 647-A", "7-AAD-A"),
  layout = FALSE
)

# Add 1D plot
cyto_plot(gs,
  parent = "CD4 T Cells",
  alias = "",
  channels = "7-AAD-A",
  density_stack = 0.6,
  layout = FALSE
)

# Signal that the plot is complete
cyto_plot_complete()
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
