% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_complete}
\alias{cyto_plot_complete}
\title{Indicate Completion of Custom cyto_plot Layout for Saving}
\usage{
cyto_plot_complete(layout = NULL, outer_margins = NULL)
}
\arguments{
\item{layout}{either a vector of the form c(nrow, ncol) defining the
dimensions of the plot or a matrix defining a more sophisticated layout
(see \code{\link[graphics]{layout}}). Vectors can optionally contain a
third element to indicate whether plots should be placed in row (1) or
column (2) order, set to row order by default.}

\item{outer_margins}{vector of length 4 indicating the number of lines of
text to place around the \code{c(bottom, left, top, right)} outer borders
of the plot, set to NULL by default to use \code{par("oma")}.}
}
\description{
Indicate Completion of Custom cyto_plot Layout for Saving
}
\examples{
library(CytoExploreRData)

# Load samples into GatingSet
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(gs))
gs <- transform(gs, trans)

# Apply gatingTemplate
gt <- Activation_gatingTemplate
gt_gating(gt, gs)

# Save custom plot
cyto_plot_save("Custom.png",
  height = 8,
  width = 16
)

# Set out plot layout
cyto_plot_layout(c(1, 2))

# Add 2D plot
cyto_plot(gs[[4]],
  parent = "CD4 T Cells",
  alias = "",
  channels = c("Alexa Fluor 647-A", "7-AAD-A"),
  layout = FALSE
)

# Add 1D plot
cyto_plot(gs,
  parent = "CD4 T Cells",
  alias = "",
  channels = "7-AAD-A",
  hist_stack = 0.6,
  layout = FALSE
)

# Signal that the plot is complete
cyto_plot_complete()
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
