% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_contour-methods.R
\name{cyto_plot_contour}
\alias{cyto_plot_contour}
\title{Contour Lines for cyto_plot}
\usage{
cyto_plot_contour(x, ...)
}
\arguments{
\item{x}{object to use for contour lines to overlay onto an existing
cyto_plot.}

\item{...}{additional method-specific arguments}
}
\description{
Contour Lines for cyto_plot
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
