% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_custom}
\alias{cyto_plot_custom}
\title{Set customised graphical parameters for cyto_plot}
\usage{
cyto_plot_custom(...)
}
\arguments{
\item{...}{graphical parameters supplied by name to be passed to
\code{\link{cyto_plot_par}} to customise the current graphics device.}
}
\description{
\code{cyto_plot_custom()} make a call to \code{cyto_plot_par()} to set
customised graphical parameters for the current graphics device.
\code{cyto_plot_custom()} will then signal to \code{cyto_plot()} to ensure
that these set graphical parameters are not overridden and to make sure that
the plots are only saved when the user calls \code{cyto_plot_complete()}.
\code{cyto_plot_custom()} will also record the current graphical parameters
before changing them, so that these can be reset when
\code{cyto_plot_complete()} is called.
}
\details{
The sequence of calls to save a customised plot are described below:
\itemize{\item \code{cyto_plot_save()} - sets a new graphics device of the
appropriate type (e.g. \code{png()}) and makes sure that \code{cyto_plot()}
does not open other graphics devices \item \code{cyto_plot_custom()} - sets
up the current graphics device with customised graphical parameters and
indicates to \code{cyto_plot()} that the user will take control over when the
plot should be saved. \item \code{cyto_plot()} - the users can then makes
calls to any of the \code{cyto_plot()} family of function to construct the
plot on the customised graphics device. \item \code{cyto_plot_complete()} -
signals that the plot is complete and ready for saving to file. Since
\code{cyto_plot_complete()} also resets a lot of graphical parameters it
should be called even if a call has not been made to \code{cyto_plot_save()}.
}
}
\examples{
\dontrun{
library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Save plot
cyto_plot_save("Test.png",
  height = 7,
  width = 14
)

# Create custom plot - 1D & 2D plot panels
cyto_plot_custom(layout = c(1, 2))
cyto_plot(fs[[32]],
  channels = "FSC-A"
)
cyto_plot(fs[[32]],
  channels = c("FSC-A", "SSC-A")
)

# Signal plot is complete and save
cyto_plot_complete()
}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
