% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_custom}
\alias{cyto_plot_custom}
\title{Create custom cyto_plot}
\usage{
cyto_plot_custom(nrow = NULL, ncol = NULL, by = "row")
}
\arguments{
\item{nrow}{number of rows to include in the panel layout.}

\item{ncol}{number of columns to include in the panel layout.}

\item{by}{chracter string indicating whether the plot should be filled by
\code{"row"} or \code{"column"}, set to \code{"row"} by default.}
}
\description{
Signal to \code{cyto_plot} that a custom plot is being created to ensure that
plots are appropraitely saved with \code{cyto_plot_save}.
\code{cyto_plot_custom} calls must be made before \code{cyto_plo_save} calls
and \code{cyto_plot} calls should be followed by a call to
\code{cyto_plot_complete} to indicate when the plot is complete and should be
saved.
}
\examples{
\dontrun{
library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Signal custom plot
cyto_plot_custom()

# Save plot
cyto_plot_save("Test.png",
               height = 7,
               width = 14)

# Create custom plot - 1D & 2D plot panels
cyto_plot_layout(1,2)
cyto_plot(fs[[32]],
          channels = "FSC-A")
cyto_plot(fs[[32]],
          channels = c("FSC-A","SSC-A"))

# Signal plot is complete and save
cyto_plot_complete()
}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
