% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_empty}
\alias{cyto_plot_empty}
\alias{cyto_plot_empty.flowFrame}
\alias{cyto_plot_empty.list}
\title{Create an empty cyto_plot}
\usage{
cyto_plot_empty(x, ...)

\method{cyto_plot_empty}{flowFrame}(x, channels, axes_trans = NA,
  overlay = NA, xlim = NA, ylim = NA, limits = "machine",
  title = NA, xlab = NA, ylab = NA, density_modal = TRUE,
  density_smooth = 1.5, density_stack = 0.5, axes_text = c(TRUE,
  TRUE), axes_text_font = 1, axes_text_size = 1,
  axes_text_col = "black", axes_label_text_font = 1,
  axes_label_text_size = 1.1, axes_label_text_col = "black",
  title_text_font = 2, title_text_size = 1.1,
  title_text_col = "black", border_line_type = 1,
  border_line_width = 1, border_line_col = "black",
  border_fill = "white", legend = FALSE, legend_text)

\method{cyto_plot_empty}{list}(x, channels, axes_trans = NA, xlim = NA,
  ylim = NA, limits = "machine", title = NA, xlab = NA,
  ylab = NA, density_modal = TRUE, density_smooth = 1.5,
  density_stack = 0.5, axes_text = c(TRUE, TRUE), axes_text_font = 1,
  axes_text_size = 1, axes_text_col = "black",
  axes_label_text_font = 1, axes_label_text_size = 1.1,
  axes_label_text_col = "black", title_text_font = 2,
  title_text_size = 1.1, title_text_col = "black",
  border_line_type = 1, border_line_width = 1,
  border_line_col = "black", border_fill = "white", legend = FALSE,
  legend_text)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}}.}

\item{axes_trans}{object of class
\code{\link[flowCore:transformList-class]{transformList}} or
\code{\link[flowWorkspace]{transformerList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} which was used to
transform the fluorescent channels of the supplied flowFrame. This
transformation object will be used internally to ensure that the axes
labels of the plot are appropriately transformed. The transformation object
will NOT be applied to the flowFrame internally and should be applied to
the flowFrame prior to plotting.}

\item{overlay}{a list of flowFrames to overlay onto the plot.}

\item{xlim}{lower and upper limits of x axis (e.g. c(0,5)).}

\item{ylim}{lower and upper limits of y axis (e.g. c(0,5)).}

\item{limits}{indicates whether the axes limits should be based on the
\code{"data"} or \code{"machine"}, set to "machine" by default to show
complete axes ranges. This argument will only alter the upper axis limits,
to modify the lower limits use \code{xlim} and \code{ylim}.}

\item{title}{title to use for the plot, set to the name of the sample by
default. Title can be removed by setting this argument to \code{NA}.}

\item{xlab}{x axis label.}

\item{ylab}{y axis label.}

\item{density_modal}{logical indicating whether density should be normalised
to mode and presented as a percentage. Set to \code{TRUE} by default.}

\item{density_smooth}{smoothing parameter passed to
\code{\link[stats:density]{density}} to adjust kernel density.}

\item{density_stack}{numeric [0,1] indicating the degree of offset for
overlaid populations, set to 0.5 by default.}

\item{axes_text}{logical vector of length 2 indicating whether axis text
should be included for the x and y axes respectively, set to
\code{c(TRUE,TRUE)} by default to display axes text on both axes.}

\item{axes_text_font}{numeric indicating the font to use for axes, set to 1
for plain font by default. See \code{\link[graphics:par]{?par}} font for
details.}

\item{axes_text_size}{character expansion for axis text.}

\item{axes_text_col}{colour of axis text.}

\item{axes_label_text_font}{numeric indicating the font to use for title, set
to 1 for plain font by default. See \code{\link[graphics:par]{?par}} font
for details.}

\item{axes_label_text_size}{character expansion for axis labels.}

\item{axes_label_text_col}{colour of axis labels.}

\item{title_text_font}{numeric indicating the font to use for title, set to 2
for bold font by default. See \code{\link[graphics:par]{?par}} font for
details.}

\item{title_text_size}{character expansion for plot title.}

\item{title_text_col}{colour for plot title.}

\item{border_line_type}{line type to use for plot border, set to 1 by default
for a sold border.}

\item{border_line_width}{line width for plot border, set to 1 by default.}

\item{border_line_col}{line colour for plot border, set to "black" by
default.}

\item{border_fill}{colour to use for the plot background, set to "white" by
default.}
}
\description{
\code{cyto_plot_empty} generates to base for cyto_plot by creating an empty
plot with border, axes, axes_text and titles. Data is subsequently added to
this base layer with \code{cyto_plot_point} or \code{cyto_plot_density}.
}
\examples{
library(CytoRSuiteData)

# Construct an empty 2D plot with black background
cyto_plot_empty(Activation[1],
                channels = c("FSC-A","SSC-A"),
                border_fill = "black)
                
# Construct an empty 1D plot
cyto_plot_empty(Activation[1],
                channels = c("FSC-A),
                overlay = Activation[2:4])

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
