% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_empty}
\alias{cyto_plot_empty}
\title{Create an empty cyto_plot}
\usage{
cyto_plot_empty(
  x,
  channels,
  axes_trans = NA,
  overlay = NA,
  gate = NA,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  axes_limits = "auto",
  axes_limits_buffer = 0.03,
  title,
  xlab,
  ylab,
  margins = c(NA, NA, NA, NA),
  hist_stat = "count",
  hist_smooth = 1,
  hist_bins = 256,
  hist_stack = 0.5,
  hist_layers = 1,
  hist_cols = NA,
  hist_fill = NA,
  hist_fill_alpha = 1,
  hist_line_type = 1,
  hist_line_width = 1,
  hist_line_col = "black",
  point_shape = ".",
  point_size = 2,
  point_col_scale = NA,
  point_cols = NA,
  point_col = NA,
  point_col_alpha = 1,
  axes_text = c(TRUE, TRUE),
  axes_text_font = 1,
  axes_text_size = 1,
  axes_text_col = "black",
  axes_label_text_font = 1,
  axes_label_text_size = 1.1,
  axes_label_text_col = "black",
  title_text_font = 2,
  title_text_size = 1.1,
  title_text_col = "black",
  border_line_type = 1,
  border_line_width = 1,
  border_line_col = "black",
  border_fill = "white",
  border_fill_alpha = 1,
  legend = FALSE,
  legend_text,
  legend_text_font = 1,
  legend_text_size = 1,
  legend_text_col = "black",
  legend_line_type = NA,
  legend_line_width = NA,
  legend_line_col = NA,
  legend_box_fill = NA,
  legend_point_col = NA,
  key = "both",
  key_size = c(1, 1),
  key_scale = "fixed",
  key_text_font = 1,
  key_text_size = 0.9,
  key_text_col = "black",
  key_text_col_alpha = 1,
  key_title = "",
  key_title_text_font = 1,
  key_title_text_size = 1,
  key_title_text_col = "black",
  key_title_text_col_alpha = 1,
  key_hist_line_type = 1,
  key_hist_line_width = 1.5,
  key_hist_line_col = "black",
  grid = TRUE,
  grid_line_type = 1,
  grid_line_width = 1,
  grid_line_col = "grey95",
  grid_line_alpha = 1,
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}}.}

\item{channels}{name of the channel(s) or marker(s) to be used to construct
the plot. The length of channels determines the type of plot to be
constructed, either a 1-D density distribution for a single channel or a
2-D scatterplot with blue-red colour scale for two channels.}

\item{axes_trans}{object of class
\code{\link[flowCore:transformList-class]{transformList}} or
\code{\link[flowWorkspace]{transformerList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} which was used to
transform the fluorescent channels of the supplied flowFrame. This
transformation object will be used internally to ensure that the axes
labels of the plot are appropriately transformed. The transformation object
will NOT be applied to the flowFrame internally and should be applied to
the flowFrame prior to plotting.}

\item{overlay}{a list of cytosets to overlay onto the plot.}

\item{gate}{list of gate objects to be plotted, used internally to ensure
gate co-ordinates are taken into account when computing axes limits.}

\item{xlim}{lower and upper limits of x axis (e.g. c(0,5)).}

\item{ylim}{lower and upper limits of y axis (e.g. c(0,5)).}

\item{axes_limits}{options include \code{"auto"}, \code{"data"} or
\code{"machine"} to use optimised, data or machine limits respectively. Set
to \code{"auto"} by default to use optimised axes ranges. Fine control over
axes limits can be obtained by altering the \code{xlim} and \code{ylim}
arguments.}

\item{axes_limits_buffer}{decimal indicating the percentage of buffering to
add to either end of the axes limits, set to 0.03 by default.}

\item{title}{title to use for the plot, set to the name of the sample by
default. Title can be removed by setting this argument to \code{NA}.}

\item{xlab}{x axis label.}

\item{ylab}{y axis label.}

\item{margins}{a vector of length 4 to control the margins around the bottom,
left, top and right of the plot, set to c(NA, NA, NA, NA) by default to let
`cyto_plot` compute optimal margins.}

\item{hist_stat}{can be either \code{"count"}, \code{"percent"} or
\code{"density"} to indicate the statistic to display on histograms, set to
\code{"percent"} by default. The \code{"percent"} option applies modal
normalisation and expresses the result as a percentage.}

\item{hist_smooth}{smoothing parameter passed to
\code{\link[stats:density]{density}} to adjust the smoothness of the kernel
density for histograms, set to \code{1} by default.}

\item{hist_bins}{number of bins to use for histograms, set to 256 by default.}

\item{hist_stack}{numeric [0,1] indicating the degree of stacking for
histograms, set to \code{0} by default.}

\item{hist_layers}{numeric indicating the number of histograms to stack in
each plot, set to all samples by default. Each plot must contain the same
number of histograms.}

\item{hist_cols}{vector colours to draw from when selecting histogram fill
colours if none are supplied to \code{hist_fill}.}

\item{hist_fill}{fill colour(s) for histograms, select from \code{hist_cols}
if not supplied.}

\item{hist_fill_alpha}{numeric [0,1] used to control histogram fill colour
transparency, set to \code{1} by default for solid colours.}

\item{hist_line_type}{line type(s) to use for histogram borders, set to 1 by
default to use solid lines. See \code{\link[graphics:par]{lty}} for
alternatives.}

\item{hist_line_width}{numeric to control line width(s) for histogram borders
lines, set to 1 by default.}

\item{hist_line_col}{colour(s) for histogram borders, set to \code{"black"}
by default.}

\item{point_shape}{shape(s) to use for points in 2-D scatterplots, set to
\code{"."} by default to maximise plotting speed.  See
\code{\link[graphics:par]{pch}} for alternatives.}

\item{point_size}{numeric to control the size of points in 2-D scatter plots
set to 2 by default.}

\item{point_col_scale}{vector of colours to use for density gradient.}

\item{point_cols}{vector colours to draw from when selecting colours for
points if none are supplied to point_col.}

\item{point_col}{colour(s) to use for points in 2-D scatter plots, set to NA
by default to use a blue-red density colour scale.}

\item{point_col_alpha}{numeric [0,1] to control point colour transparency in
2-D scatter plots, set to 1 by default to use solid colours.}

\item{axes_text}{logical vector of length 2 indicating whether axis text
should be included for the x and y axes respectively, set to
\code{c(TRUE,TRUE)} by default to display axes text on both axes.
\code{axes_text} can also be set to NA to remove the text only but keep the
axes ticks.}

\item{axes_text_font}{numeric indicating the font to use for axes, set to 1
for plain font by default. See \code{\link[graphics:par]{?par}} font for
details.}

\item{axes_text_size}{character expansion for axis text.}

\item{axes_text_col}{colour of axis text.}

\item{axes_label_text_font}{numeric indicating the font to use for title, set
to 1 for plain font by default. See \code{\link[graphics:par]{?par}} font
for details.}

\item{axes_label_text_size}{character expansion for axis labels.}

\item{axes_label_text_col}{colour of axis labels.}

\item{title_text_font}{numeric indicating the font to use for title, set to 2
for bold font by default. See \code{\link[graphics:par]{?par}} font for
details.}

\item{title_text_size}{character expansion for plot title.}

\item{title_text_col}{colour for plot title.}

\item{border_line_type}{line type to use for plot border, set to 1 by default
for a sold border.}

\item{border_line_width}{line width for plot border, set to 1 by default.}

\item{border_line_col}{line colour for plot border, set to "black" by
default.}

\item{border_fill}{colour to use for the plot background, set to "white" by
default.}

\item{border_fill_alpha}{transparency to use for border_fill colour, set to 1
by default to add no transparency.}

\item{legend}{can be either \code{"line"} or \code{"fill"} to indicate
whether a legend should be constructed based on the density \code{"line"}
or \code{"fill"}, set to FALSE by default to remove the legend.}

\item{legend_text}{vector of labels to use in the legend.}

\item{legend_text_font}{numeric to control the font of legend text, set to 1
for plain font by default. See \code{\link[graphics:par]{font}} for
alternatives.}

\item{legend_text_size}{numeric to control the size of text in the legend,
set to 1 by default.}

\item{legend_text_col}{colour(s) to use for text in legend, set to
\code{"black"} by default.}

\item{legend_line_type}{numeric to control the line type for line legends,
set to 1 by default. Refer to \code{lty} in \code{\link[graphics:par]{par}}
for alternatives.}

\item{legend_line_width}{numeric to control the line width in line legend,
set to 1 by default. Refer to \code{lwd} in \code{\link[graphics:par]{par}}
for alternatives.}

\item{legend_line_col}{colour(s) to use for the lines in 1-D plot legends
when legend is set to \code{"line"}.}

\item{legend_box_fill}{fill colour(s) to use for the boxes in 1-D plot
legends when legend is set to \code{"fill"}.}

\item{legend_point_col}{colour(s) to use for points in 2-D scatter plot
legend.}

\item{key}{options include \code{"scale"} to display only the colour scale,
\code{"both"} to display the colour scale and text to annotate
\code{counts} or \code{fluorescence units}, or \code{"none"} to remove the
key from the plot(s). Set to \code{"both"} by default to display colour
scale and associated text in the key.}

\item{key_size}{a vector of scaling factors between 0 and Inf for the weight
and height of the key, set to \code{c(1,1)} by default.}

\item{key_scale}{inherits scale limits computed within \code{cyto_plot},
should be a vector of the form \code{c(min, max)}, \code{cyto_plot_empty()}
will attempt to compute these limits based on the supplied data if
\code{key_scale} is not supplied. See the \code{key_scale} argument in
\code{\link{cyto_plot}} for details.}

\item{key_text_font}{font to use for text in the key, set to 1 by default for
plain font.}

\item{key_text_size}{numeric to control the size of text in the plot key, set
to 0.9 by default.}

\item{key_text_col}{colour to use for text in the plot key, set to
\code{"black"} by default.}

\item{key_text_col_alpha}{numeric [0, 1] to control the transparency of the
text colour in the plot key, set to 1 by default to remove transparency.}

\item{key_title}{title to place above the key set to \code{"count"} when
\code{point_col = NA} or the name of the channel/marker supplied to
\code{point_col}.}

\item{key_title_text_font}{font to use for the key title, set to 1 by default
for plain font.}

\item{key_title_text_size}{numeric to control the size of title text in teh
key, set to 1 by default.}

\item{key_title_text_col}{colour to use for the key title, set to
\code{"black"} by default.}

\item{key_title_text_col_alpha}{numeric [0, 1] to control the transparency of
the key title text, set to 1 by default to remove transparency.}

\item{key_hist_line_type}{line type(s) to use for histogram borders in the
key, set to 1 by default to use solid lines. See
\code{\link[graphics:par]{lty}} for alternatives.}

\item{key_hist_line_width}{numeric to control line width(s) for histogram
borders lines, set to 1 by default.}

\item{key_hist_line_col}{colour to use for the histogram border in the key,
set to \code{"black"} by default.}

\item{grid}{logical indicating whether to include grid lines in the plot
background, set to TRUE by default. Alternatively, users can supply a
integer to indicate the number of equally spaced quantiles to used for the
grid lines.}

\item{grid_line_type}{integer [0,6] to control the line type of grid lines,
set to \code{1} to draw solid lines by default. See
\code{\link[graphics:par]{lty}} for alternatives.}

\item{grid_line_width}{numeric to control the line width(s) of grid lines,
set to \code{1} by default.}

\item{grid_line_col}{colour to use for grid lines, set to \code{"grey95"} by
default.}

\item{grid_line_alpha}{numeric [0,1] to control the transparency of grid
lines, set to 1 by default to remove transparency.}

\item{...}{not in use.}
}
\description{
\code{cyto_plot_empty} generates a base for cyto_plot by creating an empty
plot with border, axes, axes_text and titles. Data is subsequently added to
this base layer with \code{cyto_plot_point} or \code{cyto_plot_hist}.
}
\examples{
library(CytoExploreRData)

# Construct an empty 2D plot with black background
cyto_plot_empty(Activation[[32]],
  channels = c("FSC-A", "SSC-A"),
  border_fill = "black"
)

# Construct an empty 1D plot
cyto_plot_empty(Activation[[32]],
  channels = c("FSC-A"),
  overlay = Activation[1:2]
)
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
