% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_explore.R
\name{cyto_plot_explore}
\alias{cyto_plot_explore}
\alias{cyto_plot_explore.GatingSet}
\alias{cyto_plot_explore.GatingHierarchy}
\alias{cyto_plot_explore.flowSet}
\alias{cyto_plot_explore.flowFrame}
\title{Explore Cytometry Data in Bivariate Plots}
\usage{
\method{cyto_plot_explore}{GatingSet}(
  x,
  parent = NULL,
  channels_x = NULL,
  channels_y = NULL,
  axes_trans = NA,
  layout,
  popup = FALSE,
  title = NA,
  header,
  header_text_font = 2,
  header_text_size = 1,
  header_text_col = "black",
  ...
)

\method{cyto_plot_explore}{GatingHierarchy}(
  x,
  parent = NULL,
  channels_x = NULL,
  channels_y = NULL,
  axes_trans = NA,
  layout,
  popup = FALSE,
  title = NA,
  header,
  header_text_font = 2,
  header_text_size = 1,
  header_text_col = "black",
  ...
)

\method{cyto_plot_explore}{flowSet}(
  x,
  channels_x = NULL,
  channels_y = NULL,
  axes_trans = NA,
  layout,
  popup = FALSE,
  title = NA,
  header,
  header_text_font = 2,
  header_text_size = 1,
  header_text_col = "black",
  ...
)

\method{cyto_plot_explore}{flowFrame}(
  x,
  channels_x = NULL,
  channels_y = NULL,
  axes_trans = NA,
  layout,
  popup = FALSE,
  title = NA,
  header,
  header_text_font = 2,
  header_text_size = 1,
  header_text_col = "black",
  ...
)
}
\arguments{
\item{x}{an object of class \code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierachy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}. \code{cytoset} and
\code{GatingSet} objects will be coerced to a single \code{cytoframe} prior
to plotting.}

\item{parent}{name of the parent population to plot when a
\code{GatingHierarchy} or \code{GatingSet} object is supplied, set to the
last gated population by default.}

\item{channels_x}{vector of channels to explore, set to all fluorescent
channels by default. Each channel in \code{channels_x} will be plotted
against all \code{channels_y}, with each \code{channels_x} on a separate
page.}

\item{channels_y}{vector of channels to plot each channels_x against, set to
all channels by default.}

\item{axes_trans}{object of class transformerList passed to \code{cyto_plot}
to appropriately display transformed data in plots.}

\item{layout}{vector of grid dimensions \code{c(#rows,#columns)} for each
plot.}

\item{popup}{logical indicating whether plots should be constructed in a
pop-up window.}

\item{title}{text to include above each plot, set to NA by default to remove
titles.}

\item{header}{title to use for the plots, set to the name of the sample by
default. Turn off the header by setting this argument to NA.}

\item{header_text_font}{font to use for header text, set to 2 by default.}

\item{header_text_size}{text size for header, set to 1 by default.}

\item{header_text_col}{colour for header text, set to "black" by default.}

\item{...}{additional arguments passed to \code{cyto_plot}.}
}
\description{
\code{cyto_plot_explore} is an extremely useful tool to explore all aspects
of your cytometry data. \code{cyto_plot_explore} constructs a faceted plot
for each of the supplied \code{channels_x} by plotting them against all the
supplied \code{channels_y}. Thus providing a rapid means to explore the data
in all available channels.
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
