% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_explore.R
\name{cyto_plot_explore}
\alias{cyto_plot_explore}
\title{Explore cytometry data in a series of bivariate scatterplots}
\usage{
cyto_plot_explore(
  x,
  parent = "root",
  channel_x = NULL,
  channels_y = NULL,
  select = NULL,
  merge_by = "name",
  order = "channels",
  layout,
  header,
  ...
)
}
\description{
Explore cytometry data in a series of bivariate scatterplots
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
