% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_gate-methods.R
\name{cyto_plot_gate}
\alias{cyto_plot_gate}
\title{Plot Gate Objects onto an Existing Plot}
\usage{
cyto_plot_gate(x, ...)
}
\arguments{
\item{x}{gate object of class
\code{\link[flowCore:rectangleGate-class]{rectangleGate}},
\code{\link[flowCore:polygonGate-class]{polygonGate}},
\code{\link[flowCore:ellipsoidGate-class]{ellipsoidGate}}, \code{list} or
\code{\link[flowCore:filters-class]{filters}}.}

\item{...}{additional method-specific arguments.}
}
\description{
Plot Gate Objects onto an Existing Plot
}
\seealso{
\code{\link{cyto_plot_gate,rectangleGate-method}}

\code{\link{cyto_plot_gate,polygonGate-method}}

\code{\link{cyto_plot_gate,ellipsoidGate-method}}

\code{\link{cyto_plot_gate,list-method}}

\code{\link{cyto_plot_gate,filters-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
