% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_gating_scheme.R
\name{cyto_plot_gating_scheme}
\alias{cyto_plot_gating_scheme}
\alias{cyto_plot_gating_scheme.GatingHierarchy}
\alias{cyto_plot_gating_scheme.GatingSet}
\title{Plot Cytometry Gating Strategies}
\usage{
\method{cyto_plot_gating_scheme}{GatingHierarchy}(
  x,
  gatingTemplate = NULL,
  back_gate = FALSE,
  gate_track = FALSE,
  show_all = FALSE,
  header,
  popup = FALSE,
  layout,
  point_col = NA,
  label_text_size = 0.8,
  border_line_width,
  legend = TRUE,
  legend_text,
  legend_text_size = 1.2,
  header_text_font = 2,
  header_text_size = 1,
  header_text_col = "black",
  ...
)

\method{cyto_plot_gating_scheme}{GatingSet}(
  x,
  gatingTemplate = NULL,
  group_by = "name",
  back_gate = FALSE,
  gate_track = FALSE,
  show_all = FALSE,
  header,
  popup = FALSE,
  layout = NULL,
  point_col = NA,
  label_text_size = 0.8,
  border_line_width = NULL,
  legend = TRUE,
  legend_text = NULL,
  legend_text_size = 1.2,
  header_text_font = 2,
  header_text_size = 1,
  header_text_col = "black",
  ...
)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{gatingTemplate}{name of the gatingTemplate csv file used to gate
\code{x}. If not supplied the gating scheme will be obtained directly from
the GatingSet.}

\item{back_gate}{names of the population(s) to back-gate, set to \code{FALSE}
by default to turn off back-gating. To back-gate all populations set this
argument to \code{"all"}.}

\item{gate_track}{logical indicating whether gate colour should be tracked
throughout gating scheme, set to TRUE by default.}

\item{show_all}{logical indicating whether every population should be
included in every plot in the gating scheme, set to \code{FALSE} by
default.}

\item{header}{character string to use as the header for the plot layout, set
to "Gating Scheme" by default.}

\item{popup}{logical indicating whether the gating scheme should be plotted
in a pop-up window, set to FALSE by default.}

\item{layout}{a vector of the length 2 indicating the dimensions of the grid
for plotting \code{c(#rows, #columns)}.}

\item{point_col}{colour of points in 2D plots set to NA to use default
red-blue colour scale. Control the colour of overlays by supplying multiple
colours to this argument (e.g. c("blue","red")).}

\item{label_text_size}{numeric to control the size of text in the plot
labels, set to 0.8 by default.}

\item{border_line_width}{line width for plot border, set to 3 when gate_track
is TRUE.}

\item{legend}{logical indicating whether a legend should be included when an
overlay is supplied.}

\item{legend_text}{vector of character strings to use for legend when an
overlay is supplied.}

\item{legend_text_size}{character expansion for legend text, set to 1.2 by
default.}

\item{header_text_font}{numeric to control the font of the header, set to 2
by default for bold font.}

\item{header_text_size}{numeric to control the size of the header text, set
to 1 by default.}

\item{header_text_col}{colour to use for header text, set to "black" by
default.}

\item{...}{extra arguments passed to \code{\link{cyto_plot}}.}
}
\description{
\code{cyto_plot_gating_scheme} automatically plots the entire gating scheme
and has full support for gate tracking and back-gating through
\code{gate_track} and \code{back_gate}.
}
\examples{
library(CytoExploreRData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- cyto_compensate(gs)

# Transform fluorescent channels
gs <- cyto_transform(gs)

# Gate using cyto_gate_draw
gt <- Activation_gatingTemplate
gt_gating(gt, gs)

# Gating scheme
cyto_plot_gating_scheme(gs[[4]])

# Back-gating
cyto_plot_gating_scheme(gs[[32]],
  back_gate = TRUE
)

# Gate-tracking
cyto_plot_gating_scheme(gs[1:4],
  gate_track = TRUE
)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
