% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_gating_scheme_v2.R
\name{cyto_plot_gating_scheme}
\alias{cyto_plot_gating_scheme}
\title{Visualise cytometry gating strategies}
\usage{
cyto_plot_gating_scheme(
  x,
  select = NULL,
  gatingTemplate = NULL,
  merge_by = "name",
  overlay = NA,
  order = "nodes",
  hidden = FALSE,
  layout,
  title,
  header,
  legend = TRUE,
  legend_text = NA,
  point_col = NA,
  point_cols = NA,
  gate_line_col = "red",
  border_line_col = "black",
  title_text_col = "black",
  header_text_font = 2,
  header_text_size = 1,
  header_text_col = "black",
  legend_text_font = 1,
  legend_text_size = 1.2,
  ...
)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{gatingTemplate}{name of the gatingTemplate csv file used to gate
\code{x}. If not supplied the gating scheme will be obtained directly from
the GatingHierarchy or GatingSet.}

\item{order}{can be either \code{"groups"} or \code{"nodes"} to control the
order in which the data is plotted. Setting \code{order = "nodes"} will
plot each group on a single page with a plot for each set of nodes. On the
other hand, setting \code{order = "groups"} will plot each set of nodes on
a separate page with plots for each group.}

\item{hidden}{logical indicating whether \code{hidden} nodes should be
included in the gating scheme, set to FALSE by default.}

\item{legend}{logical indicating whether a should be included when
populations are overlaid onto the plots, set to TRUE by default.
Alternatively, for gating schemes with many nodes, users can supply a
vector of integers to split the legend across multiple panels, set to 12
nodes by default.}

\item{...}{additional arguments passed to \code{cyto_plot()}.}
}
\description{
Visualise cytometry gating strategies
}
\seealso{
\link{cyto_plot}

\link{cyto_plot_gating_tree}

\link{cyto_plot_profile}

\link{cyto_plot_explore}

\link{cyto_plot_map}

\link{cyto_plot_grid}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
