% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_grid-methods.R
\name{cyto_plot_grid}
\alias{cyto_plot_grid}
\title{Grid Layout for cyto_plot}
\usage{
cyto_plot_grid(x, ...)
}
\description{
\code{cyto_plot_grid} is wrapper for cyto_plot which provides a grid layout
 without excess white space. \code{cyto_plot_grid} is particularly useful for
 visualising data grouped using \code{group_by}.
}
\details{
@param x object of class \code{flowSet} or \code{GatingSet}.
 @param ... additional method-specific arguments for \code{cyto_plot_grid}.

@author Dillon Hammill (Dillon.Hammill@anu.edu.au)

@export
}
