% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_hist.R
\name{cyto_plot_hist}
\alias{cyto_plot_hist}
\title{Add histograms to empty plots}
\usage{
cyto_plot_hist(
  x,
  parent = "root",
  channel,
  overlay = NA,
  select = NULL,
  events = 1,
  hist_stat = "count",
  hist_bins = 256,
  hist_smooth = 1,
  hist_stack = 0.5,
  hist_cols = NA,
  hist_fill = NA,
  hist_fill_alpha = 1,
  hist_line_type = 1,
  hist_line_width = 1,
  hist_line_col = "black",
  seed = 42,
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}},
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}, list of
\code{\link[flowWorkspace:cytoset]{cytosets}} or a list of
\code{\link[stats:density]{density}} objects.}

\item{parent}{name of the parent population to extract from GatingHierarchy
or GatingSet objects.}

\item{channel}{name of the channel to be used to construct the plot.}

\item{overlay}{name(s) of the populations to overlay or a \code{cytoset} or
\code{list of cytosets} containing populations to be overlaid onto the
plot(s). This argument can be set to "children" or "descendants" when a
\code{GatingSet} or \code{GatingHierarchy} to overlay all respective nodes.}

\item{select}{named list containing experimental variables to be used to
select samples using \code{\link{cyto_select}} when a \code{flowSet} or
\code{GatingSet} is supplied. Refer to \code{\link{cyto_select}} for more
details. Sample selection occurs prior to grouping with \code{merge_by}.}

\item{events}{controls the number or percentage of events to display, set to
1 by default to display all events.}

\item{hist_stat}{can be either \code{"count"}, \code{"percent"} or
\code{"density"} to indicate the statistic to display on histograms, set to
\code{"count"} by default. The \code{"percent"} option applies modal
normalisation and expresses the result as a percentage.}

\item{hist_bins}{number of bins to use for histograms, set to 256 by default.}

\item{hist_smooth}{smoothing parameter passed to
\code{\link[stats:density]{density}} to adjust the smoothness of the kernel
density for histograms, set to \code{1} by default. Only values greater or
equal to 1 are supported.}

\item{hist_stack}{numeric [0,1] indicating the degree of stacking for
histograms, set to \code{0.5} by default.}

\item{hist_cols}{vector colours to draw from when selecting histogram fill
colours if none are supplied to \code{hist_fill}.}

\item{hist_fill}{fill colour(s) for histograms, select from \code{hist_cols}
if not supplied.}

\item{hist_fill_alpha}{numeric [0,1] used to control histogram fill colour
transparency, set to \code{1} by default for solid colours.}

\item{hist_line_type}{line type(s) to use for histogram borders, set to 1 by
default to use solid lines. See \code{\link[graphics:par]{lty}} for
alternatives.}

\item{hist_line_width}{numeric to control line width(s) for histogram borders
lines, set to 1 by default.}

\item{hist_line_col}{colour(s) for histogram borders, set to \code{"black"}
by default.}

\item{seed}{numeric passed to \code{\link{set.seed}} to ensure that the same
sampling is applied with each \code{\link{cyto_plot_contour}} call, set to
an arbitrary numeric by default. This behaviour can be turned off by
setting this argument to NULL.}

\item{...}{not in use.}
}
\description{
Add histograms to empty plots
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
