% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_label-methods.R
\name{cyto_plot_label}
\alias{cyto_plot_label}
\title{Add boxed text labels to cyto_plot}
\usage{
cyto_plot_label(x, gates, ...)
}
\arguments{
\item{x}{object of class \code{"flowFrame"}.}

\item{gates}{object of class
\code{\link[flowCore:rectangleGate-class]{rectangleGate}},
\code{\link[flowCore:polygonGate-class]{polygonGate}},
\code{\link[flowCore:ellipsoidGate-class]{ellipsoidGate}}, \code{"list"} or
\code{\link[flowCore:filters-class]{filters}}.}

\item{...}{additional method-specific arguments.}
}
\value{
add a boxed text label to cyto_plot.
}
\description{
Add boxed text labels to cyto_plot
}
\seealso{
\code{\link{cyto_plot_label,flowFrame,rectangleGate-method}}

\code{\link{cyto_plot_label,flowFrame,polygonGate-method}}

\code{\link{cyto_plot_label,flowFrame,ellipsoidGate-method}}

\code{\link{cyto_plot_label,flowFrame,list-method}}

\code{\link{cyto_plot_label,flowFrame,filters-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
