% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_label.R
\name{cyto_plot_label}
\alias{cyto_plot_label}
\title{Add labels to cyto_plot}
\usage{
cyto_plot_label(
  x,
  parent = "root",
  alias = NA,
  channels = NULL,
  axes_trans = NA,
  events = 50000,
  negate = FALSE,
  gate = NA,
  overlay = NA,
  merge_by = "name",
  label_text,
  label_stat,
  label_text_x = NA,
  label_text_y = NA,
  label_text_font = 2,
  label_text_size = 0.8,
  label_text_col = "black",
  label_text_col_alpha = 1,
  label_fill = "white",
  label_fill_alpha = 0.6,
  hist_smooth = 1,
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the population to plot when a \code{GatingHierarchy} or
\code{GatingSet} object is supplied, set to the \code{"root"} node by
default.}

\item{alias}{name of the gated population(s) to gated in the plot when a
\code{GatingHierarchy} or \code{GatingSet} object is supplied. Setting
\code{alias} to "" will automatically plot any gates constructed in the
supplied channels. \code{alias} is equivalent to the \code{gate} argument
for \code{cytoset} objects.}

\item{channels}{name of the channel(s) or marker(s) to be used to construct
the plot.}

\item{axes_trans}{object of class
\code{\link[flowWorkspace:transformerList]{transformerList}} which was used
to transform the channels of the supplied data.  The transformerList is
used internally to ensure that the data is converted to a linear scale
(when required) prior to computing statistics for labels.}

\item{events}{numeric to control the number or percentage of events to
display. Values [0,1] indicate the percentage of events to display (i.e.
value of 1 will display all events), whilst values larger than 1 indicate
the number of events to display. The default value for \code{events} is set
to 50000 to display 50000 events only.}

\item{negate}{logical indicating whether a label should be included for the
negated population when gate objects are supplied, set to FALSE by default.
Setting \code{negate = TRUE} will result in the creation of a boolean
filter that contains all the events outside the gates supplied to
\code{alias} or \code{gate}. If such a boolean gate exists in the
GatingSet/GatingHierarchy it will automatically be extracted when
\code{alias = ""}. In order to prevent plotting of these boolean gates,
users will need to explicitly pass the names of the gates they want to
display to \code{alias}.}

\item{gate}{gate objects to be apply to data prior to computing statistics
for labels, can be either objects of class \code{rectangleGate},
\code{polygonGate}, \code{ellipsoidGate}, \code{quadGate} or
\code{filters}. Lists of these supported gate objects are also supported.}

\item{overlay}{name(s) of the populations to overlay or a \code{cytoset},
\code{list of cytosets} or \code{list of cytoset lists} containing
populations to be overlaid onto the plot(s). This argument can be set to
"children" or "descendants" when a \code{GatingSet} or
\code{GatingHierarchy} to overlay all respective nodes.}

\item{merge_by}{a vector of pData variables to sort and merge samples into
groups prior to plotting, set to "name" by default to prevent merging. To
merge all samples set this argument to \code{TRUE} or \code{"all"}.}

\item{label_text}{vector of population names to use in the labels. To exclude
the population names set this argument to NA.}

\item{label_stat}{indicates the type of statistic to include in the plot
labels, can be \code{"percent"}, \code{"count"}, \code{"mean"},
\code{"median"}, \code{"mode"} or \code{"geo mean"}, set to
\code{"percent"} for gated data or \code{NA} to exclude statistics for
un-gated data. Currently, only \code{"percent"} and \code{"count"} are
supported for 2-D scatter plots.}

\item{label_text_x}{vector of x co-ordinate(s) to manually adjust the
position plot label(s) on the plot. To interactively position labels set
either \code{label_text_x} or \code{label_text_y} to "select".}

\item{label_text_y}{vector of y co-ordinate(s) to manually adjust the
position plot label(s) on the plot. To interactively position labels set
either \code{label_text_x} or \code{label_text_y} to "select".}

\item{label_text_font}{numeric to control the font of text in plot labels,
set to 2 for bold font by default. See \code{\link[graphics:par]{font}} for
alternatives.}

\item{label_text_size}{numeric to control the size of text in the plot
labels, set to 1 by default.}

\item{label_text_col}{colour(s) to use for text in plot labels, set to
\code{"black"} by default.}

\item{label_text_col_alpha}{numeric [0, 1] to control the transparency of the
text colour, set to 1 by default to remove transparency.}

\item{label_fill}{fill colour(s) to use for labels, set to "white" by
default.}

\item{label_fill_alpha}{numeric to control background fill transparency of
label, set to 0.6 by default to introduce some transparency.}

\item{hist_smooth}{smoothing parameter passed to
\code{\link[stats:density]{density}} to adjust the smoothness of the kernel
density for histograms, set to \code{1} by default. Only values greater or
equal to 1 are supported.}

\item{...}{not in use.}
}
\value{
a matrix containing the xy co-ordinates of the labels.
}
\description{
\code{cyto_plot_label()} prepares the data in the same way as
\code{cyto_plot()} by allows uses to interactively add labels to plots that
have already been created.
}
\examples{
library(CytoExploreRData)

# Activation GatingSet
gs <- GatingSet(Activation)

# Call cyto_plot()
cyto_plot(gs[1], parent = "root", channels = c("FSC-A", "SSC-A"))

# Add label after plotting
cyto_plot_label(
  gs[1],
  parent = "root",
  channels = c("FSC-A", "SSC-A"),
  label_stat = "count",
  label_text = "count",
  label_text_x = 150000,
  label_text_y = 150000
)

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
