% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_label-methods.R,
%   R/cyto_plot_label_update.R
\name{cyto_plot_label}
\alias{cyto_plot_label}
\alias{cyto_plot_label.default}
\alias{cyto_plot_label.NULL}
\alias{cyto_plot_label.rectangleGate}
\alias{cyto_plot_label.polygonGate}
\alias{cyto_plot_label.ellipsoidGate}
\alias{cyto_plot_label.list}
\alias{cyto_plot_label.filters}
\title{Add boxed text labels to cyto_plot}
\usage{
cyto_plot_label(x, channels = NULL, trans = NA, display = 1,
  gate = NA, negate = FALSE, label_text = NA, label_stat = NA,
  label_text_x = NA, label_text_y = NA, label_text_font = 2,
  label_text_size = 0.8, label_text_col = "black",
  label_fill = "white", label_fill_alpha = 0.6, density_smooth = 0.6)

cyto_plot_label.default(x, gate, ...)

cyto_plot_label.NULL(x, gate, trans = NA, channels, text = NA,
  stat = NA, text_x = NA, text_y = NA, text_font = 2,
  text_size = 0.8, text_col = "black", box_alpha = 0.6,
  density_smooth = 0.6, plot = TRUE, ...)

cyto_plot_label.rectangleGate(x, gate, trans = NA, channels, text = NA,
  stat = NA, text_x = NA, text_y = NA, text_font = 2,
  text_size = 0.8, text_col = "black", box_alpha = 0.6,
  density_smooth = 0.6, plot = TRUE, ...)

cyto_plot_label.polygonGate(x, gate, channels, trans = NA, text = NA,
  stat = NA, text_x = NA, text_y = NA, text_font = 2,
  text_size = 0.8, text_col = "black", box_alpha = 0.6,
  density_smooth = 0.6, plot = TRUE, ...)

cyto_plot_label.ellipsoidGate(x, gate, channels, trans = NA, text = NA,
  stat = NA, text_x = NA, text_y = NA, text_font = 2,
  text_size = 0.8, text_col = "black", box_alpha = 0.6,
  density_smooth = 0.6, plot = TRUE, ...)

cyto_plot_label.list(x, gate, trans = NA, channels, text = NA,
  stat = NA, text_x = NA, text_y = NA, text_font = 2,
  text_size = 0.8, text_col = "black", box_alpha = 0.6,
  density_smooth = 0.6, plot = TRUE, ...)

cyto_plot_label.filters(x, gate, trans = NA, channels, text = NA,
  stat = NA, text_x = NA, text_y = NA, text_font = 2,
  text_size = 0.8, text_col = "black", box_alpha = 0.6,
  density_smooth = 0.6, plot = TRUE, ...)

cyto_plot_label(x, channels = NULL, trans = NA, display = 1,
  gate = NA, negate = FALSE, label_text = NA, label_stat = NA,
  label_text_x = NA, label_text_y = NA, label_text_font = 2,
  label_text_size = 0.8, label_text_col = "black",
  label_fill = "white", label_fill_alpha = 0.6, density_smooth = 0.6)
}
\arguments{
\item{x}{object of class \code{"flowFrame"}.}

\item{channels}{a vector indicating the fluorescent channel(s) to be used for
gating.}

\item{trans}{object of class
\code{\link[flowCore:transformList-class]{transformList}} or
\code{\link[flowWorkspace]{transformerList}} generated by
\code{\link[flowCore:logicleTransform]{estimateLogicle}} which was used to
transform the fluorescent channels of the supplied flowFrame.}

\item{display}{numeric [0,1] to control the percentage of events to be
plotted. Specifying a value for \code{display} can substantial improve
plotting speed for less powerful machines.}

\item{gate}{object of class
\code{\link[flowCore:rectangleGate-class]{rectangleGate}},
\code{\link[flowCore:polygonGate-class]{polygonGate}},
\code{\link[flowCore:ellipsoidGate-class]{ellipsoidGate}}, \code{"list"} or
\code{\link[flowCore:filters-class]{filters}}.}

\item{negate}{logical indicating whether a label should be included for the
neagted population (i.e. events outside the gate). Set to FALSE by default
to only calculate statistics for events within the gate.}

\item{label_text}{character string to include in the label above the
statistic (e.g. population name(s)).}

\item{label_stat}{indicates the type of statistic to include in the label,
can be either code{"count"}, \code{"median"}, \code{"mean"}, \code{"mode"},
\code{"geo mean"} or \code{"CV"}. Only count and percent statistics are
supported for 2D plots.}

\item{label_text_x}{vector containing the x co-ordinates for the plot labels.
Label positions can be interactively selected if no co-ordinates are
manually supplied.}

\item{label_text_y}{vector containing the x co-ordinates for the plot labels.
Label positions can be interactively selected if no co-ordinates are
manually supplied.}

\item{label_text_font}{integer [1,2,3,4] passed to \code{text} to alter the
font, set to \code{2} by default for a bold font.}

\item{label_text_size}{numeric character expansion used to control the size
of the text in the labels, set to \code{0.8} by default. See \code{?text}
for details.}

\item{label_text_col}{specify text colour in label for each gate, defaults to
\code{"black"} for all gates.}

\item{label_fill}{fill colour to use for labels, set to "white" by default.}

\item{label_fill_alpha}{numeric [0,1] controls the transparency of the fill
colour, set to \code{0.6} by default.}

\item{density_smooth}{smoothing parameter passed to
\code{\link[stats:density]{density}} to adjust kernel density for mode
calculation.}

\item{...}{passes additional list method arguments to other methods.}

\item{text}{character string to include in the label above the statistic
(e.g. population name(s)).}

\item{stat}{indicates the type of statistic to include in the label, can be
either code{"count"}, \code{"median"}, \code{"mean"}, \code{"mode"},
\code{"geo mean"} or \code{"CV"}. \code{stat} is set to \code{"median"} by
default. Statistics for fluorescent intensity are calculated for the entire
distribution. Only count and percent statistics are supported for 2D plots.}

\item{text_x}{vector containing the x co-ordinates for the plot labels. Set
to \code{NULL} by default to place labels in the center of the gates. To
interactively position labels set this argument to "select".}

\item{text_y}{vector containing the x co-ordinates for the plot labels. Set
to \code{NULL} by default to place labels in the center of the gates. To
interactively position labels set this argument to "select".}

\item{text_font}{integer [1,2,3,4] passed to \code{text} to alter the font,
set to \code{2} by default for a bold font.}

\item{text_size}{numeric character expansion used to control the size of the
text in the labels, set to \code{0.8} by default. See \code{?text} for
details.}

\item{text_col}{specify text colour in label for each gate, defaults to
\code{"black"} for all gates.}

\item{box_alpha}{numeric [0,1] controls the transparency of the background,
set to \code{0.6} by default.}

\item{plot}{logical indicating whether the labels should be added to the
plot. Can be turned off to obtain co-ordinates for the labels.}

\item{x}{object of class \code{"flowFrame"}.}

\item{channels}{a vector indicating the fluorescent channel(s) to be used for
gating.}

\item{gate}{object of class
\code{\link[flowCore:rectangleGate-class]{rectangleGate}},
\code{\link[flowCore:polygonGate-class]{polygonGate}},
\code{\link[flowCore:ellipsoidGate-class]{ellipsoidGate}},
\code{\link[flowCore:quadGate-class]{quadGate}}, \code{"list"} or
\code{\link[flowCore:filters-class]{filters}}.}

\item{trans}{object of class
\code{\link[flowWorkspace]{transformerList}} which was used to
transform the fluorescent channels of the supplied flowFrame.}

\item{density_smooth}{smoothing parameter passed to
\code{\link[stats:density]{density}} to adjust kernel density for mode
calculation.}
}
\value{
add a boxed text label to an existing plot.

add a boxed text label to an existing cyto_plot.
}
\description{
\code{cyto_plot_label} takes on a \code{flowFrame} object, population name
\code{text}, \code{channels} and a gate object to construct a text label for
the plot with the population name and frequency. In the absence of a gate,
users can manually specify the label position using the \code{text_x} and
\code{text_y} arguments. To interactively position labels through mouse click
set either \code{text_x} or \code{text_y} to "select".

Interactively label existing plots with boxed labels containing text and
statistics.
}
\examples{
#' library(CytoRSuiteData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(fs))
gs <- transform(gs, trans)

# Gate using gate_draw
gating(Activation_gatingTemplate, gs)

# Plot
cyto_plot(gs[[4]],
  parent = "T Cells",
  channels = c("Alexa Fluor 488-A", "Alexa Fluor 700-A")
)

# Labels without gates - position using text_x and text_y
cyto_plot_label(getData(gs, "T Cells")[[4]],
  gate = NULL,
  channels = c("Alexa Fluor 488-A", "Alexa Fluor 700-A"),
  trans = trans,
  text = c("CD4 T Cells", "CD8 T Cells"),
  text_x = c(3.1, 1),
  text_y = c(-0.2, 3.2)
)

#' # Plot
cyto_plot(gs[[4]],
  parent = "CD4 T Cells",
  channels = "7-AAD-A"
)

# CD69+ CD4 T Cells gate
gt <- getGate(gs, "CD69+ CD4 T Cells")[[1]]
cyto_plot_gate(gt,
  channels = "7-AAD-A"
)

# Labels
cyto_plot_label(getData(gs, "CD4 T Cells")[[4]],
  gate = gt,
  trans = trans,
  channels = "7-AAD-A",
  text = c("CD69+"),
  stat = "freq",
  text_col = "blue"
)

#' # Plot
cyto_plot(gs[[4]],
  parent = "root",
  channels = c("FSC-A", "SSC-A")
)

# Cells gate
gt <- getGate(gs, "Cells")[[1]]
cyto_plot_gate(gt,
  channels = c("FSC-A", "SSC-A")
)

# Labels
cyto_plot_label(getData(gs, "root")[[4]],
  gate = gt,
  trans = trans,
  channels = c("FSC-A", "SSC-A"),
  text = "Cells",
  stat = "freq",
  text_col = "magenta",
  text_size = 1.2
)

#' # Plot
cyto_plot(gs[[4]],
  parent = "Live Cells",
  channels = c("APC-Cy7-A", "PE-A")
)

# T Cells gate
gt <- getGate(gs, "T Cells")[[1]]
cyto_plot_gate(gt,
  channels = c("APC-Cy7-A", "PE-A")
)

# Labels
cyto_plot_label(getData(gs, "Live Cells")[[4]],
  gate = gt,
  trans = trans,
  channels = c("APC-Cy7-A", "PE-A"),
  text = "T Cells",
  stat = "count",
  text_col = "magenta",
  text_size = 1.2,
  box_alpha = 1
)

#' # Plot
cyto_plot(gs[[4]],
  parent = "Live Cells",
  channels = c("APC-Cy7-A", "PE-A")
)

# T Cells & Dendritic Cells gates
gts <- list(getGate(gs, "T Cells")[[1]], getGate(gs, "Dendritic Cells")[[1]])
cyto_plot_gate(gts,
  channels = c("APC-Cy7-A", "PE-A")
)

# Labels
cyto_plot_label(getData(gs, "Live Cells")[[4]],
  gate = gts,
  trans = trans,
  channels = c("APC-Cy7-A", "PE-A"),
  text = c("T Cells", "Dendritic Cells"),
  stat = "count",
  text_col = c("magenta", "purple"),
  text_size = 1.2,
  box_alpha = 1
)
library(CytoRSuiteData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(fs))
gs <- transform(gs, trans)

# Gate using gate_draw
gating(Activation_gatingTemplate, gs)

# Plot
cyto_plot(gs[[32]],
  parent = "CD4 T Cells",
  channels = c("CD69")
)

# Label - median fluorescent intensity
cyto_plot_label2(getData(gs, "CD4 T Cells")[[32]],
  channels = "CD69",
  label_stat = "median",
  label_text = "MedFI",
  trans = trans,
  label_text_x = 3,
  label_text_y = 50
)

# Label - no statistic
cyto_plot_label2(
  label_text = "Activated \\n Cells",
  label_text_x = 3,
  label_text_y = 25,
  label_fill = "red"
)

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)

Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
