% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_label.R
\name{cyto_plot_label}
\alias{cyto_plot_label}
\alias{cyto_plot_label.GatingHierarchy}
\alias{cyto_plot_label.flowFrame}
\title{Add labels to cyto_plot}
\usage{
\method{cyto_plot_label}{GatingHierarchy}(
  x,
  parent,
  alias = NA,
  channels = NULL,
  display = 1,
  gate = NA,
  negate = NA,
  label_text,
  label_stat,
  label_text_x = NA,
  label_text_y = NA,
  label_text_font = 2,
  label_text_size = 0.8,
  label_text_col = "black",
  label_text_col_alpha = 1,
  label_fill = "white",
  label_fill_alpha = 0.6,
  hist_smooth = 1,
  ...
)

\method{cyto_plot_label}{flowFrame}(
  x,
  channels = NULL,
  trans = NA,
  display = 1,
  gate = NA,
  negate = FALSE,
  label_text = NA,
  label_stat = NA,
  label_text_x = NA,
  label_text_y = NA,
  label_text_font = 2,
  label_text_size = 0.8,
  label_text_col = "black",
  label_text_col_alpha = 1,
  label_fill = "white",
  label_fill_alpha = 0.6,
  hist_smooth = 1,
  ...
)
}
\arguments{
\item{x}{object of class \code{flowFrame}.}

\item{parent}{name of the parent population to extract when a
\code{GatingHierarchy} object is supplied.}

\item{alias}{name of teh gated populations to label when a GatingHierarchy
object is supplied. This is equivalent to the \code{gate} argument for
\code{flowFrame} objects.}

\item{channels}{a vector indicating the fluorescent channel(s) to be used for
gating.}

\item{display}{numeric [0,1] to control the percentage of events to be
plotted. Specifying a value for \code{display} can substantial improve
plotting speed for less powerful machines.}

\item{gate}{object of class
\code{\link[flowCore:rectangleGate-class]{rectangleGate}},
\code{\link[flowCore:polygonGate-class]{polygonGate}},
\code{\link[flowCore:ellipsoidGate-class]{ellipsoidGate}},
\code{\link[flowCore:quadGate-class]{quadGate}}, \code{"list"} or
\code{\link[flowCore:filters-class]{filters}}.}

\item{negate}{logical indicating whether a label should be included for the
neagted population (i.e. events outside the gate). Set to FALSE by default
to only calculate statistics for events within the gate.}

\item{label_text}{character string to include in the label above the
statistic (e.g. population name(s)).}

\item{label_stat}{indicates the type of statistic to include in the label,
can be either code{"count"}, \code{"median"}, \code{"mean"}, \code{"mode"},
\code{"geo mean"} or \code{"CV"}. Only count and percent statistics are
supported for 2D plots.}

\item{label_text_x}{vector containing the x co-ordinates for the plot labels.
Label positions can be interactively selected if no co-ordinates are
manually supplied.}

\item{label_text_y}{vector containing the x co-ordinates for the plot labels.
Label positions can be interactively selected if no co-ordinates are
manually supplied.}

\item{label_text_font}{integer [1,2,3,4] passed to \code{text} to alter the
font, set to \code{2} by default for a bold font.}

\item{label_text_size}{numeric character expansion used to control the size
of the text in the labels, set to \code{0.8} by default. See \code{?text}
for details.}

\item{label_text_col}{specify text colour in label for each gate, defaults to
\code{"black"} for all gates.}

\item{label_text_col_alpha}{numeric [0, 1] to control the transparency of the
text colour, set to 1 by default to remove transparency.}

\item{label_fill}{fill colour to use for labels, set to "white" by default.}

\item{label_fill_alpha}{numeric [0,1] controls the transparency of the fill
colour, set to \code{0.6} by default.}

\item{hist_smooth}{smoothing parameter passed to
\code{\link[stats:density]{density}} to adjust kernel density for mode
calculation.}

\item{...}{not in use.}

\item{trans}{object of class \code{\link[flowWorkspace]{transformerList}}
which was used to transform the fluorescent channels of the supplied
flowFrame.}
}
\value{
add a boxed text label to an existing cyto_plot.
}
\description{
Interactively label existing plots with boxed labels containing text and
statistics. \code{cyto_plot_label} can only label the last plot and can only
add labels on a per sample basis (i.e. single layer only). Locations for each
label must be manually selected. More complex labelling can be achieved
directly through \code{cyto_plot}.
}
\examples{
library(CytoExploreRData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(fs))
gs <- transform(gs, trans)

# Gate using gate_draw
gt_gating(Activation_gatingTemplate, gs)

# Plot
cyto_plot(gs[[32]],
  parent = "CD4 T Cells",
  channels = c("CD69")
)

# Label - median fluorescent intensity
cyto_plot_label(cyto_data_extract(gs, "CD4 T Cells")[[1]][[32]],
  channels = "CD69",
  label_stat = "median",
  label_text = "MedFI",
  trans = trans,
  label_text_x = 3,
  label_text_y = 50
)

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
