% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_label2.R
\name{cyto_plot_label2}
\alias{cyto_plot_label2}
\title{Add boxed text labels to cyto_plot}
\usage{
cyto_plot_label2(x, channels, gate = NA, label_text = NA,
  label_stat = NA, label_text_x = NA, label_text_y = NA,
  trans = NA, negate = FALSE, display = 1, label_text_font = 2,
  label_text_size = 0.8, label_text_col = "black",
  label_fill = "white", label_fill_alpha = 0.6, density_smooth = 0.6)
}
\arguments{
\item{x}{object of class \code{"flowFrame"}.}

\item{channels}{a vector indicating the fluorescent channel(s) to be used for
gating.}

\item{gate}{object of class
\code{\link[flowCore:rectangleGate-class]{rectangleGate}},
\code{\link[flowCore:polygonGate-class]{polygonGate}},
\code{\link[flowCore:ellipsoidGate-class]{ellipsoidGate}}, \code{"list"} or
\code{\link[flowCore:filters-class]{filters}}.}

\item{label_text}{character string to include in the label above the
statistic (e.g. population name(s)).}

\item{label_stat}{indicates the type of statistic to include in the label,
can be either code{"count"}, \code{"median"}, \code{"mean"}, \code{"mode"},
\code{"geo mean"} or \code{"CV"}. Only count and percent statistics are
supported for 2D plots.}

\item{label_text_x}{vector containing the x co-ordinates for the plot labels.
Label positions can be interactively selected if no co-ordinates are
manually supplied.}

\item{label_text_y}{vector containing the x co-ordinates for the plot labels.
Label positions can be interactively selected if no co-ordinates are
manually supplied.}

\item{trans}{object of class
\code{\link[flowWorkspace]{transformerList}} which was used to
transform the fluorescent channels of the supplied flowFrame.}

\item{negate}{logical indicating whether a label should be included for the
neagted population (i.e. events outside the gate). Set to FALSE by default
to only calculate statistics for events within the gate.}

\item{display}{numeric [0,1] to control the percentage of events to be
plotted. Specifying a value for \code{display} can substantial improve
plotting speed for less powerful machines.}

\item{label_text_font}{integer [1,2,3,4] passed to \code{text} to alter the
font, set to \code{2} by default for a bold font.}

\item{label_text_size}{numeric character expansion used to control the size
of the text in the labels, set to \code{0.8} by default. See \code{?text}
for details.}

\item{label_text_col}{specify text colour in label for each gate, defaults to
\code{"black"} for all gates.}

\item{label_fill}{fill colour to use for labels, set to "white" by default.}

\item{label_fill_alpha}{numeric [0,1] controls the transparency of the fill
colour, set to \code{0.6} by default.}

\item{density_smooth}{smoothing parameter passed to
\code{\link[stats:density]{density}} to adjust kernel density for mode
calculation.}
}
\value{
add a boxed text label to an existing cyto_plot.
}
\description{
Interactively label existing plots with boxed labels containing text and
statistics.
}
\examples{
library(CytoRSuiteData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
trans <- estimateLogicle(gs[[4]], cyto_fluor_channels(fs))
gs <- transform(gs, trans)

# Gate using gate_draw
gating(Activation_gatingTemplate, gs)

# Plot
cyto_plot(gs[[32]],
  parent = "CD4 T Cells",
  channels = c("CD69")
)

# Label - median fluorescent intensity
cyto_plot_label2(getData(gs, "CD4 T Cells")[[32]],
  channels = "CD69",
  label_stat = "median",
  label_text = "MedFI",
  trans = trans,
  label_text_x = 3,
  label_text_y = 50
)

# Label - no statistic
cyto_plot_label2(
  label_text = "Activated \\n Cells",
  label_text_x = 3,
  label_text_y = 25,
  label_fill = "red"
)
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
