% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_map.R
\name{cyto_plot_map}
\alias{cyto_plot_map}
\alias{cyto_plot_map.flowSet}
\alias{cyto_plot_map.flowFrame}
\title{Plot maps coloured by parameter}
\usage{
\method{cyto_plot_map}{flowSet}(
  x,
  map = NULL,
  channels = NULL,
  select = NULL,
  group_by = "name",
  layout = NULL,
  header = NULL,
  header_text_font = 2,
  header_text_size = 1,
  header_text_col = "black",
  ...
)

\method{cyto_plot_map}{flowFrame}(
  x,
  map = NULL,
  channels = NULL,
  layout = NULL,
  header = NULL,
  header_text_font = 2,
  header_text_size = 1,
  header_text_col = "black",
  point_col_scale = NA,
  point_col_alpha = 1,
  ...
)
}
\arguments{
\item{page}{can be either "channel" or "sample" indicating whether each page
should contain for each sample or each channel.}
}
\description{
Plot maps coloured by parameter
}
