% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_map.R
\name{cyto_plot_map}
\alias{cyto_plot_map}
\title{Plot dimension reduced maps in all dimensions}
\usage{
cyto_plot_map(
  x,
  parent = "root",
  channels = NULL,
  select = NULL,
  merge_by = "name",
  overlay = NA,
  order = "channels",
  point_col = NA,
  point_col_scale = NA,
  layout,
  header = NA,
  ...
)
}
\description{
Plot dimension reduced maps in all dimensions
}
