% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_new}
\alias{cyto_plot_new}
\title{Open new graphics device for cyto_plot}
\usage{
cyto_plot_new(popup = FALSE, ...)
}
\arguments{
\item{popup}{logical indicating whether a popup graphics device should be
opened.}

\item{...}{additional arguments passed to
\code{\link[grDevices:dev.new]{dev.new}}:}
}
\description{
Open new graphics device for cyto_plot
}
\examples{
\dontrun{
# Open platform-specific graphics device
cyto_plot_new()
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
