% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_new}
\alias{cyto_plot_new}
\title{Open new graphics device for cyto_plot}
\usage{
cyto_plot_new(popup = NULL, popup_size = NULL, ...)
}
\arguments{
\item{popup}{logical indicating whether a pop-up graphics device should be
opened, set to TRUE by default if called outside of \code{cyto_plot()}.}

\item{popup_size}{a vector of length 2 to control the height and width of
pop-up graphics device in inches, set to \code{c(10,10)} by default.}

\item{...}{additional graphical parameters supplied by name to be passed to
\code{\link{cyto_plot_par}} to customise the new graphics device.}
}
\description{
\code{cyto_plot_new} opens a new RStudio or pop-up graphics device and
accepts arguments to set the graphical parameters of the new device.
\code{cyto_plot_new()} is used internally by cyto_plot to open an OS-specific
interactive garphics device to facilitate gate drawing. Mac users will need
to install \href{https://www.xquartz.org/}{XQuartz} for this functionality.
}
\examples{
\dontrun{
# Open platform-specific graphics device
cyto_plot_new()
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
