% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_new}
\alias{cyto_plot_new}
\title{Open new graphics device for cyto_plot}
\usage{
cyto_plot_new(popup = TRUE, ...)
}
\arguments{
\item{popup}{logical indicating whether a popup graphics device should be
opened, set to TRUE by default.}

\item{...}{additional arguments passed to
\code{\link[grDevices:dev]{dev.new}}:}
}
\description{
\code{cyto_plot_new} is used internally by cyto_plot to open an
OS-specific interactive garphics device to facilitate gate drawing. Mac users
will need to install \href{https://www.xquartz.org/}{XQuartz} for this
functionality.
}
\examples{
\dontrun{
# Open platform-specific graphics device
cyto_plot_new()
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
