% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_point.R
\name{cyto_plot_point}
\alias{cyto_plot_point}
\title{Add points and contour lines to empty cyto_plot}
\usage{
cyto_plot_point(
  x,
  parent = "root",
  channels,
  overlay = NA,
  select = NULL,
  events = 1,
  point_shape = ".",
  point_size = 2,
  point_col_scale = NA,
  point_col_smooth = TRUE,
  point_cols = NA,
  point_col = NA,
  point_col_alpha = 1,
  point_fast = FALSE,
  contour_lines = 0,
  contour_line_type = 1,
  contour_line_width = 1,
  contour_line_col = "black",
  contour_line_alpha = 1,
  seed = 42,
  ...
)
}
\arguments{
\item{x}{either an object of class
\code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the parent population to extract from GatingHierarchy
or GatingSet objects.}

\item{channels}{names of the channels used to construct the plot.}

\item{overlay}{name(s) of the populations to overlay or a \code{cytoset} or
\code{list of cytosets} containing populations to be overlaid onto the
plot(s). This argument can be set to "children" or "descendants" when a
\code{GatingSet} or \code{GatingHierarchy} to overlay all respective nodes.}

\item{select}{named list containing experimental variables to be used to
select samples using \code{\link{cyto_select}} when a \code{flowSet} or
\code{GatingSet} is supplied. Refer to \code{\link{cyto_select}} for more
details. Sample selection occurs prior to grouping with \code{merge_by}.}

\item{events}{controls the number or percentage of events to display, set to
1 by default to display all events.}

\item{point_shape}{shape(s) to use for points in 2-D scatterplots, set to
\code{"."} by default to maximise plotting speed.  See
\code{\link[graphics:par]{pch}} for alternatives.}

\item{point_size}{numeric to control the size of points in 2-D scatter plots
set to 2 by default.}

\item{point_col_scale}{vector of ordered colours to use for the density
colour gradient of points.}

\item{point_col_smooth}{logical indicating whether the 2D binned counts
should be smoothed using kernel density estimates prior to selecting
colours from \code{point_col_scale}, set to TRUE by default. Setting
\code{point_col_smooth} to FALSE will significantly improve plotting speed
on less powerful machines but produce more granular plots.}

\item{point_cols}{vector colours to draw from when selecting colours for
points if none are supplied to point_col.}

\item{point_col}{colour(s) to use for points in 2-D scatter plots, set to NA
by default to use a blue-red density colour scale.}

\item{point_col_alpha}{numeric [0,1] to control point colour transparency in
2-D scatter plots, set to 1 by default to use solid colours.}

\item{point_fast}{logical indicating whether points should be plotted using
the \code{scattermore} package, set to FALSE by default. This is an
optional feature so if you intend to use it, make sure that you have the
scattermore package installed \code{install.packages("scattermore")}.}

\item{contour_lines}{numeric indicating the number of levels to use for
contour lines, set to 0 by default to exclude contour lines.}

\item{contour_line_type}{type of line to use for contour lines, set to 1 by
default.}

\item{contour_line_width}{line width for contour lines, set to 2 by default.}

\item{contour_line_col}{colour to use for contour lines, set to "black" by
default.}

\item{contour_line_alpha}{numeric [0,1] to control transparency of contour
lines, set to 1 by default to remove transparency.}

\item{seed}{numeric passed to \code{\link{set.seed}} to ensure that the same
sampling is applied with each \code{\link{cyto_plot_contour}} call, set to
an arbitrary numeric by default. This behaviour can be turned off by
setting this argument to NULL.}

\item{...}{not in use.}
}
\description{
Add points and contour lines to empty cyto_plot
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
