% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_point.R
\name{cyto_plot_point}
\alias{cyto_plot_point}
\alias{cyto_plot_point.flowFrame}
\alias{cyto_plot_point.list}
\title{Add points to empty plots}
\usage{
cyto_plot_point(x, ...)

\method{cyto_plot_point}{flowFrame}(x, channels, overlay = NA,
  point_shape = ".", point_size = 2, point_col_scale = NA,
  point_cols = NA, point_col = NA, point_col_alpha = 1, ...)

\method{cyto_plot_point}{list}(x, channels, point_shape = ".",
  point_size = 2, point_col_scale = NA, point_cols = NA,
  point_col = NA, point_col_alpha = 1, ...)
}
\arguments{
\item{x}{object of class \code{flowFrame} or a list of \code{flowFrame}
onjects.}

\item{channels}{names of the channels used to construct the plot.}

\item{overlay}{optional argument if x is a flowFrame to overlay a list of
flowFrame objects.}

\item{point_shape}{shape(s) to use for points in 2-D scatterplots, set to
\code{"."} by default to maximise plotting speed.  See
\code{\link[graphics:par]{pch}} for alternatives.}

\item{point_size}{numeric to control the size of points in 2-D scatter plots
set to 2 by default.}

\item{point_col_scale}{vector of ordered colours to use for the density
colour gradient of points.}

\item{point_cols}{vector colours to draw from when selecting colours for
points if none are supplied to point_col.}

\item{point_col}{colour(s) to use for points in 2-D scatter plots, set to NA
by default to use a blue-red density colour scale.}

\item{point_col_alpha}{numeric [0,1] to control point colour transparency in
2-D scatter plots, set to 1 by default to use solid colours.}
}
\description{
Add points to empty plots
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
