% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_profile.R
\name{cyto_plot_profile}
\alias{cyto_plot_profile}
\title{Plot expression profile in all channels}
\usage{
cyto_plot_profile(
  x,
  parent = "root",
  channels = NULL,
  select = NULL,
  merge_by = "name",
  overlay = NA,
  layout = NULL,
  hist_layers = NA,
  hist_stack = 0.5,
  axes_limits = "machine",
  popup = FALSE,
  header = NULL,
  header_text_font = 2,
  header_text_size = 1,
  header_text_col = "black",
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}},
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the population to plot when a \code{GatingHierarchy} or
\code{GatingSet} object is supplied.}

\item{channels}{a vector channels to use to construct the plots, set to all
channels by default.}

\item{select}{named list containing experimental variables to be used to
select samples using \code{\link{cyto_select}} when a \code{flowSet} or
\code{GatingSet} is supplied. Refer to \code{\link{cyto_select}} for more
details. Sample selection occurs prior to grouping with \code{merge_by}.}

\item{merge_by}{a vector of pData variables to sort and merge samples into
groups prior to plotting, set to "name" by default to prevent merging. To
merge all samples set this argument to \code{TRUE} or \code{"all"}.}

\item{overlay}{name(s) of the populations to overlay or a \code{flowFrame},
\code{flowSet}, \code{list of flowFrames}, \code{list of flowSets} or
\code{list of flowFrame lists} containing populations to be overlaid onto
the plot(s). This argument can be set to "children" or "descendants" when a
\code{GatingSet} or \code{GatingHierarchy} to overlay all respective nodes.}

\item{layout}{a vector of the length 2 indicating the dimensions of the grid
for plotting \code{c(#rows, #columns)}.}

\item{hist_stack}{numeric [0,1] indicating the degree of offset for 1-D
density distributions with overlay, set to 0.5 by default.}

\item{axes_limits}{options include \code{"auto"}, \code{"data"} or
\code{"machine"} to use optimised, data or machine limits respectively. Set
to \code{"machine"} by default to complete axes ranges.}

\item{popup}{logical indicating whether a pop-up window should be opened
prior to plotting, set to \code{FALSE} by default.}

\item{header}{character string to include in the plot header, set to the
sample names by default. The header can be removed by setting this argument
to NA.}

\item{header_text_font}{numeric indicating the font to use for the header,
set to 2 by default for bold font.}

\item{header_text_size}{numeric to control the text size of the header, set
to 1 by default.}

\item{header_text_col}{colour to use for the header, set to "black" by
default.}

\item{...}{additional arguments passed to \code{\link{cyto_plot}}.}
}
\description{
Plot expression profile in all channels
}
\examples{

# Load in CytoExploreR to access data
library(CytoExploreRData)

# Load in samples
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- cyto_compensate(gs, fs[[1]]@description$SPILL)

# Transform fluorescent channels
gs <- cyto_transform(gs, trans_type = "logicle")

# Apply gates
gt <- Activation_gatingTemplate
gt_gating(gt, gs)

# Plot expression profile of T Cells in all channels
cyto_plot_profile(gs[1:9],
  parent = "T Cells"
)

# Group samples by Treatment & select highest OVAConc
 cyto_plot_profile(gs[1:9],
  parent = "CD4 T Cells",
  merge_by = "Treatment",
  select = list("OVAConc" = 500)
)

}
\seealso{
\code{\link{cyto_plot}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
