% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_profile.R
\name{cyto_plot_profile}
\alias{cyto_plot_profile}
\title{Plot expression profile in multiple channels}
\usage{
cyto_plot_profile(
  x,
  parent = "root",
  channels = NULL,
  merge_by = "name",
  overlay = NA,
  select = NULL,
  order = "channels",
  layout,
  hist_layers = NA,
  hist_stack = 0,
  header,
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the parent population to extract from \code{GatingSet}
objects, set to the \code{"root"} node by default.}

\item{channels}{names of the channels or markers in which the data should be
plotted, set to all channels except \code{"Time"} and \code{"Event-ID"} by
default.}

\item{merge_by}{a vector of pData variables to sort and merge samples into
groups prior to plotting, set to "name" by default to prevent merging. To
merge all samples set this argument to \code{TRUE} or \code{"all"}.}

\item{overlay}{name(s) of the populations to overlay or a \code{cytoset},
\code{list of cytosets} or \code{list of cytoset lists} containing
populations to be overlaid onto the plot(s). This argument can be set to
"children" or "descendants" when a \code{GatingSet} or
\code{GatingHierarchy} to overlay all respective nodes.}

\item{select}{named list containing experimental variables to be used to
select samples using \code{\link{cyto_select}} when a \code{flowSet} or
\code{GatingSet} is supplied. Refer to \code{\link{cyto_select}} for more
details. Sample selection occurs prior to grouping with \code{merge_by}.}

\item{order}{can be either \code{"channels"} or \code{"groups"} to control
the order in which the data is plotted. Setting \code{order = "channels"}
will plot each group on a single page in all the supplied channels. On the
other hand, setting \code{order = "groups"} will plot each channel on a
separate page for all groups.}

\item{layout}{a vector of the length 2 of form \code{c(#rows, #columns)} or a
matrix indicating the dimensions of the grid for plotting.}

\item{hist_layers}{numeric indicating the number of histograms to stack in
each plot when there are no overlays, set to 1 histogram per plot by
default. Each plot must contain the same number of histograms.}

\item{hist_stack}{numeric [0,1] indicating the degree of stacking for
histograms, set to \code{0} by default.}

\item{header}{vector of text to include on each page of plots, not
recommended to modify by hand, headers can be removed by setting this
argumemt to NA.}

\item{...}{additional arguments passed to \code{\link{cyto_plot}}.}
}
\value{
list of recorded plots.
}
\description{
Plot expression profile in multiple channels
}
\examples{
\dontrun{
library(CytoExploreRData)

# Activation GatingSet
gs <- cyto_load(
  system.file(
    "extdata/Activation-GatingSet",
    package = "CytoExploreRData"
  )
)

# Expression profile in all channels per sample
cyto_plot_profile(
  gs,
  parent = "Live Cells",
  order = "channels"
)

# Compare expression between samples in each channel
cyto_plot_profile(
  gs,
  parent = "Live Cells",
  order = "groups"
)
}

}
\seealso{
\code{\link{cyto_plot}}

\code{\link{cyto_plot_explore}}

\code{\link{cyto_select}}

\code{\link{cyto_merge_by}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
