% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_profile.R
\name{cyto_plot_profile}
\alias{cyto_plot_profile}
\title{Plot Expression Profile in All Fluorescent Channels}
\usage{
cyto_plot_profile(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{...}{additional method-specific arguments.}
}
\description{
Plot Expression Profile in All Fluorescent Channels
}
\seealso{
\code{\link{cyto_plot_profile,flowFrame-method}}

\code{\link{cyto_plot_profile,flowSet-method}}

\code{\link{cyto_plot_profile,GatingSet-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
