% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_save}
\alias{cyto_plot_save}
\title{Save High Resolution cyto_plot Images}
\usage{
cyto_plot_save(
  save_as,
  width = 7,
  height = 7,
  units = "in",
  res = 300,
  multiple = NULL,
  reset = FALSE,
  ...
)
}
\arguments{
\item{save_as}{name of the file to which the plot should be saved (including
the file extension). Supported file formats include png, tiff, jpeg, svg
and pdf.}

\item{width}{numeric indicating the width of exported plot in \code{units},
set to 7 by default for image with width of 7 inches.}

\item{height}{numeric indicating the height of the exported plot in
\code{units}, set to 7 by default for image with height of 7 inches.}

\item{units}{units to be used to set plot size, can be either pixels
(\code{px}), inches (\code{inches}), centimetres (\code{cm}) or millimetres
(\code{mm}). Set to \code{"in"} by default. Units cannot be altered for
\code{svg} and \code{pdf} graphics devices.}

\item{res}{resolution in ppi, set to 300 by default.}

\item{multiple}{logical indicating whether multiple pages should be saved to
separate numbered files, set to \code{TRUE} for PDF and \code{FALSE} for
other file types by default.}

\item{reset}{logical indicating whether to reset a previous call to
\code{cyto_plot_save}.}

\item{...}{additional arguments for the appropriate \code{png()},
\code{tiff()}, \code{jpeg()}, \code{svg()} or \code{pdf} graphics devices.}
}
\description{
Save High Resolution cyto_plot Images
}
\examples{
\dontrun{
library(CytoExploreRData)

# Load samples into GatingSet
fs <- Activation
gs <- GatingSet(fs)

# Apply compensation
gs <- cyto_compensate(gs)

# Transform fluorescent channels
gs <- cyto_transform(gs)

# Apply gatingTemplate
cyto_gatingTemplate_apply(gs, Activation_gatingTemplate)

# Save png image of gating scheme after plotting
cyto_plot_save("Gating-Scheme.png",
  width = 20,
  height = 16
)
cyto_plot_gating_scheme(gs[[1]])

# Save multiple pages to the same pdf file
cyto_plot_save("CD4-T-Cells.pdf",
  height = 8,
  width = 16,
  multiple = TRUE
)
cyto_plot(gs,
  parent = "CD4 T Cells",
  alias = "",
  channels = c("Alexa Fluor 647-A", "7-AAD-A"),
  layout = c(1, 2)
)
}
}
\seealso{
\code{\link[grDevices:cairo]{cairo}}

\code{\link[grDevices:png]{png}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
