% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_spectra.R
\name{cyto_plot_spectra.list}
\alias{cyto_plot_spectra.list}
\title{Plot fluorescent spectrum for each fluorochrome across detectors}
\usage{
\method{cyto_plot_spectra}{list}(
  x,
  channels,
  select = NULL,
  ylim = c(NA, NA),
  xlab = NA,
  ylab = NA,
  legend_text = NA,
  ...
)
}
\arguments{
\item{x}{object of class \code{"matrix"} or a list of matrices.}

\item{channels}{names of the channels or markers to be included in the
spectral profiles, set to all channels contained in the column names of the
supplied matrices by default.}

\item{select}{names of the dyes to extract from the rows of each matrix for
plotting, set to NULL by default to include all rows.}

\item{ylim}{limits for the y axis in percentage units, default ylim is
computed using the range across all matrices.}

\item{xlab}{label to use for the x axis, set to NA by default to remove x
axis label making more space for x axis text.}

\item{ylab}{label to use for the y axis, set to Emission percentage by
default.}

\item{legend_text}{text to use in the legend, defaults to the names of the
matrices as supplied in \code{x}.}

\item{...}{additional arguments passed to \code{\link{cyto_plot}} including
all the \code{gate_} aesthetics arguments that can be used to customise the
plotted spectra.}
}
\value{
a list containing the recorded plots.
}
\description{
Plot fluorescent spectrum for each fluorochrome across detectors
}
\seealso{
\code{\link{cyto_plot}}

\code{\link{cyto_unmix_compute}}

\code{\link{cyto_unmix}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
