% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot_spectra.R
\name{cyto_plot_spectra}
\alias{cyto_plot_spectra}
\title{Plot fluorescent spectrum for each fluorochrome across detectors}
\usage{
cyto_plot_spectra(
  x,
  parent = "root",
  channels = NULL,
  spectra_col_scale = NA,
  spectra_col = NA,
  spectra_cols = NA,
  spectra_col_alpha = 1,
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{parent}{name of the population to plot when a \code{GatingHierarchy} or
\code{GatingSet} object is supplied, set to the \code{"root"} node by
default.}

\item{channels}{names of the channels or markers to be included in the
spectral profiles, set to all channels except \code{"Time"}, \code{"FSC"}
and \code{"SSC"} by default.}

\item{spectra_col_scale}{vector of ordered colours to use for the density
colour gradient of spectra, matches \code{point_col_scale} by default.}

\item{spectra_col}{colour(s) to use for spectra, set to NA by default to use
\code{spectra_col_scale}.}

\item{spectra_cols}{vector colours to draw from when selecting colours for
spectra if none are supplied to \code{spectra_col}.}

\item{spectra_col_alpha}{numeric [0,1] to control point colour transparency
of spectra, set to 1 by default to use solid colours.}

\item{...}{additional arguments passed to \code{\link{cyto_plot}}.}
}
\value{
a list containing the recorded plots.
}
\description{
Plot fluorescent spectrum for each fluorochrome across detectors
}
\examples{
\dontrun{
# gs - transformed GatingSet of unmixed spectral data
cyto_plot_sepctra(
  gs,
  parent = "root"
)
}

}
\seealso{
\code{\link{cyto_plot}}

\code{\link{cyto_unmix_compute}}

\code{\link{cyto_unmix}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
