% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_plot-helpers.R
\name{cyto_plot_theme_reset}
\alias{cyto_plot_theme_reset}
\title{Reset an existing cyto_plot() theme to the default theme}
\usage{
cyto_plot_theme_reset(...)
}
\arguments{
\item{...}{not in use.}
}
\value{
NULL.
}
\description{
Reset an existing cyto_plot() theme to the default theme
}
\examples{
# Make all plots have a black background
cyto_plot_theme(border_fill = "black")

# Black ground with custom colour scale for points and purple gates
cyto_plot_theme(
  border_fill = "black",
  point_col_scale = c(
    "cyan",
    "green",
    "yellow",
    "orange",
    "red",
    "darkred"
  ),
  gate_line_col = "magenta"
)

# Reset to default settings
cyto_plot_theme_reset()

}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
