% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_require}
\alias{cyto_require}
\title{Load external packages for use within CytoExploreR}
\usage{
cyto_require(x, source = "CRAN", repo = NULL, version = NULL, ref = NULL, ...)
}
\arguments{
\item{x}{name of the required package.}

\item{source}{where to install the package from if it is not found on the
user's machine, options include \code{CRAN}, \code{BioC} or \code{GitHub}.
Set to \code{CRAN} by default.}

\item{repo}{name of the GitHub repository from which the package should be
installed when \code{source = GitHub}.}

\item{version}{minimal version requirement if the package is located on the
user's machine.}

\item{ref}{citation to print to the console when loading the package for use
within CytoExploreR.}

\item{...}{additional arguments passed to \code{install.packages},
\code{BiocManager::install} or \code{remotes::install_github}.}
}
\description{
Load external packages for use within CytoExploreR
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
