% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_sample}
\alias{cyto_sample}
\title{Sample a flowFrame}
\usage{
cyto_sample(x, display = 1, seed = NULL)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}}.}

\item{display}{numeric [0,1] indicating the percentage of events to keep.}

\item{seed}{value used to \code{set.seed()} internally. Setting a value for
seed will return the same result with each run.}
}
\value{
\code{\link[flowCore:flowFrame-class]{flowFrame}} restricted to
  \code{display} percentage events.
}
\description{
Sample a flowFrame
}
\examples{
library(CytoRSuiteData)

# Load in samples
fs <- Activation

# Constrict first sample by 50\%
cyto_sample(fs[[1]], 0.5)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
