% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{cyto_sample}
\alias{cyto_sample}
\title{Sample a flowFrame}
\usage{
cyto_sample(fr, display)
}
\arguments{
\item{fr}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}}.}

\item{display}{numeric [0,1] indicating the percentage of events to plot.}
}
\value{
\code{\link[flowCore:flowFrame-class]{flowFrame}} restricted to
  \code{size} events.
}
\description{
Sample a flowFrame
}
\examples{
library(CytoRSuiteData)

# Load in samples
fs <- Activation

# Constrict first sample by 50\%
cyto_sample(fs[[1]], 0.5)
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
