% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_sample}
\alias{cyto_sample}
\alias{cyto_sample.GatingHierarchy}
\alias{cyto_sample.GatingSet}
\alias{cyto_sample.flowFrame}
\alias{cyto_sample.flowSet}
\alias{cyto_sample.list}
\title{Sample a cytoframe, cytoset, GatingHierarchy or GatingSet}
\usage{
cyto_sample(x, ...)

\method{cyto_sample}{GatingHierarchy}(x, events = 1, seed = NULL, ...)

\method{cyto_sample}{GatingSet}(x, events = 1, seed = NULL, ...)

\method{cyto_sample}{flowFrame}(x, events = 1, seed = NULL, ...)

\method{cyto_sample}{flowSet}(x, events = 1, seed = NULL, ...)

\method{cyto_sample}{list}(x, events = 1, seed = NULL, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{...}{not in use.}

\item{events}{can be either a numeric [0,1] or integer to indicate the
percentage or number of events to keep respectively. For \code{cytoset} or
\code{GatingSet} objects, the same degree of sampling is applied to each
\code{cytoframe} by default. However, cytoframes will be separately sampled
if a vector the same length as the \code{cytoset} or \code{GatingSet} is
supplied. Setting \code{events} to NA will result in downsampling to the
minimum number of events across samples.}

\item{seed}{value used to \code{set.seed()} internally. Setting a value for
seed will return the same result with each run.}
}
\value{
object of class \code{\link[flowCore:flowFrame-class]{flowFrame}},
  \code{\link[flowCore:flowSet-class]{flowSet}},
  \code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
  \code{\link[flowWorkspace:GatingSet-class]{GatingSet}} sampled to
  \code{events} events.
}
\description{
\code{cyto_sample} allows restriction of a \code{cytoframe}, \code{cytoset},
\code{GatingHierarchy} or \code{GatingSet} by indicating the percentage or
number of events to retain.
}
\examples{
# Load in CytoExploreRData to access files
library(CytoExploreRData)

# Load in samples
fs <- Activation

# Restrict first sample by 50\%
cyto_sample(fs[[1]], 0.5)

# Restrict first sample to 10000 events
cyto_sample(fs[[1]], 10000)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
