% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_sample}
\alias{cyto_sample}
\alias{cyto_sample.flowFrame}
\alias{cyto_sample.flowSet}
\alias{cyto_sample.list}
\title{Sample a flowFrame or flowSet}
\usage{
cyto_sample(x, ...)

\method{cyto_sample}{flowFrame}(x, display = 1, seed = NULL)

\method{cyto_sample}{flowSet}(x, display = 1, seed = NULL)

\method{cyto_sample}{list}(x, display = 1, seed = NULL)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}} or
\code{\link[flowCore:flowSet-class]{flowSet}}.}

\item{display}{can be either a numeric [0,1] or integer to indicate the
percentage or number of events to keep respectively.}

\item{seed}{value used to \code{set.seed()} internally. Setting a value for
seed will return the same result with each run.}
}
\value{
\code{\link[flowCore:flowFrame-class]{flowFrame}} or
  \code{\link[flowCore:flowSet-class]{flowSet}} restricted to \code{display}
  percentage events.
}
\description{
\code{cyto_sample} allows restriction of a flowFrame or flowSet by indicating
the percentage or number of events to retain.
}
\details{
The list method expects a list of flowFrame objects and is used internally by
\code{cyto_plot} to ensure that overlays are appropriately sampled relative
to the base layer.
}
\examples{
library(CytoRSuiteData)

# Load in samples
fs <- Activation

# Restrict first sample by 50\%
cyto_sample(fs[[1]], 0.5)

# Restrict first sample to 10000 events
cyto_sample(fs[[1]], 10000)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
