% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_save}
\alias{cyto_save}
\title{Write samples to FCS files in new folder or save an entire GatingSet}
\usage{
cyto_save(
  x,
  save_as = ".",
  parent = NULL,
  alias = NULL,
  split = FALSE,
  id = NULL,
  names = NULL,
  overwrite = NULL,
  inverse = FALSE,
  decompensate = FALSE,
  ...
)
}
\arguments{
\item{x}{object of class \code{cytoframe}, \code{cytoset},
\code{GatingHierarchy} or \code{GatingSet}.}

\item{save_as}{name of the folder to which the written FCS files should be
saved, set to NULL by default to save the files to the current working
directory. To prevent files being overwritten, it is recommended that
\code{save_as} directory not be manually created before running
\code{cyto_save}.}

\item{parent}{name of the parent population to extract when a
\code{GatingHierarchy} or \code{GatingSet} object is supplied. If the name
of the parent is supplied the samples will be written to FCS files in
separate named folders within the \code{save_as} directory. Otherwise the
entire \code{GatingSet} or \code{GatingHierarchy} will be saved to the
specified \code{save_as} directory.}

\item{alias}{an alternative way to \code{parent} to specify the names of the
populations to be saved.}

\item{split}{logical indicating whether samples merged using
\code{cyto_merge_by} should be split prior to writing FCS files, set to
FALSE by default.}

\item{id}{passed to \code{cyto_split},  can be either the name of a channel
or marker containing the IDs for samples or a vector containing an index
for each event within \code{x}, set to \code{Sample-ID} channel by default.}

\item{names}{a vector of file names for the new FCS files. \code{names} is
also passed to \code{cyto_split()} if \code{split = TRUE} to provides the
original names for the split samples prior to writing new FCS files.}

\item{overwrite}{logical flag to control how \code{save_as} should be handled
if files already exist in this directory, users will be asked interactively
if \code{overwrite} is not manually supplied here.}

\item{inverse}{logical indicating whether inverse transformations should be
applied to the data prior to export to obtain data on the original linear
scale, set to FALSE by default.}

\item{decompensate}{logical indicating whether applied compensation should be
removed from the data prior to extraction, set to FALSE by default.
Optionally, users can supply a named of list of spillover matrices when a
\code{cytoset} is supplied to remove compensation from the extracted data.}

\item{...}{additional arguments passed to \code{\link{cyto_data_extract}} to
allow control over how the data is formatted prior to saving.}
}
\value{
a GatingHierarchy, GatingSet or a list of cytosets.
}
\description{
Write samples to FCS files in new folder or save an entire GatingSet
}
\examples{

\dontrun{
# Load in CytoExploreRData to access data
library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Save each flowFrame to file
cyto_save(fs, save_as = "Samples")
}

}
\seealso{
\code{\link{cyto_split}}

\code{\link{cyto_data_extract}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
