% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_save}
\alias{cyto_save}
\title{Write samples to FCS files in new folder or save an entire GatingSet}
\usage{
cyto_save(
  x,
  save_as = ".",
  parent = NULL,
  alias = NULL,
  split = FALSE,
  names = NULL,
  id = NULL,
  overwrite = NULL,
  ...
)
}
\arguments{
\item{x}{object of class \code{cytoframe}, \code{cytoset},
\code{GatingHierarchy} or \code{GatingSet}.}

\item{save_as}{name of the folder to which the written FCS files should be
saved, set to NULL by default to save the files to the current working
directory. To prevent files being overwritten, it is recommended that
\code{save_as} directory not be manually created before running
\code{cyto_save}.}

\item{parent}{name of the parent population to extract when a
\code{GatingHierarchy} or \code{GatingSet} object is supplied. If the name
of the parent is supplied the samples will be written to FCS files in
separate named folders within the \code{save_as} directory. Otherwise the
entire \code{GatingSet} or \code{GatingHierarchy} will be saved to the
specified \code{save_as} directory.}

\item{split}{logical indicating whether samples merged using
\code{cyto_merge_by} should be split prior to writing FCS files, set to
FALSE by default.}

\item{names}{original names of the samples prior to merging using
\code{cyto_merge_by}, only required when split is TRUE. These names will be
re-assigned to each of split flowFrames.}

\item{id}{passed to \code{cyto_split},  can be either the name of a channel
or marker containing the IDs for samples or a vector containing an index
for each event within \code{x}, set to \code{Sample-ID} channel by default.}

\item{overwrite}{logical flag to control how \code{save_as} should be handled
if files already exist in this directory, users will be asked interactively
if \code{overwrite} is not manually supplied here.}

\item{...}{additional arguments passed to \code{\link{cyto_data_extract}} to
allow control over how the data is formatted prior to saving. Setting
\code{inverse = TRUE} will export data on the linear scale but be sure to
set \code{copy = TRUE} as well if you only want to inverse data
transformations for export but leave your data unchanged.}

\item{inverse}{logical indicating whether the data should be inverse
transformed prior to writing FCS files, set to FALSE by default. Inverse
transformations of \code{flowFrame} or \code{flowSet} objects requires
passing of transformers through the \code{trans} argument.}

\item{trans}{object of class \code{transformerList} containing the
transformation definitions applied to the supplied data. Used internally
when \code{inverse_transform} is TRUE, to inverse the transformations prior
to writing FCS files.}
}
\value{
a GatingHierarchy, GatingSet or a list of cytosets.
}
\description{
Write samples to FCS files in new folder or save an entire GatingSet
}
\examples{

\dontrun{
# Load in CytoExploreRData to access data
library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Save each flowFrame to file
cyto_save(fs, save_as = "Samples")
}

}
\seealso{
\code{\link{cyto_split}}

\code{\link{cyto_data_extract}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
