% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_save}
\alias{cyto_save}
\alias{cyto_save.GatingSet}
\alias{cyto_save.GatingHierarchy}
\alias{cyto_save.flowSet}
\alias{cyto_save.flowFrame}
\title{Write samples to fcs files in new folder}
\usage{
cyto_save(x, ...)

\method{cyto_save}{GatingSet}(
  x,
  parent = "root",
  split = FALSE,
  names = NULL,
  save_as = NULL,
  trans = NULL
)

\method{cyto_save}{GatingHierarchy}(
  x,
  parent = "root",
  split = FALSE,
  names = NULL,
  save_as = NULL,
  trans = NULL
)

\method{cyto_save}{flowSet}(x, split = FALSE, names = NULL, save_as = NULL, trans = NULL)

\method{cyto_save}{flowFrame}(x, split = FALSE, names = NULL, save_as = NULL, trans = NULL)
}
\arguments{
\item{x}{object of class \code{flowFrame}, \code{flowSet},
\code{GatingHierarchy} or \code{GatingSet}.}

\item{parent}{name of the parent population to extract when a
\code{GatingHierarchy} or \code{GatingSet} object is supplied, defaults to
the "root" node if not supplied.}

\item{split}{logical indicating whether samples merged using
\code{cyto_merge_by} should be split prior to writing fcs files, set to
FALSE by default.}

\item{names}{original names of the samples prior to merging using
\code{cyto_merge_by}, only required when split is TRUE. These names will be
re-assigned to each of split flowFrames and included in the file names.}

\item{save_as}{name of the folder to which the written fcs files should be
saved, set to NULL by default to save the files to the current working
directory.}

\item{trans}{object of class \code{transformerList} containg the
transformation definitions applied to the supplied data. If transformations
are supplied, the data will be inverse transformed prior to saving to
return the data on the original linear scale.}
}
\value{
flowFrame or list of flowFrames that were saved to fcs files.
}
\description{
Write samples to fcs files in new folder
}
\examples{

\dontrun{
# Load in CytoExploreRData to access data
library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Save each flowFrame to file
cyto_save(fs)
}

}
\seealso{
\code{\link{cyto_split}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
