% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_select}
\alias{cyto_select}
\title{Select samples based on experiment variables}
\usage{
cyto_select(x, exclude = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{exclude}{logical indicating whether samples matching the input
selection criteria should be excluded from the returned \code{flowSet} or
\code{GatingSet}.}

\item{...}{named list containing experimental variables to be used to select
samples or named arguments containing the levels of the variables to
select. See below examples for use cases.}
}
\value{
\code{flowSet} or \code{GatingSet} restricted to samples which meet
  the designated selection criteria.
}
\description{
Select samples based on experiment variables
}
\examples{

# Load in CytoExploreRData to access data
library(CytoExploreRData)

# Look at experiment details
cyto_details(Activation)

# Select Stim-C samples with 0 and 0.5 OVA concentrations
fs <- cyto_select(Activation,
  Treatment = "Stim-C",
  OVAConc = c(0, 0.5)
)

# Select Stim-A and Stim-C treatment groups
fs <- cyto_select(
  Activation,
  list("Treatment" = c("Stim-A", "Stim-C"))
)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
