% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_setup}
\alias{cyto_setup}
\title{Load.fcs files into GatingSet and annotate with experiment details}
\usage{
cyto_setup(path = ".", ...)
}
\arguments{
\item{path}{points to the location of the .fcs files to read in.}

\item{...}{additional arguments passed to read.ncdfFlowSet.}
}
\value{
\code{GatingSet}
}
\description{
\code{cyto_setup} takes care of all the data loading and annotation steps to
prepare your cytometry data for downstream analyses. The .fcs files are first
read into a \code{ncdfFlowSet} using \code{cyto_load} which is then added to
a \code{GatingSet}. Calls are then made to \code{cyto_markers} and
\code{cyto_annotate} to update the GtaingSet with the details of the
experiment. These details can be modified later with additional calls to
\code{cyto_markers} and/or \code{cyto_annotate}.
}
\examples{
library(CytoRSuiteData)

# Get path to Activation .fcs files in CytoRSuiteData
datadir <- system.file("extdata", package = "CytoRSuiteData")
path <- paste0(datadir,"/Activation")

# Load in .fcs files into an annotated GatingSet
fs <- cyto_setup(path)

# Markers have been assigned
cyto_extract(gs, "root")[[1]]

# Experiment details have been updated
cyto_details(gs)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
