% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_setup}
\alias{cyto_setup}
\title{Load.fcs files into GatingSet and annotate with experiment details}
\usage{
cyto_setup(path = ".", gatingTemplate = NULL, ...)
}
\arguments{
\item{path}{points to the location of the .fcs files to read in (e.g. name of
a folder in current working directory).}

\item{gatingTemplate}{name of a gatingTemplate csv file to be used for gate
saving.}

\item{...}{additional arguments passed to
\code{\link[flowWorkspace:load_cytoset_from_fcs]{load_cytoset_from_fcs}}.}
}
\value{
object of class
  \code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.
}
\description{
\code{cyto_setup} takes care of all the data loading and annotation steps to
prepare your cytometry data for downstream analyses. The .fcs files are first
read into a \code{\link[flowWorkspace:cytoset]{cytoset}} using
\code{\link{cyto_load}} which is then added to a
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}. Calls are then made
to \code{\link{cyto_markers_edit}} and \code{\link{cyto_details_edit}} to
update the GatingSet with the details of the experiment. These details can be
modified later with additional calls to \code{\link{cyto_markers_edit}}
and/or \code{\link{cyto_details_edit}}. Users can optionally provide a
name for a gatingTemplate csv file which will be created if necessary and
assigned as the active gatingTemplate.
}
\examples{

\dontrun{
# Load in CytoExploreRData to access data
library(CytoExploreRData)

# Get path to Activation .fcs files in CytoExploreRData
datadir <- system.file("extdata", package = "CytoExploreRData")
path <- paste0(datadir, "/Activation")

# Load in .fcs files into an annotated GatingSet
gs <- cyto_setup(path)

# Markers have been assigned
cyto_extract(gs, "root")[[1]]

# Experiment details have been updated
cyto_details(gs)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
