% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_spillover_compute.R
\name{cyto_spillover_compute}
\alias{cyto_spillover_compute}
\title{Compute Spillover Matrix}
\usage{
cyto_spillover_compute(
  x,
  parent = "root",
  select = NULL,
  channels = NULL,
  gate = "draw",
  type = "CytoDecode",
  save_as = NULL,
  gatingTemplate = NULL,
  axes_trans = NA,
  axes_limits = "machine",
  spillover = NULL,
  heatmap = TRUE,
  events = 500,
  model = "rlm",
  max_iter = 20,
  trim = 0.001,
  tol = 1e-06,
  grid_size = 100,
  resid = "both",
  details = NULL,
  rerun = FALSE,
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} containing
transformed and gated compensation controls.}

\item{parent}{name of the population to use for the spillover calculation
when a GatingSet object is supplied, set to the last node of the GatingSet
by default (e.g. "Single Cells"). For greater flexibility, users can
specify a parent population for each control, which will be extracted for
the spillover calculation (e.g. Lymphocytes for CD4 APC or Myeloid Cells
for CD11b FITC). The parent populations for each control can also be
specified in a \code{parent} column in the channel match CSV file or in
\code{cyto_details}.}

\item{select}{passed to \code{\link{cyto_select}} to select the samples
required to compute the spillover matrix.}

\item{channels}{names of the channels or markers for which spillover
coefficients should be computed, set to all area fluorescent channels by
default.}

\item{gate}{indicates whether to \code{draw} or use \code{auto} gating
methods to gate the negative and positive populations for each control.
Gating is optional for \code{"Autospill"} and \code{"CytoDecode"} methods
but is recommended to achieve better results and ensure compatibility with
\code{cyto_panel_design()}. Set to \code{"draw"} by default to allow for
more flexibility over gating. For \code{gate = "draw"} users can also
specify whether to gate the 1D histograms or 2D scatter plots 1D or 2D
suffix to \code{gate}. For example, to gate in 2D set \code{gate =
"draw-2D"} which is the new default manual gating method.}

\item{type}{options include \code{"Bagwell" or "autocomp"}, \code{"Roca" or
"autospill"} or \code{"CytoDecode"} to indicate which method to use when
computing the spillover matrix, set to \code{"CytoDecode"} by default.}

\item{save_as}{name of a csv file to which the computed spillover matrix
should be written, set to \code{Spillover-Matrix.csv} prefixed with the
date by default.}

\item{gatingTemplate}{name of \code{gatingTemplate} csv file to which the
\code{gatingTemplate} entries for the \code{GatingSet} method should be
saved, set to \code{cyto_gatingTemplate_active()} by default.}

\item{axes_trans}{object of class
\code{\link[flowWorkspace:transformerList]{transformerList}} generated by a
\code{cyto_transformer} which contains the transformer definitions used to
transform the data. Transformer definitions are only required when a
flowSet object is supplied.}

\item{axes_limits}{options include \code{"auto"}, \code{"data"} or
\code{"machine"} to use optimised, data or machine limits respectively. Set
to \code{"machine"} by default to use entire axes ranges. Fine control over
axes limits can be obtained by altering the \code{xlim} and \code{ylim}
arguments.}

\item{spillover}{superseded by the \code{save_as} argument above, only
included for backwards compatibility with older versions of CytoExploreR.}

\item{heatmap}{logical indicating whether the computed spillover matrix
should be displayed in a heatmap, set to TRUE by default.}

\item{events}{number of events to extract from each control prior to fitting
RLM models, set to 500 events by default.}

\item{model}{indicates the type of model to use when running autospill,
options include \code{"rlm"} for robust linear models or \code{"vwqr"} for
variance weighted quantile regression used in \code{CytoDecode}, set to
\code{"rlm"} by default for speed and backwards compatibility.}

\item{max_iter}{indicates the maximum number of allowable iterations for
refining the spillover coefficients, set to 20 by default.}

\item{trim}{proportion of events to exclude from the top and bottom of each
scale when fitting models in \code{AutoSpill} and \code{CytoDecode}
methods, set to 0.001 by default.}

\item{tol}{tolerance level for convergence of vwqr model in \code{CytoDecode}
method, set to \code{1e-6} by default.}

\item{grid_size}{size of the grid to use for wvqr model in \code{CytoDecode}
method, set to 100 by default. Smaller grid sizes will provide a speed
boost at the cost of accuracy.}

\item{resid}{indicates whether to use \code{"lower"}, \code{"upper"} or
\code{"both"} residuals when fitting variance models in vwqr used in
\code{CytoDecode}, set to \code{"both"} by default.}

\item{details}{name of a CSV file to which the details of the single colour
controls should be saved, set to NULL by default to use
\code{date-Control-Details.csv}. Setting this argument to \code{NA} will
prevent details from being written to a CSV file.}

\item{rerun}{logicla indicating whether to bypass gating and simply rerun the
algorithm using existing gates, set to FALSE by default.}

\item{...}{additional arguments passed to \code{\link{cyto_plot}}.}
}
\value{
spillover matrix and write spillover matrix to csv file named in
  accordance with \code{save_as}.
}
\description{
\code{cyto_spillover_compute} automatically computes the fluorescent
spillover matrix for a set of single colour compensation controls using
either the method proposed by Bagwell et al. (1993) or the more recent method
proposed by Roca et al. (2021).
}
\details{
The method proposed by Roca et al. (2021) is now the default method, and has
been implemented in \code{cyto_spillover_compute} as a fully automated
algorithm to compute the spillover matrix. Users simply need to ensure that
the channel associated with each control is annotated in the
\code{cyto_details()} of the supplied data.

The method proposed by Bagwell et al. (1993) is also supported but is
implemented as a semi-automated approach as it requires gating of the
positive and negative signal within each single colour control to accurately
compute the spillover coefficients. Gating is performed on a per channel
basis with the user manually drawing interval gates around the positive and
negative signal for each control. If a universal unstained control is
supplied, users are only asked to gate the positive signal for each control
and the unstained control will be used as the reference for the negative
signal in each channel.

The computed spillover is returned as a matrix and exported as a csv file
specified by the \code{save_as} argument. To assess the quality of the
compensation, refer to \code{\link{cyto_spillover_edit}} and
\code{\link{cyto_plot_compensation}}. Users can apply their computed
spillover matrices to their samples using \code{\link{cyto_compensate}}.
}
\examples{
\dontrun{
library(CytoExploreRData)

# Load in compensation controls
fs <- Compensation
gs <- GatingSet(Compensation)

# Gate using cyto_gate_draw
gt <- Compensation_gatingTemplate
cyto_gatingTemplate_apply(gs, gt)

# Channel match fille
cmfile <- system.file("extdata",
  "Compensation-Channels.csv",
  package = "CytoExploreRData"
)

# Compute fluorescent spillover matrix
spill <- cyto_spillover_compute(
  gs,
  save_as = "Example-spillover.csv"
)

# Compensate samples
gs <- cyto_compensate(gs, spill)

}

}
\references{
C. B. Bagwell & E. G. Adams (1993). Fluorescence spectral overlap
  compensation for any number of flow cytometry parameters. in: Annals of the
  New York Academy of Sciences, 677:167-184.

Roca et al. (2021). AutoSpill is a principled framework that
  simplifies the analysis of multichromatic flow cytometry data. Nature
  Communications 12(2890).
}
\seealso{
\code{\link{cyto_channel_match}}

\code{\link{cyto_spillover_edit}}

\code{\link{cyto_plot_compensation}}

\code{\link{cyto_compensate}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
