% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_spillover_compute.R
\name{cyto_spillover_compute}
\alias{cyto_spillover_compute}
\alias{cyto_spillover_compute.GatingSet}
\alias{cyto_spillover_compute.flowSet}
\title{Compute Spillover Matrix}
\usage{
\method{cyto_spillover_compute}{GatingSet}(x, parent = NULL,
  axes_trans = NULL, channel_match = NULL, spillover = NULL, ...)

\method{cyto_spillover_compute}{flowSet}(x, axes_trans = NULL,
  channel_match = NULL, spillover = NULL, ...)
}
\arguments{
\item{x}{object of class \code{flowSet} or \code{GatingSet} containing
un-transformed pre-gated compensation controls.}

\item{parent}{name of the population to use for the spillover calculation
when a GatingSet object is supplied, set to the last node of the GatingSet
by default (e.g. "Single Cells").}

\item{axes_trans}{object of class
\code{\link[flowWorkspace:transformerList-class]{transformerList}}
generated by a \code{cyto_transformer} which contains the transformer
definitions to be used internally to transform the fluorescent channels of
the samples.}

\item{channel_match}{name of csv file to associate a fluorescent channel with
each of the compensation controls. The \code{channel_match} file should
contain a "name" column with the names of the compensation controls and a
"channel" column to associate a fluorescent channel with each compensation
control. Users need not generate this file by hand as it will be created
following the channel selection process.}

\item{spillover}{name of the output spillover csv file, set to
\code{"Spillover-Matrix.csv"} by default.}

\item{...}{additional arguments passed to \code{\link{cyto_plot}}.}
}
\value{
spillover matrix and write spillover matrix to csv file named in
  accordance with \code{spillover}.
}
\description{
\code{cyto_spillover_compute} uses the method described by Bagwell & Adams
1993 to automatically calculate the fluorescent spillover matrix using single
stain compensation controls.
}
\details{
\code{cyto_spillover_compute} supports spillover matrix calculation for both
internal or universal unstained reference populations based on channel
selection. \code{cyto_spillover_compute} expects the fluorescent channels of
the samples to be un-transformed, biexponential transformations will be
applied internally when appropriate to allow proper visualisation of the
data. To use your own custom transformations, simply suppply a
transformerList object to the \code{axes_trans} argument.

\code{cyto_spillover_compute} begins by the user selecting which fluorescent
channel is associated with each compensation control from a dropdown menu.
The results of these selections are saved to a csv file called
"Compensation-Channels.csv" which can be passed to the \code{channel_match}
argument on subsequent runs to bypass the channel selection process.

Following channel selection, \code{cyto_spillover_compute} runs through each
compensation control and plots the density distribution in the associated
channel. If an universal "Unstained" compensation control is supplied, the
unstained compensation control will be overlaid onto the plot as a reference
for gating. Users can then gate the positive signal for spillover calculation
using an interval gate. If no universal unstained compensation control is
supplied, users are expected to gate the negative and then the positive
signal for each compensation control.

The percentage spillover is calculated based on the median fluorescent
intensities (MedFI) of the positive populations relative to that of the
reference negative population(s). The calculated spillover matrix is returned
and written to a named csv file for future use.
}
\examples{
library(CytoRSuiteData)

# Bypass directory check for external files
options("CytoExploreR_wd_check" = FALSE)

# Load in compensation controls
fs <- Compensation
gs <- GatingSet(Compensation)

# Gate using cyto_gate_draw
gt <- Compensation_gatingTemplate
gating(gt, gs)

# Channel match fille
cmfile <- system.file("extdata",
  "Compensation-Channels.csv",
  package = "CytoRSuiteData"
)

# Compute fluorescent spillover matrix
spill <- spillover_compute(getData(gs, "Single Cells"),
  channel_match = cmfile,
  spillover = "Example-spillover.csv"
)

# Compensate samples
gs <- compensate(gs, spill)

# Return CytoExploreR_wd_check to default
options("CytoExploreR_wd_check" = TRUE)
}
\references{
C. B. Bagwell \& E. G. Adams (1993). Fluorescence spectral
  overlap compensation for any number of flow cytometry parameters. in:
  Annals of the New York Academy of Sciences, 677:167-184.
}
\seealso{
\code{\link{cyto_spillover_edit}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
