% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_spillover_extract}
\alias{cyto_spillover_extract}
\title{Extract spillover matrix from cytometry object}
\usage{
cyto_spillover_extract(x)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}
}
\value{
list of spillover matrices or NULL.
}
\description{
Extract spillover matrix from cytometry object
}
\examples{
library(CytoExploreRData)

# Activation Gatingset
gs <- load_gs(system.file("extdata/Activation-GatingSet",
                          package = "CytoExploreRData"))

# Extract spillover matrices
spill <- cyto_spillover_extract(gs)
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
