% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_spillover_spread_compute.R
\name{cyto_spillover_spread_compute}
\alias{cyto_spillover_spread_compute}
\alias{cyto_spillover_spread_compute.GatingSet}
\alias{cyto_spillover_spread_compute.flowSet}
\title{Compute Spillover Spreading Matrix}
\usage{
\method{cyto_spillover_spread_compute}{GatingSet}(x, parent = NULL,
  axes_trans = NULL, channel_match = NULL, compensated = FALSE,
  spillover = NULL, spillover_spread = NULL, ...)

\method{cyto_spillover_spread_compute}{flowSet}(x, axes_trans = NULL,
  channel_match = NULL, compensated = FALSE, spillover = NULL,
  spillover_spread = NULL, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} containing pre-gated
single stain compensation controls.}

\item{parent}{name of the population to use for downstream calculations when
a \code{GatingSet} object is supplied, set to the last gated population by
default (e.g. "Single Cells").}

\item{axes_trans}{object of class
\code{\link[flowWorkspace:transformerList-class]{transformerList}}
generated by a \code{cyto_transformer} which contains the transformer
definitions to be used internally to transform the fluorescent channels of
the samples as required.}

\item{channel_match}{name of csv file to associate a fluorescent channel with
each of the compensation controls. The \code{channel_match} file should
contain a "name" column with the names of the compensation controls and a
"channel" column to associate a fluorescent channel with each compensation
control. Users need not generate this file by hand as it will be created
following the channel selection process.}

\item{compensated}{logical indicating whether the supplied compensation
controls have been compensated. Compensation controls must be compensated
prior tro calculation of spillover spreading matrix. If \code{compensated}
is set to FALSE, the compensation controls will be compensated internally
using the supplied \code{spillover}.}

\item{spillover}{name of the output spillover matrix csv file to be used
internally to compensate the compensation controls. If no spillover matrix
is supplied \code{spillover_spread_compute} will look for a
"Spillover-Matrix.csv" file or as a last resort use the spillover matrix
attached to compensation controls.}

\item{spillover_spread}{name of the csv file to which the spillover spreading
matrix will be saved, set to \code{"date-Spillover-Spread-Matrix.csv"} by
default.}

\item{...}{additional arguments passed to \code{\link{cyto_plot}}.}
}
\description{
\code{spillover_spread} computes the spillover spreading matrix as described
by Nguyen et al. (2013) using gated and compensated compensation controls.
}
\references{
Nguyen R, Perfetto S, Mahnke YD, Chattopadhyay P & Roederer M,
  (2013). Quantifying spillover spreading for comparing instrument
  performance and aiding in multicolor panel design. Cytometry A,
  83(3):306-15.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
