% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_split}
\alias{cyto_split}
\title{Split samples merged with cyto_merge_by()}
\usage{
cyto_split(x, id = NULL, names = NULL)
}
\arguments{
\item{x}{object of class \code{cytoframe} or \code{cytoset}.}

\item{id}{can be either the name of a channel or marker containing the IDs
for samples or a vector containing an index for each event within \code{x},
set to \code{Sample-ID} channel by default.}

\item{names}{vector of names to assign to each of the extracted cytoframes
when saving the split files. Names should be supplied in the order used
prior to merging (i.e. in Sample-ID order).}
}
\value{
a cytoset containing the split cytoframes.
}
\description{
Extract individual samples merged using \code{cyto_merge_by()} or
\code{cyto_coerce()} based on \code{"Sample-ID"} column created by
\code{cyto_barcode()}.
}
\examples{

# Load CytoExploreRData to access data
library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Merge samples
fr <- cyto_merge_by(fs, "all")[[1]]

# Split merged samples
fr_list <- cyto_split(fr, names = cyto_names(fs))
}
\seealso{
\code{\link{cyto_merge_by}}

\code{\link{cyto_barcode}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
