% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_stats_compute.R
\name{cyto_stats_compute}
\alias{cyto_stats_compute}
\title{Compute Statistics}
\usage{
cyto_stats_compute(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{...}{additional method-specific arguments.}
}
\description{
Calculate and export flow cytometry statistics.
}
\seealso{
\code{\link{cyto_stats_compute,flowFrame-method}}

\code{\link{cyto_stats_compute,flowSet-method}}

\code{\link{cyto_stats_compute,GatingSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
