% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_transform}
\alias{cyto_transform}
\alias{cyto_transform.flowSet}
\alias{cyto_transform.GatingSet}
\title{Apply logicle transformation to channels of flowSet or GatingSet}
\usage{
cyto_transform(x, ...)

\method{cyto_transform}{flowSet}(x, channels, select = NULL, ...)

\method{cyto_transform}{GatingSet}(x, channels, select = NULL, ...)
}
\arguments{
\item{x}{object of class \code{flowSet} or \code{GatingSet}.}

\item{...}{additional arguments passed to
\code{\link[flowCore:estimateLogicle]{estimateLogicle}}.}

\item{channels}{name(s) of the channels to transform. A call will be made to
\code{\link{cyto_fluor_channels}} to get the names of the fluorescent
channels if no channels are supplied.}

\item{select}{a named list of experiment variables passed to
\code{\link{cyto_select}} to select particular samples to use for logicle
parameter estimation.}
}
\value{
transformed \code{flowSet} or \code{GatingSet}.
}
\description{
A convenient wrapper for
\code{\link[flowCore:estimateLogicle]{estimateLogicle}} and
\code{\link[flowCore:transform]{transform}} which coerces samples prior to
logicle parameter estimation. For large \code{flowSet} or \code{GatingSet}
objects it is recommended that users select a subset of samples for parameter
estimation using the \code{select} argument.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
