% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_transform}
\alias{cyto_transform}
\alias{cyto_transform.default}
\alias{cyto_transform.transformList}
\alias{cyto_transform.transformerList}
\title{Apply Transformations to Cytometry Data}
\usage{
cyto_transform(x, trans = NULL, ...)

\method{cyto_transform}{default}(x, trans = NULL,
  trans_type = "logicle", channels = NULL, parent = "root",
  select = NULL, inverse = FALSE, plot = TRUE, popup = FALSE, ...)

\method{cyto_transform}{transformList}(x, trans = NULL, plot = TRUE,
  popup = FALSE, ...)

\method{cyto_transform}{transformerList}(x, trans = NULL,
  inverse = FALSE, plot = TRUE, popup = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{flowFrame}, \code{flowSet},
\code{GatingHierarchy} or \code{GatingSet}.}

\item{trans}{object of class \code{transformerList} containing the
transformation definitions to apply to \code{x}.}

\item{...}{additional arguments passed to \code{cyto_transform_arcsinh},
\code{cyto_transform_biex} or \code{cyto_transform_logicle} when no
\code{trans} object is supplied.}

\item{trans_type}{type of transformation to apply when no trans object is
supplied, options include "arcsinh", "biex" and "logicle".}

\item{channels}{names of the channels to transform. Only required when no
\code{trans} object is supplied.}

\item{parent}{name of the parent population of \code{GatingHierarchy} or
\code{GatingSet} objects used to visualise the transformations.}

\item{select}{list of selection criteria passed to \code{cyto_select} to
select a subset of samples for visualising the transformations.}

\item{inverse}{logical indicating whether the inverse transformations should
be applied. Currently only supported for \code{flowFrame} and
\code{flowSet} objects.}

\item{plot}{logicla indicating whether the result of the transformations
should be plotted using \code{cyto_plot}.}

\item{popup}{logical indicating whether plots should be constructed in a
popup window, set to FALSE by default.}
}
\description{
Apply Transformations to Cytometry Data
}
\examples{

library(CytoRSuite)

# Use Activation flowSet
fs <- Activation

# Automatically transform flowSet
fs_trans <- cyto_transform(fs, trans_type = "arcsinh")

# Manually construct & apply transformations
trans <- cyto_transform_biex(fs)
fs_trans <- cyto_transform(fs, trans)

# Add fs to GatingSet
gs <- GatingSet(fs)

# Automatically transform GatingSet
gs_trans <- cyto_transform(gs, trans_type = "logicle")

# Manually construct & apply transformations
trans <- cyto_transform_logicle(gs)
gs_trans <- cyto_transform(gs, trans)

}
\seealso{
\code{\link{cyto_transform_arcsinh}}

\code{\link{cyto_transform_biex}}

\code{\link{cyto_transform_logicle}}

\code{\link{cyto_transform_combine}}

\code{\link{cyto_transform}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
