% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_transform}
\alias{cyto_transform}
\alias{cyto_transform.default}
\alias{cyto_transform.transformList}
\alias{cyto_transform.transformerList}
\title{Apply Transformations to Cytometry Data}
\usage{
cyto_transform(x, trans = NULL, ...)

\method{cyto_transform}{default}(
  x,
  trans = NULL,
  type = "logicle",
  channels = NULL,
  parent = "root",
  select = NULL,
  inverse = FALSE,
  plot = TRUE,
  popup = FALSE,
  axes_limits = "machine",
  ...
)

\method{cyto_transform}{transformList}(
  x,
  trans = NULL,
  plot = TRUE,
  popup = FALSE,
  axes_limits = "machine",
  ...
)

\method{cyto_transform}{transformerList}(
  x,
  trans = NULL,
  inverse = FALSE,
  plot = TRUE,
  popup = FALSE,
  axes_limits = "machine",
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{trans}{object of class
\code{\link[flowWorkspace:transformerList]{transformerList}} containing the
transformation definitions to apply to \code{x}.}

\item{...}{additional arguments passed to \code{\link{cyto_transformer_log}},
\code{\link{cyto_transformer_arcsinh}}, \code{\link{cyto_transformer_biex}}
or \code{\link{cyto_transformer_logicle}}, when no \code{trans} object is
supplied.}

\item{type}{type of transformation to apply when no trans object is supplied,
options include \code{"log"}, \code{"arcsinh"}, \code{"biex"} and
\code{"logicle"}.}

\item{channels}{names of the channels to transform, set to all the
fluorescent channels by default. Only required when no \code{trans} object
is supplied.}

\item{parent}{name of the parent population of \code{GatingHierarchy} or
\code{GatingSet} objects used to visualise the transformations.}

\item{select}{list of selection criteria passed to \code{\link{cyto_select}}
to select a subset of samples for visualising the transformations.}

\item{inverse}{logical indicating whether the inverse transformations should
be applied. Currently only supported for \code{flowFrame} and
\code{flowSet} objects.}

\item{plot}{logical indicating whether the result of the transformations
should be plotted using \code{\link{cyto_plot}}.}

\item{popup}{logical indicating whether plots should be constructed in a
popup window, set to FALSE by default.}

\item{axes_limits}{options include \code{"auto"}, \code{"data"} or
\code{"machine"} to use optimised, data or machine limits respectively. Set
to \code{"machine"} by default to use entire axes ranges. Fine control
over axes limits can be obtained by altering the \code{xlim} and
\code{ylim} arguments.}
}
\value{
object of class \code{flowFrame}, \code{flowSet},
  \code{GatingHierarchy} or \code{GatingSet} with transformations applied.
}
\description{
Apply Transformations to Cytometry Data
}
\examples{

# Load in CytoExploreRData to access data
library(CytoExploreRData)

# Activation flowSet
fs <- Activation

# Automatically transform flowSet
fs_trans <- cyto_transform(fs, type = "arcsinh")

# Manually construct & apply transformations
trans <- cyto_transformer_biex(fs)
fs_trans <- cyto_transform(fs, trans)

# Add fs to GatingSet
gs <- GatingSet(fs)

# Automatically transform GatingSet
gs_trans <- cyto_transform(gs, type = "logicle")

# Manually construct & apply transformations
trans <- cyto_transformer_logicle(gs)
gs_trans <- cyto_transform(gs, trans)
}
\seealso{
\code{\link{cyto_transformer_log}}

\code{\link{cyto_transformer_arcsinh}}

\code{\link{cyto_transformer_biex}}

\code{\link{cyto_transformer_logicle}}

\code{\link{cyto_transformer_combine}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
