% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto-helpers.R
\name{cyto_transform_extract}
\alias{cyto_transform_extract}
\title{Extract Transformations from TransformerList}
\usage{
cyto_transform_extract(x, inverse = FALSE)
}
\arguments{
\item{x}{object of class
\code{\link[flowWorkspace]{transformerList}}.}

\item{inverse}{logical indicating whether the returned transformList should
contain the inverse transformations.}
}
\value{
A \code{\link[flowCore:transformList-class]{transformList}} containing
  the desired transformations.
}
\description{
Extract Transformations from TransformerList
}
\examples{
library(CytoRSuiteData)

# Load in samples to flowSet
fs <- Activation

# Add fs to GatingSet
gs <- GatingSet(fs)

# Convert transformerList into transformList
trans <- estimateLogicle(gs[[32]], cyto_fluor_channels(gs))
trans_list <- cyto_transfrm_extract(trans)

# Convert transformerList into inverse transformList
inv <- cyto_transform_extract(trans, inverse = TRUE)

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
