% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_transform.R
\name{cyto_transform_log}
\alias{cyto_transform_log}
\alias{cyto_transform_log.flowFrame}
\alias{cyto_transform_log.flowSet}
\alias{cyto_transform_log.GatingHierarchy}
\alias{cyto_transform_log.GatingSet}
\title{Definition(s) of Log Transformation(s)}
\usage{
cyto_transform_log(x, ...)

\method{cyto_transform_log}{flowFrame}()

\method{cyto_transform_log}{flowSet}()

\method{cyto_transform_log}{GatingHierarchy}()

\method{cyto_transform_log}{GatingSet}()
}
\description{
Definition(s) of Log Transformation(s)
}
