% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_transformers.R
\name{cyto_transformer_arcsinh}
\alias{cyto_transformer_arcsinh}
\alias{cyto_transformer_arcsinh.GatingSet}
\alias{cyto_transformer_arcsinh.GatingHierarchy}
\alias{cyto_transformer_arcsinh.flowSet}
\alias{cyto_transformer_arcsinh.flowFrame}
\title{Definition(s) of arcsinh transformation(s)}
\usage{
cyto_transformer_arcsinh(x, ...)

\method{cyto_transformer_arcsinh}{GatingSet}(x, channels = NULL,
  parent = "root", select = NULL, plot = TRUE, popup = FALSE, ...)

\method{cyto_transformer_arcsinh}{GatingHierarchy}(x, channels = NULL,
  parent = "root", select = NULL, plot = TRUE, popup = FALSE, ...)

\method{cyto_transformer_arcsinh}{flowSet}(x, channels = NULL,
  select = NULL, plot = TRUE, popup = FALSE, ...)

\method{cyto_transformer_arcsinh}{flowFrame}(x, channels = NULL,
  select = NULL, plot = TRUE, popup = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{flowFrame}, \code{flowSet},
\code{GatingHierarchy} or \code{GatingSet}.}

\item{...}{additional arguments passed to
\code{\link[flowWorkspace:asinh_Gml2]{asinh_Gml2}}.}

\item{channels}{name(s) of channel(s)/marker(s) for which transformation
functions must be generated. Set to all the fluorescent channels by default
if no channels/markers are supplied.}

\item{parent}{name of the parent population to use for generating the
transformation(s).}

\item{select}{list of selection criteria passed to \code{cyto_select} to
select a subset of samples for \code{cyto_plot}.}

\item{plot}{logical indicating whether the results of transformations should
be plotted using \code{cyto_plot}.}

\item{popup}{logical indicating whether the plots should be constructed in a
pop-up window.}
}
\value{
a \code{transformerList} object.
}
\description{
CytoExploreR implementation of \code{flowWorkspace}
\code{\link[flowWorkspace:asinh_Gml2]{asinh_Gml2}} transformation which
always returns a \code{transformerList} object and displays the result of the
transformation(s) using \code{cyto_plot}. To combine different types of
transformations have a look at \code{cyto_transformer_combine}.
\code{cyto_transform} should be used to apply the transformations to the
data.
}
\seealso{
\code{\link[flowWorkspace:asinh_Gml2]{asinh_Gml2}}

\code{\link{cyto_transformer_log}}

\code{\link{cyto_transformer_biex}}

\code{\link{cyto_transformer_logicle}}

\code{\link{cyto_transformer_combine}}

\code{\link{cyto_transform}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
