% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_transformers.R
\name{cyto_transformer_combine}
\alias{cyto_transformer_combine}
\title{Combine cyto_transformer definitions}
\usage{
cyto_transformer_combine(...)
}
\arguments{
\item{...}{objects of class
\code{\link[flowWorkspace:transformerList]{transformerList}} to be
combined into a single \code{transformerList} object.}
}
\description{
\code{cyto_transformer_combine} makes it easy to combine transformation
definitions obtained from \code{\link{cyto_transformer_log}},
\code{\link{cyto_transformer_arcsinh}}, \code{\link{cyto_transformer_biex}}
and/or \code{\link{cyto_transformer_logicle}} prior to applying these
transformations to the data using \code{\link{cyto_transform}}.
}
\seealso{
\code{\link{cyto_transformer_log}}

\code{\link{cyto_transformer_arcsinh}}

\code{\link{cyto_transformer_biex}}

\code{\link{cyto_transformer_logicle}}

\code{\link{cyto_transform}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
