% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_spectra_unmix.R
\name{cyto_unmix}
\alias{cyto_unmix}
\title{Apply spectral unmixing matrix to spectral data}
\arguments{
\item{x}{an object of class \code{\link[flowWorkspace:cytoframe]{cytoframe}},
\code{\link[flowWorkspace:cytoset]{cytoset}},
\code{\link[flowWorkspace:GatingHierarchy-class]{GatingHierarchy}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} containing unmixed
spectral data.}

\item{parent}{name of the parental population in the supplied
\code{GatingHierarchy} or \code{GatingSet} to unmix, set to \code{"root"}
by default.}

\item{unmix}{an unmixing matrix or the name of a CSV file containing the
unmixing matrix.}

\item{select}{sample selection criteria to select a subset of samples for
unmixing, see \code{\link{cyto_select}} for details.}

\item{copy}{logical indicating whether unmixing should be applied to a copy
of the original data, set to FALSE by default.}

\item{drop}{logical indicating whether the original channels used compute the
unmixing matrix should be dropped from the unmixed data, set to TRUE by
default.}

\item{trans}{object of class \code{transformerList} containing the
definitions of transformers already applied to the input \code{cytoframe}
or \code{cytoset}, set to NA by default.}

\item{inverse}{logical to indicate whether inverse data transformations
should be applied to the data prior to unmixing, set to TRUE by default.}

\item{...}{not in use.}
}
\value{
unmixed \code{cytoframe}, \code{cytoset}, \code{GatingHierarchy} or
  \code{GatingSet} with fluorescent values on the linear scale.
}
\description{
Apply spectral unmixing matrix to spectral data
}
\examples{
\dontrun{
# Compute unmixing matrix
unmix <- cyto_unmix_compute(
  gs,
  save_as = "Spectral-Unmixing-Matrix.csv"
)
# Apply unmixing matrix
gs <- cyto_unmix(
  gs,
  unmix = unmix
)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
