% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_spectra_unmix.R
\name{cyto_unmix_compute}
\alias{cyto_unmix_compute}
\title{Compute spectral unmixing matrix}
\usage{
cyto_unmix_compute(
  x,
  parent = "root",
  select = NULL,
  channels = NULL,
  gate = "draw-2d",
  type = "CytoDecode",
  stat = "geomedian",
  save_as = NULL,
  gatingTemplate = NULL,
  axes_trans = NA,
  axes_limits = "machine",
  heatmap = TRUE,
  events = 500,
  model = "rlm",
  max_iter = 20,
  trim = 0.001,
  tol = 1e-06,
  grid_size = 100,
  resid = "both",
  details = NULL,
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} containing single
colour controls.}

\item{parent}{name of the population to use for the unmixing calculation
when a GatingSet object is supplied, set to the last node of the GatingSet
by default (e.g. "Single Cells"). For greater flexibility, users can
specify a parent population for each control, which will be extracted for
the unmixing calculation (e.g. Lymphocytes for CD4 APC or Myeloid Cells for
CD11b FITC). The parent populations for each control can also be specified
in a \code{parent} column in the channel match CSV file or in
\code{cyto_details}.}

\item{select}{passed to \code{\link{cyto_select}} to select the samples
required to compute the spillover matrix.}

\item{channels}{names of the channels/markers for which unmixing coefficients
should be computed, set to all fluorescent channels by default.}

\item{gate}{indicates whether to \code{draw} or use \code{auto} gating
methods to gate the negative and positive populations for each control.
Gating is optional for \code{"Autospill"} and \code{"CytoDecode"} methods
but is recommended to achieve better results and ensure compatibility with
\code{cyto_panel_design()}. Set to \code{"draw"} by default to allow for
more flexibility over gating. For \code{gate = "draw"} users can also
specify whether to gate the 1D histograms or 2D scatter plots 1D or 2D
suffix to \code{gate}. For example, to gate in 2D set \code{gate =
"draw-2D"} which is the new default manual gating method.}

\item{type}{options include \code{"Bagwell" or "autocomp"}, \code{"Roca" or
"autospill"} or \code{"CytoDecode"} to indicate which method to use when
computing the spillover matrix, set to \code{"CytoDecode"} by default.}

\item{stat}{indicates whether to use \code{"median"} or \code{"geomedian"} as
the statistic of choice for the \code{"Bagwell"} method, set to
\code{"geomedian"} by default.}

\item{save_as}{name of a CSV to which the computed unmixing matrix should be
written, set to \code{Spectral-Unmixing-Matrix.csv} prefixed with the date
by default. Set this argument to NA if you don't want to write the unmixing
matrix to file.}

\item{gatingTemplate}{name of \code{gatingTemplate} csv file to which the
\code{gatingTemplate} entries for the \code{GatingSet} method should be
saved, set to \code{cyto_gatingTemplate_active()} by default.}

\item{axes_trans}{object of class \code{transformerList} containing the
transformer definitions for transformations applied to the supplied data,
only required when \code{cytoset} objects are supplied.}

\item{axes_limits}{options include \code{"auto"}, \code{"data"} or
\code{"machine"} to use optimised, data or machine limits respectively. Set
to \code{"machine"} by default to use entire axes ranges. Fine control over
axes limits can be obtained by altering the \code{xlim} and \code{ylim}
arguments.}

\item{heatmap}{logical indicating whether the computed spectral unmixing
matrix should be displayed in a heatmap, set to TRUE by default.}

\item{events}{number of events to extract from each control prior to fitting
RLM models, set to 500 events by default.}

\item{model}{indicates the type of model to use when running autospill,
options include \code{"rlm"} for robust linear models or \code{"vwqr"} for
variance weighted quantile regression, set to \code{"rlm"} by default.}

\item{max_iter}{indicates the maximum number of allowable iterations for
refining the spillover coefficients, set to 20 by default.}

\item{trim}{proportion of events to exclude from the top and bottom of each
scale when fitting models in \code{AutoSpill} and \code{CytoDecode}
methods, set to 0.001 by default.}

\item{tol}{tolerance level for convergence of vwqr model in \code{CytoDecode}
method, set to \code{1e-6} by default.}

\item{grid_size}{size of the grid to use for wvqr model in \code{CytoDecode}
method, set to 100 by default. Smaller grid sizes will provide a speed
boost at the cost of accuracy.}

\item{resid}{indicates whether to use \code{"lower"}, \code{"upper"} or
\code{"both"} residuals when fitting variance models in vwqr used in
\code{CytoDecode}, set to \code{"both"} by default.}

\item{details}{name of a CSV file to which the details of the compensation
controls should be saved, set to NULL by default to use
\code{date-Control-Details.csv}. Setting this argument to \code{NA}
will prevent details from being written to a CSV file.}

\item{...}{additional arguments passed to \code{cyto_plot()} to allow
customisation of plots used for gating.}
}
\value{
unmixing matrix and write unmixing matrix to csv file named in
  accordance with \code{save_as}.
}
\description{
\code{cyto_unmix_compute()} will guide the user through the steps to required
to compute the spectral unmixing matrix using single colour control samples.
}
\details{
\code{cyto_unmix_compute()} does not perform any gating internally so samples
should be gated beforehand to isolate homogeneous populations of cells and/or
beads. \code{cyto_unmix_compute()} also requires annotation of some
additional information in \code{cyto_details(x)} prior to computing the
unmixing matrix. This includes \itemize{\item{group}{indicates the type of
each sample (i.e. cells or beads)}\item{parent}{name of the population to
extract from each sample (e.g. single cells or single beads) when a
GatingHierarchy or Gatingset is supplied}\item{label}{a name for the new
unmixed parameter for each control (e.g. CD4 FITC) - labels for unstained
controls should be set to \code{"Unstained"}}}. Each \code{group} must
contain a reference unstained population in order to compute the unmixing
matrix.
}
\examples{
\dontrun{
# GatingSet of controls
unmix <- cyto_unmix_compute(
  gs,
  auto = "cells",
  save_as = "Spectral-Unmixing-Matrix.csv"
)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
