% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_spectra_unmix.R
\name{cyto_unmix_compute}
\alias{cyto_unmix_compute}
\title{Compute spectral unmixing matrix}
\usage{
cyto_unmix_compute(
  x,
  parent = "root",
  channels = NULL,
  select = NULL,
  auto = TRUE,
  save_as = NULL,
  axes_trans = NA,
  axes_limits = "machine",
  heatmap = TRUE,
  ...
)
}
\arguments{
\item{x}{object of class \code{\link[flowWorkspace:cytoset]{cytoset}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} containing single
colour controls.}

\item{parent}{name of the population to extract from each control to compute
the spectral unmixing coefficients, set to \code{"root"} by default.}

\item{channels}{names of the channels/markers for which unmixing coefficients
should be computed, set to all fluorescent channels by default.}

\item{select}{passed to \code{cyto_select()} to extract a subset of samples
for computation of unmixing matrix.}

\item{auto}{logical indicating whether autofluorescence subtraction should be
performed, set to TRUE by default. Users will be asked which group to use
for autofluorescence subtraction or the name of the group can be supplied
manually here.}

\item{save_as}{name of a CSV to which the computed unmixing matrix should be
written, set to \code{Spectral-Unmixing-Matrix.csv} prefixed with the date
by default. Set this argument to NA if you don't want to write the unmixing
matrix to file.}

\item{axes_trans}{object of class \code{transformerList} containing the
transformer definitions for transformations applied to the supplied data,
only required when \code{cytoset} objects are supplied.}

\item{axes_limits}{options include \code{"auto"}, \code{"data"} or
\code{"machine"} to use optimised, data or machine limits respectively. Set
to \code{"machine"} by default to use entire axes ranges. Fine control over
axes limits can be obtained by altering the \code{xlim} and \code{ylim}
arguments.}

\item{heatmap}{logical indicating whether the computed spectral unmixing
matrix should be displayed in a heatmap, set to TRUE by default.}

\item{...}{additional arguments passed to \code{cyto_plot()} to allow
customisation of plots used for gating.}
}
\value{
computed unmixing coefficients in the form of a matrix.
}
\description{
\code{cyto_unmix_compute()} will guide the user through the steps to required
to compute the spectral unmixing matrix using single colour control samples.
}
\details{
\code{cyto_unmix_compute()} does not perform any gating internally so samples
should be gated beforehand to isolate homogeneous populations of cells and/or
beads. \code{cyto_unmix_compute()} also requires annotation of some
additional information in \code{cyto_details(x)} prior to computing the
unmixing matrix. This includes \itemize{\item{group}{indicates the type of
each sample (i.e. cells or beads)}\item{parent}{name of the population to
extract from each sample (e.g. single cells or single beads) when a
GatingHierarchy or Gatingset is supplied}\item{label}{a name for the new
unmixed parameter for each control (e.g. CD4 FITC) - labels for unstained
controls should be set to \code{"Unstained"}}}. Each \code{group} must
contain a reference unstained population in order to compute the unmixing
matrix.
}
\examples{
\dontrun{
# GatingSet of controls
unmix <- cyto_unmix_compute(
  gs,
  auto = "cells",
  save_as = "Spectral-Unmixing-Matrix.csv"
)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
