% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_editor.R
\name{data_editor}
\alias{data_editor}
\title{Interactive data editor}
\usage{
data_editor(
  x,
  title = "Data Editor",
  menu = FALSE,
  options = NULL,
  save_as = NULL,
  viewer = TRUE
)
}
\arguments{
\item{x}{object of class matrix or data.frame will colnames specified,
matrices will be coerced to data.frames.}

\item{title}{title to include in above the table.}

\item{menu}{logical indicating whether the last column should contain
dropdown menus, set to FALSE by default.}

\item{options}{vector of options to use in dropdpwn menus when menu is TRUE.}

\item{save_as}{name of a csv file to which the edited table should be saved.}

\item{viewer}{logical indicating whether the table editor should be launched
in the RStudio viewer pane, set to TRUE by default. Mke sure to use the
"Save & Close" to close the data editor or else the changes will not be
saved.}
}
\description{
Interactive data editor
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
