% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawGate-methods.R
\docType{methods}
\name{drawGate,GatingSet-method}
\alias{drawGate,GatingSet-method}
\title{drawGate GatingSet Method}
\usage{
\S4method{drawGate}{GatingSet}(x, pData = NULL, parent = "root",
  alias = NULL, channels, gate_type = "polygon", subSample = 250000,
  template = NULL, file = "gatingTemplate.csv", axis = "x",
  adjust = 1.5, plot = TRUE, labs = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{GatingSet}.}

\item{pData}{vector of indicating \code{c(column,row)} of \code{pData(gs)} used extract particular samples for gating.}

\item{parent}{name of the \code{parent} population to extract for gating.}

\item{alias}{the name(s) of the populations to be gated. If multiple population names are supplied (e.g. \code{c("CD3,"CD4)}) multiple 
gates will be returned. \code{alias} is \code{NULL} by default which will halt the gating routine.}

\item{channels}{a vector indicating the fluorescent channel(s) to be used for gating. If a single channel is supplied, a histogram of
of the kernel density will be constructed.}

\item{gate_type}{vector of gate type names used to construct the gates. Multiple \code{gate_types} are supported but should be accompanied wuth
an \code{alias} argument of the same length (i.e. one \code{gate_type} per \code{alias}). Supported \code{gate_types} are \code{polygon, rectangle,
ellipse, threshold, boundary, interval and quadrant} which can be abbreviated as upper or lower case first letters as well.}

\item{subSample}{numeric indicating the number of events to plot to speed up plotting. If \code{subSample} is greater than the total
number of events, all events will be plotted. \code{subSample} is set to 250 000 events by default.}

\item{template}{the name of an existing R object in global environment constructed using \code{openCyto::add_pop()} (e.g. "template").}

\item{file}{name of \code{gatingTemplate} csv file if \code{template = TRUE}.}

\item{axis}{indicates whether the \code{"x"} or \code{"y"} axis should be gated for 2-D interval gates.}

\item{adjust}{smooothing factor passed to \code{density()} for 1D plots (defaults to 1.5).}

\item{plot}{logical indicating whether the data should be plotted. This feature allows for constructing gates of different types over 
existing plots which may already contain a different gate type. For demonstration of this feature refer to the package vignette.}

\item{labs}{logical indicating whether to include \code{plotLabels} for the gated population(s), \code{TRUE} by default.}

\item{...}{additional arguments for \code{plotDens}.}
}
\value{
object of class \code{GatingSet} with gates applied. The \code{gatingTemplate} can also be optionally written to .csv file
and entries add to the R object called \code{Template} generated use the \code{openCyto::add_pop} API.
}
\description{
drawGate GatingSet Method
}
