% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawGate-methods.R
\name{drawGate}
\alias{drawGate}
\title{drawGate Root}
\usage{
drawGate(x, ...)
}
\arguments{
\item{x}{object of class \code{flowFrame}, \code{flowSet} or \code{GatingSet}.}
}
\description{
drawGate Root
}
