% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawGate-methods.R
\name{drawGate}
\alias{drawGate}
\title{drawGate}
\usage{
drawGate(x, ...)
}
\arguments{
\item{x}{object of class \code{flowFrame}, \code{flowSet} or \code{GatingSet}.}
}
\description{
\code{drawGate} is a convenient wrapper for the gating functions shipped with \code{cytoSuite} to facilitate gate drawing. Using \code{drawGate} you can specify the type of gate(s) to be constructed and \code{drawGate}
will automatically make calls to the relevant gating function(s) and handle plotting appropriately.
}
\details{
Pools samples
Extract subsample for plotting
make call to gating functions
return gates as a list of filters

GatingSet method has additional functionality - gates are applied to GatingSet & gates saved in gatingTemplate for future use
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
