% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating-functions.R
\name{drawWeb}
\alias{drawWeb}
\title{Draw Web to Gate Multiple Flow Cytometry Populations.}
\usage{
drawWeb(fr, channels, alias = NULL, plot = TRUE, labels = TRUE, ...)
}
\arguments{
\item{fr}{a \code{\link[flowCore:flowFrame-class]{flowFrame}} object
containing the flow cytometry data for plotting and gating.}

\item{channels}{vector of channel names to use for plotting, can be of length
1 for 1-D density histogram or length 2 for 2-D scatter plot.}

\item{alias}{the name(s) of the populations to be gated. If multiple
population names are supplied (e.g. \code{c("CD3","CD4")}) multiple gates
will be returned. \code{alias} is \code{NULL} by default which will halt
the gating routine. Recommended for 3 or more populations.}

\item{plot}{logical indicating whether the data should be plotted. This
feature allows for constructing gates of different types over existing
plots which may already contain a different gate type.}

\item{labels}{logical indicating whether to include \code{\link{plotLabels}}
for the gated population(s), \code{TRUE} by default.}

\item{...}{additional arguments for \code{\link{plotCyto,flowFrame-method}}.}
}
\value{
a\code{\link[flowCore:filters-class]{filters}} list containing the
  constructed \code{\link[flowCore:polygonGate]{polygonGate}}
  object(s).
}
\description{
\code{drawWeb} is a variation of drawQuadrant which allows more flexibility
with gate co-ordinates (angled lines) and supports any number of gates as
indicated by the \code{alias} argument. To construct the gate simply select
the center point and surrounding divider points on plot edge. \code{drawWeb}
will contruct the \code{\link[flowCore:polygonGate]{polygonGate}}
objects and store them in a \code{\link[flowCore:filters-class]{filters}}
list.
}
\seealso{
\code{\link{plotCyto1d,flowFrame-method}}

\code{\link{plotCyto2d,flowFrame-method}}

\code{\link{drawGate}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
\keyword{draw,}
\keyword{drawWeb}
\keyword{gating,}
\keyword{manual,}
\keyword{openCyto,}
\keyword{polygonGate,}
