% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingTemplate-modifiers.R
\name{editGate}
\alias{editGate}
\title{Edit Existing Gate(s).}
\usage{
editGate(x, select = NULL, parent = NULL, alias = NULL,
  overlay = NULL, type = NULL, gtfile = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{GatingSet}.}

\item{select}{vector containing the indicies of samples within gs to use for
plotting.}

\item{parent}{name of the parental population.}

\item{alias}{name(s) of the gate to edit (e.g. "Single Cells").}

\item{overlay}{name(s) of the population(s) to overlay onto the plot.}

\item{type}{vector of gate type names used to construct the gates. Multiple
\code{gate_types} are supported but should be accompanied with an
\code{alias} argument of the same length (i.e. one \code{type} per
\code{alias}). Supported \code{gate_types} are \code{polygon, rectangle,
ellipse, threshold, boundary, interval, quadrant and web} which can be
abbreviated as upper or lower case first letters as well. Default
\code{type} is \code{"polygon"}.}

\item{gtfile}{name of the \code{gatingTemplate} csv file (e.g.
"gatingTemplate.csv") where the gate is saved.}

\item{...}{additional arguments passed to plotCyto, see ?plotCyto for
details.}
}
\value{
an object of calss \code{GatingSet} with edited gate applied, as well
  as gatingTemplate file with editied gate saved.
}
\description{
Edit Existing Gate(s).
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
