% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editSpillover-methods.R
\docType{methods}
\name{editSpillover,flowSet-method}
\alias{editSpillover,flowSet-method}
\title{Edit Spillover Matrix flowSet}
\usage{
\S4method{editSpillover}{flowSet}(x, spfile = NULL, gtfile = NULL)
}
\arguments{
\item{x}{an object of class \code{flowSet}.}

\item{spfile}{name of spillover csv file to be edited (defaults to NULL). Spillover extracted from samples if not supplied.}

\item{gtfile}{name of gatingTemplate csv file to apply to flowSet prior to plotting (defaults to NULL). If not supplied samples will be gated FSC-A/SSC-A and
SSC-W/SSC-H to gate single cells.}
}
\description{
Edit Spillover Matrix flowSet
}
