% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editSpillover-methods.R
\name{editSpillover}
\alias{editSpillover}
\title{Edit Spillover Matrix}
\usage{
editSpillover(x, ...)
}
\arguments{
\item{x}{object of class \code{flowSet} or \code{GatingSet}.}
}
\description{
Edit spillover matrix using single stain compensation controls and an unstained control.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
