% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editSpillover-methods.R
\name{editSpillover}
\alias{editSpillover}
\title{Edit Spillover Matrix}
\usage{
editSpillover(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowCore:flowSet-class]{flowSet}}.}

\item{...}{additional method-specific arguments.}
}
\description{
Edit spillover matrices in real-time using a shiny interface.
}
\details{
\code{editSpillover} provides an interactive shiny interface for editing
fluorescent spillover matrices. \code{editSpillover} takes on either a
\code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} containing
untransformed single stain compensation controls and a universal unstained
control. It is recommended that samples be pre-gated based on FSC and SSC
parameters to obtain a homogeneous population for calculation of fluorescent
spillover. Users begin by selecting the unstained control and a stained
control from dropdown menus of sample names. \code{editSpillover} leverages
\code{plotCyto} to plot the stained sample and overlay the unstained control
in black. Users should then select the channel associated with the selected
control on the \code{x axis} and go through all other channels on the \code{y
axis}. The displayed spillover matrix is extracted directly from the
\code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} unless another
spillover matrix is supplied through the spfile argument. To edit the
spillover matrix simply modify the appropriate cell in the the table. The new
spillover matrix will be re-applied to the samples with each edit and
automatically re-plotted so you can track changes in real-time. To aid in
selection of an appropriate spillover value, the median fluorescent intensity
of the unstained control is indicated by a red line and median fluorescent
intensity of the stained control is tracked with a purple line. These
features can be turned off by de-selecting the check boxes. Changes to the
spillover matrix are automatically saved to a csv file called
\code{"Spillover Matrix.csv"} in the case where the \code{spfile} is not
specified or to the same name as the specified \code{spfile}.
\code{editSpillover} has methods for both
\code{\link[flowCore:flowSet-class]{flowSet}} and
\code{\link[flowCore:flowSet-class]{flowSet}} objects refer to their
respective help pages for more information.
}
\seealso{
\code{\link{editSpillover,flowSet-method}}

\code{\link{editSpillover,GatingSet-method}}

\code{\link{plotCyto1d,flowFrame-method}}

\code{\link{plotCyto2d,flowFrame-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
