% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editSpillover.R
\name{editSpillover}
\alias{editSpillover}
\title{Edit Spillover Matrix}
\usage{
editSpillover(x, parent = "root", spfile = NULL)
}
\arguments{
\item{x}{object of class \code{GatingSet}.}

\item{parent}{name of the population to use for plotting (defaults to "root").}

\item{spfile}{name of spillover matrix csv file including .csv file extension to use as a starting point for editing.}
}
\value{
write edited spillover matrix to csv file called \code{"Spillover Matrix.csv"} for later use.
}
\description{
Edit Spillover Matrix
}
