% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editSpillover-methods.R
\name{editSpillover}
\alias{editSpillover}
\title{Edit Spillover Matrix}
\usage{
editSpillover(x, ...)
}
\arguments{
\item{x}{object of class \code{flowSet} or \code{GatingSet}.}
}
\description{
Edit Spillover Matrix
}
