% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIt-SNE.R
\name{fftRtsne}
\alias{fftRtsne}
\title{FIt-SNE Based on Kluger Lab FIt-SNE}
\usage{
fftRtsne(
  X,
  dims = 2,
  perplexity = 30,
  theta = 0.5,
  max_iter = 750,
  fft_not_bh = TRUE,
  ann_not_vptree = TRUE,
  stop_early_exag_iter = 250,
  exaggeration_factor = 12,
  no_momentum_during_exag = FALSE,
  start_late_exag_iter = -1,
  late_exag_coeff = 1,
  mom_switch_iter = 250,
  momentum = 0.5,
  final_momentum = 0.8,
  learning_rate = "auto",
  n_trees = 50,
  search_k = -1,
  rand_seed = -1,
  nterms = 3,
  intervals_per_integer = 1,
  min_num_intervals = 50,
  K = -1,
  sigma = -30,
  initialization = "pca",
  max_step_norm = 5,
  load_affinities = NULL,
  fast_tsne_path = NULL,
  nthreads = 0,
  perplexity_list = NULL,
  get_costs = FALSE,
  df = 1
)
}
\arguments{
\item{X}{a matrix containing the data to be mapped.}

\item{dims}{dimensionality of the embedding, set to 2 by default.}

\item{perplexity}{used to determine the bandwidth of the Gaussian kernel in
the input space, set to 30 by default.}

\item{theta}{set to 0 for exact t-SNE. If non-zero, then will use either
Barnes Hut or FIt-SNE based on nbody_algo. If Barnes Hut, then this
determines the accuracy of BH approximation. Set to 0.5 by default.}

\item{max_iter}{number of iterations of t-SNE to run, set to 750 by default.}

\item{fft_not_bh}{if theta is nonzero, this determines whether to use FIt-SNE
or Barnes Hut approximation, set to TRUE by default for FIt-SNE.}

\item{ann_not_vptree}{use vp-trees (as in bhtsne) or approximate nearest
neighbors (default). Set to be TRUE for approximate nearest neighbours.}

\item{stop_early_exag_iter}{when to switch off early exaggeration, set to 250
by default.}

\item{exaggeration_factor}{coefficient for early exaggeration (>1), set to 12
by default.}

\item{no_momentum_during_exag}{set to 0 to use momentum and other
optimization tricks. Can be set to 1 to do plain, vanilla gradient descent
(useful for testing large exaggeration coefficients).}

\item{start_late_exag_iter}{when to start late exaggeration, set to -1 by
default to not use late exaggeration.}

\item{late_exag_coeff}{late exaggeration coefficient, set to 1 by default to
not use late exaggeration.}

\item{mom_switch_iter}{iteration number to switch from momentum to
final_momentum, set to 250 by default.}

\item{momentum}{initial value of momentum, set to 0.5 by default.}

\item{final_momentum}{value of momentum to use later in the optimisation, set
to 0.8 by default.}

\item{learning_rate}{set to desired learning rate or 'auto', which sets
learning rate to N/exaggeration_factor where N is the sample size, or to
200 if N/exaggeration_factor < 200.}

\item{n_trees}{when using Annoy, the number of search trees to use, set to 50
by default.}

\item{search_k}{When using Annoy, the number of nodes to inspect during
search. Default is 3*perplexity*n_trees (or K*n_trees when using fixed
sigma).}

\item{rand_seed}{seed for random initialisation, set to -1 by default to
initialise random number generator with current time.}

\item{nterms}{if using FIt-SNE, this is the number of interpolation points
per sub-interval, set to 3 by default.}

\item{intervals_per_integer}{see min_num_intervals.}

\item{min_num_intervals}{let maxloc = ceil(max(max(X))) and minloc =
floor(min(min(X))). i.e. the points are in a [minloc]^no_dims by
[maxloc]^no_dims interval/square. The number of intervals in each dimension
is either min_num_intervals or ceil((maxloc -
minloc)/intervals_per_integer), whichever is larger. min_num_intervals must
be an integer >0, and intervals_per_integer must be >0. Defaults are
min_num_intervals=50 and intervals_per_integer = 1.}

\item{K}{number of nearest neighbours to get when using fixed sigma, set to
-1 by default.}

\item{sigma}{fixed sigma value to use when perplexity==-1, set to -30 by
default.}

\item{initialization}{'pca', 'random', or N x no_dims array to intialize the
solution, set to 'pca' by default.}

\item{max_step_norm}{maximum distance that a point is allowed to move on one
iteration. Larger steps are clipped to this value. This prevents possible
instabilities during gradient descent. Set to -1 to switch it off.
Set to 5 by default.}

\item{load_affinities}{if 1, input similarities are loaded from a file and
not computed. If 2, input similarities are saved into a file. If 0,
affinities are neither saved nor loaded.}

\item{fast_tsne_path}{path to FItSNE executable.}

\item{nthreads}{number of threads to use, set to use all available threads by
default.}

\item{perplexity_list}{if perplexity==0 then perplexity combination will be
used with values taken from perplexity_list. Default: NULL df - Degree of
freedom of t-distribution, must be greater than 0. Values smaller than 1
correspond to heavier tails, which can often resolve substructure in the
embedding. See Kobak et al. (2019) for details. Default is 1.0.}

\item{get_costs}{logical indicating whether the KL-divergence costs computed
every 50 iterations should be returned, set to FALSE by default.}

\item{df}{positive numeric that controls the degree of freedom of
t-distribution. The actual degree of freedom is 2*df-1. The standard t-SNE
choice of 1 degree of freedom corresponds to df=1. Large df approximates
Gaussian kernel. df<1 corresponds to heavier tails, which can often resolve
substructure in the embedding. See Kobak et al. (2019) for details. Default
is 1.0.}
}
\description{
Modified version of \url{https://github.com/KlugerLab/FIt-SNE/}
implementation to expose argument names and defaults within CytoExploreR.
This function should not be used directly, data should instead be mapped
using \code{\link{cyto_map}}.
}
\references{
Linderman, G., Rachh, M., Hoskins, J., Steinerberger, S.,
  Kluger., Y. (2019). Fast interpolation-based t-SNE for improved
  visualization of single-cell RNA-seq data. Nature Methods.
  \url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6402590/}.
}
\seealso{
\code{\link{cyto_map}}
}
