% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openCyto-plugins.R
\name{gate_draw}
\alias{gate_draw}
\title{drawGate Plugin for openCyto}
\usage{
gate_draw(fr, pp_res, channels, alias, subSample, gate_range = NULL,
  min = NULL, max = NULL, gate_type = c("polygon", "Polygon", "p", "P",
  "rectangle", "Rectangle", "r", "R", "interval", "Interval", "i", "I",
  "threshold", "Threshold", "t", "T", "boundary", "Boundary", "b", "B",
  "ellipse", "Ellipse", "e", "E", "quadrant", "Quadrant", "q", "Q"),
  axis = c("x", "y"), adjust = 1.5, plot = TRUE, labs = TRUE, ...)
}
\arguments{
\item{fr}{a \code{flowFrame} object containing the flow cytometry data for gating.}

\item{pp_res}{output of preprocessing function.}

\item{channels}{a vector of length 2 indicating the channels used to construct the 2D plot.}

\item{gate_range}{range in which gate should be constructed (only needed for autogating functions).}

\item{min}{argument passed to \code{truncate_flowFrame} to restrict data to values > \code{min}.}

\item{max}{argument passed to \code{truncate_flowFrame} to restrict data to values < \code{max}.}

\item{gate_type}{type of gate to be constructed, supported types include 
\code{c("polygon", "rectangle", "interval", "threshold", "ellipse", "quadrant")}.}

\item{axis}{indicates the axis to use for gating for \code{gate_type="interval"} when 2 fluorescent channel are supplied.}

\item{adjust}{numeric smoothing factor used for 1D density plots.}

\item{...}{additional arguments passsed to \code{DrawGate}.}

\item{filterId}{gate name assigned by openCyto from the \code{gatingTemplate}.}

\item{N}{an integer indicating the number of gates to construct, set to 1 by default.}
}
\value{
a \code{polygonGate} constructed from coordinates supplied by \code{DrawGate}.
}
\description{
\code{drawGate} allows the user to draw polygon gates directly onto plots of flow cytometry data.
Simply left click to gate and right click to close the gate.
}
\examples{
{
library(openCyto)
library(cytoTools)

fs <- read.flowSet(path = "Samples", pattern = ".fcs") # load in .fcs files

gs <- GatingSet(fs) # add flowSet to GatingSet

template <- add_pop(
gs, alias = "Lymphocytes", pop = "+", parent = "root", dims = "FSC-A,SSC-A", gating_method = "drawGate", gating_args = "gate_type='ellipse'",
gating_args = "subSample=10000", collapseDataForGating = TRUE, groupBy = 2
)

# gating window will open to construct gate left click vertices on plot and close gate by right click and selecting "stop".

ggcyto(gs[[1]], subset = "root", aes(x = "FSC-A",y = "SSC-A")) + geom_hex(bins = 100) + geom_stats()

}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
\keyword{gating,}
\keyword{manual,}
\keyword{polygon,}
\keyword{polygonGate}
