% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_draw-methods.R
\name{gate_draw}
\alias{gate_draw}
\title{gate_draw}
\usage{
gate_draw(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}}.}

\item{...}{additional method-specific arguments.}
}
\description{
Manually draw gates around populations for analysis of flow cytometry data.
}
\details{
\code{gate_draw} is a convenient wrapper for the gating functions shipped
with \code{cytoRSuite} to facilitate analysis of flow cytometry by gate
drawing. Using \code{gate_draw} users can specify the type of gate(s) to be
constructed through the \code{type} argument and \code{gate_draw} will
automatically handle plotting the data and make calls to the relevant gating
function(s) to construct the gates around populations of interest.
\code{gate_draw} has methods for
\code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}} and
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} objects, refer to
their respective help pages for more information. The flowFrame and flowSet
methods simply return the constructed gates as a list of
\code{\link[flowCore:filters-class]{filters}}, whilst the GatingSet method
automatically applies the constructed gates to the GatingSet and saves the
constructed gates in an \code{openCyto}
\code{\link[openCyto:gatingTemplate-class]{gatingTemplate}}for future use.
See \code{\link{gate_edit}} and \code{\link{gate_remove}} to manipulate
constructed gates and modify their entries in the gatingTemplate.
}
\seealso{
\code{\link{gate_draw,flowFrame-method}}

\code{\link{gate_draw,flowSet-method}}

\code{\link{gate_draw,GatingSet-method}}
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
