% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\docType{methods}
\name{getChannels,flowSet-method}
\alias{getChannels,flowSet-method}
\title{Extract Fluorescent Channels - flowSet Method}
\usage{
\S4method{getChannels}{flowSet}(x)
}
\arguments{
\item{x}{object \code{flowSet}}
}
\value{
vector of fluorescent channels
}
\description{
Extract Fluorescent Channels - flowSet Method
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
