% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\docType{methods}
\name{getChannels,flowSet-method}
\alias{getChannels,flowSet-method}
\title{Extract fluorescent parameters from a flowSet}
\usage{
\S4method{getChannels}{flowSet}(x)
}
\arguments{
\item{x}{object \code{flowSet}}
}
\value{
vector of fluorescent channels
}
\description{
Extract fluorescent parameters from a flowSet
}
