% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\docType{methods}
\name{getChannels,GatingSet-method}
\alias{getChannels,GatingSet-method}
\title{Extract Fluorescent Channels - GatingSet Method}
\usage{
\S4method{getChannels}{GatingSet}(x)
}
\arguments{
\item{x}{object \code{GatingSet}}
}
\value{
vector of fluorescent channels
}
\description{
Extract Fluorescent Channels - GatingSet Method
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
