% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatingTemplate-modifiers.R
\name{getGateType}
\alias{getGateType}
\title{Get Gate Type from Saved Gate.}
\usage{
getGateType(gates)
}
\arguments{
\item{gates}{an object of class \code{filters} containing the gates from
which the \code{type(s)} must be obtained.}
}
\value{
vector of gate type names for supplied gates.
}
\description{
Get Gate Type from Saved Gate.
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
