% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plotGates}
\alias{plotGates}
\title{Plot gates stored in filters list on an existing plot.}
\usage{
plotGates(gates, col = "magenta")
}
\arguments{
\item{gates}{an object of class \code{filters} containing the gate(s) for plotting.}

\item{col}{indicates the colour of the gate to be constructed, set to \code{"red"} by default.}
}
\description{
Plot gates stored in filters list on an existing plot.
}
