% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLabels-methods.R
\name{plotLabels}
\alias{plotLabels}
\title{Plot Labels for Gated Populations}
\usage{
plotLabels(x, gates, ...)
}
\arguments{
\item{x}{object of class \code{"flowFrame"}.}

\item{gates}{object of class
\code{\link[flowCore:rectangleGate]{rectangleGate}},
\code{\link[flowCore:polygonGate]{polygonGate}},
\code{\link[flowCore:ellipsoidGate]{ellipsoidGate}}, \code{"list"} or
\code{\link[flowCore:filters-class]{filters}}.}

\item{...}{additional method-specific arguments.}
}
\value{
add label to existing plot with information about gated population.
}
\description{
Plot Labels for Gated Populations
}
\seealso{
\code{\link{plotLabels,flowFrame,rectangleGate-method}}

\code{\link{plotLabels,flowFrame,polygonGate-method}}

\code{\link{plotLabels,flowFrame,ellipsoidGate-method}}

\code{\link{plotLabels,flowFrame,list-method}}

\code{\link{plotLabels,flowFrame,filters-method}}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
