% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{removeGate}
\alias{removeGate}
\title{Remove Gate(s) and Edit gatingTemplate .csv File}
\usage{
removeGate(gs, alias = NULL, gtfile = NULL)
}
\arguments{
\item{gs}{an object of class \code{GatingSet}.}

\item{alias}{name(s) of the population(s) to remove (e.g. "Single Cells"). By default all descendant populations will be removed as well.}

\item{gtfile}{name of the \code{gatingTemplate} csv file (e.g. "gatingTemplate.csv").}
}
\value{
an object of class \code{gatingSet} with gate and children removed, as well as gatingTemplate file with population removed.
}
\description{
Remove Gate(s) and Edit gatingTemplate .csv File
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
