% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{selectChannels}
\alias{selectChannels}
\title{Select Fluorescent Channel for Single Stain Controls}
\usage{
selectChannels(fs)
}
\arguments{
\item{x}{object of class \code{flowSet} or \code{GatingSet} containing compensation controls}
}
\value{
vector of channels in order of \code{flowSet} or \code{GatingSet} samples.
}
\description{
Select Fluorescent Channel for Single Stain Controls
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
