% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{selectChannels}
\alias{selectChannels}
\title{Select Fluorescent Channel for Compensation Controls}
\usage{
selectChannels(x)
}
\arguments{
\item{x}{object of class \code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}} or
\code{\link[flowWorkspace:GatingSet-class]{GatingSet}} containing
compensation controls.}
}
\value{
vector of channels in order of compensation Control samples.
}
\description{
Select Fluorescent Channel for Compensation Controls
}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
